% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/all_methods.R
\docType{methods}
\name{aggregate_replicates}
\alias{aggregate_replicates}
\alias{aggregate_replicates,matrix-method}
\alias{aggregate_replicates,DESeqDataSet-method}
\title{aggregate_replicates ====
Aggregate lane-split technical replicates}
\usage{
aggregate_replicates(object, ...)

\S4method{aggregate_replicates}{matrix}(object, pattern = "_L\\\\d+")

\S4method{aggregate_replicates}{DESeqDataSet}(object, pattern = "_L\\\\d+")
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough.}

\item{pattern}{Grep pattern to match lane identifiers in file name.}

\item{object}{Object.}
}
\value{
Object of same class, with pooled technical replicates.
}
\description{
Frequently RNA-seq experiments are performed with technical replicates
split across flow cell lanes. This generic facilitates quick aggregation
of counts across the flow cells.
}
\seealso{
A new \link{DESeqDataSet} is returned using \code{\link[=DESeqDataSetFromMatrix]{DESeqDataSetFromMatrix()}}.
}
\author{
Michael Steinbaugh
}
