% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-aggregateReplicates.R
\docType{methods}
\name{aggregateReplicates}
\alias{aggregateReplicates}
\alias{aggregateReplicates,DESeqDataSet-method}
\title{Aggregate Lane-Split Technical Replicates}
\usage{
\S4method{aggregateReplicates}{DESeqDataSet}(object, pattern = "_L\\\\d+")
}
\arguments{
\item{object}{Object.}

\item{pattern}{Grep pattern to match lane identifiers in sample name.}
}
\value{
Object of same class, with pooled technical replicates.
}
\description{
Frequently RNA-seq experiments are performed with technical replicates
split across flow cell lanes. This generic facilitates quick aggregation
of counts across the flow cells.
}
\note{
\link{DESeqDataSet} is returned using \code{\link[=DESeqDataSetFromMatrix]{DESeqDataSetFromMatrix()}}.
}
\examples{
load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioRNASeq"))

# bcbioRNASeq
\dontrun{
aggregateReplicates(bcb)
}

# DESeqDataSet
\dontrun{
aggregateReplicates(dds)
}

# matrix
\dontrun{
counts <- counts(bcb)
aggregateReplicates(counts)
}
}
