% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/basejump.R
\name{aggregate_replicates}
\alias{aggregate_replicates}
\alias{assay}
\alias{assays}
\alias{colData}
\alias{design}
\alias{metadata}
\alias{rowData}
\title{S4 generics}
\usage{
assay(x, i, ...)

assays(x, ..., withDimnames = TRUE)

colData(x, ...)

design(object, ...)

metadata(x, ...)

rowData(x, ...)
}
\description{
S4 generics

colData
}
\keyword{internal}
