% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{AllGenerics}
\alias{AllGenerics}
\title{S4 Generics}
\arguments{
\item{object}{Object.}

\item{i}{An integer or numeric scalar.}

\item{withDimnames}{A \code{logical}, indicating whether dimnames should be
applied to extracted assay elements.}

\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
No value.
}
\description{
S4 Generics
}
\keyword{internal}
