% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-annotable.R
\docType{methods}
\name{annotable}
\alias{annotable}
\alias{annotable,bcbioRNASeqANY-method}
\title{Ensembl Annotations}
\usage{
\S4method{annotable}{bcbioRNASeqANY}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\link{data.frame}
}
\description{
Ensembl Annotations
}
\author{
Michael Steinbaugh
}
