% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-annotables.R
\name{.annotable}
\alias{.annotable}
\title{\href{http://www.ensembl.org/}{Ensembl} annotations from \link{annotables}}
\usage{
.annotable(genome_build, format = "gene")
}
\arguments{
\item{genome_build}{Genome build. Consult the \link{annotables} documentation
for a list of currently supported genomes.}

\item{format}{Desired table format, either \code{gene} or \code{tx2gene}.}
}
\value{
Modified annotable with unique rows (.
}
\description{
Quickly access gene annotations and transcript-to-gene (tx2gene) mappings
pre-compiled from \href{http://www.ensembl.org/}{Ensembl} with the
\href{http://bioconductor.org/packages/release/bioc/html/biomaRt.html}{biomaRt}
package. For gene annotables, the Entrez identifier is removed, to allow
for unique Ensembl gene identifiers.
}
\author{
Michael Steinbaugh
}
\keyword{internal}
