% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basejump.R
\name{comp}
\alias{comp}
\alias{kables}
\alias{make_names}
\alias{make_names_snake}
\alias{revcomp}
\alias{save_data}
\alias{set_names_dot}
\alias{set_names_snake}
\title{Utility functions from basejump package.}
\usage{
comp(dna)

kables(list, captions = NULL)

make_names(character)

make_names_snake(character)

revcomp(dna)

save_data(..., dir = "data")

set_names_dot(data)

set_names_snake(data)
}
\arguments{
\item{dna}{DNA sequence (ATGC nucleotides).}

\item{list}{List of tables (e.g. data frame, matrix).}

\item{captions}{Caption character vector.}

\item{character}{Character vector.}

\item{dir}{Output directory.}

\item{data}{Data type that supports name assignments.}
}
\description{
Complement DNA sequence.

Handle multiple kables in a single RMarkdown chunk.

Make syntactically valid names out of character vectors.

Make names in snake_case.

Reverse complement DNA sequence.

Quick save to \code{data} directory.

Set names in dot notation.

Set names in snake_case.
}
\author{
Michael Steinbaugh
}
\keyword{internal}
