% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bcb}
\alias{bcb}
\title{\emph{M. musculus} dummy data}
\format{An object of class \code{bcbioRNADataSet} with 505 rows and 4 columns.}
\usage{
bcb
}
\description{
\emph{M. musculus} dummy data
}
\author{
Lorena Pantano
}
\keyword{datasets}
\keyword{internal}
