% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bcb_small}
\alias{bcb_small}
\title{Minimal \code{bcbioRNASeq} Example}
\format{An object of class \code{bcbioRNASeq} with 100 rows and 6 columns.}
\usage{
bcb_small
}
\description{
GSE65267
}
\author{
Michael Steinbaugh, Lorena Pantano
}
\keyword{internal}
