% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-bcbio-legacy.R
\docType{methods}
\name{bcbio-legacy}
\alias{bcbio-legacy}
\alias{bcbio,bcbioRNADataSet-method}
\alias{bcbio<-,bcbioRNADataSet-method}
\title{bcbioRNASeq Additional Run Data Accessor (Legacy)}
\usage{
\S4method{bcbio}{bcbioRNADataSet}(object, type)

\S4method{bcbio}{bcbioRNADataSet}(object, type) <- value
}
\arguments{
\item{object}{Object.}

\item{type}{Type of data to retrieve.}

\item{value}{Value to assign.}
}
\value{
\link{slot} object.
}
\description{
Package versions prior to 0.0.27 used \code{callers} to define the extra bcbio
slot. The structure of the object is otherwise the same.
}
\details{
Additional count matrices of interest include:
\itemize{
\item Aligned counts generated by \href{https://github.com/alexdobin/STAR}{STAR} and
\href{http://bioinf.wehi.edu.au/featureCounts/}{featureCounts}.
\item Transcript-level gene expression, for
\href{http://pachterlab.github.io/sleuth}{sleuth} analysis.
\item miRNA expression counts.
}

Since these matrices don't necessarily contain the same number of columns and
rows as the counts imported with \code{\link[=tximport]{tximport()}}, it is preferable to store this
information as alternative callers.
}
\keyword{internal}
