% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{bcbioRNADataSet-class}
\alias{bcbioRNADataSet-class}
\alias{bcbioRNADataSet}
\title{\code{bcbioRNADataSet}}
\description{
This class will be deprecated in favor of \link{bcbioRNASeq} in a future release.
}
\section{Slots}{

\describe{
\item{\code{callers}}{\link{SimpleList} containing additional bcbio run data with
dimensions that don't match the count matrix.}
}}

\author{
Lorena Pantano, Michael Steinbaugh
}
\keyword{internal}
