% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{bcbioRnaDataSet-class}
\alias{bcbioRnaDataSet-class}
\alias{bcbioRnaDataSet}
\title{bcbioRnaDataSet}
\description{
S4 class containing RNA-seq counts and metadata generated by
\href{https://bcbio-nextgen.readthedocs.org/}{bcbio-nextgen}.
}
\details{
\code{bcbioRnaDataSet} is a subclass of \link{SummarizedExperiment} designed to store
an RNA-seq analysis. This class contains experiment metadata, raw counts,
normalilzed counts, and summary statistics for each sample analyzed.

Methods for this objects ...

\code{metadata} contains ...
}
\keyword{internal}
