% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\name{bcbioRnaDataSetFromFolder}
\alias{bcbioRnaDataSetFromFolder}
\title{\code{bcbioRnaDataSetFromFolder} loads bcbio-nextgen output}
\usage{
bcbioRnaDataSetFromFolder(upload_dir = "final", analysis = "rnaseq",
  intgroup = "description", metadata_file = NULL, organism = NULL,
  read_counts = TRUE, ...)
}
\arguments{
\item{upload_dir}{Path to final upload directory. This path is set when
running \code{bcbio_nextgen -w template}.}

\item{analysis}{Analysis type (e.g. \code{rnaseq} or \code{srnaseq})}

\item{intgroup}{Character vector of interesting groups. First entry is used
for plot colors during quality control (QC) analysis. Entire vector is used
for PCA and heatmap QC functions.}

\item{metadata_file}{Custom metadata file to import. Otherwise defaults to
sample metadata saved in the YAML file.}

\item{organism}{Organism name, following Ensembl/Biomart conventions. Must be
lowercase and one word (e.g. hsapiens). This will be detected automatically
for common reference genomes and normally does not need to be set.}

\item{read_counts}{Automatically read in the count data using
\code{read_bcbio_counts()}}
}
\value{
bcbio-nextgen run object
}
\description{
This function parses
output of bcbio-nextgen tool to allow (small) RNAseq analysis of samples
in different groups such as
characterization, differential expression and clustering. It creates an
\code{\link[bcbioRnaseq]{bcbioRnaDataSetFromFolder}} object.
}
\author{
Michael Steinbaugh

Lorena Pantano
}
