% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{bcbioRNAResults-class}
\alias{bcbioRNAResults-class}
\alias{bcbioRNAResults}
\title{\code{bcbioRNAResults}}
\description{
S4 class containing contrast-level results of a differential expression
analysis.
}
\author{
Michael Steinbaugh
}
