% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{bcbioRNASeq-class}
\alias{bcbioRNASeq-class}
\title{\code{bcbioRNASeq} Class}
\description{
\code{bcbioRNASeq} is an S4 class that extends \code{RangedSummarizedExperiment}, and
is designed to store a \href{https://bcbio-nextgen.readthedocs.org}{bcbio} RNA-seq
analysis.
}
\note{
\code{bcbioRNASeq} extended \code{SummarizedExperiment} prior to v0.2.0, where we
migrated to \code{RangedSummarizedExperiment}.
}
\examples{
uploadDir <- system.file("extdata/bcbio", package = "bcbioRNASeq")
x <- bcbioRNASeq(uploadDir)
print(x)
}
\seealso{
Other S4 Object: \code{\link{DESeqAnalysis-class}},
  \code{\link{DESeqAnalysis}},
  \code{\link{DESeqResultsTables-class}},
  \code{\link{DESeqResultsTables}}, \code{\link{coerce}},
  \code{\link{extract}}, \code{\link{show}},
  \code{\link{updateObject}}
}
\author{
Michael Steinbaugh, Lorena Pantano, Rory Kirchner, Victor Barrera
}
\concept{S4 Object}
