% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbioRnaseq.R
\docType{package}
\name{bcbioRnaseq}
\alias{bcbioRnaseq}
\alias{bcbioRnaseq-package}
\title{\code{bcbioRnaseq} package}
\description{
Quality control and differential expression for bcbio-nextgen RNA-seq
experiments.
}
\details{
See the README on \href{https://github.com/roryk/bcbioRnaseq}{GitHub}.
}
