% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{bcbioRNASeqCacheURL}
\alias{bcbioRNASeqCacheURL}
\title{Cache URL}
\format{An object of class \code{character} of length 1.}
\usage{
bcbioRNASeqCacheURL
}
\description{
Cache URL
}
\examples{
bcbioRNASeqCacheURL
}
\keyword{internal}
