% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{bcbioRnaSeqTestsUrl}
\alias{bcbioRnaSeqTestsUrl}
\title{Cache URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
bcbioRnaSeqTestsUrl
}
\description{
Cache URL
}
\examples{
bcbioRnaSeqTestsUrl
}
\keyword{internal}
