% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_bcbio.R, R/check_dds.R, R/check_dt.R,
%   R/check_res.R
\name{check_bcbio}
\alias{check_bcbio}
\alias{check_dds}
\alias{check_dt}
\alias{check_res}
\title{Perform object integrity checks}
\usage{
check_bcbio(bcbio)

check_dds(dds, stop = TRUE)

check_dt(dt, stop = TRUE)

check_res(res, stop = TRUE)
}
\arguments{
\item{bcbio}{bcbio run list object}

\item{dds}{DESeqDataSet}

\item{stop}{Stop if class doesn't match}

\item{dt}{DESeqTransform}

\item{res}{DESeqResults}
}
\description{
Perform object integrity checks
}
\author{
Michael Steinbaugh
}
