% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_bcbio.R
\name{check_bcbio}
\alias{check_bcbio}
\title{Perform integrity checks on a bcbio run object}
\usage{
check_bcbio(bcbio)
}
\arguments{
\item{bcbio}{bcbio run object}
}
\value{
Silent, otherwise stop execution if integrity check fails
}
\description{
Perform integrity checks on a bcbio run object
}
\author{
Michael Steinbaugh
}
\keyword{internal}
