% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/contrastName-methods.R
\docType{methods}
\name{contrastName}
\alias{contrastName}
\alias{contrastName,DESeqAnalysis-method}
\alias{contrastName,DESeqResults-method}
\alias{contrastName,DESeqResultsTables-method}
\title{Contrast Name}
\usage{
contrastName(object, ...)

\S4method{contrastName}{DESeqAnalysis}(object, results)

\S4method{contrastName}{DESeqResults}(object)

\S4method{contrastName}{DESeqResultsTables}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{results}{\code{scalar}. Position or name of \code{DESeqResults}.}
}
\value{
\code{string}. Contrast name.
}
\description{
Contrast Name
}
\examples{
data(deseq_small)

## DESeqAnalysis ====
contrastName(deseq_small, results = 1L)

## DESeqResults ====
object <- deseq_small@results[[1L]]
contrastName(object)
}
\author{
Michael Steinbaugh
}
