% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_heatmap.R
\name{correlation_heatmap}
\alias{correlation_heatmap}
\title{Correlation matrix heatmap}
\usage{
correlation_heatmap(run, counts, method = "pearson")
}
\arguments{
\item{run}{\code{bcbio-nextgen} run object}

\item{counts}{RNA-Seq counts as \code{DESeqDataSet} or \code{DESeqTransform}
object. If input as \code{DESeqDataSet}, an \code{rlog} transformation will
be applied.}

\item{method}{Correlation coefficient (or covariance) to be computed.
Defaults to \code{pearson} but \code{spearman} can also be used.}
}
\value{
Heatmap
}
\description{
Correlation matrix heatmap
}
\author{
Michael Steinbaugh
}
