% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basejump.R, R/methods-counts.R
\docType{methods}
\name{counts}
\alias{counts}
\alias{counts,bcbioRNADataSet-method}
\title{Count matrix accessors}
\usage{
\S4method{counts}{bcbioRNADataSet}(object, normalized = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional parameters.}

\item{normalized}{Select raw counts (\code{FALSE}), DESeq2 normalized counts
(\code{TRUE}), or additional normalization methods:
\itemize{
\item \code{tpm}: Transcripts per million.
\item \code{tmm}: Trimmed mean of M-values (edgeR).
\item \code{rlog}: Regularized log transformation (\code{\link[DESeq2:rlog]{DESeq2::rlog()}}).
\item \code{vst}: Variance stabilizing transformation
(\code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}}).
}}
}
\value{
Counts matrix
}
\description{
By default, \code{\link[=counts]{counts()}} returns the raw counts. This method will return
transcripts per million (TPM) by default when \code{normalized = TRUE}. This can
be overriden by requesting the normalization method directly.
}
\examples{
\dontrun{
# Raw counts
counts(bcb, normalized = FALSE)

# DESeq2 normalized counts
counts(bcb, normalized = TRUE)

# TPM
counts(bcb, normalized = "tpm")

# TMM
counts(bcb, normalized = "tmm")

# rlog
counts(bcb, normalized = "rlog")

# VST
counts(bcb, normalized = "vst")
}
}
\author{
Michael Steinbaugh
}
