% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_local_project.R
\name{create_local_project}
\alias{create_local_project}
\title{Create project directory structure}
\usage{
create_local_project()
}
\description{
Create project directory structure
}
\author{
Michael Steinbaugh
}
