% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_create_new_project.R
\name{create_new_project}
\alias{create_new_project}
\title{Create new project directory structure}
\usage{
create_new_project()
}
\description{
Generate the necessary directories for an RMarkdown report in RStudio
}
\author{
Michael Steinbaugh
}
\keyword{internal}
