% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{Create project directory structure}
\usage{
create_project()
}
\description{
Create project directory structure
}
\author{
Michael Steinbaugh
}
\keyword{internal}
