% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-files.R
\name{.custom_metadata}
\alias{.custom_metadata}
\title{Read custom metadata file}
\usage{
.custom_metadata(file, pattern = NULL, pattern_col = "description",
  lanes = NULL)
}
\arguments{
\item{file}{Metadata file. CSV and XLSX formats are supported.}

\item{pattern}{Apply grep pattern matching to samples}

\item{pattern_col}{Column in data frame used for pattern subsetting}

\item{lanes}{Number of lanes used to split the samples into technical
replicates (\code{_LXXX}) suffix.}
}
\value{
\link{DataFrame}.
}
\description{
Read custom metadata file
}
\author{
Michael Steinbaugh
}
\keyword{internal}
