% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/david.R
\name{run_david}
\alias{run_david}
\alias{david_table}
\title{Run \code{\link[RDAVIDWebService]{DAVIDWebService}}.}
\usage{
run_david(res_tbl, direction = "both", alpha = 0.05, count = 3,
  write_results = TRUE)

david_table(david)
}
\arguments{
\item{res_tbl}{\code{\link{res_tables}} list object.}

\item{direction}{Gene direction: up, down, or both.}

\item{alpha}{Alpha level cutoff.}

\item{count}{Minimum hit count.}

\item{write_results}{Write results as tab separated values (TSV) files to
disk.}

\item{david}{DAVID results list generated by \code{\link{run_david}}.}
}
\value{
List of \code{\link[RDAVIDWebService]{DAVIDWebService}} returns.

Cutoff chart \code{\link[knitr]{kable}}.
}
\description{
Wrapper that runs gene set enrichment analysis (GSEA) with
  \code{\link[RDAVIDWebService]{DAVIDWebService}}, using simplified input
  options.

Generate a DAVID results table for knit report.
}
\details{
Query the \href{https://david.ncifcrf.gov}{DAVID} website API with
\href{https://www.bioconductor.org/packages/release/bioc/html/RDAVIDWebService.html}{RDAVIDWebService}.
}
\seealso{
\itemize{
  \item \code{\link[RDAVIDWebService]{DAVIDWebService}}
  \item \code{\link[knitr]{kable}}
}
}
\author{
Michael Steinbaugh
}
