% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/david_table.R
\name{david_table}
\alias{david_table}
\title{Prepare a DAVID results table for knit report}
\usage{
david_table(david)
}
\arguments{
\item{david}{DAVID results list generated by \code{run_david}}
}
\value{
Cutoff chart kable
}
\description{
Prepare a DAVID results table for knit report
}
\author{
Michael Steinbaugh
}
