% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dds}
\alias{dds}
\title{Example \link{DESeqDataSet}}
\format{An object of class \code{DESeqDataSet} with 505 rows and 4 columns.}
\usage{
dds
}
\description{
Derived from \code{bcb}, the example \link{bcbioRNASeq} run.
}
\keyword{internal}
