% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dds_small}
\alias{dds_small}
\title{Minimal \code{DESeqDataSet} Example}
\format{An object of class \code{DESeqDataSet} with 500 rows and 6 columns.}
\usage{
dds_small
}
\description{
Minimal \code{DESeqDataSet} Example
}
\seealso{
Other GSE65267: \code{\link{bcb_small}},
  \code{\link{res_small}}, \code{\link{rld_small}}
}
\author{
Michael Steinbaugh, Lorena Pantano
}
\keyword{internal}
