% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_heatmap.R
\name{de_heatmap}
\alias{de_heatmap}
\title{Differential expression heatmap}
\usage{
de_heatmap(bcbio, dds, contrast, alpha = 0.1, lfc = 0)
}
\arguments{
\item{bcbio}{bcbio list object}

\item{dds}{\code{DESeqDataSet} object}

\item{contrast}{Contrast}

\item{alpha}{Alpha level cutoff}

\item{lfc}{Log fold change ratio (base 2) cutoff}
}
\value{
Gene heatmap
}
\description{
Differential expression heatmap
}
\author{
Michael Steinbaugh
}
