% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ma.R, R/plot_volcano.R
\name{plot_ma}
\alias{plot_ma}
\alias{plot_volcano}
\title{Differential expression plots}
\usage{
plot_ma(res, ylim = 2)

plot_volcano(bcbio, res, lfc = 1, text_labels = 20)
}
\arguments{
\item{res}{\code{DESeqResults} object}

\item{ylim}{Y-axis maximum (single integer)}

\item{lfc}{Log fold change ratio (base 2) cutoff for coloring}

\item{text_labels}{Number of text labels to plot}

\item{res}{DESeqResults}
}
\value{
MA plot

Volcano plot
}
\description{
Wrapper for \code{DESeq2::plotMA()} that enforces a symmetric
  y-axis and an automatic title

Volcano plot
}
\author{
Michael Steinbaugh
}
