% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_results.R
\name{de_results}
\alias{de_results}
\title{Differential gene expression results}
\usage{
de_results(bcbio, dds, alpha = 0.1, lfc = 0)
}
\arguments{
\item{bcbio}{bcbio list object}

\item{dds}{\code{DESeqDataSet} object}

\item{alpha}{Alpha level cutoff. Used for Wald test BH-adjusted P values.}

\item{lfc}{Log fold change ratio (base 2) cutoff. Does not apply to
statistical hypothesis testing, only gene filtering in the results tables.
See \code{?DESeq2::results} for additional information about using
\code{lfcThreshold} and \code{altHypothesis} to set an alternative
hypothesis based on expected fold changes.}
}
\value{
Results list
}
\description{
Differential gene expression results
}
\examples{
\dontrun{
de_results(bcbio, dds, alpha = 0.1, lfc = 0.25)
}
}
\author{
Michael Steinbaugh
}
