% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{defunct}
\alias{defunct}
\alias{plotGeneHeatmap}
\alias{txi}
\alias{resultsTables}
\title{Defunct Functions}
\usage{
plotGeneHeatmap(...)

txi(...)

resultsTables(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
\code{\link[=.Defunct]{.Defunct()}}.
}
\description{
Defunct Functions
}
\author{
Michael Steinbaugh
}
\keyword{internal}
