% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_heatmap.R
\name{deg_heatmap}
\alias{deg_heatmap}
\title{Differentially expressed gene (DEG) heatmap}
\usage{
deg_heatmap(bcbio, dds, contrast, alpha = 0.1, lfc = 0)
}
\arguments{
\item{bcbio}{bcbio list object}

\item{dds}{\code{DESeqDataSet} object}

\item{contrast}{Contrast}

\item{alpha}{Alpha level cutoff}

\item{lfc}{Log fold change ratio (base 2) cutoff}
}
\value{
Gene heatmap
}
\description{
Differentially expressed gene (DEG) heatmap
}
\author{
Michael Steinbaugh
}
