% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_tables.R
\name{deg_tables}
\alias{deg_tables}
\title{Differentially expressed gene (DEG) tables}
\usage{
deg_tables(bcbio, res, lfc = 0, write_csv = TRUE)
}
\arguments{
\item{bcbio}{bcbio list object}

\item{lfc}{Log fold change ratio (base 2) cutoff. Does not apply to
statistical hypothesis testing, only gene filtering in the results tables.
See \code{?DESeq2::results} for additional information about using
\code{lfcThreshold} and \code{altHypothesis} to set an alternative
hypothesis based on expected fold changes.}

\item{write_csv}{Write CSV files}

\item{dds}{\code{DESeqDataSet} object}
}
\value{
Results list
}
\description{
Differentially expressed gene (DEG) tables
}
\examples{
\dontrun{
deg_tables(bcbio, res, lfc = 0.25)
}
}
\author{
Michael Steinbaugh
}
