% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{download}
\title{Deprecated Functions}
\usage{
download(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
No value.
}
\description{
Deprecated Functions
}
\keyword{internal}
