% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deseq_small}
\alias{deseq_small}
\title{\code{DESeqAnalysis} Example}
\format{An object of class \code{DESeqAnalysis} of length 1.}
\usage{
deseq_small
}
\description{
Minimal DESeq2 differential expression analysis.
}
\examples{
deseq_small
}
\seealso{
Other GSE65267: \code{\link{bcb_small}}
}
\author{
Michael Steinbaugh, Lorena Pantano
}
\concept{GSE65267}
\keyword{datasets}
