% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DESeqAnalysis-class}
\alias{DESeqAnalysis-class}
\title{DESeq2 Differential Expression Analysis}
\description{
Class containing all elements generated during differential expression
analysis with DESeq2.
}
\section{Slots}{
\preformatted{- `data`: `DESeqDataSet`.
- `transform`: `DESeqTransform`.
- `results`: `list` containing one or more `DESeqResults`.
}
}

\examples{
x <- new(
    "DESeqAnalysis",
    data = dds_small,
    transform = vst_small,
    results = list(res_small)
)
class(x)
slotNames(x)
}
