% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DESeqResultsTables-class}
\alias{DESeqResultsTables-class}
\title{DESeq2 Differential Expression Results Tables}
\description{
\code{DESeqResults} object with \code{DataFrame} subsets and file path information.
}
\section{Slots}{

\describe{
\item{\code{all}}{\code{DESeqResults}. Original unmodified \code{DESeqResults}. Should contain
all genes, including those with \code{NA} adjusted \emph{P} values.}

\item{\code{deg}}{\code{DataFrame}. Subset containing genes that pass adjusted \emph{P} value
and log2 fold change cutoffs.}

\item{\code{degUp}}{\code{DataFrame}. Directional subset containing only upregulated
genes.}

\item{\code{degDown}}{\code{DataFrame}. Directional subset containing only downregulated
genes.}

\item{\code{localFiles}}{\code{list}. Local file paths.}

\item{\code{dropboxFiles}}{\code{list}. Dropbox file paths.}
}}

\seealso{
\code{\link[=DESeqResultsTables]{DESeqResultsTables()}}.

Other S4 Classes: \code{\link{DESeqAnalysis-class}},
  \code{\link{bcbioRNASeq-class}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Classes}
