% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerators.R
\name{DESeqResultsTables}
\alias{DESeqResultsTables}
\title{\code{DESeqResultsTables} Generator}
\usage{
DESeqResultsTables(object)
}
\arguments{
\item{object}{\code{DESeqResults}.}
}
\value{
\code{DESeqResultsTables}.
}
\description{
\code{DESeqResultsTables} Generator
}
\examples{
object <- deseq_small@results[[1L]]
class(object)
x <- DESeqResultsTables(object)
print(x)
}
\seealso{
Other S4 Generators: \code{\link{DESeqAnalysis}},
  \code{\link{bcbioRNASeq}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Generators}
