% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-design.R
\docType{methods}
\name{design}
\alias{design}
\alias{design,bcbioRNASeq-method}
\alias{design<-,bcbioRNASeq,formula-method}
\title{Design Formula Accessor}
\usage{
\S4method{design}{bcbioRNASeq}(object)

\S4method{design}{bcbioRNASeq,formula}(object) <- value
}
\arguments{
\item{object}{a \code{DESeqDataSet} object}

\item{value}{a \code{formula} used for estimating dispersion
and fitting Negative Binomial GLMs}
}
\description{
The design holds the R \code{formula} which expresses how the
counts depend on the variables in \code{colData}.
See \code{\link{DESeqDataSet}} for details.
}
\examples{
dds <- makeExampleDESeqDataSet(m=4)
design(dds) <- formula(~ 1)
}
\seealso{
\itemize{
\item \link[BiocGenerics:design]{BiocGenerics::design}.
\item \link[DESeq2:design]{DESeq2::design}.
}
}
\author{
Michael Steinbaugh
}
