% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-projectDir.R
\name{.logFile}
\alias{.logFile}
\title{Read Log File}
\usage{
.logFile(file)
}
\arguments{
\item{file}{Log file.}
}
\value{
Character vector.
}
\description{
Read Log File
}
\author{
Michael Steinbaugh
}
\keyword{internal}
