% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-prepareSummarizedExperiment.R
\name{.prepareSummarizedExperiment}
\alias{.prepareSummarizedExperiment}
\title{Prepare \code{SummarizedExperiment()}}
\usage{
.prepareSummarizedExperiment(assays, rowData, colData, metadata = NULL)
}
\arguments{
\item{assays}{List containing RNA-seq count matrices with matching
dimensions. Counts can be passed in either dense (\code{matrix}) or sparse
(\code{dgCMatrix}, \code{dgTMatrix}) format.}

\item{rowData}{Object describing assay matrix rows. Must support
\code{\link[base:dim]{base::dim()}}.}

\item{colData}{Object describing assay matrix columns. Must support
\code{\link[base:dim]{base::dim()}}.}

\item{metadata}{\emph{Optional}. Metadata list.}
}
\value{
\link{SummarizedExperiment}.
}
\description{
This is a utility wrapper for \code{SummarizedExperiment()} that provides
automatic subsetting for \code{colData} and \code{rowData}.
}
\details{
This function also provides automatic metadata slotting of multiple useful
environment parameters:
\itemize{
\item \code{date}: Today's date.
\item \code{wd}: Working directory.
\item \code{utilsSessionInfo}: \code{\link[utils:sessionInfo]{utils::sessionInfo()}} return.
\item \code{devtoolsSessionInfo}: \code{\link[devtools:session_info]{devtools::session_info()}} return.
}
}
\examples{
mat <- matrix(
    seq(1L:16L),
    nrow = 4L,
    ncol = 4L,
    dimnames = list(
        c("ENSMUSG00000000001",
          "ENSMUSG00000000003",
          "ENSMUSG00000000028",
          "ENSMUSG00000000031"),
        c("sample_1",
          "sample_2",
          "sample_3",
          "sample_4")))
assays <- list(assay = mat)
rowData <- data.frame(
    ensgene = c("ENSMUSG00000000001",
                "ENSMUSG00000000003",
                "ENSMUSG00000000028",
                "ENSMUSG00000000031"),
    biotype = c("coding",
                "coding",
                "coding",
                "pseudogene"),
    row.names = rownames(mat))
colData <- data.frame(
    genotype = c("wildtype",
                 "wildtype",
                 "knockout",
                 "knockout"),
    age = c(3L, 6L, 3L, 6L),
    row.names = colnames(mat))
.prepareSummarizedExperiment(
    assays = assays,
    rowData = rowData,
    colData = colData)
}
\seealso{
\itemize{
\item \code{\link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment::SummarizedExperiment()}}.
\item \code{\link[base:Sys.Date]{base::Sys.Date()}}.
\item \code{\link[base:getwd]{base::getwd()}}.
\item \code{\link[utils:sessionInfo]{utils::sessionInfo()}}.
}
}
\concept{bcbio Utilities}
\keyword{internal}
