% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-readSampleMetadataFile.R
\name{.readSampleMetadataFile}
\alias{.readSampleMetadataFile}
\title{Read Custom Metadata File}
\usage{
.readSampleMetadataFile(file, pattern = NULL, patternCol = "sampleName",
  lanes = 1L)
}
\arguments{
\item{file}{Metadata file. CSV and XLSX formats are supported.}

\item{pattern}{\emph{Optional}. Grep pattern to match against sample names.}

\item{patternCol}{\emph{Optional}. Column in data frame used for pattern
subsetting.}

\item{lanes}{\emph{Optional}. Number of lanes used to split the samples into
technical replicates (\code{_LXXX}) suffix.}
}
\value{
\link{tibble} grouped by \code{sampleName}.
}
\description{
Read Custom Metadata File
}
\author{
Michael Steinbaugh
}
\keyword{internal}
