% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-DESeq2.R
\name{.resContrastName}
\alias{.resContrastName}
\title{\code{DESeqResults} Contrast Name}
\usage{
.resContrastName(res)
}
\arguments{
\item{res}{\link{DESeqResults}.}
}
\value{
Contrast name string.
}
\description{
\code{DESeqResults} Contrast Name
}
\author{
Michael Steinbaugh
}
\keyword{internal}
