% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-tx2gene.R
\name{.tx2gene}
\alias{.tx2gene}
\title{Transcript to Gene Annotations}
\usage{
.tx2gene(projectDir, organism, release = "current")
}
\arguments{
\item{projectDir}{Project directory.}

\item{release}{Ensembl release version.}

\item{genomeBuild}{Genome build.}
}
\value{
\link{data.frame} with unique rownames.
}
\description{
Transcript to Gene Annotations
}
\author{
Michael Steinbaugh
}
\keyword{internal}
