% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-download.R
\docType{methods}
\name{download}
\alias{download}
\alias{download}
\alias{download,missing-method}
\alias{download,character-method}
\title{Download Dependency File}
\usage{
download(object)

\S4method{download}{missing}()

\S4method{download}{character}(object)
}
\arguments{
\item{object}{\emph{Optional}. File name. If \code{NULL} (default), download the
default dependency files for a new experiment.}
}
\value{
No value.
}
\description{
If the required dependency file isn't present, download latest version from
the \href{http://bioinformatics.sph.harvard.edu/}{HBC website}.
}
\details{
File download utility for RMarkdown knit reports.
}
\examples{
download("setup.R")
}
\author{
Michael Steinbaugh
}
