% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-ensembl.R
\name{.ensembl}
\alias{.ensembl}
\alias{.annotables}
\title{Ensembl annotations}
\usage{
.ensembl(genome_build)

.annotables(genome_build)
}
\arguments{
\item{run}{\link{bcbioRnaDataSet}.}
}
\value{
Tibble grouped by \code{ensembl_gene_id} with nested
\code{ensembl_transcript_id}.
}
\description{
Transcript-level annotations are downloaded from
\href{http://www.ensembl.org/}{Ensembl} and saved as a tibbled grouped by
\code{ensembl_gene_id} with nested \code{ensembl_transcript_id} in the bcbio run
object. We also define broad classes, which are used in quality control
analysis.
}
\author{
Michael Steinbaugh

Rory Kirchner
}
\keyword{internal}
