% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_annotations.R
\name{ensembl_annotations}
\alias{ensembl_annotations}
\title{Get gene annotations from Ensembl}
\usage{
ensembl_annotations(values = NULL, filters = "ensembl_gene_id", organism)
}
\arguments{
\item{values}{Ensembl gene identifier values. Optional but will run faster if
specified.}

\item{filters}{biomaRt filters. See \code{biomaRt::listFilters()}.}

\item{organism}{Organism identifier}
}
\value{
Data frame
}
\description{
Get gene annotations from Ensembl
}
\examples{
\dontrun{
ensembl_annotations("ENSMUSG00000000001", organism = "mmusculus")
}
}
\author{
Michael Steinbaugh
}
\keyword{internal}
