% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{examples}
\alias{examples}
\title{Example Output from \link{bcbioRNASeq} Workflow}
\format{An object of class \code{list} of length 5.}
\usage{
examples
}
\description{
Example Output from \link{bcbioRNASeq} Workflow
}
\author{
Lorena Pantano, Michael Steinbaugh
}
\keyword{internal}
