% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-methods.R
\name{extract}
\alias{extract}
\alias{[,bcbioRNASeq,ANY,ANY,ANY-method}
\title{Extract or replace parts of an object}
\usage{
\S4method{[}{bcbioRNASeq,ANY,ANY,ANY}(x, i, j, drop = FALSE, recalculate = TRUE)
}
\arguments{
\item{x}{Object.}

\item{i}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:

\if{html}{\out{<div class="sourceCode">}}\preformatted{help(topic = "Extract", package = "base")
}\if{html}{\out{</div>}}}

\item{j}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:

\if{html}{\out{<div class="sourceCode">}}\preformatted{help(topic = "Extract", package = "base")
}\if{html}{\out{</div>}}}

\item{drop}{For matrices and arrays.  If \code{TRUE} the result is
    coerced to the lowest possible dimension (see the examples).  This
    only works for extracting elements, not for the replacement.  See
    \code{\link[base]{drop}} for further details.
  }

\item{recalculate}{\code{logical(1)}.
Recalculate DESeq2 normalized counts and variance-stabilizing
transformations defined in \code{assays}.\cr
Recommended by default, but can take a long time for large datasets.\cr\
If \code{FALSE}, these assays will be removed automatically: \code{normalized},
\code{rlog}, \code{vst}.}
}
\value{
\code{bcbioRNASeq}.
}
\description{
Extract genes by row and samples by column.
}
\details{
Internal count transformations are rescaled automatically, if defined. DESeq2
transformations will only be updated when \code{recalculate = TRUE}.
}
\note{
Updated 2020-12-21.
}
\examples{
## bcbioRNASeq ====
data(bcb)
object <- bcb

## Minimum of 100 genes, 2 samples.
genes <- head(rownames(object), 100L)
head(genes)
samples <- head(colnames(object), 2L)
head(samples)

## Extract by sample name.
object[, samples]

## Extract by gene list.
object[genes, ]

## Extract by both genes and samples.
x <- object[genes, samples]
print(x)

## Fast subsetting, by skipping DESeq2 recalculations.
## Note that `normalized`, `rlog`, and `vst` assays will be removed.
x <- object[, samples, recalculate = FALSE]
print(x)
}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988)
  \emph{The New S Language}.
  Wadsworth & Brooks/Cole.
}
\author{
Michael Steinbaugh, Lorena Pantano
}
