% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-read_file.R
\name{.file}
\alias{.file}
\title{Read file from bcbio-nextgen final upload directory}
\usage{
.file(parent_dir, file_name, column_to_rownames = NULL, ...)
}
\arguments{
\item{parent_dir}{Parent directory.}

\item{file_name}{File name.}

\item{column_to_rownames}{Column identifier to use for row names.}

\item{...}{Passthrough parameters for \code{\link[=read_csv]{read_csv()}} or \code{\link[=read_tsv]{read_tsv()}}.}
}
\value{
Miscellaneous data frame.
}
\description{
Read file from bcbio-nextgen final upload directory
}
\author{
Michael Steinbaugh
}
\keyword{internal}
