% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gender_markers_mmusculus}
\alias{gender_markers_mmusculus}
\title{\emph{M. musculus} gender marker genes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 4 columns.}
\usage{
gender_markers_mmusculus
}
\description{
\emph{M. musculus} gender marker genes
}
\author{
Michael Steinbaugh
}
\keyword{datasets}
\keyword{internal}
