% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{genderMarkers}
\alias{genderMarkers}
\title{Gender Marker Genes}
\format{An object of class \code{list} of length 2.}
\usage{
genderMarkers
}
\description{
Gender Marker Genes
}
\author{
Michael Steinbaugh
}
\keyword{internal}
