% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{genderMarkersMmusculus}
\alias{genderMarkersMmusculus}
\title{\emph{M. musculus} Gender Marker Genes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 4 columns.}
\usage{
genderMarkersMmusculus
}
\description{
\emph{M. musculus} Gender Marker Genes
}
\author{
Michael Steinbaugh
}
\keyword{internal}
