% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-gene2symbol.R
\docType{methods}
\name{gene2symbol}
\alias{gene2symbol}
\alias{gene2symbol,bcbioRNASeq-method}
\title{Gene to Symbol Mappings}
\usage{
\S4method{gene2symbol}{bcbioRNASeq}(object)
}
\arguments{
\item{object}{Object.}
}
\description{
Gene to Symbol Mappings
}
\examples{
bcb <- examples[["bcb"]]
gene2symbol(bcb) \%>\% head()
}
\author{
Michael Steinbaugh
}
