% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_heatmap.R
\name{gene_heatmap}
\alias{gene_heatmap}
\title{Gene heatmap}
\usage{
gene_heatmap(bcbio, dds, alpha = 0.1, lfc = 1, de = TRUE)
}
\arguments{
\item{bcbio}{bcbio list object}

\item{dds}{\code{DESeqDataSet} object}

\item{alpha}{Alpha level cutoff}

\item{lfc}{Log fold change ratio (base 2) cutoff}

\item{de}{Plot differentially expressed (DE) genes. If \code{FALSE}, will
plot a heatmap of all genes.}
}
\value{
Gene heatmap
}
\description{
Gene heatmap
}
\author{
Michael Steinbaugh
}
