% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-general.R
\name{general}
\alias{general}
\title{General Arguments}
\arguments{
\item{i}{An integer or numeric scalar.}

\item{object}{Object.}

\item{title}{Title.}

\item{value}{Value to assign.}

\item{withDimnames}{A \code{logical}, indicating whether dimnames should be
applied to extracted assay elements.}

\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
No value.
}
\description{
General Arguments
}
\keyword{internal}
