% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_config_metadata.R
\name{import_config_metadata}
\alias{import_config_metadata}
\title{Metadata for bcbio run}
\usage{
import_config_metadata(bcbio, group = "treatment", lane_split = FALSE)
}
\arguments{
\item{bcbio}{bcbio run object}

\item{group}{Sample grouping. Used later in pipline for plot colors.}

\item{lane_split}{Whether samples were split across flow cell lanes.}
}
\value{
Metadata data frame
}
\description{
Metadata for bcbio run
}
\examples{
\dontrun{
import_config_metadata(bcbio, group = "treatment")
}
}
\author{
Michael Steinbaugh
}
