% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_file.R
\name{import_file}
\alias{import_file}
\title{Import a bcbio project data file}
\usage{
import_file(bcbio, file, rownames = NULL, ...)
}
\arguments{
\item{bcbio}{bcbio run object}

\item{file}{File name}

\item{rownames}{Column identifier to use for rownames}

\item{...}{Optional pass through to \code{readr} package}
}
\value{
bcbio data
}
\description{
Import a bcbio project data file
}
\examples{
\dontrun{
import_file(bcbio, file = "tx2gene.csv", rownames = "id")
}
}
\author{
Michael Steinbaugh
}
