% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_rnaseq_data.R
\name{import_rnaseq_data}
\alias{import_rnaseq_data}
\title{Import RNA-Seq data from a bcbio project}
\usage{
import_rnaseq_data(bcbio)
}
\arguments{
\item{bcbio}{bcbio run object}
}
\value{
List containing RNA-Seq data
}
\description{
Import RNA-Seq data from a bcbio project
}
\examples{
\dontrun{
import_rnaseq_data(bcbio)
}
}
\author{
Michael Steinbaugh
}
