% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sailfish.R
\name{import_sailfish}
\alias{import_sailfish}
\title{Import sailfish counts}
\usage{
import_sailfish(bcbio, summary)
}
\arguments{
\item{bcbio}{bcbio run object}

\item{summary}{Summary data frame}
}
\value{
txi \code{tximport} count data generated by \code{sailfish}
}
\description{
Import \code{bcbio-rnaseq} \code{sailfish} count data run using
\code{tximport}
}
\examples{
\dontrun{
import_sailfish(bcbio, summary)
}
}
\author{
Michael Steinbaugh
}
