% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_salmon.R
\name{import_salmon}
\alias{import_salmon}
\title{Import salmon counts}
\usage{
import_salmon(bcbio, summary)
}
\arguments{
\item{bcbio}{bcbio run object}

\item{summary}{Summary data frame}
}
\value{
txi \code{tximport} count data generated by \code{salmon}
}
\description{
Import \code{bcbio-rnaseq} \code{salmon} count data run using
\code{tximport}
}
\examples{
\dontrun{
import_sailfish(bcbio, summary)
}
}
\author{
Michael Steinbaugh
}
