% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_summary.R
\name{import_summary}
\alias{import_summary}
\title{Import a bcbio project summary}
\usage{
import_summary(bcbio, metadata)
}
\arguments{
\item{bcbio}{bcbio run object}

\item{metadata}{Metadata data frame}
}
\description{
Import a bcbio project summary
}
\examples{
\dontrun{
import_summary(bcbio)
}
}
\author{
Michael Steinbaugh
}
