% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-integrity.R
\name{.check_dds}
\alias{.check_dds}
\alias{.check_dt}
\alias{.check_res}
\alias{.check_sample_dirs}
\title{Object integrity checks}
\usage{
.check_dds(dds, stop = TRUE)

.check_dt(dt, stop = TRUE)

.check_res(res, stop = TRUE)

.check_sample_dirs(sample_dirs)
}
\arguments{
\item{dds}{\link{DESeqDataSet}.}

\item{stop}{Stop upon check failure.}

\item{dt}{\link{DESeqTransform}.}

\item{res}{\link{DESeqResults}.}

\item{sample_dirs}{Named character vector of sample directory paths.}
}
\description{
Object integrity checks
}
\author{
Michael Steinbaugh
}
\keyword{internal}
