% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{check_run}
\alias{check_run}
\alias{check_dds}
\alias{check_dt}
\alias{check_res}
\title{Perform object integrity checks}
\usage{
check_run(run)

check_dds(dds, stop = TRUE)

check_dt(dt, stop = TRUE)

check_res(res, stop = TRUE)
}
\arguments{
\item{run}{\link{bcbioRnaDataSet}.}

\item{dds}{\link{DESeqDataSet}.}

\item{stop}{Stop upon check failure.}

\item{dt}{\link{DESeqTransform}.}

\item{res}{\link{DESeqResults}.}
}
\description{
Perform object integrity checks
}
\author{
Michael Steinbaugh
}
\keyword{internal}
