% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-interestingGroups.R
\docType{methods}
\name{interestingGroups}
\alias{interestingGroups}
\alias{interestingGroups,bcbioRNASeq-method}
\alias{interestingGroups<-,bcbioRNASeq,character-method}
\alias{interestingGroups<-,bcbioRNASeq,NULL-method}
\title{Interesting Groups}
\usage{
\S4method{interestingGroups}{bcbioRNASeq}(object)

\S4method{interestingGroups}{bcbioRNASeq,character}(object) <- value

\S4method{interestingGroups}{bcbioRNASeq,`NULL`}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}
}
\value{
Character vector.
}
\description{
Interesting Groups
}
\examples{
bcb <- examples[["bcb"]]
interestingGroups(bcb)

# Assignment support
interestingGroups(bcb) <- "sampleName"
interestingGroups(bcb)
}
