% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-projectDir.R
\name{.dataVersions}
\alias{.dataVersions}
\title{Read Data Versions}
\usage{
.dataVersions(projectDir)
}
\arguments{
\item{projectDir}{Project directory.}
}
\value{
\link{data.frame}.
}
\description{
Read Data Versions
}
\author{
Michael Steinbaugh
}
\keyword{internal}
