% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-project_dir.R
\name{.program_versions}
\alias{.program_versions}
\title{Read program versions}
\usage{
.program_versions(project_dir)
}
\arguments{
\item{project_dir}{Project directory.}
}
\description{
Read program versions
}
