% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-smallrna.R
\name{.readSmallRNACounts}
\alias{.readSmallRNACounts}
\title{Create \code{isomiRs} Object from \code{bcbio} Output}
\usage{
.readSmallRNACounts(rna)
}
\arguments{
\item{rna}{\link{bcbioRNADataSet}.}
}
\description{
Read \code{bcbio} sample information from YAML to get \code{isomiR} object.
}
\author{
Lorena Patano
}
\keyword{internal}
