% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-sampleDirs.R
\name{.sampleDirs}
\alias{.sampleDirs}
\title{Detect Sample Directories}
\usage{
.sampleDirs(uploadDir)
}
\arguments{
\item{uploadDir}{Upload directory.}
}
\value{
Named character vector containing sample directory paths. Function
will \code{\link[=stop]{stop()}} if no complete sample directories match.
}
\description{
Detect Sample Directories
}
\author{
Michael Steinbaugh
}
\keyword{internal}
