% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{intgroup_as_factor}
\alias{intgroup_as_factor}
\title{Return metadata intgroups as factor}
\usage{
intgroup_as_factor(run)
}
\arguments{
\item{run}{bcbio-nextgen run.}
}
\description{
Return metadata intgroups as factor
}
\author{
Michael Steinbaugh
}
\keyword{internal}
