% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_metadata.R
\name{knit_metadata}
\alias{knit_metadata}
\title{Knit metadata table}
\usage{
knit_metadata(bcbio, metadata)
}
\arguments{
\item{bcbio}{\code{bcbio-nextgen} run object}

\item{metadata}{Metadata data frame}
}
\description{
Knit metadata table
}
\author{
Michael Steinbaugh
}
