% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_bcbio_run.R
\name{load_bcbio_run}
\alias{load_bcbio_run}
\title{Load bcbio run}
\usage{
load_bcbio_run(template, parent_dir = NULL, config_dir = NULL,
  final_dir = NULL)
}
\arguments{
\item{template}{Template name (YAML filename, without the extension)
originally called by \code{bcbio_nextgen.py}}

\item{parent_dir}{Parent directory containing the run folder}

\item{config_dir}{Set the config output directory, if non-standard}

\item{final_dir}{Set the final output directory, if non-standard}
}
\value{
List with directory paths
}
\description{
We recommend loading the \code{bcbio-nextgen} run as a remote connection over
\code{sshfs}. This requires setting \code{parent_dir} in the function.
}
\examples{
\dontrun{
load_bcbio_run("illumina_rnaseq")
}
}
\author{
Michael Steinbaugh
}
