% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_bcbio_run.R
\name{load_bcbio_run_from_yaml}
\alias{load_bcbio_run_from_yaml}
\title{Load bcbio run using YAML file}
\usage{
load_bcbio_run_from_yaml(yaml_file, intgroup = "description",
  organism = NULL, metadata = NULL, srnaseq = FALSE)
}
\arguments{
\item{yaml_file}{Path to final output yaml_file This file is inside
the final directory used when running \code{bcbio_nextgen -w template}.}

\item{intgroup}{Character vector of interesting groups. First entry is used
for plot colors during quality control (QC) analysis. Entire vector is used
for PCA and heatmap QC functions.}

\item{organism}{Organism name, following Ensembl/Biomart conventions. Must be
lowercase and one word (e.g. hsapiens). This will be detected automatically
for common reference genomes.}

\item{metadata}{Optional custom metadata file to import}

\item{srnaseq}{Oprional boolean indicating if the analysis was small RNAseq}
}
\value{
\code{bcbio-nextgen} run object
}
\description{
We recommend loading the \code{bcbio-nextgen} run as a remote connection over
\code{sshfs}. This requires setting \code{parent_dir} in the function.
}
\author{
Michael Steinbaugh
}
