% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_run.R
\name{load_run}
\alias{load_run}
\title{Load \code{bcbio} Run Data}
\usage{
load_run(upload_dir = "final", analysis = "rnaseq",
  interesting_groups = "sample_name", sample_metadata_file = NULL, ...)
}
\arguments{
\item{upload_dir}{Path to final upload directory. This path is set when
running \code{bcbio_nextgen -w template}.}

\item{analysis}{Analysis type. Supports RNA-seq (\code{rnaseq}; \strong{default}) or
small RNA-seq (\code{srnaseq}).}

\item{interesting_groups}{Character vector of interesting groups. First entry
is used for plot colors during quality control (QC) analysis. Entire vector
is used for PCA and heatmap QC functions.}

\item{sample_metadata_file}{\emph{Optional}. Custom metadata file containing
sample information. Otherwise defaults to sample metadata saved in the YAML
file.}

\item{...}{Additional arguments, slotted into the \code{\link[=metadata]{metadata()}} accessor.}
}
\value{
\link{bcbioRNADataSet}.
}
\description{
Simply point to the final upload directory output by
\href{https://bcbio-nextgen.readthedocs.io/}{bcbio}, and this function will take
care of the rest. It automatically imports RNA-seq counts, metadata, and
program versions used.
}
\note{
When working in RStudio, we recommend connecting to the bcbio-nextgen
run directory as a remote connection over
\href{https://github.com/osxfuse/osxfuse/wiki/SSHFS}{sshfs}.
}
\examples{
extra_dir <- system.file("extra", package = "bcbioRnaseq")
upload_dir <- file.path(extra_dir, "bcbio")
sample_metadata_file <- file.path(extra_dir, "sample_metadata.csv")
bcb <- load_run(
    upload_dir,
    interesting_groups = "group",
    sample_metadata_file = sample_metadata_file)
}
\author{
Michael Steinbaugh, Lorena Patano
}
