% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRNASeq.R
\name{loadRNASeq}
\alias{loadRNASeq}
\title{Load bcbio RNA-Seq Data}
\usage{
loadRNASeq(uploadDir, level = c("genes", "transcripts"),
  caller = c("salmon", "kallisto", "sailfish"), samples = NULL,
  sampleMetadataFile = NULL, interestingGroups = "sampleName", organism,
  ensemblRelease = NULL, genomeBuild = NULL, isSpike = NULL,
  gffFile = NULL, transformationLimit = 50L, ...)
}
\arguments{
\item{uploadDir}{Path to final upload directory. This path is set when
running \code{bcbio_nextgen -w template}.}

\item{level}{Import counts as "\code{genes}" (default) or "\code{transcripts}".}

\item{caller}{Expression caller. Supports "\code{salmon}" (default), "\code{kallisto}",
or "\code{sailfish}".}

\item{samples}{\emph{Optional.} Specify a subset of samples to load. The names
must match the \code{description} specified in the bcbio YAML metadata. If a
\code{sampleMetadataFile} is provided, that will take priority for sample
selection. Typically this can be left \code{NULL}.}

\item{sampleMetadataFile}{\emph{Optional.} Custom metadata file containing
sample information. Otherwise defaults to sample metadata saved in the YAML
file. Remote URLs are supported. Typically this can be left \code{NULL}.}

\item{organism}{Organism name. Use the full latin name (e.g.
"Homo sapiens"), since this will be input downstream to
AnnotationHub and ensembldb, unless \code{gffFile} is set.}

\item{ensemblRelease}{\emph{Optional.} Ensembl release version. If \code{NULL},
defaults to current release, and does not typically need to be
user-defined. Passed to AnnotationHub for \code{EnsDb} annotation matching,
unless \code{gffFile} is set.}

\item{genomeBuild}{\emph{Optional.} Ensembl genome build name (e.g. "GRCh38").
This will be passed to AnnotationHub for \code{EnsDb} annotation matching,
unless \code{gffFile} is set.}

\item{isSpike}{\emph{Optional.} Gene names corresponding to FASTA spike-in
sequences (e.g. ERCCs, EGFP, TDTOMATO).}

\item{gffFile}{\emph{Optional, not recommended.} By default, we recommend leaving
this \code{NULL} for genomes that are supported on Ensembl. In this case, the
row annotations (\code{\link[=rowRanges]{rowRanges()}}) will be obtained automatically from Ensembl
by passing the \code{organism}, \code{genomeBuild}, and \code{ensemblRelease} arguments to
AnnotationHub and ensembldb. For a genome that is not supported on Ensembl
and/or AnnotationHub, a GFF/GTF (General Feature Format) file is required.
Generally, we recommend using a GTF (GFFv2) file here over a GFF3 file if
possible, although all GFF formats are supported. The function will
internally generate a \code{TxDb} containing transcript-to-gene mappings and
construct a \code{GRanges} object containing the genomic ranges (\code{\link[=rowRanges]{rowRanges()}}).}

\item{transformationLimit}{Maximum number of samples to calculate
\code{\link[DESeq2:rlog]{DESeq2::rlog()}} and \code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}} matrix.
For large datasets, DESeq2 is slow to apply variance stabilization. In this
case, we recommend loading up the dataset in a high-performance computing
environment. Use \code{Inf} to always apply and \code{-Inf} to always skip.}

\item{...}{Additional arguments, slotted into the \code{\link[=metadata]{metadata()}} accessor.}
}
\value{
\code{bcbioRNASeq}.
}
\description{
Simply point to the final upload directory output by
\href{https://bcbio-nextgen.readthedocs.io/}{bcbio}, and this function will take
care of the rest. It automatically imports RNA-seq counts, metadata, and the
program versions used.
}
\details{
When the number of samples is bigger than the \code{transformationLimit}, \code{rlog}
and \code{vst} counts will not be slotted into \code{assays()}. In this case, we
recommend visualization using \code{\link[=tmm]{tmm()}} counts, which are automatically
calculated using edgeR.
}
\note{
When working in RStudio, we recommend connecting to the bcbio-nextgen
run directory as a remote connection over
\href{https://github.com/osxfuse/osxfuse/wiki/SSHFS}{sshfs}.
}
\examples{
uploadDir <- system.file("extdata/bcbio", package = "bcbioRNASeq")

# Gene level
loadRNASeq(
    uploadDir = uploadDir,
    level = "genes",
    organism = "Mus musculus"
)

# Transcript level
loadRNASeq(
    uploadDir = uploadDir,
    level = "transcripts",
    organism = "Mus musculus"
)
}
\seealso{
\code{help("bcbioRNASeq-class", "bcbioRNASeq")}.
}
\author{
Michael Steinbaugh, Lorena Pantano
}
