% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-loadRNASeqRun.R
\docType{methods}
\name{loadRNASeqRun}
\alias{loadRNASeqRun}
\alias{loadRNASeqRun}
\alias{loadRNASeqRun,character-method}
\title{Load bcbio RNA-Seq Run}
\usage{
loadRNASeqRun(object, ...)

\S4method{loadRNASeqRun}{character}(object = "final",
  interestingGroups = "sampleName", sampleMetadataFile = NULL, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments, slotted into the \code{\link[=metadata]{metadata()}} accessor.}

\item{interestingGroups}{Character vector of interesting groups. First entry
is used for plot colors during quality control (QC) analysis. Entire vector
is used for PCA and heatmap QC functions.}

\item{sampleMetadataFile}{\emph{Optional}. Custom metadata file containing
sample information. Otherwise defaults to sample metadata saved in the YAML
file.}

\item{uploadDir}{Path to final upload directory. This path is set when
running \code{bcbio_nextgen -w template}.}
}
\value{
\link{bcbioRNADataSet}.
}
\description{
Simply point to the final upload directory output by
\href{https://bcbio-nextgen.readthedocs.io/}{bcbio}, and this function will take
care of the rest. It automatically imports RNA-seq counts, metadata, and
program versions used.
}
\note{
When working in RStudio, we recommend connecting to the bcbio-nextgen
run directory as a remote connection over
\href{https://github.com/osxfuse/osxfuse/wiki/SSHFS}{sshfs}.
}
\examples{
extraDir <- system.file("extra", package = "bcbioRnaseq")
uploadDir <- file.path(extraDir, "bcbio")
sampleMetadataFile <- file.path(extraDir, "sample_metadata.csv")
bcb <- loadRNASeqRun(
    uploadDir,
    interestingGroups = "group",
    sampleMetadataFile = sampleMetadataFile)
}
\author{
Michael Steinbaugh, Lorena Pantano
}
