% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-project_dir.R
\name{.log_file}
\alias{.log_file}
\title{Read Log File}
\usage{
.log_file(file)
}
\arguments{
\item{file}{Log file.}
}
\description{
Read Log File
}
\author{
Michael Steinbaugh
}
\keyword{internal}
