% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-rmarkdown.R
\name{.mdResultsTables}
\alias{.mdResultsTables}
\title{Markdown List of Results Files}
\usage{
.mdResultsTables(resTbl, dir)
}
\arguments{
\item{resTbl}{List of results tables generated by \code{\link[=resultsTables]{resultsTables()}}.}

\item{dir}{Output directory.}
}
\value{
\code{\link[=writeLines]{writeLines()}}.
}
\description{
Enables looping of results contrast file links for RMarkdown.
}
\author{
Michael Steinbaugh
}
\keyword{internal}
