% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{melt_log10}
\alias{melt_log10}
\title{Melt RNA-seq count data to long format and log10 transform}
\usage{
melt_log10(bcb, counts = NULL)
}
\arguments{
\item{counts}{Counts matrix.}

\item{run}{\link{bcbioRnaDataSet}.}
}
\value{
log10 melted data frame.
}
\description{
Melt RNA-seq count data to long format and log10 transform
}
\author{
Michael Steinbaugh
}
\keyword{internal}
