% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-meltLog10.R
\docType{methods}
\name{meltLog10}
\alias{meltLog10}
\alias{meltLog10}
\alias{meltLog10,bcbioRNASeq-method}
\alias{meltLog10,DESeqDataSet-method}
\alias{meltLog10,DESeqTransform-method}
\title{Melt Count Matrix to Long Format and log10 Transform}
\usage{
meltLog10(object, ...)

\S4method{meltLog10}{bcbioRNASeq}(object, normalized = TRUE)

\S4method{meltLog10}{DESeqDataSet}(object, normalized = TRUE)

\S4method{meltLog10}{DESeqTransform}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{normalized}{Select normalized counts (\code{TRUE}), raw counts (\code{FALSE}),
or specifically request TMM-normalized counts (\code{tmm}).}
}
\value{
log10 melted \code{data.frame}.
}
\description{
Melt Count Matrix to Long Format and log10 Transform
}
\examples{
# bcbioRNASeq ====
meltLog10(bcb_small) \%>\% glimpse()

# DESeqDataSet ====
meltLog10(dds_small) \%>\% glimpse()

# DESeqTransform ====
meltLog10(rld_small) \%>\% glimpse()
}
\seealso{
\code{\link[reshape2:melt]{reshape2::melt()}}.
}
\author{
Michael Steinbaugh
}
