% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metadata.R
\docType{methods}
\name{metadata}
\alias{metadata}
\alias{metadata<-,bcbioRNASeq-method}
\title{Metadata}
\usage{
\S4method{metadata}{bcbioRNASeq}(x) <- value
}
\description{
Metadata
}
\examples{
# bcbioRNASeq ====
metadata(bcb_small)[["stash"]] <- "test"
metadata(bcb_small)[["stash"]]
}
\author{
Michael Steinbaugh
}
\keyword{internal}
