% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_kable.R
\name{metadata_kable}
\alias{metadata_kable}
\title{Metadata kable}
\usage{
metadata_kable(run, metadata)
}
\arguments{
\item{run}{\code{bcbio-nextgen} run}

\item{metadata}{Metadata data frame}
}
\value{
Data frame containing only the columns of interest
}
\description{
Returns a subset of metadata columns of interest used for knit reports. These
"interesting group" columns are defined as \code{intgroup} in the
\code{bcbio-nextgen} \code{run} object.
}
\author{
Michael Steinbaugh
}
