% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{metadata_table}
\alias{metadata_table}
\title{Metadata table}
\usage{
metadata_table(bcb)
}
\arguments{
\item{bcb}{\link{bcbioRnaDataSet}.}
}
\value{
Data frame containing only the columns of interest.
}
\description{
Returns a subset of metadata columns of interest used for knit reports. These
"interesting group" columns are defined as \code{groups_of_interest} in the
\link{bcbioRnaDataSet} object.
}
\author{
Michael Steinbaugh
}
