% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-metadataTable.R
\docType{methods}
\name{metadataTable}
\alias{metadataTable}
\alias{metadataTable}
\alias{metadataTable,bcbioRNADataSet-method}
\alias{metadataTable,DESeqDataSet-method}
\alias{metadataTable,DESeqTransform-method}
\title{Metadata Table}
\usage{
metadataTable(object, ...)

\S4method{metadataTable}{bcbioRNADataSet}(object, ...)

\S4method{metadataTable}{DESeqDataSet}(object, ...)

\S4method{metadataTable}{DESeqTransform}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional parameters, passed to \code{\link[=kable]{kable()}}.}
}
\value{
\link{data.frame} containing only the columns of interest.
}
\description{
Returns a subset of metadata columns of interest used for knit reports. These
"interesting group" columns are defined as \code{interestingGroups} in the
\link{bcbioRNADataSet} object.
}
\examples{
data(bcb, dds, rld)

# bcbioRNADataSet
metadataTable(bcb)

# DESeqDataSet
metadataTable(dds)

# DESeqTransform
metadataTable(dds)
}
\author{
Michael Steinbaugh
}
