% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-metrics.R
\docType{methods}
\name{metrics}
\alias{metrics}
\alias{metrics,bcbioRnaDataSet-method}
\title{Sample metrics}
\usage{
metrics(object)

\S4method{metrics}{bcbioRnaDataSet}(object)
}
\arguments{
\item{object}{\link{bcbioRnaDataSet} object.}
}
\description{
Sample metrics
}
