% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_kables.R
\name{multiple_kables}
\alias{multiple_kables}
\title{Knit multiple kables in a single chunk}
\usage{
multiple_kables(list)
}
\arguments{
\item{list}{List of tables (e.g. data frame, matrix)}
}
\value{
Knit tables, using \code{kable()}
}
\description{
Knit multiple kables in a single chunk
}
\examples{
multiple_kables(list(iris, mtcars))
}
\author{
Michael Steinbaugh
}
\keyword{internal}
