% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plot53Bias.R
\docType{methods}
\name{plot53Bias}
\alias{plot53Bias}
\alias{plot53Bias,bcbioRNADataSet-method}
\alias{plot53Bias,data.frame-method}
\title{Plot 5'->3' Bias}
\usage{
plot53Bias(object, ...)

\S4method{plot53Bias}{bcbioRNADataSet}(object, interestingGroup,
  warnLimit = 2L, flip = TRUE)

\S4method{plot53Bias}{data.frame}(object, interestingGroup = "sampleName",
  warnLimit = 2L, flip = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{interestingGroup}{Category to use to group samples (color and shape).
If unset, this is automatically determined by the metadata set inside the
\link{bcbioRNADataSet}.}

\item{warnLimit}{Threshold to plot warning color marker.}

\item{flip}{Flip X and Y axes.}
}
\value{
\link{ggplot}.
}
\description{
Plot 5'->3' Bias
}
\examples{
data(bcb)

# bcbioRNADataSet
plot53Bias(bcb)
plot53Bias(bcb, interestingGroup = "group")

\dontrun{
# data.frame
metrics(bcb) \%>\% plot53Bias
}
}
\seealso{
Other Quality Control Plots: \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotGenesDetected}}, \code{\link{plotGene}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh
}
\concept{Quality Control Plots}
