% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-plot_5_3_bias.R
\docType{methods}
\name{plot_5_3_bias}
\alias{plot_5_3_bias}
\alias{.plot_5_3_bias}
\alias{plot_5_3_bias,bcbioRNADataSet-method}
\alias{plot_5_3_bias,data.frame-method}
\title{Plot 5'->3' Bias}
\usage{
plot_5_3_bias(object, ...)

.plot_5_3_bias(metrics, interesting_group = "sample_name", warn_limit = 2L,
  flip = TRUE)

\S4method{plot_5_3_bias}{bcbioRNADataSet}(object, ...)

\S4method{plot_5_3_bias}{data.frame}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{metrics}{Metrics \link{data.frame} generated by \code{\link[=metrics]{metrics()}}.}

\item{interesting_group}{\emph{Optional}. Category to use to group samples (color
and shape). If unset, this is automatically determined by the metadata set
inside the \link{bcbioRNADataSet}.}

\item{warn_limit}{Threshold to plot warning color marker.}

\item{flip}{Flip X and Y axes.}
}
\value{
\link{ggplot}.
}
\description{
Plot 5'->3' Bias
}
\examples{
data(bcb)

# bcbioRNADataSet
plot_5_3_bias(bcb)

# data.frame
metrics <- metrics(bcb)
plot_5_3_bias(metrics)
}
\seealso{
Other Quality Control Plots: \code{\link{plot_count_density}},
  \code{\link{plot_counts_per_gene}},
  \code{\link{plot_exonic_mapping_rate}},
  \code{\link{plot_gene_detection_saturation}},
  \code{\link{plot_genes_detected}},
  \code{\link{plot_intronic_mapping_rate}},
  \code{\link{plot_mapped_reads}},
  \code{\link{plot_mapping_rate}},
  \code{\link{plot_rrna_mapping_rate}},
  \code{\link{plot_total_reads}}
}
\author{
Michael Steinbaugh
}
