% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_counts_per_gene.R
\name{plot_counts_per_gene}
\alias{plot_counts_per_gene}
\title{Counts per gene plot}
\usage{
plot_counts_per_gene(counts, metadata, flip = TRUE, print = TRUE)
}
\arguments{
\item{counts}{Counts matrix}

\item{metadata}{Metadata data frame}

\item{flip}{Flip x and y axes}

\item{print}{Print plot}
}
\value{
Boxplot
}
\description{
Counts per gene plot
}
\examples{
\dontrun{
plot_counts_per_gene(counts, metadata)
}
}
\author{
Michael Steinbaugh
}
