% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-plot_deg_heatmap.R
\docType{methods}
\name{plot_deg_heatmap}
\alias{plot_deg_heatmap}
\alias{.plot_deg_heatmap}
\alias{plot_deg_heatmap,DESeqResults,DESeqTransform-method}
\title{Differentially Expressed Gene Heatmap}
\usage{
plot_deg_heatmap(object, counts, ...)

.plot_deg_heatmap(object, counts, alpha = 0.05, lfc = 0L, ...)

\S4method{plot_deg_heatmap}{DESeqResults,DESeqTransform}(object, counts)
}
\arguments{
\item{object}{Object.}

\item{counts}{Secondary object containing normalized counts.}

\item{...}{Additional arguments.}

\item{alpha}{Alpha level cutoff.}

\item{lfc}{\link{log2} fold change ratio cutoff.}
}
\value{
Graphical output only.
}
\description{
This function is a simplified version of \code{\link[=plot_gene_heatmap]{plot_gene_heatmap()}} that is
optimized for handling a \link{DESeqResults} object rather a gene vector. All of
the optional parameters for \code{\link[=plot_gene_heatmap]{plot_gene_heatmap()}} are also available to this
function.
}
\examples{
data(bcb)
dds <- DESeqDataSetFromTximport(
    txi = txi(bcb),
    colData = colData(bcb),
    design = formula(~group)) \%>\%
    DESeq
res <- results(dds)
rld <- rlog(dds)
plot_deg_heatmap(res, rld)
}
\seealso{
Other Heatmaps: \code{\link{plot_correlation_heatmap}},
  \code{\link{plot_gene_heatmap}}
}
\author{
Michael Steinbaugh
}
