% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_deseq_pca.R
\name{plot_deseq_pca}
\alias{plot_deseq_pca}
\title{Plot PCA}
\usage{
plot_deseq_pca(run, dt, label = TRUE)
}
\arguments{
\item{run}{\code{bcbio-nextgen} run}

\item{dt}{\code{DESeqTransform} generated from \code{rlog()} or \code{vst()}
on a \code{DESeqDataSet} object. We prefer to pipe the transformed data in
to this function because the operation is CPU intensive, and is slow when
applied to multiple PCA plot operations.}

\item{label}{Superimpose sample text labels on the plot}
}
\value{
PCA plot
}
\description{
Wrapper for \code{DESeq2::plotPCA()} that improves PCA sample coloring and
labeling
}
