% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gender_markers.R
\name{plot_gender_markers}
\alias{plot_gender_markers}
\title{Graph sexually dimorphic marker genes}
\usage{
plot_gender_markers(counts, organism)
}
\arguments{
\item{counts}{Counts matrix. Transcripts per million (TPM) or normalized
counts are preferred.}

\item{organism}{Organism. \code{hsapiens} and \code{mmusculus} are supported.}
}
\value{
Scatterplot
}
\description{
Graph sexually dimorphic marker genes
}
\examples{
\dontrun{
plot_gender_markers(tpm, "mmusculus")
}
}
\author{
Michael Steinbaugh
}
