% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot_gene.R
\docType{methods}
\name{plot_gene,bcbioRNADataSet-method}
\alias{plot_gene,bcbioRNADataSet-method}
\title{Plot an individual gene}
\usage{
\S4method{plot_gene}{bcbioRNADataSet}(object, gene, format = "symbol")
}
\arguments{
\item{object}{Object.}

\item{gene}{Gene identifier. Can input multiple genes as a character vector.}

\item{format}{Ensembl identifier format. Defaults to the gene symbol (a.k.a.
\code{external_gene_name}).}
}
\description{
Plot an individual gene
}
\author{
Michael Steinbaugh
}
