% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-plot_gene_heatmap.R
\docType{methods}
\name{plot_gene_heatmap}
\alias{plot_gene_heatmap}
\alias{.plot_gene_heatmap}
\alias{plot_gene_heatmap,bcbioRNADataSet-method}
\alias{plot_gene_heatmap,DESeqDataSet-method}
\alias{plot_gene_heatmap,DESeqTransform-method}
\title{Gene Heatmap}
\usage{
plot_gene_heatmap(object, ...)

.plot_gene_heatmap(counts, genes = NULL, cluster_rows = TRUE,
  cluster_cols = TRUE, scale = "row", annotation_col = NULL, ...)

\S4method{plot_gene_heatmap}{bcbioRNADataSet}(object, ..., symbol = FALSE)

\S4method{plot_gene_heatmap}{DESeqDataSet}(object, ...)

\S4method{plot_gene_heatmap}{DESeqTransform}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments, passed to \code{\link[=pheatmap]{pheatmap()}}.}

\item{counts}{Counts matrix.}

\item{genes}{Character vector of specific gene identifiers to plot.}

\item{cluster_rows}{Use hierarchical clustering to arrange rows.}

\item{cluster_cols}{Use hierarchical clustering to arrange columns.}

\item{scale}{Character indicating if the values should be centered and scaled
in either the \code{row} direction, \code{column} direction, or \code{none}.}

\item{annotation_col}{\emph{Optional}. Alternative annotation to use. Useful when
plotting more than one column.}

\item{symbol}{Match against Ensembl gene symbols.}
}
\value{
Graphical output only.
}
\description{
These functions facilitate heatmap plotting of a specified set of genes. By
default, row- and column-wise hierarchical clustering is performed using the
Ward method, but this behavior can be overrided by setting \code{cluster_rows} or
\code{cluster_cols} to \code{FALSE}. When column clustering is disabled, the columns
are sorted by the interesting groups (\code{interesting_groups}) specified in the
\link{bcbioRNADataSet} and then the sample names.
}
\examples{
data(bcb)
genes <- counts(bcb) \%>\% rownames \%>\% .[1L:50L]

# bcbioRNADataSet
plot_gene_heatmap(bcb)
plot_gene_heatmap(bcb, genes, symbol = FALSE)

# DESeqDataSet
dds <- DESeqDataSetFromTximport(
    txi = txi(bcb),
    colData = colData(bcb),
    design = formula(~group)) \%>\%
    DESeq
plot_gene_heatmap(dds)

# DESeqTransform
rld <- rlog(dds)
plot_gene_heatmap(rld)
}
\seealso{
\code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}.

Other Heatmaps: \code{\link{plot_correlation_heatmap}},
  \code{\link{plot_deg_heatmap}}
}
\author{
Michael Steinbaugh
}
