% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genes_detected.R
\name{plot_genes_detected}
\alias{plot_genes_detected}
\title{Genes detected plot}
\usage{
plot_genes_detected(raw_counts, summary)
}
\arguments{
\item{raw_counts}{Raw counts matrix}

\item{summary}{\code{bcbio-rnaseq} summary report}
}
\description{
Genes detected plot
}
\examples{
\dontrun{
plot_genes_detected(raw_counts, summary)
}
}
\author{
Michael Steinbaugh
}
