% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot_intronic_mapping_rate.R
\name{plot_intronic_mapping_rate}
\alias{plot_intronic_mapping_rate}
\title{Plot Intronic Mapping Rate}
\usage{
plot_intronic_mapping_rate(bcb, warn_limit = 20L, interesting_groups = NULL)
}
\description{
Plot Intronic Mapping Rate
}
