% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ma.R
\name{plot_ma}
\alias{plot_ma}
\title{Plot MA}
\usage{
plot_ma(res, ylim = 2)
}
\arguments{
\item{res}{\code{DESeqResults} object}

\item{ylim}{Y-axis maximum (single integer)}
}
\value{
MA plot
}
\description{
Wrapper function for \code{DESeq2::plotMA()} that enforces a symmetric y-axis
and an automatic title
}
\author{
Michael Steinbaugh
}
