% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mapping_rate.R
\name{plot_mapping_rate}
\alias{plot_mapping_rate}
\title{Mapping rate plot}
\usage{
plot_mapping_rate(summary)
}
\arguments{
\item{summary}{\code{bcbio-rnaseq} summary report}
}
\description{
Mapping rate plot
}
\examples{
\dontrun{
plot_mapping_rate(summary)
}
}
\author{
Michael Steinbaugh
}
