% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-plot_mean_sd.R
\docType{methods}
\name{plot_mean_sd}
\alias{plot_mean_sd}
\alias{plot_mean_sd,bcbioRNADataSet-method}
\title{Plot Row Standard Deviations vs. Row Means}
\usage{
plot_mean_sd(object, ...)

\S4method{plot_mean_sd}{bcbioRNADataSet}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\link{ggplot} grid.
}
\description{
\code{\link[vsn:meanSdPlot]{vsn::meanSdPlot()}} wrapper that plots \code{\link[=log2]{log2()}}, \code{\link[=rlog]{rlog()}}, and
\code{\link[=varianceStabilizingTransformation]{varianceStabilizingTransformation()}} normalized counts.
}
\examples{
data(bcb)
plot_mean_sd(bcb)
}
\seealso{
Other DESeq2 Utilities: \code{\link{plot_dispersion}}
}
\author{
Michael Steinbaugh, Lorena Patano
}
