% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-plot_mirna_counts.R
\docType{methods}
\name{plot_mirna_counts}
\alias{plot_mirna_counts}
\alias{plot_mirna_counts,bcbioRNADataSet-method}
\title{Plot of Total miRNA Counts}
\usage{
plot_mirna_counts(object, ...)

\S4method{plot_mirna_counts}{bcbioRNADataSet}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\link{ggplot}.
}
\description{
Plot of Total miRNA Counts
}
\author{
Lorena Pantano, Michael Steinbaugh
}
