% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot principal component analysis (PCA)}
\usage{
plot_pca(run, dt, label = TRUE, ntop = 500)
}
\arguments{
\item{run}{bcbio-nextgen run.}

\item{dt}{\linkS4class{DESeqTransform} generated from \code{\link[DESeq2:rlog]{DESeq2::rlog()}} or
\code{\link[DESeq2:vst]{DESeq2::vst()}} on a \linkS4class{DESeqDataSet}.}

\item{label}{Superimpose sample text labels on the plot.}

\item{ntop}{number of genes to use for PCA.
To plot all gene variation, use \code{ntop = nrow(.)}}
}
\value{
ggplot2 object.
}
\description{
Wrapper for \code{\link[DESeq2:plotPCA]{DESeq2::plotPCA()}} that improves PCA sample coloring and
labeling.
}
\seealso{
\code{\link[DESeq2]{plotPCA}}
}
