% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-plot_pca_covariates.R
\docType{methods}
\name{plot_pca_covariates}
\alias{plot_pca_covariates}
\alias{plot_rrna_mapping_rate}
\alias{plot_pca_covariates,bcbioRNADataSet-method}
\title{Find Correlation Between Principal Components (PCs) and Covariates}
\usage{
plot_pca_covariates(object, ...)

plot_rrna_mapping_rate(object, ...)

\S4method{plot_pca_covariates}{bcbioRNADataSet}(object, transform = "rlog",
  use = NULL, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[DEGreport:degCovariates]{DEGreport::degCovariates()}}.}

\item{transform}{String specifying \link{rlog} (\strong{recommended}) or \link{vst}
\link{DESeqTransform} slotted inside the \link{bcbioRNADataSet}.}

\item{use}{\emph{Optional}. Character vector of columns to use.}
}
\value{
\link{ggplot}.
}
\description{
\code{\link[DEGreport:degCovariates]{DEGreport::degCovariates()}} wrapper supporting a \link{bcbioRNADataSet}.
}
\seealso{
\code{\link[DEGreport:degCovariates]{DEGreport::degCovariates()}}.
}
\author{
Lorena Pantano, Michael Steinbaugh
}
