% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot_rrna_mapping_rate.R
\name{plot_rrna_mapping_rate}
\alias{plot_rrna_mapping_rate}
\title{Plot rRNA Mapping Rate}
\usage{
plot_rrna_mapping_rate(bcb, warn_limit = 10L, interesting_groups = NULL)
}
\description{
Plot rRNA Mapping Rate
}
\author{
Michael Steinbaugh
}
