% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R,
%   R/methods-plot_size_distribution.R
\docType{methods}
\name{plot_size_distribution}
\alias{plot_size_distribution}
\alias{plot_size_distribution,bcbioRNADataSet-method}
\title{Plot Size Distribution of Small RNA-Seq Data}
\usage{
plot_size_distribution(object, ...)

\S4method{plot_size_distribution}{bcbioRNADataSet}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\link{ggplot}.
}
\description{
Plot Size Distribution of Small RNA-Seq Data
}
\seealso{
Other Small RNA-Seq Utilities: \code{\link{plot_srna_clusters}}
}
\author{
Lorena Pantano, Michael Steinbaugh
}
