% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-plot_volcano.R
\docType{methods}
\name{plot_volcano}
\alias{plot_volcano}
\alias{.plot_volcano}
\alias{plot_volcano,DESeqResults-method}
\title{Plot Volcano}
\usage{
plot_volcano(object, ...)

.plot_volcano(res_df, padj = TRUE, alpha = 0.05, lfc = 1L,
  text_labels = 0L, direction = "both", shade_color = "green",
  shade_alpha = 0.25, point_color = "gray", point_alpha = 0.75,
  point_outline_color = "darkgray", grid = TRUE)

\S4method{plot_volcano}{DESeqResults}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{res_df}{\link{DESeqResults} coerced to a \link{data.frame}. By inputting a
\link{data.frame}, we allow for increased flexibility, since a user may want to
subset the results prior to plotting.}

\item{padj}{Use P values adjusted for multiple comparisions.}

\item{lfc}{Log fold change ratio (base 2) cutoff for coloring.}

\item{direction}{Plot \code{up}, \code{down}, or \code{both} (\strong{default}) directions.}

\item{shade_color}{Shading color for bounding box.}

\item{shade_alpha}{Shading transparency alpha.}

\item{point_color}{Point color.}

\item{point_alpha}{Point transparency alpha.}

\item{point_outline_color}{Point outline color.}

\item{grid}{Arrange plots into grid.}

\item{ntop}{Number of top genes to label.}

\item{title}{\emph{Optional}. Plot title.}
}
\value{
Volcano plot arranged as grid (\code{grid = TRUE}), or \code{\link[=show]{show()}}
individual \link{ggplot} (\code{grid = FALSE}).
}
\description{
Plot Volcano
}
\examples{
data(bcb)
dds <- DESeqDataSetFromTximport(
    txi = txi(bcb),
    colData = colData(bcb),
    design = formula(~group)) \%>\%
    DESeq
res <- results(dds)
plot_volcano(res)
plot_volcano(res, padj = FALSE, lfc = 2L)
plot_volcano(res, text_labels = 10L)
}
\seealso{
This function is an updated variant of
\code{CHBUtils::volcano_density_plot()}.

Other Differential Expression Plots: \code{\link{plot_ma}}
}
\author{
John Hutchinson, Michael Steinbaugh, Lorena Pantano
}
