% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotCountDensity.R
\docType{methods}
\name{plotCountDensity}
\alias{plotCountDensity}
\alias{plotCountDensity}
\alias{plotCountDensity,bcbioRNADataSet-method}
\alias{plotCountDensity,data.frame-method}
\title{Plot Count Density}
\usage{
plotCountDensity(object, ...)

\S4method{plotCountDensity}{bcbioRNADataSet}(object, normalized = "tmm")

\S4method{plotCountDensity}{data.frame}(object,
  interestingGroup = "sampleName")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{normalized}{Count normalization method. See \code{\link[=counts]{counts()}} documentation
for more information.}

\item{interestingGroup}{\emph{Optional}. Category to use to group samples (color
and shape). If unset, this is automatically determined by the metadata set
inside the \link{bcbioRNADataSet}.}
}
\value{
\link{ggplot}.
}
\description{
Plot Count Density
}
\examples{
data(bcb)

# bcbioRNADataSet
plotCountDensity(bcb)

# data.frame
meltLog10(bcb) \%>\% plotCountDensity
}
\seealso{
Other Quality Control Plots: \code{\link{plot53Bias}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotGenesDetected}}, \code{\link{plotGene}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
