% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCounts-methods.R
\docType{methods}
\name{plotCounts}
\alias{plotCounts}
\alias{plotCounts,bcbioRNASeq-method}
\title{Plot counts}
\usage{
plotCounts(object, ...)

\S4method{plotCounts}{bcbioRNASeq}(object, genes, normalized = c("tpm",
  "sf", "fpkm", "vst", "rlog", "tmm", "rle"), ...)
}
\arguments{
\item{object}{Object.}

\item{genes}{\code{character}.
Gene identifiers.}

\item{normalized}{\code{character(1)} or \code{logical(1)}.
Normalization method to apply:
\itemize{
\item \code{FALSE}: Raw counts.
When using a \href{https://bioconductor.org/packages/tximport/}{tximport}-compatible caller, these are length scaled
by default (see \code{countsFromAbundance} argument).
When using a \href{http://bioinf.wehi.edu.au/featureCounts/}{featureCounts}-compatible caller, these are \code{integer}.
}

\href{https://bioconductor.org/packages/tximport/}{tximport} caller-specific normalizations:
\itemize{
\item \code{"tpm"}: \strong{T}ranscripts \strong{p}er \strong{m}illion.
}

Additional gene-level-specific normalizations:
\itemize{
\item \code{TRUE} / \code{"sf"}: \strong{S}ize \strong{f}actor (i.e. library size) normalized
counts.\cr
See \code{\link[DESeq2:sizeFactors]{DESeq2::sizeFactors()}} for details.
\item \code{"fpkm"}: \strong{F}ragments \strong{p}er \strong{k}ilobase per \strong{m}illion mapped
fragments.\cr
Requires \code{fpkm = TRUE} in \code{\link[=bcbioRNASeq]{bcbioRNASeq()}} call and gene annotations in
\code{\link[SummarizedExperiment:rowRanges]{rowRanges()}} with defined
\code{\link[BiocGenerics:width]{width}}.\cr
See \code{\link[DESeq2:fpkm]{DESeq2::fpkm()}} for details.
\item \code{"vst"}: \strong{V}ariance-\strong{s}tabilizing \strong{t}ransformation (log2).\cr
Requires \code{vst = TRUE} to be set during \code{\link[=bcbioRNASeq]{bcbioRNASeq()}} call.\cr
See \code{DESeq2::varianceStabilizingTransformation} for more information.
\item \code{"rlog"}: \strong{R}egularized \strong{log} transformation (log2).\cr
Requires \code{rlog = TRUE} to be set during \code{\link[=bcbioRNASeq]{bcbioRNASeq()}} call.\cr
See \code{\link[DESeq2:rlog]{DESeq2::rlog()}} for details.
\item \code{"tmm"}: \strong{T}rimmed \strong{m}ean of \strong{M}-values.\cr
Calculated on the fly.\cr
See \code{\link[edgeR:calcNormFactors]{edgeR::calcNormFactors()}} for details.
\item \code{"rle"}: \strong{R}elative \strong{l}og \strong{e}xpression transformation.\cr
Calculated on the fly.\cr
See \code{\link[=relativeLogExpression]{relativeLogExpression()}} for details.

Note that \code{logical(1)} support only applies to \code{counts()}. Other
functions in the package require \code{character(1)} and use
\code{\link[base:match.arg]{match.arg()}} internally.
}}

\item{...}{Passthrough to \code{SummarizedExperiment} method defined in acidplots.
See \code{\link[acidplots:plotCounts]{acidplots::plotCounts()}} for details.}
}
\value{
\itemize{
\item \code{style = "facet"}: \code{ggplot} grouped by \code{sampleName}, with
\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} applied to panel the samples.
\item \code{style = "wide"}: \code{ggplot} in wide format, with genes on the x-axis.
}
}
\description{
Plot counts
}
\note{
Updated 2019-09-16.
}
\section{Methods (by class)}{

\itemize{
\item \code{DESeqDataSet}: Automatically plots normalized counts. Arguments pass
through to \code{SummarizedExperiment} method, but \code{assay} and \code{countAxisLabel}
are automatic.
}
}

\examples{
data(bcb)

g2s <- basejump::Gene2Symbol(bcb)
geneIDs <- head(g2s[["geneID"]])
print(geneIDs)
geneNames <- head(g2s[["geneName"]])
print(geneNames)

plotCounts(bcb, genes = geneIDs, style = "facet")
plotCounts(bcb, genes = geneNames, style = "wide")
}
\seealso{
\code{DESeq2::plotCounts()}.
}
\author{
Michael Steinbaugh
}
