% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCountsPerFeature-methods.R
\docType{methods}
\name{plotCountsPerFeature}
\alias{plotCountsPerFeature}
\alias{plotCountsPerFeature,bcbioRNASeq-method}
\alias{plotCountDensity}
\title{Plot counts per feature}
\usage{
plotCountsPerFeature(object, ...)

\S4method{plotCountsPerFeature}{bcbioRNASeq}(object,
  normalized = c("tmm", "vst", "rlog", "rle", "tpm", "fpkm"), ...)

plotCountDensity(object, ...)
}
\arguments{
\item{object}{Object.}

\item{normalized}{\code{character(1)}.
Type of normalized counts to visualize:
\itemize{
\item \code{tpm} Transcripts per million.
\item \code{tmm}: edgeR trimmed mean of M-values.
\item \code{rlog}: DESeq2 log2 regularized log transformation.
\item \code{vst}: DESeq2 log2 variance stabilizing transformation.
}}

\item{...}{Passthrough to \code{SummarizedExperiment} method defined in acidplots.
See \code{\link[acidplots:plotCountsPerFeature]{acidplots::plotCountsPerFeature()}} for details.}
}
\value{
\code{ggplot}.
}
\description{
Generally, we expect similar count spreads for all genes between samples
unless the library sizes or total RNA expression are different.
}
\note{
Updated 2019-09-15.
}
\section{Trimmed Mean of M-Values}{


We recommend visualizing counts normalized with the \strong{T}rimmed \strong{M}ean of
\strong{M}-Values (TMM) method here. TMM normalization equates the overall
expression levels of genes between samples under the assumption that the
majority of them are not differentially expressed. Therefore, by normalizing
for total RNA expression by sample, we expect the spread of the
TMM-normalized counts per gene to be similar for every sample.
}

\section{Methods (by class)}{

\itemize{
\item \code{SingleCellExperiment}: Applies \code{\link[=aggregateCellsToSamples]{aggregateCellsToSamples()}}
calculation to summarize at sample level prior to plotting.
}
}

\examples{
data(bcb)
plotCountsPerFeature(bcb)
}
\references{
TMM: Robinson, et al., 2010.
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
