% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEGPCA-methods.R
\docType{methods}
\name{plotDEGPCA}
\alias{plotDEGPCA}
\alias{plotDEGPCA,DESeqResults-method}
\alias{plotDEGPCA,DESeqAnalysis-method}
\title{Plot DEG PCA}
\usage{
\S4method{plotDEGPCA}{DESeqResults}(object, counts, direction = c("both",
  "up", "down"), interestingGroups = NULL,
  color = getOption("basejump.color", NULL),
  label = getOption("basejump", FALSE), return = c("ggplot",
  "DataFrame"))

\S4method{plotDEGPCA}{DESeqAnalysis}(object, counts = NULL, results,
  direction = c("both", "up", "down"), interestingGroups = NULL,
  color = getOption("basejump.color", NULL),
  label = getOption("basejump", FALSE), return = c("ggplot",
  "DataFrame"))
}
\arguments{
\item{object}{Object.}

\item{counts}{\code{DESeqTransform}.}

\item{direction}{\code{string}. Plot "\code{both}", "\code{up}", or "\code{down}" directions.}

\item{interestingGroups}{\code{character} or \code{NULL}. Groups of interest that
define the samples. If left unset, defaults to \code{sampleName}.}

\item{color}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{ggplot2::scale_color_manual()}. To set the discrete color
palette globally, use
\code{options(basejump.discrete.color = ggplot2::scale_color_viridis_d())}.}

\item{label}{\code{boolean}. Superimpose sample text labels on the plot.}

\item{return}{\code{string}. Return type. Uses \code{base::match.arg()}
internally and defaults to the first argument in the \code{character} vector.}

\item{results}{\code{scalar}. Position or name of \code{DESeqResults}.}
}
\value{
\code{ggplot}.
}
\description{
Plot DEG PCA
}
\details{
Wrapper for \code{\link[DESeq2:plotPCA]{DESeq2::plotPCA()}} that improves principal component analysis
(PCA) sample coloring and labeling. Always plots using normalized counts.

PCA (Jolliffe, et al., 2002) is a multivariate technique that allows us to
summarize the systematic patterns of variations in the data. PCA takes the
expression levels for genes and transforms it in principal component space,
reducing each sample into one point. Thereby, we can separate samples by
expression variation, and identify potential sample outliers. The PCA plot is
a way to look at how samples are clustering.
}
\examples{
data(deseq)

## DESeqAnalysis ====
plotDEGPCA(deseq)

## DESeqResults ====
plotDEGPCA(
    object = as(deseq, "DESeqResults"),
    counts = as(deseq, "DESeqTransform")
)
}
\references{
Jolliffe, et al., 2002.
}
\seealso{
\itemize{
\item \code{\link[DESeq2:plotPCA]{DESeq2::plotPCA()}}.
\item \code{getMethod("plotPCA", "DESeqTransform")}
}
}
\author{
Michael Steinbaugh
}
