% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotDispersion.R
\docType{methods}
\name{plotDispersion}
\alias{plotDispersion}
\alias{plotDispersion}
\alias{plotDispersion,bcbioRNADataSet-method}
\alias{plotDispersion,DESeqDataSet-method}
\title{Plot Dispersion Estimates}
\usage{
plotDispersion(object)

\S4method{plotDispersion}{bcbioRNADataSet}(object)

\S4method{plotDispersion}{DESeqDataSet}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Dispersion plot \link{ggplot}.
}
\description{
\code{\link[DESeq2:plotDispEsts]{DESeq2::plotDispEsts()}} wrapper supporting a \link{bcbioRNADataSet}.
}
\examples{
data(bcb, dds)

# bcbioRNADataSet
plotDispersion(bcb)

# DESeqDataSet
plotDispersion(dds)
}
\seealso{
\code{\link[DESeq2:plotDispEsts]{DESeq2::plotDispEsts()}}.

Other DESeq2 Utilities: \code{\link{plotMeanSD}}
}
\author{
Michael Steinbaugh
}
