% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotDispEsts.R
\docType{methods}
\name{plotDispEsts}
\alias{plotDispEsts}
\alias{plotDispEsts,bcbioRNASeq-method}
\title{Plot Dispersion Estimates}
\usage{
\S4method{plotDispEsts}{bcbioRNASeq}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[DESeq2:plotDispEsts]{DESeq2::plotDispEsts()}}.}
}
\value{
\link{ggplot}.
}
\description{
Plot Dispersion Estimates
}
\examples{
plotDispEsts(bcb)

\dontrun{
plotDispEsts(
    bcb,
    genecol = "gray",
    fitcol = "purple",
    finalcol = "orange")
}
}
\seealso{
\code{\link[DESeq2:plotDispEsts]{DESeq2::plotDispEsts()}}.

Other Differential Expression Utilities: \code{\link{alphaSummary}},
  \code{\link{plotMeanSD}}
}
\author{
Michael Steinbaugh
}
\concept{Differential Expression Utilities}
