% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/plotFeatureSaturation-methods.R
\name{plotFeatureSaturation}
\alias{plotFeatureSaturation}
\alias{plotFeatureSaturation,bcbioRNASeq-method}
\title{Plot feature (gene) detection saturation}
\usage{
plotFeatureSaturation(object, ...)

\S4method{plotFeatureSaturation}{bcbioRNASeq}(
  object,
  interestingGroups = NULL,
  minCounts = 1L,
  perMillion = TRUE,
  trendline = FALSE,
  label = getOption(x = "acid.label", default = FALSE),
  labels = list(title = "Gene saturation", subtitle = NULL, x = "mapped reads", y =
    "gene count")
)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{minCounts}{\code{integer(1)}.
Minimum number of counts per feature (i.e. gene).}

\item{perMillion}{\code{logical(1)}.
Display as counts per million.}

\item{trendline}{\code{logical(1)}.
Include trendline on plot.}

\item{label}{\code{logical(1)}.
Superimpose sample text labels on the plot.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{ggplot2::labs()} for details.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
We should observe a linear trend in the number of genes detected with the
number of mapped reads, which indicates that the sample input was not
overloaded.
}
\note{
Updated 2023-10-05.
}
\examples{
data(bcb)

## bcbioRNASeq ====
plotFeatureSaturation(bcb, label = FALSE)
plotFeatureSaturation(bcb, label = TRUE)
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
