% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGene-methods.R
\docType{methods}
\name{plotGene}
\alias{plotGene}
\alias{plotGene,bcbioRNASeq-method}
\alias{plotGene,DESeqAnalysis-method}
\alias{plotGene,DESeqTransform-method}
\alias{plotGene,DESeqDataSet-method}
\title{Plot Gene Expression}
\usage{
\S4method{plotGene}{bcbioRNASeq}(object, genes, normalized = c("vst",
  "rlog", "tmm", "tpm", "rle"), interestingGroups = NULL,
  medianLine = TRUE, color = getOption("basejump.discrete.color",
  NULL), legend = getOption("basejump.legend", TRUE), headerLevel = 2L,
  return = c("facet", "wide"))

\S4method{plotGene}{DESeqAnalysis}(object, genes,
  interestingGroups = NULL, medianLine = TRUE,
  color = getOption("basejump.discrete.color", NULL),
  legend = getOption("basejump.legend", TRUE), headerLevel = 2L,
  return = c("facet", "wide"))

\S4method{plotGene}{DESeqTransform}(object, genes,
  interestingGroups = NULL, medianLine = TRUE,
  color = getOption("basejump.discrete.color", NULL),
  legend = getOption("basejump.legend", TRUE), headerLevel = 2L,
  return = c("facet", "wide"))

\S4method{plotGene}{DESeqDataSet}(object, genes,
  interestingGroups = NULL, medianLine = TRUE,
  color = getOption("basejump.discrete.color", NULL),
  legend = getOption("basejump.legend", TRUE), headerLevel = 2L,
  return = c("facet", "wide"))
}
\arguments{
\item{object}{Object.}

\item{genes}{\code{character}. Gene identifiers.}

\item{interestingGroups}{\code{character}. Groups of interest.}

\item{medianLine}{\code{boolean}. Include median line for each group. Disabled if
samples are colored by sample name.}

\item{headerLevel}{\code{scalar integer}. Markdown header level (1-7).}

\item{return}{Object class to return. Uses \code{\link[=match.arg]{match.arg()}} internally and picks
the first item in the vector by default.}
}
\value{
\itemize{
\item "\code{facet}": \code{ggplot} grouped by \code{sampleName}, with \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}
applied to panel the samples.
\item "\code{wide}": Show \code{ggplot} in wide format, with genes on the x-axis.
}
}
\description{
Plot Gene Expression
}
\examples{

g2s <- gene2symbol(bcb_small)
geneIDs <- head(g2s[["geneID"]])
print(geneIDs)
geneNames <- head(g2s[["geneName"]])
print(geneNames)

# bcbioRNASeq ====
object <- bcb_small
plotGene(
    object = object,
    genes = geneIDs,
    normalized = "vst",
    return = "facet"
)
plotGene(
    object = object,
    genes = geneNames,
    normalized = "vst",
    return = "wide"
)

# DESeqAnalysis ====
object <- deseq_small
plotGene(object, genes = geneIDs, return = "facet")
plotGene(object, genes = geneNames, return = "wide")
}
\seealso{
\code{\link[DESeq2:plotCounts]{DESeq2::plotCounts()}}.

Other Plot Functions: \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotHeatmap}},
  \code{\link{plotQuantileHeatmap}}
}
\author{
Michael Steinbaugh
}
\concept{Gene Expression Functions}
