% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotGeneHeatmap.R
\docType{methods}
\name{plotGeneHeatmap}
\alias{plotGeneHeatmap}
\alias{plotGeneHeatmap,bcbioRNASeq-method}
\alias{plotGeneHeatmap,DESeqDataSet-method}
\alias{plotGeneHeatmap,DESeqTransform-method}
\alias{plotGeneHeatmap,matrix-method}
\title{Gene Heatmap}
\usage{
plotGeneHeatmap(object, ...)

\S4method{plotGeneHeatmap}{bcbioRNASeq}(object, genes = NULL, title = NULL,
  color = inferno(256), legendColor = viridis)

\S4method{plotGeneHeatmap}{DESeqDataSet}(object, genes = NULL,
  annotationCol = NULL, title = NULL, color = inferno(256),
  legendColor = viridis)

\S4method{plotGeneHeatmap}{DESeqTransform}(object, genes = NULL,
  annotationCol = NULL, title = NULL, color = inferno(256),
  legendColor = viridis)

\S4method{plotGeneHeatmap}{matrix}(object, genes = NULL,
  annotationCol = NULL, title = NULL, color = inferno(256),
  legendColor = viridis)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{genes}{Character vector of specific gene identifiers to plot.}

\item{title}{\emph{Optional}. Plot title.}

\item{color}{Colors to use for plot. Defaults to \code{\link[=inferno]{inferno()}} palette.}

\item{legendColor}{Colors to use for legend labels. Defaults to \code{\link[=viridis]{viridis()}}
palette.}

\item{annotationCol}{\link{data.frame} that specifies the annotations shown on the
right side of the heatmap. Each row of this \link{data.frame} defines the
features of the heatmap columns.}
}
\value{
Graphical output only.
}
\description{
These functions facilitate heatmap plotting of a specified set of genes. By
default, row- and column-wise hierarchical clustering is performed using the
Ward method, but this behavior can be overrided by setting \code{cluster_rows} or
\code{cluster_cols} to \code{FALSE}. When column clustering is disabled, the columns
are sorted by the interesting groups (\code{interestingGroups}) specified in the
\link{bcbioRNASeq} and then the sample names.
}
\examples{
# Genes as Ensembl identifiers
genes <- counts(bcb)[1:20, ] \%>\% rownames()
plotGeneHeatmap(bcb, genes = genes)

# Flip the plot and legend palettes
plotGeneHeatmap(
    bcb,
    genes = genes,
    color = viridis(256),
    legendColor = inferno)

# Transcriptome heatmap
\dontrun{
plotGeneHeatmap(bcb)
}

# Use default pheatmap color palette
\dontrun{
plotGeneHeatmap(
    bcb,
    color = NULL,
    legendColor = NULL)
}

# DESeqDataSet
\dontrun{
plotGeneHeatmap(dds)
}

# DESeqTransform
\dontrun{
plotGeneHeatmap(rld)
}
}
\seealso{
\code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}.

Other Heatmaps: \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotDEGHeatmap}}
}
\author{
Michael Steinbaugh
}
\concept{Heatmaps}
