% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotGeneSaturation.R
\docType{methods}
\name{plotGeneSaturation}
\alias{plotGeneSaturation}
\alias{plotGeneSaturation,bcbioRNASeq,missing-method}
\alias{plotGeneSaturation,data.frame,matrix-method}
\title{Plot Gene Detection Saturation}
\usage{
plotGeneSaturation(object, counts, ...)

\S4method{plotGeneSaturation}{bcbioRNASeq,missing}(object, interestingGroups,
  normalized = "tmm", minCounts = 0, color = scale_color_viridis(discrete
  = TRUE))

\S4method{plotGeneSaturation}{data.frame,matrix}(object, counts,
  interestingGroups = "sampleName", minCounts = 0,
  color = scale_color_viridis(discrete = TRUE))
}
\arguments{
\item{object}{Object.}

\item{counts}{Object containing a count matrix.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{interestingGroups}{Category to use to group samples (color and shape).
If unset, this is automatically determined by the metadata set inside the
\link{bcbioRNASeq} object.}

\item{normalized}{Count normalization method. See \code{\link[=counts]{counts()}} documentation
for more information.}

\item{minCounts}{Numeric value for filtering the counts matrix before
plotting.}

\item{color}{Desired ggplot color scale. Defaults to
\code{\link[viridis:scale_color_viridis]{viridis::scale_color_viridis()}}. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using
\code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}
}
\value{
\link{ggplot}.
}
\description{
Plot Gene Detection Saturation
}
\examples{
plotGeneSaturation(bcb)

\dontrun{
plotGeneSaturation(bcb, interestingGroups = "group")
}

# data.frame, matrix
\dontrun{
plotGeneSaturation(metrics(bcb), assay(rld))
}
}
\seealso{
Other Quality Control Plots: \code{\link{plot53Bias}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGenesDetected}}, \code{\link{plotGene}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
\concept{Quality Control Plots}
