% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotGenesDetected.R
\docType{methods}
\name{plotGenesDetected}
\alias{plotGenesDetected}
\alias{plotGenesDetected}
\alias{plotGenesDetected,bcbioRNADataSet,missing-method}
\alias{plotGenesDetected,data.frame,DESeqDataSet-method}
\alias{plotGenesDetected,data.frame,matrix-method}
\title{Plot Genes Detected}
\usage{
plotGenesDetected(object, counts, ...)

\S4method{plotGenesDetected}{bcbioRNADataSet,missing}(object, counts, ...)

\S4method{plotGenesDetected}{data.frame,DESeqDataSet}(object, counts, ...)

\S4method{plotGenesDetected}{data.frame,matrix}(object, counts,
  interestingGroup = "sampleName", passLimit = 20000L, warnLimit = 15000L,
  minCounts = 0L, flip = TRUE)
}
\arguments{
\item{object}{Object.}

\item{counts}{Counts matrix.}

\item{...}{Additional arguments.}

\item{interestingGroup}{\emph{Optional}. Category to use to group samples (color
and shape). If unset, this is automatically determined by the metadata set
inside the \link{bcbioRNADataSet}.}

\item{passLimit}{Threshold to plot pass color marker.}

\item{warnLimit}{Threshold to plot warning color marker.}

\item{minCounts}{Numeric value for filtering the counts matrix before
plotting.}

\item{flip}{Flip X and Y axes.}
}
\value{
\link{ggplot}.
}
\description{
Plot Genes Detected
}
\examples{
data(bcb, dds)

# bcbioRNADataSet
plotGenesDetected(bcb, passLimit = NULL, warnLimit = NULL)

# data.frame + DESeqDataSet
plotGenesDetected(metrics(bcb), dds,
                  passLimit = NULL, warnLimit = NULL)

# data.frame + matrix
plotGenesDetected(metrics(bcb), assay(dds),
                  passLimit = NULL, warnLimit = NULL)
}
\seealso{
Other Quality Control Plots: \code{\link{plot53Bias}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGeneSaturation}}, \code{\link{plotGene}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
