% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotMA.R
\docType{methods}
\name{plotMA}
\alias{plotMA}
\alias{plotMA,DESeqResults-method}
\title{Plot Mean Average}
\usage{
\S4method{plotMA}{DESeqResults}(object, genes = NULL, gene2symbol = NULL,
  pointColor = "darkgray", sigPointColor = "red", labelColor = "black",
  title = TRUE)
}
\arguments{
\item{object}{Object.}

\item{genes}{Gene identifiers (rownames) to plot. These must be the stable
identifiers (e.g. "ENSG00000000003") used on Ensembl and not the gene
symbols.}

\item{gene2symbol}{Apply gene identifier to symbol mappings. A gene2symbol
\code{data.frame} can be passed in, and must contain the columns \code{ensgene} and
\code{symbol}. then the Ensembl gene identifiers will be labeled in place of
gene symbols.}

\item{pointColor}{Default point color for the plot.}

\item{sigPointColor}{Color for points corresponding to significant genes that
have passed alpha level cutoffs.}

\item{labelColor}{Text label color.}

\item{title}{Title.}
}
\value{
\link{ggplot}.
}
\description{
Plot Mean Average
}
\examples{
load(system.file("extdata/bcb.rda", package = "bcbioRNASeq"))
load(system.file("extdata/res.rda", package = "bcbioRNASeq"))
gene2symbol <- gene2symbol(bcb)

# DESeqResults ====
genes <- head(rownames(res), 4L)
plotMA(res, genes = genes, gene2symbol = NULL)
plotMA(res, genes = genes, gene2symbol = gene2symbol)
}
\seealso{
Other Differential Expression Plots: \code{\link{plotVolcano}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
