% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotMappedReads.R
\docType{methods}
\name{plotMappedReads}
\alias{plotMappedReads}
\alias{plotMappedReads,bcbioRNASeqANY-method}
\alias{plotMappedReads,data.frame-method}
\title{Plot Mapped Reads}
\usage{
plotMappedReads(object, ...)

\S4method{plotMappedReads}{bcbioRNASeqANY}(object, interestingGroups,
  passLimit = 20, warnLimit = 10, flip = TRUE)

\S4method{plotMappedReads}{data.frame}(object,
  interestingGroups = "sampleName", passLimit = 20, warnLimit = 10,
  flip = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{interestingGroups}{Category to use to group samples (color and shape).
If unset, this is automatically determined by the metadata set inside the
\link{bcbioRNASeq} object.}

\item{passLimit}{Threshold to plot pass color marker.}

\item{warnLimit}{Threshold to plot warning color marker.}

\item{flip}{Flip X and Y axes.}
}
\value{
\link{ggplot}.
}
\description{
Plot Mapped Reads
}
\examples{
data(bcb)

# bcbioRNASeq
plotMappedReads(bcb)

\dontrun{
plotMappedReads(bcb, interestingGroups = "group")

# data.frame
metrics(bcb) \%>\%
    plotMappedReads()
}
}
\seealso{
Other Quality Control Plots: \code{\link{plot53Bias}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotGenesDetected}}, \code{\link{plotGene}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappingRate}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
\concept{Quality Control Plots}
