% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotMeanAverage-methods.R
\docType{methods}
\name{plotMeanAverage}
\alias{plotMeanAverage}
\alias{plotMeanAverage}
\alias{plotMeanAverage,DESeqResults-method}
\title{Plot Mean Average}
\usage{
plotMeanAverage(object, ...)

\S4method{plotMeanAverage}{DESeqResults}(object, alpha, lfcThreshold = 0L,
  genes = NULL, gene2symbol = NULL, ntop = 0L, direction = c("both",
  "up", "down"), pointColor = "gray50", sigPointColor = c(upregulated =
  "purple", downregulated = "orange"), return = c("ggplot", "data.frame"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{alpha}{Adjusted P value ("alpha") cutoff.}

\item{lfcThreshold}{Log fold change ratio (base 2) cutoff threshold.}

\item{genes}{Character vector of genes to include. These must match the
rownames of the object. It is best practice to use the stable gene
identifiers from Ensembl (e.g. "ENSG00000000003") and not the gene symbols.}

\item{gene2symbol}{\code{data.frame} containing gene-to-symbol mappings. Must
contain the columns \code{geneID} and \code{geneName}.}

\item{ntop}{Number of top genes to label.}

\item{direction}{Plot "\code{both}", "\code{up}", or "\code{down}" directions.}

\item{pointColor}{Default point color for the plot.}

\item{sigPointColor}{\code{character} vector containing color names for labeling
upregulated and downregulated genes. Also supports a character string for
labeling DEGs with the same color, regardless of direction.}

\item{return}{Object class to return. Uses \code{\link[=match.arg]{match.arg()}} internally and picks
the first item in the vector by default.}
}
\value{
\code{ggplot}.
}
\description{
Plot Mean Average
}
\examples{
gene2symbol <- gene2symbol(bcb_small)

# DESeqResults ====
# Color DEGs in each direction separately
plotMeanAverage(
    object = res_small,
    sigPointColor = c(
        upregulated = "purple",
        downregulated = "orange"
    )
)

# Label DEGs with a single color
plotMeanAverage(res_small, sigPointColor = "purple")

# Directional support
plotMeanAverage(
    object = res_small,
    direction = "up",
    ntop = 5L,
    gene2symbol = gene2symbol
)
plotMeanAverage(
    object = res_small,
    direction = "down",
    ntop = 5L,
    gene2symbol = gene2symbol
)

# Label genes manually
plotMeanAverage(
    object = res_small,
    genes = head(rownames(res_small)),
    gene2symbol = gene2symbol
)
}
\seealso{
\code{\link[DESeq2:plotMA]{DESeq2::plotMA()}}.

Other Differential Expression Functions: \code{\link{alphaSummary}},
  \code{\link{contrastName}}, \code{\link{plotDEGHeatmap}},
  \code{\link{plotDEGPCA}}, \code{\link{plotVolcano}},
  \code{\link{resultsTables}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
