% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotMIRNACounts.R
\docType{methods}
\name{plotMIRNACounts}
\alias{plotMIRNACounts}
\alias{plotMIRNACounts}
\alias{plotMIRNACounts,bcbioRNADataSet-method}
\title{Plot Total miRNA Counts}
\usage{
plotMIRNACounts(object, ...)

\S4method{plotMIRNACounts}{bcbioRNADataSet}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\link{ggplot}.
}
\description{
Plot Total miRNA Counts
}
\author{
Lorena Pantano, Michael Steinbaugh
}
