% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotPattern.R
\docType{methods}
\name{plotPattern}
\alias{plotPattern}
\alias{plotPattern,bcbioRNADataSet-method}
\title{Make Groups of Genes Using Expression Profile}
\usage{
plotPattern(object, ...)

\S4method{plotPattern}{bcbioRNADataSet}(object, res, fdr = 0.1, n = NULL,
  ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional parameters, passed to \code{\link[DEGreport:degPatterns]{DEGreport::degPatterns()}}.}

\item{res}{Table with padj as column and gene names as \link{rownames}.}

\item{fdr}{Float cutoff to consider genes significant.}

\item{n}{Integer maximum number of genes to consider.}
}
\value{
\link{ggplot}.
}
\description{
\code{\link[DEGreport:degPatterns]{DEGreport::degPatterns()}} wrapper supporting a \link{bcbioRNADataSet}.
}
\seealso{
\code{\link[DEGreport:degPatterns]{DEGreport::degPatterns()}}.
}
\author{
Lorena Pantano, Michael Steinbaugh
}
