% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotPCA.R
\docType{methods}
\name{plotPCA}
\alias{plotPCA}
\alias{plotPCA,bcbioRNASeq-method}
\title{Sample PCA Plot for Transformed Data}
\usage{
\S4method{plotPCA}{bcbioRNASeq}(object, normalized = c("rlog", "vst", "tmm",
  "tpm"), genes = NULL, censorSamples = NULL, interestingGroups,
  color = scale_color_viridis(discrete = TRUE), label = FALSE,
  title = "pca", return = c("ggplot", "data.frame"), ...)
}
\arguments{
\item{object}{Object.}

\item{normalized}{Logical or character indicating which normalization
method to apply:
\itemize{
\item \code{FALSE}: Raw counts (tximport).
\item \code{TRUE}: DESeq2 normalized counts. Calculated on the fly.
\item "\code{tpm}": Transcripts per million (tximport).
\item "\code{tmm}": edgeR trimmed mean of M-values. Calculated on the fly.
\item "\code{rlog}": DESeq2 \strong{log2} regularized log transformation.
\item "\code{vst}": DESeq2 \strong{log2} variance stabilizing transformation.
}}

\item{genes}{\emph{Optional.} Character vector of gene identifiers to use.}

\item{censorSamples}{\emph{Optional.} Samples to exclude from the PCA plot.
Supply them as a character vector. These should match \code{colnames(object)}.}

\item{interestingGroups}{\emph{Optional.} Interesting groups to use for point
appearance. If missing, color defaults to all \code{interestingGroups}
parameters set in the \link{bcbioRNASeq} object.}

\item{color}{Desired ggplot color scale. Defaults to
\code{\link[=scale_color_viridis]{scale_color_viridis()}}. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[=scale_color_manual]{scale_color_manual()}}.}

\item{label}{\emph{Optional.} Superimpose sample text labels on the plot.}

\item{title}{Plot title.}

\item{return}{Return \code{ggplot} or \code{data.frame}.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot} or \code{data.frame}.
}
\description{
Wrapper for \code{\link[DESeq2:plotPCA]{DESeq2::plotPCA()}} that improves principal component analysis
(PCA) sample coloring and labeling.
}
\examples{
# bcbioRNASeq ====
plotPCA(bcb_small, label = FALSE)
plotPCA(bcb_small, label = TRUE)

# Censor samples
plotPCA(
    object = bcb_small,
    censorSamples = head(colnames(bcb_small), 2L),
    label = TRUE
)
}
\seealso{
\itemize{
\item \code{\link[DESeq2:plotPCA]{DESeq2::plotPCA()}}.
\item \code{getMethod("plotPCA", "DESeqTransform")}
}

Other Quality Control Functions: \code{\link{metrics}},
  \code{\link{plot5Prime3PrimeBias}},
  \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotDispEsts}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotGenesDetected}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}}, \code{\link{plotMeanSD}},
  \code{\link{plotPCACovariates}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh
}
