% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotPCACovariates.R
\docType{methods}
\name{plotPCACovariates}
\alias{plotPCACovariates}
\alias{plotPCACovariates,bcbioRNASeqANY-method}
\title{Find Correlation Between Principal Components (PCs) and Covariates}
\usage{
plotPCACovariates(object, ...)

\S4method{plotPCACovariates}{bcbioRNASeqANY}(object, transform = "rlog",
  metrics = TRUE, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments, passed to \code{\link[DEGreport:degCovariates]{DEGreport::degCovariates()}}.}

\item{transform}{String specifying \link{DESeqTransform} slotted inside the
\link{bcbioRNASeq} object:
\itemize{
\item \code{rlog} (\strong{recommended}).
\item \code{vst}: variance stabilizing transformation.
}}

\item{metrics}{Include sample summary metrics as covariates. Defaults to
include all metrics columns (\code{TRUE}), but desired columns can be specified
here as a character vector.}
}
\value{
\link{ggplot}.
}
\description{
\code{\link[DEGreport:degCovariates]{DEGreport::degCovariates()}} wrapper supporting a \link{bcbioRNASeq} object.
}
\examples{
data(bcb)
plotPCACovariates(bcb, metrics = TRUE)
plotPCACovariates(bcb, metrics = c("exonicRate", "intronicRate"))

# Return `NULL` on example data, since only `sampleName` is interesting
plotPCACovariates(bcb, metrics = FALSE)
}
\seealso{
\itemize{
\item \code{\link[DEGreport:degCovariates]{DEGreport::degCovariates()}}.
\item \code{\link[DESeq2:rlog]{DESeq2::rlog()}}.
\item \code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}}.
}
}
\author{
Lorena Pantano, Michael Steinbaugh
}
