% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/plotPseudoVsAlignedCounts-methods.R
\docType{methods}
\name{plotPseudoVsAlignedCounts}
\alias{plotPseudoVsAlignedCounts}
\alias{plotPseudoVsAlignedCounts,bcbioRNASeq-method}
\title{Compare pseudoaligned counts to aligned counts.}
\usage{
plotPseudoVsAlignedCounts(object, ...)

\S4method{plotPseudoVsAlignedCounts}{bcbioRNASeq}(object, genes = NULL,
  title = "pseudo vs. aligned counts",
  return = c("plotCorrelationHeatmap", "cor"), ...)
}
\arguments{
\item{...}{Passthrough arguments.
When \code{genes} argument is undefined, passes through to
\code{\link[acidplots:plotCorrelationHeatmap]{plotCorrelationHeatmap()}}.}

\item{genes}{\code{character}.
Specific genes to plot. Currently limited to a max of 10 genes.}

\item{return}{\code{character(1)}.
Return type. Uses \code{\link[base:match.arg]{match.arg()}} internally. Only applies
when \code{genes} argument is not defined.}
}
\value{
\code{pheatmap} or \code{ggplot}.
Returns \code{ggplot} when \code{genes} argument is defined, otherwise \code{pheatmap}.
}
\description{
Compare pseudoaligned counts to aligned counts.
}
\note{
Currently supported for salmon or kallisto. The function will
intentionally error for datasets containing aligned counts in the primary
\code{counts} assay.
}
\examples{
data(bcb)

## Correlation heatmap.
plotPseudoVsAlignedCounts(bcb)

## Individual genes.
## Checking the most expressed aligned genes here.
genes <- assays(bcb) \%>\%
    .[["aligned"]] \%>\%
    rowSums() \%>\%
    sort() \%>\%
    tail(n = 2) \%>\%
    names()
plotPseudoVsAlignedCounts(bcb, genes = genes)
}
