% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQC-methods.R
\docType{methods}
\name{plotQC}
\alias{plotQC}
\alias{plotQC,bcbioRNASeq-method}
\title{Plot Quality Control}
\usage{
\S4method{plotQC}{bcbioRNASeq}(object)
}
\arguments{
\item{object}{Object.}
}
\description{
Plot Quality Control
}
\examples{
data(bcb_small)
plotQC(bcb_small)
}
\seealso{
Other Plots: \code{\link{plotCountsPerBiotype}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGenesDetected}}, \code{\link{plotGene}},
  \code{\link{plotHeatmap}}, \code{\link{plotTotalCounts}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh
}
