% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotRRNAMappingRate.R
\docType{methods}
\name{plotRRNAMappingRate}
\alias{plotRRNAMappingRate}
\alias{plotRRNAMappingRate}
\alias{plotRRNAMappingRate,bcbioRNADataSet-method}
\alias{plotRRNAMappingRate,data.frame-method}
\title{Plot Ribosomal RNA (rRNA) Mapping Rate}
\usage{
plotRRNAMappingRate(object, ...)

\S4method{plotRRNAMappingRate}{bcbioRNADataSet}(object, ...)

\S4method{plotRRNAMappingRate}{data.frame}(object,
  interestingGroup = "sampleName", warnLimit = 10L, flip = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{interestingGroup}{\emph{Optional}. Category to use to group samples (color
and shape). If unset, this is automatically determined by the metadata set
inside the \link{bcbioRNADataSet}.}

\item{warnLimit}{Threshold to plot warning color marker.}

\item{flip}{Flip X and Y axes.}
}
\value{
\link{ggplot}.
}
\description{
Plot Ribosomal RNA (rRNA) Mapping Rate
}
\examples{
data(bcb)

# bcbioRNADataSet
plotRRNAMappingRate(bcb)

# data.frame
metrics(bcb) \%>\% plotRRNAMappingRate
}
\seealso{
Other Quality Control Plots: \code{\link{plot53Bias}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotGenesDetected}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
