% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-heatmap.R
\name{plot_correlation_heatmap}
\alias{plot_correlation_heatmap}
\alias{plot_gene_heatmap}
\alias{plot_deg_heatmap}
\title{Heatmaps}
\usage{
plot_correlation_heatmap(bcb, dt, method = "pearson",
  clustering_method = "ward.D2", genes = NULL, interesting_groups = NULL,
  annotation = NULL, title = NULL, ...)

plot_gene_heatmap(bcb, dt, genes, interesting_groups = NULL,
  annotation = NULL, title = NULL, clustering_method = "ward.D2",
  cluster_rows = TRUE, cluster_cols = TRUE, scale = "row", ...)

plot_deg_heatmap(bcb, dt, res, lfc = 1L, title = NULL, ...)
}
\arguments{
\item{bcb}{\link{bcbioRNADataSet}.}

\item{dt}{\link{DESeqTransform}. Transformed counts generated by \code{\link[=rlog]{rlog()}}
(\strong{recommended}) or \code{\link[=vst]{vst()}} on a \link{DESeqDataSet} are supported.}

\item{method}{Correlation coefficient (or covariance) to be computed.
Defaults to \code{pearson} but \code{spearman} can also be used. Consult the
\code{\link[stats:cor]{stats::cor()}} documentation for more information.}

\item{clustering_method}{Clustering method. Accepts the same values as
\code{\link[stats:hclust]{stats::hclust()}}.}

\item{genes}{\emph{Optional}. Character vector of specific gene identifiers to
plot.}

\item{interesting_groups}{\emph{Optional}. Interesting groups to label with bars
above heatmap. If \code{NULL}, defaults to \code{interesting_groups} defined in the
\link{bcbioRNADataSet}.}

\item{annotation}{\emph{Optional}. Alternative annotation to use. Useful when
plotting more than one column.}

\item{title}{\emph{Optional}. Text to include in plot title.}

\item{...}{Additional parameters, passed to \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}.}

\item{cluster_rows}{Whether rows should be clustered using \code{\link[stats:hclust]{stats::hclust()}}.}

\item{cluster_cols}{Whether columns should be clustered using \code{\link[stats:hclust]{stats::hclust()}}.}

\item{scale}{Character indicating if the values should be centered and scaled
in either the \code{row} direction, \code{column} direction, or \code{none}.}

\item{res}{\link{DESeqResults}.}

\item{lfc}{log2 fold change ratio.}
}
\value{
\code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}} return.
}
\description{
These functions are wrappers to \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}} that have been
optimized to work with a \link{bcbioRNADataSet} object. They also depend on a
\link{DESeqTransform} object containing transformed counts, which are ideal for
heatmap plots.
}
\details{
\code{\link[=plot_correlation_heatmap]{plot_correlation_heatmap()}} calculates a correlation matrix based on gene
expression per sample. By default, this function processes all gene counts
per sample to calculate the corrlation matrix. This behavior can be overrided
with the input of \code{gene} identifier vector. In this case, only the expression
of the desired genes will be used to calculate the correlation matrix.

\code{\link[=plot_gene_heatmap]{plot_gene_heatmap()}} facilities heatmap plotting of a specified set of
genes. The function by default performs row- and column-wise hierarhical
clustering using the Ward method, but this behavior can be overrided by
setting \code{cluster_rows} or \code{cluster_cols} to \code{FALSE}. When column clustering
is disabled, the function sorts columns by the interesting groups
(\code{interesting_groups}) specified in the \link{bcbioRNADataSet} and then the sample
names.

\code{\link[=plot_deg_heatmap]{plot_deg_heatmap()}} is a simplified version of \code{\link[=plot_gene_heatmap]{plot_gene_heatmap()}} that is
optimized for handling a \link{DESeqResults} object rather a gene vector. All of
the optional parameters for \code{\link[=plot_gene_heatmap]{plot_gene_heatmap()}} are also available to this
function.
}
\seealso{
\itemize{
\item \code{\link[S4Vectors:cor]{S4Vectors::cor()}}.
\item \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}.
}
}
\author{
Michael Steinbaugh
}
