% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotSizeDistribution.R
\docType{methods}
\name{plotSizeDistribution}
\alias{plotSizeDistribution}
\alias{plotSizeDistribution}
\alias{plotSizeDistribution,bcbioRNADataSet-method}
\title{Plot Size Distribution of Small RNA-Seq Data}
\usage{
plotSizeDistribution(object, ...)

\S4method{plotSizeDistribution}{bcbioRNADataSet}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\link{ggplot}.
}
\description{
Plot Size Distribution of Small RNA-Seq Data
}
\examples{
\dontrun{
data(bcb)
plotSizeDistribution(bcb)
}
}
\seealso{
Other Small RNA-Seq Utilities: \code{\link{plotSmallRNAClusters}}
}
\author{
Lorena Pantano, Michael Steinbaugh
}
