% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotSmallRNAClusters.R
\docType{methods}
\name{plotSmallRNAClusters}
\alias{plotSmallRNAClusters}
\alias{plotSmallRNAClusters}
\alias{plotSmallRNAClusters,bcbioRNADataSet-method}
\title{Cluster Small RNA Samples}
\usage{
plotSmallRNAClusters(object, ...)

\S4method{plotSmallRNAClusters}{bcbioRNADataSet}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\link{ggplot}.
}
\description{
Cluster Small RNA Samples
}
\examples{
data(bcb)
plotSmallRNAClusters(bcb)
}
\seealso{
Other Small RNA-Seq Utilities: \code{\link{plotSizeDistribution}}
}
\author{
Lorena Pantano, Michael Steinbaugh
}
