% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotTotalCounts-methods.R
\docType{methods}
\name{plotTotalCounts}
\alias{plotTotalCounts}
\alias{plotTotalCounts,bcbioRNASeq-method}
\title{Plot Total Counts}
\usage{
plotTotalCounts(object, ...)

\S4method{plotTotalCounts}{bcbioRNASeq}(object, normalized = FALSE,
  interestingGroups = NULL, fill = getOption("bcbio.discrete.fill",
  NULL), flip = getOption("bcbio.flip", TRUE), title = "total counts")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{normalized}{\code{string} or \code{boolean}. Which normalization method to apply:
- \code{FALSE}: Raw counts (tximport).
- \code{TRUE}: DESeq2 normalized counts. Calculated on the fly.
- "\code{tpm}": Transcripts per million (tximport).
- "\code{vst}": DESeq2 \strong{log2} variance stabilizing transformation.
- "\code{rlog}": DESeq2 \strong{log2} regularized log transformation.
- "\code{tmm}": edgeR trimmed mean of M-values. Calculated on the fly.
- "\code{rle}": Relative log expression transformation.
Transcript-level counts support only \code{FALSE} and "\code{tpm}".}

\item{interestingGroups}{\code{character}. Groups of interest that define the
samples. If left unset, defaults to \code{sampleName}.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 fill scale.
Must supply discrete values. When set to \code{NULL}, the default ggplot2 color
palette will be used. If manual color definitions are desired, we recommend
using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.
To set the discrete fill palette globally, use
\code{options(bcbio.discrete.fill = scale_fill_viridis_d())}.}

\item{flip}{\code{boolean}. Flip x and y axes. Recommended for quality control
plots containing many samples.}

\item{title}{\code{string} or \code{NULL}. Plot title.}
}
\value{
\code{ggplot}.
}
\description{
Plot Total Counts
}
\examples{
plotTotalCounts(bcb_small)
}
\seealso{
Other Quality Control Functions: \code{\link{plot5Prime3PrimeBias}},
  \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotDispEsts}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotGenesDetected}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}}, \code{\link{plotMeanSD}},
  \code{\link{plotPCACovariates}}, \code{\link{plotPCA}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
\concept{Quality Control Functions}
