% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2.R, R/counts.R
\name{pool_dds}
\alias{pool_dds}
\alias{pool_counts}
\title{Lane-split FASTQ aggregation functions}
\usage{
pool_dds(dds, pattern = "_L\\\\d+", save = FALSE)

pool_counts(counts, pattern = "_L\\\\d+")
}
\arguments{
\item{dds}{\link{DESeqDataSet}.}

\item{pattern}{Grep pattern to match lane identifiers in file name.}

\item{save}{Whether to assign into environment and write CSV files.}

\item{counts}{Raw counts matrix.}
}
\value{
Object of same class, with pooled technical replicates.
}
\description{
Pool lane-split counts matrix or \link{DESeqDataSet} using \code{\link[=rowSums]{rowSums()}}.
}
\seealso{
A new \link{DESeqDataSet} is returned from \code{\link[=DESeqDataSetFromMatrix]{DESeqDataSetFromMatrix()}}.
}
\author{
Michael Steinbaugh
}
