% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_lane_split_counts.R
\name{pool_lane_split_counts}
\alias{pool_lane_split_counts}
\title{Pool lane split counts}
\usage{
pool_lane_split_counts(counts, lane_grep = "_L\\\\d+$")
}
\arguments{
\item{counts}{Counts matrix}

\item{lane_grep}{Grep string to match lane identifiers in file name}
}
\value{
Pooled counts matrix
}
\description{
Pool lane split counts
}
\author{
Michael Steinbaugh
}
