% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_lane_split_dds.R
\name{pool_lane_split_dds}
\alias{pool_lane_split_dds}
\title{Pool DESeqDataSet counts for technical replicates}
\usage{
pool_lane_split_dds(bcbio, dds, save_counts = TRUE)
}
\arguments{
\item{bcbio}{bcbio list object}

\item{dds}{DESeqDataSet object}

\item{save_counts}{Whether to save and export counts}
}
\value{
DESeqDataSet object using pooled technical replicates
}
\description{
Extract lane split technical replicate counts from a \code{DESeqDataSet}
object, perform \code{rowSums}, and create a new \code{DESeqDataSet} using
\code{DESeqDataSetFromMatrix()}
}
\author{
Michael Steinbaugh
}
