% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_lanes.R
\name{pool_lane_split_counts}
\alias{pool_lane_split_counts}
\alias{pool_lane_split_dds}
\title{Lane-split FASTQ aggregation functions}
\usage{
pool_lane_split_counts(raw_counts, lane_grep = "_L\\\\d+$")

pool_lane_split_dds(run, dds, save = TRUE)
}
\arguments{
\item{raw_counts}{Raw counts matrix}

\item{lane_grep}{Grep string to match lane identifiers in file name}

\item{run}{bcbio-nextgen run object}

\item{dds}{DESeqDataSet object}

\item{save}{Whether to save and export counts}
}
\value{
Pooled raw counts matrix

DESeqDataSet object using pooled technical replicates
}
\description{
Pool lane split counts

Pool DESeqDataSet counts for technical replicates. Extract lane
  split technical replicate counts from a \code{DESeqDataSet} object, perform
  \code{rowSums}, and create a new \code{DESeqDataSet} using
  \code{DESeqDataSetFromMatrix()}
}
\author{
Michael Steinbaugh
}
