% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-prepareRNASeqTemplate.R
\name{prepareRNASeqTemplate}
\alias{prepareRNASeqTemplate}
\title{Prepare RNA-Seq R Markdown Template}
\usage{
prepareRNASeqTemplate()
}
\value{
No value.
}
\description{
Prepare RNA-Seq R Markdown Template
}
\examples{
prepareRNASeqTemplate()

# Clean up
unlink(c(
    "_footer.Rmd",
    "_header.Rmd",
    "_output.yaml",
    "_setup.R",
    "bibliography.bib"
))
}
\author{
Michael Steinbaugh
}
