% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_table.R
\name{print_table}
\alias{print_table}
\title{Print table}
\usage{
print_table(...)
}
\arguments{
\item{...}{Pass through to \code{knitr::kable()}}
}
\description{
Wrapper function for knit reports
}
\author{
Michael Steinbaugh
}
\keyword{internal}
