% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-project_dir.R
\name{.programs}
\alias{.programs}
\title{Read program versions}
\usage{
.programs(project_dir)
}
\arguments{
\item{project_dir}{Project directory.}
}
\description{
Read program versions
}
\author{
Michael Steinbaugh
}
\keyword{internal}
