% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-projectDir.R
\name{.programs}
\alias{.programs}
\title{Read Program Versions}
\usage{
.programs(projectDir)
}
\arguments{
\item{projectDir}{Project directory.}
}
\description{
Read Program Versions
}
\author{
Michael Steinbaugh
}
\keyword{internal}
