% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/pseudoVsAlignedCounts-methods.R
\docType{methods}
\name{pseudoVsAlignedCounts}
\alias{pseudoVsAlignedCounts}
\alias{pseudoVsAlignedCounts,bcbioRNASeq-method}
\title{Compare pseudoaligned counts to aligned counts.}
\usage{
pseudoVsAlignedCounts(object, ...)

\S4method{pseudoVsAlignedCounts}{bcbioRNASeq}(object,
  title = "pseudo vs. aligned counts", ...)
}
\arguments{
\item{...}{Passthrough arguments to
\code{\link[acidplots:plotCorrelationHeatmap]{plotCorrelationHeatmap()}}.}
}
\value{
\code{ggplot}.
}
\description{
Compare pseudoaligned counts to aligned counts.
}
\note{
Currently supported for salmon or kallisto. The function will
intentionally error for datasets containing aligned counts in the primary
\code{counts} assay.
}
\examples{
data(bcb)
pseudoVsAlignedCounts(bcb)
}
