% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_count_density.R, R/plot_counts_per_gene.R,
%   R/plot_deseq_pca.R, R/plot_exonic_mapping_rate.R, R/plot_gender_markers.R,
%   R/plot_gene_detection_saturation.R, R/plot_genes_detected.R,
%   R/plot_intronic_mapping_rate.R, R/plot_mapped_reads.R,
%   R/plot_mapping_rate.R, R/plot_rrna_mapping_rate.R, R/plot_total_reads.R
\name{plot_count_density}
\alias{plot_count_density}
\alias{plot_counts_per_gene}
\alias{plot_deseq_pca}
\alias{plot_exonic_mapping_rate}
\alias{plot_gender_markers}
\alias{plot_gene_detection_saturation}
\alias{plot_genes_detected}
\alias{plot_intronic_mapping_rate}
\alias{plot_mapped_reads}
\alias{plot_mapping_rate}
\alias{plot_rrna_mapping_rate}
\alias{plot_total_reads}
\title{Quality control plots}
\usage{
plot_count_density(bcbio, counts)

plot_counts_per_gene(bcbio, normalized_counts)

plot_deseq_pca(bcbio, dt, label = TRUE)

plot_exonic_mapping_rate(bcbio)

plot_gender_markers(bcbio, normalized_counts)

plot_gene_detection_saturation(bcbio, raw_counts)

plot_genes_detected(bcbio, raw_counts)

plot_intronic_mapping_rate(bcbio)

plot_mapped_reads(bcbio)

plot_mapping_rate(bcbio)

plot_rrna_mapping_rate(bcbio)

plot_total_reads(bcbio)
}
\arguments{
\item{bcbio}{\code{bcbio-nextgen} run object}

\item{normalized_counts}{Normalized counts matrix. Can be obtained from
\code{DESeqDataSet} by running \code{counts(normalized = TRUE)}.
Transcripts per million (TPM) are also acceptable.}

\item{dt}{\code{DESeqTransform} generated from \code{rlog()} or \code{vst()}
on a \code{DESeqDataSet} object. We prefer to pipe the transformed data in
to this function because the operation is CPU intensive, and is slow when
applied to multiple PCA plot operations.}

\item{label}{Superimpose sample text labels on the plot}

\item{raw_counts}{Raw counts matrix. Can be obtained from \code{DESeqDataSet}
by running \code{counts(normalized = FALSE)}.}
}
\value{
Density plot

Boxplot

PCA plot

Bar plot

Scatter plot

Smooth plot

Bar plot

Bar plot

Bar plot

Bar plot

Bar plot

Bar plot
}
\description{
Plot count density

Counts per gene plot

Wrapper for \code{DESeq2::plotPCA()} that improves PCA sample
  coloring and labeling

Exonic mapping rate plot

Plot sexually dimorphic gender marker genes

Genes detection saturation plot

Genes detected plot

Intronic mapping rate plot

Mapped reads plot

Mapping rate plot

rRNA contamination mapping rate

Plot total reads
}
\author{
Michael Steinbaugh
}
