% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{qcPlots}
\alias{qcPlots}
\title{Quality Control Plots}
\arguments{
\item{counts}{Counts matrix.}

\item{flip}{Flip X and Y axes.}

\item{interestingGroup}{\emph{Optional}. Category to use to group samples (color
and shape). If unset, this is automatically determined by the metadata set
inside the \link{bcbioRNADataSet}.}

\item{minCounts}{Numeric value for filtering the counts matrix before
plotting.}

\item{normalized}{Count normalization method. See \code{\link[=counts]{counts()}} documentation
for more information.}

\item{passLimit}{Threshold to plot pass color marker.}

\item{warnLimit}{Threshold to plot warning color marker.}
}
\value{
\link{ggplot}.
}
\description{
Quality Control Plots
}
\keyword{internal}
