% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{query_ensembl}
\alias{query_ensembl}
\title{Query Ensembl}
\usage{
query_ensembl(run)
}
\arguments{
\item{run}{bcbio run.}
}
\value{
Tibble grouped by \code{ensembl_gene_id} with nested
\code{ensembl_transcript_id}.
}
\description{
Query Ensembl
}
\author{
Michael Steinbaugh

Rory Kirchner
}
\keyword{internal}
