% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/all_methods.R
\docType{methods}
\name{raw_counts}
\alias{raw_counts}
\alias{raw_counts,bcbioRnaDataSet-method}
\title{Raw counts}
\usage{
raw_counts(object)

\S4method{raw_counts}{bcbioRnaDataSet}(object)
}
\description{
Raw counts
}
