% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bcbio.R
\name{read_metadata}
\alias{read_metadata}
\alias{read_qc_summary}
\alias{read_counts}
\alias{read_file}
\title{Read data from bcbio-nextgen run}
\usage{
read_metadata(run, grep = NULL, save = FALSE, pool = FALSE)

read_qc_summary(run, save = FALSE)

read_counts(run, type = "salmon", grep = NULL, samples = NULL)

read_file(run, file, row_names = NULL, ...)
}
\arguments{
\item{run}{\code{bcbio-nextgen}}

\item{grep}{Apply grep pattern matching to samples}

\item{save}{Save data}

\item{pool}{Pool lane split samples}

\item{type}{The type of software used to generate the abundances. Follows the
conventions of \code{tximport}.}

\item{samples}{Specify the names of samples in bcbio final directory to
input. If \code{NULL} (default), all samples will be imported.}

\item{file}{File name}

\item{row_names}{Column identifier to use for row names}

\item{...}{Optional parameters for \code{readr}}
}
\value{
Metadata data frame

Summary data frame

txi \code{tximport} list

bcbio run data
}
\description{
Import metadata

Import project quality control summary statistics

Import RNA-Seq counts using \code{tximport}

Import a project data file
}
\author{
Michael Steinbaugh

Rory Kirchner
}
