% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_read_data_raw.R
\name{read_data_raw}
\alias{read_data_raw}
\title{Read and save files in data-raw directory}
\usage{
read_data_raw()
}
\description{
Read and save files in data-raw directory
}
\author{
Michael Steinbaugh
}
\keyword{internal}
