% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-files.R
\name{.read_file}
\alias{.read_file}
\title{Read file with \link{readr} package automatically}
\usage{
.read_file(file, column_to_rownames = NULL, ...)
}
\arguments{
\item{file}{File path.}

\item{column_to_rownames}{Column identifier to use for row names.}

\item{...}{Passthrough parameters to \link{readr} package. Consult the
\code{\link[=read_csv]{read_csv()}} documentation for more information.}
}
\value{
\link{DataFrame}.
}
\description{
Supports automatic loading of \code{.csv}, \code{.tsv}, \code{.xlsx}, and \code{.counts} files.
}
\author{
Michael Steinbaugh
}
\keyword{internal}
