% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Read metadata}
\usage{
read_metadata(csv, pattern = NULL, pattern_col = "description")
}
\arguments{
\item{csv}{Comma separated values file}

\item{pattern}{Apply grep pattern matching to samples}

\item{pattern_col}{Column in data frame used for pattern subsetting}
}
\value{
Metadata data frame
}
\description{
Read metadata
}
