% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_smallrna_counts}
\alias{read_smallrna_counts}
\title{Create isomiRs object from bcbio output}
\usage{
read_smallrna_counts(rna)
}
\arguments{
\item{rna}{\link{bcbioRnaDataSet}.}
}
\description{
Read bcbio sample information from YAML to get isomiR object.
}
\author{
Lorena Patano
}
\keyword{internal}
