% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-smallrna.R
\name{.read_smallrna_counts}
\alias{.read_smallrna_counts}
\title{Create \code{isomiRs} Object from \code{bcbio} Output}
\usage{
.read_smallrna_counts(rna)
}
\arguments{
\item{rna}{\link{bcbioRNADataSet}.}
}
\description{
Read \code{bcbio} sample information from YAML to get \link{isomiR} object.
}
\author{
Lorena Patano
}
\keyword{internal}
