% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_yaml.R
\name{read_yaml}
\alias{read_yaml}
\title{Read YAML file}
\usage{
read_yaml(yaml_file)
}
\arguments{
\item{yaml_file}{YAML file}
}
\value{
List of parsed YAML data if file found, otherwise \code{NULL}
}
\description{
Read YAML file
}
