% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\name{references}
\alias{references}
\title{Download BibTex references}
\usage{
references(bibtex_file = "bcbioRnaseq.bib")
}
\arguments{
\item{bibtex_file}{BibTex library file.}
}
\description{
If file isn't present, download latest version from the \href{http://bioinformatics.sph.harvard.edu/}{HBCwebsite}.
}
\author{
Michael Steinbaugh
}
