% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/relativeLogExpression-methods.R
\docType{methods}
\name{relativeLogExpression}
\alias{relativeLogExpression}
\alias{relativeLogExpression,matrix-method}
\alias{relativeLogExpression,SummarizedExperiment-method}
\title{Relative Log Expression}
\usage{
relativeLogExpression(object, ...)

\S4method{relativeLogExpression}{matrix}(object)

\S4method{relativeLogExpression}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
Relative Log Expression
}
\examples{
data(bcb)
relativeLogExpression(bcb)
}
\references{
Anders and Huber (2010).
}
\seealso{
\code{\link[edgeR:calcNormFactors]{edgeR::calcNormFactors()}}.
}
\author{
Lorena Pantano, Michael Steinbaugh
}
