% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{res}
\alias{res}
\title{Example \link{DESeqResults}}
\format{An object of class \code{DESeqResults} with 505 rows and 6 columns.}
\usage{
res
}
\description{
Derived from \code{dds}, the example \link{DESeqDataSet}.
}
\keyword{internal}
