% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-deseq.R
\name{.res_contrast_name}
\alias{.res_contrast_name}
\title{\code{DESeqResults} Contrast Name}
\usage{
.res_contrast_name(res)
}
\arguments{
\item{res}{\link{DESeqResults}.}
}
\value{
Contrast name string.
}
\description{
\code{DESeqResults} Contrast Name
}
\author{
Michael Steinbaugh
}
\keyword{internal}
