% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{res_small}
\alias{res_small}
\title{Minimal \code{DESeqResults} Example}
\format{An object of class \code{DESeqResults} with 100 rows and 6 columns.}
\usage{
res_small
}
\description{
GSE65267
}
\author{
Michael Steinbaugh, Lorena Pantano
}
\keyword{internal}
