% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_tables.R
\name{results_results}
\alias{results_results}
\title{Differential gene expression results tables}
\usage{
results_results(bcbio, res, lfc = 0)
}
\arguments{
\item{bcbio}{bcbio list object}

\item{lfc}{Log fold change ratio (base 2) cutoff. Does not apply to
statistical hypothesis testing, only gene filtering in the results tables.
See \code{?DESeq2::results} for additional information about using
\code{lfcThreshold} and \code{altHypothesis} to set an alternative
hypothesis based on expected fold changes.}

\item{dds}{\code{DESeqDataSet} object}
}
\value{
Results list
}
\description{
Differential gene expression results tables
}
\examples{
\dontrun{
results_tables(bcbio, res, lfc = 0.25)
}
}
\author{
Michael Steinbaugh
}
