% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-resultsTables.R
\docType{methods}
\name{resultsTables}
\alias{resultsTables}
\alias{resultsTables}
\alias{resultsTables,DESeqResults-method}
\title{Differential Expression Results Tables}
\usage{
resultsTables(object, ...)

\S4method{resultsTables}{DESeqResults}(object, lfc = 0, annotable,
  write = TRUE, summary = TRUE, headerLevel = 3,
  dir = file.path("results", "differential_expression"), quiet = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{lfc}{Log fold change ratio (base 2) cutoff. Does not apply to
statistical hypothesis testing, only gene filtering in the results tables.
See \code{\link[=results]{results()}} for additional information about using \code{lfcThreshold} and
\code{altHypothesis} to set an alternative hypothesis based on expected fold
changes.}

\item{annotable}{\emph{Optional}. Use a pre-saved Ensembl annotations
\link{data.frame}. If left missing, the function will attempt to load from
Ensembl by matching the organism from the first gene identifier in the
rownames. Alternatively, this can be left \code{NULL}, and no gene annotations
will be added to the results.}

\item{write}{Write CSV files to disk.}

\item{summary}{Show summary statistics as a Markdown list.}

\item{headerLevel}{Markdown header level.}

\item{dir}{Directory path where to write files.}

\item{quiet}{If \code{TRUE}, suppress any status messages and/or progress bars.}
}
\value{
Results \link{list}.
}
\description{
Differential Expression Results Tables
}
\examples{
bcb <- examples[["bcb"]]
res <- examples[["res"]]
annotable <- annotable(bcb)
resTbl <- resultsTables(
    res,
    lfc = 0.25,
    annotable = annotable,
    write = FALSE)
names(resTbl)
}
\author{
Michael Steinbaugh
}
