% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rld}
\alias{rld}
\title{Example \link{DESeqTransform}}
\format{An object of class \code{DESeqTransform} with 505 rows and 4 columns.}
\usage{
rld
}
\description{
\code{\link[=rlog]{rlog()}} transformation of \code{dds}, the example \link{DESeqDataSet}.
}
\keyword{internal}
