% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-rowData.R
\docType{methods}
\name{rowData}
\alias{rowData}
\alias{rowData,bcbioRNASeq-method}
\title{Row Data}
\usage{
\S4method{rowData}{bcbioRNASeq}(x, return = c("DataFrame", "data.frame",
  "AsIs"))
}
\arguments{
\item{x}{Object.}

\item{return}{Return as "\code{DataFrame}", "\code{data.frame}", or unmodified
("\code{AsIs}").}
}
\value{
Data describing the rows of the object.
}
\description{
Row Data
}
\examples{
load(system.file("extdata/bcb.rda", package = "bcbioRNASeq"))

# Return as data.frame
rowData(bcb, return = "data.frame")
}
\seealso{
\code{help("rowData", "SummarizedExperiment")}
}
