% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_david.R
\name{run_david}
\alias{run_david}
\title{DAVID gene set enrichment analysis}
\usage{
run_david(...)
}
\arguments{
\item{...}{Pass through to \code{basejump::runDavid()}}
}
\description{
DAVID gene set enrichment analysis
}
\author{
Michael Steinbaugh
}
