% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_deseq_pooled.R
\name{run_deseq_pooled}
\alias{run_deseq_pooled}
\title{Run DESeq2 with pooled technical replicates}
\usage{
run_deseq_pooled(txi, intgroup, design)
}
\arguments{
\item{txi}{tximport list object}

\item{intgroup}{intgroup}

\item{design}{Design formula}
}
\value{
DESeqDataSet object
}
\description{
Run DESeq2 with pooled technical replicates
}
\author{
Michael Steinbaugh
}
