% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/runGSEA-methods.R
\docType{methods}
\name{runGSEA}
\alias{runGSEA}
\alias{runGSEA,DESeqAnalysis-method}
\title{Run Gene Set Enrichment Analysis (GSEA)}
\usage{
runGSEA(object, ...)

\S4method{runGSEA}{DESeqAnalysis}(object, gmtFile)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{gmtFile}{\code{string}. MSigDB GMT file path.}
}
\description{
Run Gene Set Enrichment Analysis (GSEA)
}
\author{
Michael Steinbaugh
}
