% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-sample_dirs.R, R/methods-sample_dirs.R
\docType{methods}
\name{.sample_dirs}
\alias{.sample_dirs}
\alias{sample_dirs,bcbioRNADataSet-method}
\title{Detect sample directories}
\usage{
.sample_dirs(upload_dir)

\S4method{sample_dirs}{bcbioRNADataSet}(object)
}
\arguments{
\item{upload_dir}{Upload directory.}

\item{object}{Primary object.}
}
\value{
Named character vector containing sample directory paths. Function
will \link{stop} if no complete sample directories match.

Folders where samples are kept.
}
\description{
Detect sample directories

This method will be used to access folders where sample information is kept.
}
\author{
Michael Steinbaugh
}
\keyword{internal}
