% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-sample_metadata_file.R
\name{.sample_metadata_file}
\alias{.sample_metadata_file}
\title{Read custom metadata file}
\usage{
.sample_metadata_file(file, pattern = NULL, pattern_col = "description",
  lanes = NULL)
}
\arguments{
\item{file}{Metadata file. CSV and XLSX formats are supported.}

\item{pattern}{\emph{Optional}. Grep pattern to match against sample names.}

\item{pattern_col}{\emph{Optional}. Column in data frame used for pattern
subsetting.}

\item{lanes}{\emph{Optional}. Number of lanes used to split the samples into
technical replicates (\code{_LXXX}) suffix.}
}
\value{
\link{data.frame} with \code{description} as rowname.
}
\description{
Read custom metadata file
}
\author{
Michael Steinbaugh
}
\keyword{internal}
