% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleData-methods.R
\docType{methods}
\name{sampleData}
\alias{sampleData}
\alias{sampleData,bcbioRNASeq-method}
\title{Sample Data}
\usage{
\S4method{sampleData}{bcbioRNASeq}(object, clean = FALSE)
}
\arguments{
\item{object}{Object.}

\item{clean}{\code{boolean}. Only return \code{factor} columns not defined in
\link[bcbioBase:metadataBlacklist]{bcbioBase::metadataBlacklist}. This removes metrics columns used for
quality control analysis, which are often not informative as sample
metadata.}
}
\value{
\code{DataFrame}.
}
\description{
Metadata that describes the samples.
}
\details{
This is a complement to the standard \code{\link[=colData]{colData()}} function, but improves
support for accessing sample metadata for datasets where multiple items in
the columns map to a single sample (e.g. cells for a single-cell RNA-seq
experiment).
}
\examples{
data(bcb_small)
sampleData(bcb_small, clean = TRUE) \%>\% lapply(head)

## Assignment support.
x <- bcb_small
sampleData(x)[["test"]] <- seq_len(ncol(x))
## `test` column should be now defined.
"test" \%in\% colnames(sampleData(x))
}
\seealso{
Other Metadata Functions: \code{\link{detectHPC}},
  \code{\link{detectLanes}}, \code{\link{detectOrganism}},
  \code{\link{interestingGroups}},
  \code{\link{microplate}},
  \code{\link{minimalSampleData}}, \code{\link{organism}},
  \code{\link{sampleNames}},
  \code{\link{uniteInterestingGroups}}

Other SummarizedExperiment Functions: \code{\link{counts}},
  \code{\link{interestingGroups}},
  \code{\link{makeSummarizedExperiment}},
  \code{\link{mapGenes}}, \code{\link{meltCounts}},
  \code{\link{metrics}}, \code{\link{sampleNames}},
  \code{\link{selectSamples}}, \code{\link{tpm}}
}
\author{
Michael Steinbaugh
}
