% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-sampleMetadata.R
\docType{methods}
\name{sampleMetadata}
\alias{sampleMetadata}
\alias{sampleMetadata,bcbioRNASeqANY-method}
\alias{sampleMetadata,DESeqANY-method}
\title{Sample Metadata}
\usage{
\S4method{sampleMetadata}{bcbioRNASeqANY}(object, ...)

\S4method{sampleMetadata}{DESeqANY}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
\link{data.frame}.
}
\description{
Returns the sample metadata.
}
\examples{
data(bcb, dds, rld)

# bcbioRNASeq
sampleMetadata(bcb)

# DESeqDataSet
sampleMetadata(dds)

# DESeqTransform
sampleMetadata(dds)
}
