% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-DESeq2.R
\name{.select_groups_of_interest}
\alias{.select_groups_of_interest}
\title{\link{DESeq} \link{colData}}
\usage{
.select_groups_of_interest(dds, groups_of_interest)
}
\arguments{
\item{dds}{\link{DESeqDataSet} or \link{DESeqTransform}.}

\item{groups_of_interest}{Groups of interest.}
}
\value{
\code{\link[=colData]{colData()}} data frame.
}
\description{
Selects \code{groups_of_interest}-defined interesting groups automatically.
}
\author{
Michael Steinbaugh
}
\keyword{internal}
