% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-DESeq2.R
\name{.select_interesting_groups}
\alias{.select_interesting_groups}
\title{\link{DESeq} \link{colData}}
\usage{
.select_interesting_groups(dds, interesting_groups)
}
\arguments{
\item{dds}{\link{DESeqDataSet} or \link{DESeqTransform}.}

\item{interesting_groups}{Groups of interest.}
}
\value{
\code{\link[=colData]{colData()}} data frame.
}
\description{
Selects \code{interesting_groups}-defined interesting groups automatically.
}
\author{
Michael Steinbaugh
}
\keyword{internal}
