% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{select_intgroup_coldata}
\alias{select_intgroup_coldata}
\title{\link{DESeq} \link{colData}}
\usage{
select_intgroup_coldata(dds, intgroup)
}
\arguments{
\item{dds}{\link{DESeqDataSet} or \link{DESeqTransform}.}

\item{intgroup}{Interesting groups.}
}
\value{
\code{\link[=colData]{colData()}} data frame.
}
\description{
Selects \code{intgroup}-defined interesting groups automatically.
}
\author{
Michael Steinbaugh
}
\keyword{internal}
