% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-selectSamples.R
\docType{methods}
\name{selectSamples}
\alias{selectSamples}
\alias{selectSamples,bcbioRNASeq-method}
\alias{selectSamples,DESeqDataSet-method}
\title{Select Samples}
\usage{
\S4method{selectSamples}{bcbioRNASeq}(object, ..., transform = TRUE)

\S4method{selectSamples}{DESeqDataSet}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{transform}{Apply CPU-intensive DESeq2 transformations. This can
take a long time for large datasets.}
}
\value{
\code{bcbioRNASeq}.
}
\description{
Select Samples
}
\examples{
# bcbioRNASeq ====
x <- selectSamples(bcb_small, day = 7L)
show(x)
colnames(x)

# DESeqDataSet ====
x <- selectSamples(dds_small, day = 7L)
show(x)
colnames(x)
}
\seealso{
Other Data Functions: \code{\link{colData}},
  \code{\link{counts}}, \code{\link{rowData}},
  \code{\link{tmm}}, \code{\link{tpm}}
}
\author{
Michael Steinbaugh
}
