% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-show.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,bcbioRNASeq-method}
\title{Show an Object}
\usage{
\S4method{show}{bcbioRNASeq}(object)
}
\arguments{
\item{object}{Any R object}
}
\value{
\code{show} returns an invisible \code{NULL}.
}
\description{
Show an Object
}
\details{
Objects from an S4 class (a class defined by a call to
  \code{\link{setClass}}) will be displayed automatically is if by a
  call to \code{show}.  S4 objects that occur as attributes of S3
objects will also be displayed in this form; conversely, S3 objects
encountered as slots in S4 objects will be printed using the S3
convention, as if by a call to \code{\link{print}}.

Methods defined for \code{show} will only be inherited  by simple
inheritance, since otherwise the method would not receive the
complete, original object, with misleading results.  See the
\code{simpleInheritanceOnly} argument to \code{\link{setGeneric}} and
the discussion in \code{\link{setIs}} for the general concept.
}
\examples{
show(bcb_small)
}
\seealso{
Other S4 Class Definition: \code{\link{bcbioRNASeq-class}},
  \code{\link{coerce}}, \code{\link{subset}},
  \code{\link{updateObject}}
}
\author{
Michael Steinbuagh
}
