% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/slotAlignedCounts-methods.R
\name{slotAlignedCounts}
\alias{slotAlignedCounts}
\alias{slotAlignedCounts,bcbioRNASeq-method}
\title{Slot aligned counts}
\usage{
slotAlignedCounts(object, ...)

\S4method{slotAlignedCounts}{bcbioRNASeq}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
}
\description{
Slot aligned counts
}
\details{
This function loads aligned counts (e.g. STAR, HISAT2) from the bcbio final
output directory into the bcbioRNASeq object, so we can visually inspect
correlations with the primary pseudoaligned counts.
}
\note{
Updated 2020-12-22.
}
\examples{
uploadDir <- system.file("extdata/bcbio", package = "bcbioRNASeq")
## Fast mode skips import of aligned counts.
bcb <- bcbioRNASeq(uploadDir, fast = TRUE)
"aligned" \%in\% assayNames(bcb)
bcb <- slotAlignedCounts(bcb)
"aligned" \%in\% assayNames(bcb)
}
\author{
Michael Steinbaugh
}
