% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallrna_qc_plots.R
\name{plot_size_distribution}
\alias{plot_size_distribution}
\title{Small RNA-seq quality control plots}
\usage{
plot_size_distribution(run)
}
\arguments{
\item{run}{\code{bcbio-nextgen} run}
}
\value{
ggplot figure
}
\description{
Plot size distribution of small RNA-seq data
}
\author{
Lorena Pantano
}
