% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-subset.R
\docType{methods}
\name{subset}
\alias{subset}
\alias{[,bcbioRNASeq,ANY,ANY,ANY-method}
\title{Bracket-Based Subsetting}
\usage{
\S4method{[}{bcbioRNASeq,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{
    object from which to extract element(s) or in which to replace element(s).
  }

\item{i}{
    indices specifying elements to extract or replace.  Indices are
    \code{numeric} or \code{character} vectors or empty (missing) or
    \code{NULL}.  Numeric values are coerced to integer as by
    \code{\link{as.integer}} (and hence truncated towards zero).
    Character vectors will be matched to the \code{\link{names}} of the
    object (or for matrices/arrays, the \code{\link{dimnames}}):
    see \sQuote{Character indices} below for further details.

    For \code{[}-indexing only: \code{i}, \code{j}, \code{\dots} can be
    logical vectors, indicating elements/slices to select.  Such vectors
    are recycled if necessary to match the corresponding extent.
    \code{i}, \code{j}, \code{\dots} can also be negative integers,
    indicating elements/slices to leave out of the selection.

    When indexing arrays by \code{[} a single argument \code{i} can be a
    matrix with as many columns as there are dimensions of \code{x}; the
    result is then a vector with elements corresponding to the sets of
    indices in each row of \code{i}.

    An index value of \code{NULL} is treated as if it were \code{integer(0)}.
  }

\item{j}{
    indices specifying elements to extract or replace.  Indices are
    \code{numeric} or \code{character} vectors or empty (missing) or
    \code{NULL}.  Numeric values are coerced to integer as by
    \code{\link{as.integer}} (and hence truncated towards zero).
    Character vectors will be matched to the \code{\link{names}} of the
    object (or for matrices/arrays, the \code{\link{dimnames}}):
    see \sQuote{Character indices} below for further details.

    For \code{[}-indexing only: \code{i}, \code{j}, \code{\dots} can be
    logical vectors, indicating elements/slices to select.  Such vectors
    are recycled if necessary to match the corresponding extent.
    \code{i}, \code{j}, \code{\dots} can also be negative integers,
    indicating elements/slices to leave out of the selection.

    When indexing arrays by \code{[} a single argument \code{i} can be a
    matrix with as many columns as there are dimensions of \code{x}; the
    result is then a vector with elements corresponding to the sets of
    indices in each row of \code{i}.

    An index value of \code{NULL} is treated as if it were \code{integer(0)}.
  }

\item{...}{Additional arguments.}

\item{drop}{For matrices and arrays.  If \code{TRUE} the result is
    coerced to the lowest possible dimension (see the examples).  This
    only works for extracting elements, not for the replacement.  See
    \code{\link{drop}} for further details.
  }
}
\value{
\code{bcbioRNASeq}.
}
\description{
Extract genes by row and samples by column from a \code{bcbioRNASeq} object. The
internal \code{DESeqDataSet} and count transformations are rescaled automatically.
DESeq2 transformations can be disabled on large subset operations by setting
\code{transform = FALSE}.
}
\examples{
genes <- rownames(bcb_small)[1:50]
head(genes)
samples <- colnames(bcb_small)[1:2]
head(samples)

# Subset by sample name
bcb_small[, samples]

# Subset by gene list
bcb_small[genes, ]

# Subset by both genes and samples
subset <- bcb_small[genes, samples]
print(subset)
assayNames(subset)

# Skip DESeq2 variance stabilizing transformations
subset <- bcb_small[genes, samples, transform = FALSE]
print(subset)
assayNames(subset)
}
\seealso{
\code{help("[", "base")}.
}
\author{
Lorena Pantano, Michael Steinbaugh
}
