% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-summarized_experiment.R
\name{.summarized_experiment}
\alias{.summarized_experiment}
\alias{.row_data}
\title{Package RNA-seq counts into \link{SummarizedExperiment}}
\usage{
.summarized_experiment(assays, col_data, row_data, metadata = NULL)
}
\arguments{
\item{assays}{assays.}

\item{col_data}{Sample metadata.}

\item{row_data}{\href{http://www.ensembl.org/}{Ensembl} gene annotations.}

\item{metadata}{Custom metadata.}
}
\value{
\link{SummarizedExperiment}.
}
\description{
Package RNA-seq counts into \link{SummarizedExperiment}
}
\author{
Michael Steinbaugh
}
\keyword{internal}
