% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-SummarizedExperiment.R
\name{.SummarizedExperiment}
\alias{.SummarizedExperiment}
\title{RNA-Seq SummarizedExperiment}
\usage{
.SummarizedExperiment(txi, colData, rowData, metadata)
}
\arguments{
\item{txi}{tximport list.}

\item{colData}{Sample metadata.}

\item{rowData}{Ensembl gene annotations.}

\item{metadata}{Optional list of arbitrary content.}
}
\value{
\link{SummarizedExperiment}.
}
\description{
Contains counts generated by \code{\link[=tximport]{tximport()}}, along with bcbio-nextgen
run metadata.
}
\author{
Michael Steinbaugh
}
\keyword{internal}
