% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_table.R
\name{top_table}
\alias{top_table}
\title{Print top table}
\usage{
top_table(res_tables)
}
\arguments{
\item{res_tables}{Results tables generated by \code{res_tables()}}
}
\value{
Data frame or kable, depending on the call
}
\description{
Print top table
}
\author{
Michael Steinbaugh
}
