% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topTables.R
\name{topTables}
\alias{topTables}
\title{Top Tables of Differential Expression Results}
\usage{
topTables(resTbl, n = 50L, coding = FALSE)
}
\arguments{
\item{resTbl}{DESeq2 results tables list generated by \code{\link[=resultsTables]{resultsTables()}}.}

\item{n}{Number genes to report.}

\item{coding}{Whether to only return coding genes.}
}
\value{
Top table kables, for knit report.
}
\description{
Top Tables of Differential Expression Results
}
\author{
Michael Steinbaugh
}
