% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tpm-methods.R
\docType{methods}
\name{tpm}
\alias{tpm}
\alias{tpm,bcbioRNASeq-method}
\title{Transcripts Per Million (TPM)}
\usage{
tpm(object)

\S4method{tpm}{bcbioRNASeq}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{matrix}.
}
\description{
Transcripts Per Million (TPM)
}
\note{
Both gene- and transcript-level counts are supported, as long as they
were imported using a tximport caller (e.g. salmon, kallisto).
}
\examples{
x <- tpm(bcb_small)
summary(x)
}
\seealso{
Other Data Functions: \code{\link{counts}},
  \code{\link{metrics}}, \code{\link{sampleData}},
  \code{\link{tmm}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
