% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-tx2gene.R
\name{.tx2gene}
\alias{.tx2gene}
\title{Transcript to Gene Annotations}
\usage{
.tx2gene(project_dir, genome_build)
}
\arguments{
\item{project_dir}{Project directory.}

\item{genome_build}{Genome build.}
}
\value{
\link{data.frame} with unique rownames.
}
\description{
Transcript to Gene Annotations
}
\author{
Michael Steinbaugh
}
\keyword{internal}
