% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_methods.R
\docType{methods}
\name{txi}
\alias{txi}
\alias{txi<-}
\alias{txi}
\alias{txi.bcbioRnaDataSet}
\alias{txi,bcbioRnaDataSet-method}
\alias{txi<-,bcbioRnaDataSet,matrix-method}
\alias{txi,bcbioRnaDataSet-method}
\alias{"<-txi"}
\title{Accessors for the count matrix of a bcbioRnaDataSet object.}
\usage{
txi(object, ...)

txi(object, ...) <- value

txi.bcbioRnaDataSet(object)

\S4method{txi}{bcbioRnaDataSet}(object)

\S4method{txi}{bcbioRnaDataSet,matrix}(object) <- value
}
\arguments{
\item{object}{a \code{bcbioRnaDataSet} object}

\item{...}{matrix count data}

\item{value}{an integer matrix}
}
\value{
\code{\link[base]{matrix}} with raw count data.
}
\description{
Accessors for the count matrix of a bcbioRnaDataSet object.
}
\author{
Lorena Pantano
}
