% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-tximport.R
\docType{methods}
\name{tximport}
\alias{tximport}
\alias{tximport}
\alias{tximport,bcbioRNASeq-method}
\title{Access tximport Data}
\usage{
tximport(object)

\S4method{tximport}{bcbioRNASeq}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{\link[tximport:tximport]{tximport::tximport()}} list.
}
\description{
Access tximport Data
}
\examples{
load(system.file(
    file.path("inst", "extdata", "bcb.rda"),
    package = "bcbioRNASeq"))

# bcbioRNASeq
tximport(bcb) \%>\% names()
}
\author{
Michael Steinbaugh
}
\keyword{internal}
