% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-updateObject.R
\docType{methods}
\name{updateObject}
\alias{updateObject}
\alias{updateObject,bcbioRNASeq-method}
\title{Update Object}
\usage{
\S4method{updateObject}{bcbioRNASeq}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{bcbioRNASeq}.
}
\description{
Update old objects created by the bcbioRNASeq package. The session
information metadata is preserved from the time when the bcbio data was
originally loaded into R.
}
\section{Legacy \code{bcbioRNADataSet} class}{

Support for \code{bcbioRNADataSet} objects was dropped in v0.2.0 of the package.
If you need to load one of these objects, please install an older release.
}

\section{Legacy objects created with \code{bcbioRnaseq}}{

The previous \code{bcbioRnaseq} package (note case) must be reinstalled to load
objects from versions <= 0.0.20. We changed the name of the package to
\code{bcbioRNASeq} starting in v0.0.21.
}

\examples{
loadRemoteData("http://bcbiornaseq.seq.cloud/f1000v1/bcb.rda")
metadata(bcb)[["version"]]
updated <- updateObject(bcb)
metadata(updated)[["version"]]
metadata(updated)[["previousVersion"]]
}
\author{
Michael Steinbaugh
}
