% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_density_plot.R
\name{volcano_density_plot}
\alias{volcano_density_plot}
\title{Create volcano plot from log2FC and adjusted pvalues data frame}
\usage{
volcano_density_plot(stats, side = "both",
  title = "volcano plot with marginal distributions", pval_cutoff = 0.05,
  lfc_cutoff = 1, shade.colour = "green", shade_alpha = 0.25,
  point_color = "gray", point_alpha = 0.75,
  point_outline_color = "darkgray", line_color = "gray", plot_text = NULL)
}
\arguments{
\item{stats}{data.frame with two columns: logFC and Adjusted.Pvalue}

\item{side}{plot UP, DOWN or BOTH de-regulated points}

\item{title}{title for the figure}

\item{pval_cutoff}{cutoff for the adjusted pvalue. Default 0.05}

\item{lfc_cutoff}{cutoff for the log2FC. Default 1}

\item{shade_alpha}{transparency value. Default 0.25}

\item{point_color}{colours for points. Default gray}

\item{point_alpha}{transparency for points. Default 0.75}

\item{point_outline_color}{Default darkgray}

\item{line_color}{Defaul gray}

\item{plot_text}{Data frame similar to stats table, with a third column being
the text you want to add next to the point in the volcano plot.}
}
\description{
Create volcano plot from log2FC and adjusted pvalues data frame
}
