% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_counts.R
\name{write_counts}
\alias{write_counts}
\title{Write gzipped counts to the results directory}
\usage{
write_counts(counts)
}
\arguments{
\item{counts}{Counts matrix}
}
\description{
Write gzipped counts to the results directory
}
\author{
Michael Steinbaugh
}
