% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-yaml.R
\name{.yaml}
\alias{.yaml}
\alias{.yaml_metadata}
\alias{.yaml_metrics}
\title{YAML utilities}
\usage{
.yaml(yaml, ...)

.yaml_metadata(yaml)

.yaml_metrics(yaml)
}
\arguments{
\item{yaml}{Project summary YAML.}

\item{...}{Nested operator keys supplied as dot objects.}
}
\value{
\link{DataFrame}.
}
\description{
YAML utilities
}
\note{
Metrics are only generated for a standard RNA-seq run with aligned
counts. Fast RNA-seq mode with lightweight counts (pseudocounts) doesn't
output the same metrics into the YAML.
}
\author{
Michael Steinbaugh
}
\keyword{internal}
