% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-aggregateReplicates.R
\docType{methods}
\name{aggregateReplicates}
\alias{aggregateReplicates}
\alias{aggregateReplicates,bcbioRNASeq-method}
\title{Aggregate Replicates}
\usage{
\S4method{aggregateReplicates}{bcbioRNASeq}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{DESeqDataSet}.
}
\description{
Aggregate Replicates
}
\seealso{
Other Data Functions: \code{\link{counts}},
  \code{\link{metrics}}, \code{\link{tmm}},
  \code{\link{tpm}}
}
\author{
Michael Steinbaugh
}
