% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateReplicates-methods.R
\docType{methods}
\name{aggregateReplicates}
\alias{aggregateReplicates}
\alias{aggregateReplicates,bcbioRNASeq-method}
\title{Aggregate Replicates}
\usage{
aggregateReplicates(object, ...)

\S4method{aggregateReplicates}{bcbioRNASeq}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{RangedSummarizedExperiment}.
}
\description{
Aggregate Replicates
}
\examples{
bcb <- bcb_small
colnames(bcb)
# Assign groupings into `aggregate` column of `colData()`
aggregate <- as.factor(sub("^([a-z]+)_.*", "\\\\1", colnames(bcb)))
names(aggregate) <- colnames(bcb)
aggregate
bcb$aggregate <- aggregate
aggregateReplicates(bcb)
}
\seealso{
Other Data Functions: \code{\link{counts}},
  \code{\link{metrics}}, \code{\link{sampleData}},
  \code{\link{tmm}}, \code{\link{tpm}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
