% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/basejump.R
\name{all_generics}
\alias{all_generics}
\alias{assay}
\alias{assays}
\alias{colData}
\alias{design}
\alias{metadata}
\alias{rowData}
\title{S4 Generics}
\usage{
assay(x, i, ...)

assays(x, ..., withDimnames = TRUE)

colData(x, ...)

design(object, ...)

metadata(x, ...)

rowData(x, ...)
}
\arguments{
\item{x}{Object.}

\item{i}{An integer or numeric scalar.}

\item{...}{Additional arguments.}

\item{withDimnames}{A \code{logical}, indicating whether dimnames should be
applied to extracted assay elements.}

\item{object}{Object.}
}
\description{
S4 Generics

colData
}
\keyword{internal}
