% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-alpha_summary.R
\docType{methods}
\name{alpha_summary}
\alias{alpha_summary}
\alias{alpha_summary,DESeqDataSet-method}
\title{Print Summary Statistics of Alpha Level Cutoffs}
\usage{
alpha_summary(object, ...)

\S4method{alpha_summary}{DESeqDataSet}(object, alpha = c(0.1, 0.05, 0.01,
  0.001, 1e-06), contrast = NULL)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{alpha}{Numeric vector of desired alpha cutoffs.}

\item{contrast}{Character vector to use with \link{results} function.}
}
\value{
\link{kable}.
}
\description{
Print Summary Statistics of Alpha Level Cutoffs
}
\examples{
data(bcb)
dds <- DESeqDataSetFromTximport(
    txi = txi(bcb),
    colData = colData(bcb),
    design = formula(~group)) \%>\%
    DESeq
alpha_summary(dds)
}
\author{
Michael Steinbaugh, Lorena Patano
}
