% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_summary.R
\name{alpha_summary}
\alias{alpha_summary}
\title{Print summary statistics of alpha level cutoffs.}
\usage{
alpha_summary(dds, alpha = c(0.1, 0.05, 0.01, 0.001, 1e-06))
}
\arguments{
\item{dds}{DESeqDataSet.}

\item{alpha}{alpha Numeric vector of alpha levels to check.}
}
\value{
Printed summary.
}
\description{
Print summary statistics of alpha level cutoffs.
}
\author{
Michael Steinbaugh
}
