% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{alpha_summary}
\alias{alpha_summary}
\title{Print Summary Statistics of Alpha Level Cutoffs}
\usage{
alpha_summary(dds, alpha = c(0.1, 0.05, 0.01, 0.001, 1e-06),
  contrast = NULL)
}
\arguments{
\item{dds}{\link{DESeqDataSet}.}

\item{alpha}{Numeric vector of desired alpha cutoffs.}

\item{contrast}{Character vector to use with \link{results} function.}
}
\value{
\link{kable}.
}
\description{
Print Summary Statistics of Alpha Level Cutoffs
}
\author{
Michael Steinbaugh, Lorena Patano
}
