% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{alpha_summary}
\alias{alpha_summary}
\title{Print summary statistics of alpha level cutoffs}
\usage{
alpha_summary(dds, alpha = c(0.1, 0.05, 0.01, 0.001, 1e-06), ...)
}
\arguments{
\item{dds}{\link{DESeqDataSet}.}

\item{alpha}{Numeric vector of desired alpha cutoffs.}

\item{...}{Passthrough parameters to \code{\link[=results]{results()}}.}
}
\value{
Printed summary.
}
\description{
Print summary statistics of alpha level cutoffs
}
\author{
Michael Steinbaugh
}
