% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-alphaSummary.R
\docType{methods}
\name{alphaSummary}
\alias{alphaSummary}
\alias{alphaSummary}
\alias{alphaSummary,DESeqDataSet-method}
\alias{alphaSummary,bcbioRNASeq-method}
\title{Alpha Level Cutoff Summary Statistics}
\usage{
alphaSummary(object, ...)

\S4method{alphaSummary}{DESeqDataSet}(object, alpha = c(0.1, 0.05, 0.01,
  0.001, 1e-06), caption = NULL, ...)

\S4method{alphaSummary}{bcbioRNASeq}(object, alpha = c(0.1, 0.05, 0.01, 0.001,
  1e-06), caption = NULL, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Optional.} Passthrough arguments to \code{\link[DESeq2:results]{DESeq2::results()}}. Use
either \code{contrast} or \code{name} arguments to define the desired contrast.}

\item{alpha}{Numeric vector of desired alpha cutoffs.}

\item{caption}{\emph{Optional.} Character vector to add as caption to the table.}
}
\value{
\code{kable}.
}
\description{
Alpha Level Cutoff Summary Statistics
}
\note{
\code{bcbioRNASeq} class does currently support contrast definitions, since
the object contains an internal \code{DESeqDataSet} with an empty design
formula.
}
\examples{
load(system.file("extdata/bcb.rda", package = "bcbioRNASeq"))
load(system.file("extdata/dds.rda", package = "bcbioRNASeq"))

# bcbioRNASeq ====
alphaSummary(bcb)

# DESeqDataSet ====
resultsNames(dds)
alphaSummary(dds)
alphaSummary(dds, contrast = c("day", "7", "0"))
alphaSummary(dds, name = "day_7_vs_0")
}
\seealso{
\code{\link[DESeq2:results]{DESeq2::results()}}.

Other Differential Expression Utilities: \code{\link{plotDispEsts}},
  \code{\link{plotMeanSD}}
}
\author{
Michael Steinbaugh, Lorena Patano
}
