% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/alphaSummary-methods.R
\docType{methods}
\name{alphaSummary}
\alias{alphaSummary}
\alias{alphaSummary}
\alias{alphaSummary,DESeqDataSet-method}
\alias{alphaSummary,bcbioRNASeq-method}
\title{Alpha Level Cutoff Summary Statistics}
\usage{
alphaSummary(object, ...)

\S4method{alphaSummary}{DESeqDataSet}(object, alpha = c(0.1, 0.05, 0.01,
  0.001, 1e-06), caption = NULL, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Optional.} Passthrough arguments to \code{\link[DESeq2:results]{DESeq2::results()}}. Use
either \code{contrast} or \code{name} arguments to define the desired contrast.}

\item{alpha}{Numeric vector of multiple alpha cutoffs.}

\item{caption}{\emph{Optional.} Character string to use as a caption.}
}
\value{
\code{kable}.
}
\description{
Alpha Level Cutoff Summary Statistics
}
\note{
\code{bcbioRNASeq} class does currently support contrast definitions, since
the object contains an internal \code{DESeqDataSet} with an empty design
formula.
}
\examples{
# DESeqDataSet ====
design(dds_small)
resultsNames(dds_small)
alphaSummary(dds_small, contrast = c("treatment", "folic_acid", "control"))
alphaSummary(dds_small, name = "treatment_folic_acid_vs_control")
}
\seealso{
\code{\link[DESeq2:results]{DESeq2::results()}}.

Other Differential Expression Functions: \code{\link{contrastName}},
  \code{\link{plotDEGHeatmap}}, \code{\link{plotDEGPCA}},
  \code{\link{plotMeanAverage}}, \code{\link{plotVolcano}},
  \code{\link{resultsTables}}
}
\author{
Michael Steinbaugh, Lorena Patano
}
