% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-alphaSummary.R
\docType{methods}
\name{alphaSummary}
\alias{alphaSummary}
\alias{alphaSummary,DESeqDataSet-method}
\title{Print Summary Statistics of Alpha Level Cutoffs}
\usage{
alphaSummary(object, ...)

\S4method{alphaSummary}{DESeqDataSet}(object, alpha = c(0.1, 0.05, 0.01,
  0.001, 1e-06), contrast = NULL)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{alpha}{Numeric vector of desired alpha cutoffs.}

\item{contrast}{Character vector to use with \link{results} function.}
}
\value{
\link{kable}.
}
\description{
Print Summary Statistics of Alpha Level Cutoffs
}
\examples{
data(bcb)
dds <- DESeqDataSetFromTximport(
    txi = txi(bcb),
    colData = colData(bcb),
    design = formula(~group)) \%>\%
    DESeq
alphaSummary(dds)
}
\author{
Michael Steinbaugh, Lorena Patano
}
