% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaSummary-methods.R
\docType{methods}
\name{alphaSummary}
\alias{alphaSummary}
\alias{alphaSummary,DESeqDataSet-method}
\title{Alpha level cutoff summary statistics}
\usage{
alphaSummary(object, ...)

\S4method{alphaSummary}{DESeqDataSet}(object, alpha = c(0.1, 0.05, 0.01,
  0.001, 1e-06), caption = NULL, ...)
}
\arguments{
\item{object}{Object.}

\item{alpha}{\code{numeric}. Multiple alpha cutoffs.}

\item{caption}{\code{string} or \code{NULL}. Table caption. If set \code{NULL}, will be
generated automatically from the contrast used.}

\item{...}{Passthrough arguments to \code{\link[DESeq2:results]{DESeq2::results()}}. Use either
\code{contrast} or \code{name} arguments to define the desired contrast.}
}
\value{
\code{kable}.
}
\description{
Quickly generate a summary table of various alpha level cutoffs.
}
\note{
\code{bcbioRNASeq} class does currently support contrast definitions, since
the object contains an internal \code{DESeqDataSet} with an empty design
formula.
}
\examples{
# DESeqDataSet ====
design(dds_small)
resultsNames(dds_small)
alphaSummary(dds_small, contrast = c("treatment", "folic_acid", "control"))
alphaSummary(dds_small, name = "treatment_folic_acid_vs_control")
}
\seealso{
\code{\link[DESeq2:results]{DESeq2::results()}}.

Other Differential Expression Functions: \code{\link{contrastName}},
  \code{\link{plotDEGHeatmap}}, \code{\link{plotDEGPCA}},
  \code{\link{plotMA}}, \code{\link{plotVolcano}},
  \code{\link{resultsTables}}
}
\author{
Michael Steinbaugh, Lorena Patano
}
\concept{Differential Expression Functions}
