% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-methods.R
\name{as}
\alias{as}
\alias{coerce}
\alias{coerce,bcbioRNASeq,DESeqDataSet-method}
\alias{coerce,bcbioRNASeq,DESeqTransform-method}
\title{Force an Object to Belong to a Class}
\description{
Coerce an object to a given class.
}
\section{bcbioRNASeq to DESeqDataSet}{

\enumerate{
\item Coerces to \code{RangedSummarizedExperiment}.
\item Rounds raw counts to \code{integer matrix}.
\item Subsets \code{\link[SummarizedExperiment:colData]{SummarizedExperiment::colData()}} to include only clean factor
columns. See \code{\link[basejump:sampleData]{basejump::sampleData()}} for more details.
\item Simplifies \code{\link[S4Vectors:metadata]{S4Vectors::metadata()}} to include only relevant information
and updates \code{sessionInfo}.
}

Note that gene-level counts are required. Alternatively
\code{tximport::summarizeToGene()} can be called to convert transcript-level
counts to gene-level. By default, we're using length-scaled TPM, so a
corresponding average transcript length matrix isn't necessary. The average
transcript length matrix is only necessary when raw counts matrix isn't
scaled during tximport call (see \code{countsFromAbundance} in tximport).
}

\section{bcbioRNASeq to DESeqTransform}{

\enumerate{
\item Coerces to \code{DESeqDataSet}.
\item Calls \code{\link[DESeq2:DESeq]{DESeq2::DESeq()}}.
\item Calls \code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}}.
}
}

\section{Description}{

    \code{as(object)}
      returns the version of this object coerced to be the given
      \code{Class}.  When used in the replacement form on the left of
      an assignment, the portion of the object corresponding to
      \code{Class} is replaced by \code{value}.

      The operation of \code{as()} in either form depends on the
      definition of coerce methods.  Methods are defined automatically
      when the two classes are related by inheritance; that is, when
      one of the classes is a subclass of the other.

      Coerce methods are also predefined for basic classes (including all
      the types of vectors, functions and a few others).

      Beyond these two sources of methods, further methods are defined
      by calls to the \code{\link{setAs}} function.  See that
      documentation also for details of how coerce methods work. Use
      \code{showMethods(coerce)} for a list of all currently defined methods, as in the
      example below.

 
  
}

\section{Basic Coercion Methods}{

  Methods are pre-defined for coercing any object to one of the basic
  datatypes.  For example, \code{as(x, "numeric")} uses the existing
  \code{as.numeric} function.  These and all other existing methods
  can be listed as shown in the example.

}

\examples{
data(bcb)

## bcbioRNASeq to DESeqDataSet ====
x <- as(bcb, "DESeqDataSet")
names(S4Vectors::mcols(x))
class(x)
show(x)

## bcbioRNASeq to RangedSummarizedExperiment ====
x <- as(bcb, "RangedSummarizedExperiment")
slotNames(x)
show(x)

## bcbioRNASeq to SummarizedExperiment ====
## Coerce to RangedSummarizedExperiment first.
x <- as(bcb, "RangedSummarizedExperiment")
x <- as(x, "SummarizedExperiment")
class(x)
slotNames(x)
show(x)
}
\references{
Chambers, John M. (2016)
 \emph{Extending R},
  Chapman & Hall.
(Chapters 9 and 10.)
}
\seealso{
\itemize{
\item \code{tximport::tximport}
\item \code{DESeq2::DESeqDataSetFromTximport()}.
}
}
\author{
Michael Steinbaugh
}
