% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basejump.R
\name{kables}
\alias{kables}
\alias{make_names}
\alias{make_names_snake}
\alias{set_names_snake}
\title{Utility functions imported from basejump package}
\usage{
kables(list, captions = NULL)

make_names(character)

make_names_snake(character)

set_names_snake(data)
}
\arguments{
\item{list}{List of tables (e.g. data frame, matrix)}

\item{captions}{Caption character vector}

\item{character}{Character vector}

\item{data}{Data type that supports name assignments}
}
\description{
Handle multiple kables in a single RMarkdown chunk.
Import of \href{https://github.com/steinbaugh/basejump/blob/master/R/kables.R}{kables}.

Make syntactically valid names out of character vectors.
Variant of \href{https://github.com/steinbaugh/basejump/blob/master/R/makeNames.R}{makeNames}.

Make names in snake_case.
Variant of \href{https://github.com/steinbaugh/basejump/blob/master/R/makeNames.R}{makeNamesSnake}.

Set names in snake_case.
Variant of \href{https://github.com/steinbaugh/basejump/blob/master/R/setNames.R}{setNamesSnake}.
}
\details{
snake_case variants for function naming consistency
}
\author{
Michael Steinbaugh
}
