% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bcb}
\alias{bcb}
\title{Example \link{bcbioRNASeq}}
\format{An object of class \code{bcbioRNADataSet} with 505 rows and 4 columns.}
\usage{
bcb
}
\description{
Example \link{bcbioRNASeq}
}
\author{
Lorena Pantano
}
\keyword{internal}
