% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcb}
\alias{bcb}
\title{bcbio RNA-seq example data set}
\format{
An object of class \code{bcbioRNASeq} with 100 rows and 6 columns.
}
\usage{
data(bcb)
}
\description{
bcbio RNA-seq example data set
}
\examples{
data(bcb)
bcb
}
\author{
Michael Steinbaugh, Lorena Pantano
}
\keyword{datasets}
