% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcb_small}
\alias{bcb_small}
\title{bcbio RNA-Seq Example Data}
\format{An object of class \code{bcbioRNASeq} with 500 rows and 6 columns.}
\usage{
bcb_small
}
\description{
bcbio RNA-Seq Example Data
}
\examples{
bcb_small
}
\seealso{
Other GSE65267: \code{\link{deseq_small}}
}
\author{
Michael Steinbaugh, Lorena Pantano
}
\concept{GSE65267}
\keyword{datasets}
