% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcb_small}
\alias{bcb_small}
\title{\code{bcbioRNASeq} example}
\format{An object of class \code{bcbioRNASeq} with 500 rows and 6 columns.}
\usage{
bcb_small
}
\description{
\code{bcbioRNASeq} example
}
\seealso{
Other GSE65267: \code{\link{dds_small}},
  \code{\link{res_small}}
}
\author{
Lorena Pantano, Michael Steinbaugh
}
\concept{GSE65267}
\keyword{datasets}
