% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_methods.R
\docType{methods}
\name{bcbcols}
\alias{bcbcols}
\alias{bcbcols,bcbioRnaDataSet}
\alias{bcbcols,bcbioRnaDataSet-method}
\title{Accessors samples dir of a \link{bcbioRnaDataSet} object}
\usage{
bcbcols(object)

\S4method{bcbcols}{bcbioRnaDataSet}(object)
}
\arguments{
\item{object}{\link{bcbioRnaDataSet} object.}
}
\value{
folders where samples are kept
}
\description{
This method will be used to access folders where
sample information is kept
}
