% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/all_methods.R
\docType{methods}
\name{bcbio}
\alias{bcbio}
\alias{bcbio,bcbioRnaDataSet-method}
\alias{bcbio<-,bcbioRnaDataSet,matrix-method}
\alias{"bcbio<-"}
\alias{bcbio,bcbioRnaDataSet-method}
\alias{bcbio<-,bcbioRnaDataSet,matrix-method}
\title{\link{bcbioRnaDataSet} count matrix accessors}
\usage{
bcbio(object, ...)

bcbio(object, ...) <- value

\S4method{bcbio}{bcbioRnaDataSet}(object, type = "counts")

\S4method{bcbio}{bcbioRnaDataSet,matrix}(object, type = "counts") <- value
}
\arguments{
\item{object}{\link{bcbioRnaDataSet} object.}

\item{...}{Matrix count data.}

\item{value}{An integer matrix or other object.}

\item{type}{type of count data to retrieve}
}
\value{
Count matrix.
}
\description{
This method will be used to access all different count matrix from
the object. Gene expression, transcript expression, miRNA expression...
}
