% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-bcbio.R
\docType{methods}
\name{bcbio}
\alias{bcbio}
\alias{bcbio<-}
\alias{bcbio,bcbioRNADataSet-method}
\alias{bcbio<-,bcbioRNADataSet-method}
\title{\code{bcbioRNADataSet} Caller Slot Accessor}
\usage{
bcbio(object, ...)

bcbio(object, ...) <- value

\S4method{bcbio}{bcbioRNADataSet}(object, type)

\S4method{bcbio}{bcbioRNADataSet}(object, type) <- value
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{value}{Value to assign.}

\item{type}{Type of data to retrieve.}
}
\value{
\link{slot} object.
}
\description{
This method is used to access alternative count matrices with differing
dimensions from the primary counts stored as a \code{SummarizedExperiment}.
}
\details{
Additional count matrices of interest include:
\itemize{
\item Aligned counts generated by \href{https://github.com/alexdobin/STAR}{STAR} and
\href{http://bioinf.wehi.edu.au/featureCounts/}{featureCounts}.
\item Transcript-level gene expression, for
\href{http://pachterlab.github.io/sleuth}{sleuth} analysis.
\item miRNA expression counts.
}

Since these matrices don't necessarily contain the same number of columns and
rows as the counts imported with \code{\link[=tximport]{tximport()}}, it is preferable to store this
information as alternative callers.
}
\note{
By default, we store the quasi-aligned (a.k.a. lightweight) gene
expression counts generated by
\href{https://combine-lab.github.io/salmon/}{salmon} into in a
\link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment::SummarizedExperiment} using \code{\link[=tximport]{tximport()}}. These
counts should be accessed using the \code{counts()} generic. Any additional
transformations of those original counts, including \code{\link[=tpm]{tpm()}}, \code{\link[=rlog]{rlog()}}, or
\code{\link[=tmm]{tmm()}} are stored alongside the raw counts.
}
\examples{
# tximport list
txi <- bcbio(bcb, "tximport")
class(txi)
length(txi)
names(txi)

# featureCounts matrix
fc <- bcbio(bcb, "featureCounts")
class(fc)
dim(fc)
colnames(fc)
}
\author{
Lorena Pantano, Michael Steinbaugh
}
