% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_methods.R
\docType{methods}
\name{bcbio}
\alias{bcbio}
\alias{bcbio,bcbioRnaDataSet-method}
\alias{bcbio<-,bcbioRnaDataSet,matrix-method}
\alias{bcbio.bcbioRnaDataSet}
\alias{bcbio,bcbioRnaDataSet-method}
\alias{bcbio<-}
\alias{"<-bcbio"}
\title{Accessors for the count matrix of a \linkS4class{bcbioRnaDataSet} object}
\usage{
bcbio(object, ...)

bcbio.bcbioRnaDataSet(object, type = "counts")

\S4method{bcbio}{bcbioRnaDataSet}(object, type = "counts")

bcbio(object, ...) <- value

\S4method{bcbio}{bcbioRnaDataSet,matrix}(object, type = "counts") <- value
}
\arguments{
\item{object}{\linkS4class{bcbioRnaDataSet} object.}

\item{...}{Matrix count data.}

\item{type}{type of count data to retrieve}

\item{value}{An integer matrix.}
}
\value{
\link[base:matrix]{base::matrix} with raw count data.
}
\description{
This method will be used to access all different count matrix from
the object. Gene expression, transcript expression, miRNA expression...
}
