% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses-legacy.R
\docType{class}
\name{bcbioRNADataSet-class}
\alias{bcbioRNADataSet-class}
\alias{bcbioRNADataSet}
\title{bcbioRNADataSet Object Class (Legacy)}
\description{
This class has been superceded by \code{bcbioRNASeq} and will be formally
deprecated in a future release.
}
\section{Slots}{

\describe{
\item{\code{callers}}{\code{SimpleList} containing additional bcbio run data with
dimensions that don't match the count matrix.}
}}

\author{
Lorena Pantano, Michael Steinbaugh
}
\keyword{internal}
