% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{bcbioRNADataSet-class}
\alias{bcbioRNADataSet-class}
\alias{bcbioRNADataSet}
\title{\code{bcbioRNADataSet}}
\description{
S4 class containing RNA-seq counts and metadata generated by
\href{https://bcbio-nextgen.readthedocs.org/}{bcbio-nextgen}.
}
\details{
\link{bcbioRNADataSet} is a subclass of \link{SummarizedExperiment} designed to store
an RNA-seq analysis. This class contains raw read counts and transcripts per
million (TPM) generated by \code{\link[=tximport]{tximport()}}.

The \code{metadata} accessor contains:
\itemize{
\item Sample summary statistics (a.k.a. metrics).
\item Ensembl annotations.
\item Server run paths.
\item R local environment information, including \code{\link[=sessionInfo]{sessionInfo()}}.
}
}
\examples{
path <- system.file("extra", package = "bcbioRnaseq")
bcb <- load_run(file.path(path, "bcbio"))
}
\author{
Lorena Patano, Michael Steinbaugh
}
