% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{bcbioRNADataSet-class}
\alias{bcbioRNADataSet-class}
\alias{bcbioRNADataSet}
\alias{[,bcbioRNADataSet-method}
\alias{[,bcbioRNADataSet,ANY,ANY,ANY-method}
\title{\code{bcbioRNADataSet}}
\description{
S4 class containing RNA-seq counts and metadata generated by
\href{https://bcbio-nextgen.readthedocs.org/}{bcbio}.
}
\details{
\link{bcbioRNADataSet} is a subclass of \code{SummarizedExperiment} designed to store
an RNA-seq analysis. This class contains raw read counts and transcripts per
million (TPM) generated by \code{\link[=tximport]{tximport()}}.

The \code{metadata} accessor contains:
\itemize{
\item Sample summary statistics (a.k.a. metrics).
\item Ensembl annotations.
\item Server run paths.
\item R local environment information, including \code{\link[=sessionInfo]{sessionInfo()}}.
}
}
\examples{
path <- system.file("extra", package = "bcbioRnaseq")
bcb <- loadRNASeqRun(file.path(path, "bcbio"))
}
\seealso{
\link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment::SummarizedExperiment}.
}
\author{
Michael Steinbaugh, Lorena Pantano
}
