% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{SummarizedExperiment-class}
\alias{SummarizedExperiment-class}
\alias{bcbioRnaDataSet-class}
\title{Class that contains bcbio run information.}
\description{
bcbioRnaDataSet is a subclass of \link{SummarizedExperiment} used to store the
raw data, intermediate calculations and results of an miRNA/isomiR analysis.
This class stores all raw isomiRs data for each sample, processed
information, summary for each isomiR type, raw counts, normalized counts, and
table with experimental information for each sample.
}
\details{
\code{\link[=bcbioRnaDataSetFromFolder]{bcbioRnaDataSetFromFolder()}} creates this object using bcbio output files.

Methods for this objects ...

\code{metadata} contains ...
}
