% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{SummarizedExperiment-class}
\alias{SummarizedExperiment-class}
\alias{bcbioRnaDataSet-class}
\title{Class that contains bcbio run information}
\description{
\code{bcbioRnaDataSet} is a subclass of \link{SummarizedExperiment} designed to store
an RNA-seq analysis. This class contains experiment metadata, raw counts,
normalilzed counts, and summary statistics for each sample analyzed.
}
\details{
Methods for this objects ...

\code{metadata} contains ...
}
\keyword{internal}
