% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{bcbioRNASeq-class}
\alias{bcbioRNASeq-class}
\alias{bcbioRNASeq}
\title{bcbioRNASeq Class}
\description{
S4 class containing RNA-seq counts and metadata generated by
\href{https://bcbio-nextgen.readthedocs.org}{bcbio}.
}
\details{
\code{bcbioRNASeq} extends \code{RangedSummarizedExperiment} and is designed to store a
bcbio RNA-seq analysis. This class contains raw read counts and length-scaled
transcripts per million (TPM) generated by \code{\link[tximport:tximport]{tximport::tximport()}}. Counts can
be loaded at gene (default) or transcript level.

DESeq2 is run automatically when \code{\link[=loadRNASeq]{loadRNASeq()}} is called, and variance
stabilized counts are slotted into \code{\link[=assays]{assays()}}.

The \code{\link[=metadata]{metadata()}} accessor contains:
\itemize{
\item Sample quality control metrics.
\item Ensembl annotations.
\item Server run paths.
\item R session information (e.g. \code{\link[utils:sessionInfo]{utils::sessionInfo()}}).
}
}
\note{
\code{bcbioRNASeq} extended \code{SummarizedExperiment} prior to v0.2.0, where we
migrated to \code{RangedSummarizedExperiment}.
}
\examples{
show(bcb_small)
validObject(bcb_small)
}
\seealso{
\itemize{
\item \code{\link[=loadRNASeq]{loadRNASeq()}}.
\item \code{\link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment::SummarizedExperiment()}}.
\item \code{\link[methods:initialize]{methods::initialize()}}.
\item \code{\link[methods:validObject]{methods::validObject()}}.
\item \code{\link[BiocGenerics:updateObject]{BiocGenerics::updateObject()}}.
\item \code{.S4methods(class = "bcbioRNASeq")}.
}

Other S4 Class Definition: \code{\link{coerce}},
  \code{\link{colData}}, \code{\link{rowData}},
  \code{\link{show}}, \code{\link{subset}},
  \code{\link{updateObject}}
}
\author{
Michael Steinbaugh, Lorena Pantano
}
