% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{bcbioRNASeq-class}
\alias{bcbioRNASeq-class}
\alias{bcbioRNASeq}
\title{bcbioRNASeq Object Class}
\description{
S4 class containing RNA-seq counts and metadata generated by
\href{https://bcbio-nextgen.readthedocs.org/}{bcbio}.
}
\details{
\code{bcbioRNASeq} is a subclass of \code{SummarizedExperiment} designed to store
an RNA-seq analysis. This class contains raw read counts and transcripts per
million (TPM) generated by \code{\link[=tximport]{tximport()}}.

The \code{metadata} accessor contains:
\itemize{
\item Sample summary statistics (a.k.a. metrics).
\item Ensembl annotations.
\item Server run paths.
\item R local environment information, including \code{\link[=sessionInfo]{sessionInfo()}}.
}
}
\section{Slots}{

\describe{
\item{\code{bcbio}}{\code{SimpleList} containing additional bcbio run data with dimensions
that don't match the count matrix. This is currently used to store STAR
featureCounts aligned counts.}
}}

\examples{
uploadDir <- system.file("extdata/bcbio", package = "bcbioRNASeq")

bcb <- loadRNASeq(uploadDir)
print(bcb)

# Check validity
validObject(bcb)
}
\seealso{
\itemize{
\item \link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment::SummarizedExperiment}.
\item \code{.S4methods(class = "bcbioRNASeq")}.
}
}
\author{
Lorena Pantano, Michael Steinbaugh
}
