% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{bcbioRNASeq-class}
\alias{bcbioRNASeq-class}
\title{\code{bcbioRNASeq} Class}
\description{
Inherits from \code{RangedSummarizedExperiment}.
}
\examples{
uploadDir <- system.file("extdata/bcbio", package = "bcbioRNASeq")
x <- bcbioRNASeq(uploadDir)
print(x)
}
\seealso{
\code{\link[=bcbioRNASeq]{bcbioRNASeq()}}.

Other S4 Object: \code{\link{DESeqAnalysis-class}},
  \code{\link{DESeqResultsTables-class}},
  \code{\link{bcbioRNASeq}}, \code{\link{coerce}},
  \code{\link{extract}}, \code{\link{show}},
  \code{\link{updateObject}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Object}
