% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{bcbioRNASeq-class}
\alias{bcbioRNASeq-class}
\title{bcbio RNA-Seq data set}
\description{
\code{bcbioRNASeq} is an S4 class that extends \code{RangedSummarizedExperiment}, and
is designed to store a \href{https://bcbio-nextgen.readthedocs.org}{bcbio} RNA-seq
analysis.
}
\note{
\code{bcbioRNASeq} extended \code{SummarizedExperiment} prior to v0.2.0, where we
migrated to \code{RangedSummarizedExperiment}.

Updated 2022-03-07.
}
\section{Automatic metadata}{


The \code{metadata()} slot always contains:
\itemize{
\item Object version.
\item bcbio data provenance information.
\item File paths and timestamps.
\item R session information.
}
}

\author{
Michael Steinbaugh, Lorena Pantano
}
