% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbioRNASeq-package.R
\docType{package}
\name{bcbioRNASeq-package}
\alias{bcbioRNASeq}
\alias{bcbioRNASeq-package}
\title{bcbioRNASeq}
\description{
Quality control and differential expression for
\href{http://bcbio-nextgen.readthedocs.io}{bcbio} RNA-seq experiments.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://bioinformatics.sph.harvard.edu/bcbioRNASeq}
  \item \url{https://github.com/hbc/bcbioRNASeq}
  \item Report bugs at \url{https://github.com/hbc/bcbioRNASeq/issues}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com}

Authors:
\itemize{
  \item Lorena Pantano \email{lorena.pantano@gmail.com}
  \item Rory Kirchner \email{roryk@alum.mit.edu}
  \item Victor Barrera \email{barrera.vic@gmail.com}
  \item Mary Piper \email{mary.piper@gmail.com}
  \item John Hutchinson \email{jnhutchinson@gmail.com}
}

Other contributors:
\itemize{
  \item Harvard Chan Bioinformatics Core \email{bioinformatics@hsph.harvard.edu} [copyright holder, funder]
}

}
