% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbioRnaseq.R
\docType{package}
\name{bcbioRnaseq-package}
\alias{bcbioRnaseq}
\alias{bcbioRnaseq-package}
\title{bcbioRnaseq}
\description{
Quality control and differential expression for bcbio-nextgen RNA-seq
experiments.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://bioinformatics.sph.harvard.edu/bcbioRnaseq/}
  \item \url{https://github.com/hbc/bcbioRnaseq}
  \item Report bugs at \url{https://github.com/hbc/bcbioRnaseq/issues}
}

}
\author{
\strong{Maintainer}: Michael J. Steinbaugh \email{mike@steinbaugh.com}

Authors:
\itemize{
  \item Rory Kirchner \email{roryk@alum.mit.edu}
  \item Lorena Pantano \email{lorena.pantano@gmail.com}
  \item Victor Barrera \email{barrera.vic@gmail.com}
}

Other contributors:
\itemize{
  \item Harvard Chan Bioinformatics Core \email{bioinformatics@hsph.harvard.edu} [copyright holder, funder]
}

}
