% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{bcbioRNASeq-package}
\title{bcbioRNASeq}
\description{
Import and analyze \href{https://bcbio-nextgen.readthedocs.io/}{bcbio} RNA-seq
data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bioinformatics.sph.harvard.edu/bcbioRNASeq/}
  \item \url{https://github.com/hbc/bcbioRNASeq/}
  \item Report bugs at \url{https://github.com/hbc/bcbioRNASeq/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (0000-0002-2403-2221)

Authors:
\itemize{
  \item Lorena Pantano \email{lorena.pantano@gmail.com} (0000-0002-3859-3249)
  \item Rory Kirchner \email{roryk@alum.mit.edu} (0000-0003-4814-5885)
  \item Victor Barrera \email{barrera.vic@gmail.com} (0000-0003-0590-4634)
  \item Brad Chapman \email{chapmanb@fastmail.com} (0000-0002-3026-1856)
  \item Mary Piper \email{mary.piper@gmail.com} (0000-0003-2699-3840)
  \item Meeta Mistry \email{meeta.mistry@gmail.com} (0000-0002-5830-9479)
  \item Radhika Khetani \email{rkhetani@hsph.harvard.edu} (0000-0003-2430-2970)
  \item Kayleigh Rutherford \email{kayrut1@gmail.com} (0000-0001-9699-2363)
  \item Oliver Hofmann \email{ohofmann72@gmail.com} (0000-0002-7738-1513)
  \item John Hutchinson \email{jnhutchinson@gmail.com} (0000-0002-7804-7576)
  \item Shannan Ho Sui \email{shosui@hsph.harvard.edu} (0000-0002-6191-4709)
}

Other contributors:
\itemize{
  \item Harvard Chan Bioinformatics Core \email{bioinformatics@hsph.harvard.edu} [copyright holder, funder]
}

}
\keyword{internal}
