% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{bcbioRNASeq-package}
\title{bcbioRNASeq}
\description{
Import and analyze \href{https://bcbio-nextgen.readthedocs.io/}{bcbio} RNA-seq
data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r.acidgenomics.com/packages/bcbiornaseq/}
  \item \url{https://github.com/acidgenomics/r-bcbiornaseq/}
  \item Report bugs at \url{https://github.com/acidgenomics/r-bcbiornaseq/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})

Authors:
\itemize{
  \item Lorena Pantano \email{lorena.pantano@gmail.com} (\href{https://orcid.org/0000-0002-3859-3249}{ORCID})
  \item Rory Kirchner \email{roryk@alum.mit.edu} (\href{https://orcid.org/0000-0003-4814-5885}{ORCID})
  \item Victor Barrera \email{barrera.vic@gmail.com} (\href{https://orcid.org/0000-0003-0590-4634}{ORCID})
  \item Brad Chapman \email{chapmanb@fastmail.com} (\href{https://orcid.org/0000-0002-3026-1856}{ORCID})
  \item Mary Piper \email{mary.piper@gmail.com} (\href{https://orcid.org/0000-0003-2699-3840}{ORCID})
  \item Meeta Mistry \email{meeta.mistry@gmail.com} (\href{https://orcid.org/0000-0002-5830-9479}{ORCID})
  \item Radhika Khetani \email{rkhetani@hsph.harvard.edu} (\href{https://orcid.org/0000-0003-2430-2970}{ORCID})
  \item Kayleigh Rutherford \email{kayrut1@gmail.com} (\href{https://orcid.org/0000-0001-9699-2363}{ORCID})
  \item Oliver Hofmann \email{ohofmann72@gmail.com} (\href{https://orcid.org/0000-0002-7738-1513}{ORCID})
  \item John Hutchinson \email{jnhutchinson@gmail.com} (\href{https://orcid.org/0000-0002-7804-7576}{ORCID})
  \item Shannan Ho Sui \email{shosui@hsph.harvard.edu} (\href{https://orcid.org/0000-0002-6191-4709}{ORCID})
}

Other contributors:
\itemize{
  \item Harvard Chan Bioinformatics Core \email{bioinformatics@hsph.harvard.edu} [copyright holder, funder]
}

}
\keyword{internal}
