% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{bcbioRNASeq-package}
\alias{bcbioRNASeq-package}
\title{bcbioRNASeq}
\description{
Quality control and differential expression for
\href{http://bcbio-nextgen.readthedocs.io}{bcbio} RNA-seq experiments.
}
