% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbioRnaseq.R
\docType{package}
\name{bcbioRnaseq}
\alias{bcbioRnaseq}
\alias{bcbioRnaseq-package}
\title{bcbioRnaseq.}
\description{
Quality control and differential expression for bcbio-nextgen RNA-seq
experiments. Consult the
\href{http://bioinformatics.sph.harvard.edu/bcbioRnaseq}{package website} for
additional information.
}
\keyword{internal}
