% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/bcbioRNASeq.R
\docType{class}
\name{bcbioRNASeq}
\alias{bcbioRNASeq}
\alias{bcbioRNASeq-class}
\title{\code{bcbioRNASeq} object and constructor}
\usage{
bcbioRNASeq(uploadDir, level = c("genes", "transcripts"),
  caller = c("salmon", "kallisto", "sailfish", "star", "hisat2"),
  interestingGroups = "sampleName", samples = NULL,
  censorSamples = NULL, sampleMetadataFile = NULL, organism = NULL,
  genomeBuild = NULL, ensemblRelease = NULL, gffFile = NULL,
  transgeneNames = NULL, spikeNames = NULL, vst = TRUE,
  rlog = FALSE, ...)
}
\arguments{
\item{uploadDir}{\code{string}. Path to final upload directory. This path is set
when running "\code{bcbio_nextgen -w template}".}

\item{level}{\code{string}. Import counts at gene level ("\code{genes}"; \emph{default}) or
transcript level ("\code{transcripts}"; \emph{advanced use}).}

\item{caller}{\code{string}. Expression caller:
\itemize{
\item "\code{salmon}" (\emph{default}): \href{https://combine-lab.github.io/salmon}{Salmon}
alignment-free, quasi-mapped counts.
\item "\code{kallisto}". \href{https://pachterlab.github.io/kallisto}{Kallisto}
alignment-free, pseudo-aligned counts.
\item "\code{sailfish}".
\href{http://www.cs.cmu.edu/~ckingsf/software/sailfish}{Sailfish}
alignment-free, lightweight counts.
\item "\code{star}": \href{https://github.com/alexdobin/STAR}{STAR}
(Spliced Transcripts Alignment to a Reference) aligned counts.
\item "\code{hisat2}": \href{https://ccb.jhu.edu/software/hisat2}{HISAT2}
(Hierarchical Indexing for Spliced Alignment of Transcripts) graph-based
aligned counts.
}}

\item{interestingGroups}{\code{character}. Groups of interest that define the
samples. If left unset, defaults to \code{sampleName}.}

\item{samples}{\code{character} or \code{NULL}. Specify a subset of samples to load.
The names must match the \code{description} specified in the bcbio YAML
metadata. If a \code{sampleMetadataFile} is provided, that will take priority
for sample selection. Typically this can be left unset.}

\item{censorSamples}{\code{character} or \code{NULL}. Samples to exclude from the
analysis.}

\item{sampleMetadataFile}{\code{string} or \code{NULL}. Custom metadata file containing
sample information. Otherwise defaults to sample metadata saved in the YAML
file. Remote URLs are supported. Typically this can be left unset.}

\item{organism}{\code{string} or \code{NULL}. Organism name. Use the full Latin name
(e.g. "Homo sapiens"), since this will be input downstream to AnnotationHub
and ensembldb, unless \code{gffFile} is set. If left \code{NULL} (\emph{not recommended}),
the function call will skip loading gene/transcript-level annotations into
\code{\link[=rowRanges]{rowRanges()}}. This can be useful for poorly annotation genomes or
experiments involving multiple genomes.}

\item{genomeBuild}{\code{string} or \code{NULL}. Ensembl genome build name (e.g.
"GRCh38"). This will be passed to AnnotationHub for \code{EnsDb} annotation
matching, unless \code{gffFile} is set.}

\item{ensemblRelease}{\code{scalar integer} or \code{NULL}. Ensembl release version. If
unset, defaults to current release, and does not typically need to be
user-defined. Passed to AnnotationHub for \code{EnsDb} annotation matching,
unless \code{gffFile} is set.}

\item{gffFile}{\code{string} or \code{NULL}. By default, we recommend leaving this
\code{NULL} for genomes that are supported on Ensembl. In this case, the row
annotations (\code{\link[=rowRanges]{rowRanges()}}) will be obtained automatically from Ensembl by
passing the \code{organism}, \code{genomeBuild}, and \code{ensemblRelease} arguments to
AnnotationHub and ensembldb. For a genome that is not supported on Ensembl
and/or AnnotationHub, a GFF/GTF (General Feature Format) file is required.
Generally, we recommend using a GTF (GFFv2) file here over a GFF3 file if
possible, although all GFF formats are supported. The function will
internally generate a \code{TxDb} containing transcript-to-gene mappings and
construct a \code{GRanges} object containing the genomic ranges (\code{\link[=rowRanges]{rowRanges()}}).}

\item{transgeneNames}{\code{character}.
Vector indicating which assay rows denote transgenes (e.g. EGFP, TDTOMATO).}

\item{spikeNames}{\code{character}.
Vector indicating which assay rows denote spike-in sequences (e.g. ERCCs).}

\item{vst}{\code{boolean}. Calculate variance-stabilizing transformation using
\code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}}. Recommended by default
for visualization.}

\item{rlog}{\code{boolean}. Calcualte regularized log transformation using
\code{\link[DESeq2:rlog]{DESeq2::rlog()}}. This calculation is slow for large datasets and now
discouraged by default for visualization.}

\item{...}{Additional arguments, slotted into the \code{\link[=metadata]{metadata()}} accessor.}
}
\value{
\code{bcbioRNASeq}.
}
\description{
\code{bcbioRNASeq} is an S4 class that extends \code{RangedSummarizedExperiment}, and
is designed to store a \href{https://bcbio-nextgen.readthedocs.org}{bcbio} RNA-seq
analysis.
}
\details{
Simply point to the final upload directory generated by
\href{https://bcbio-nextgen.readthedocs.io/}{bcbio}, and this constructor function
will take care of the rest. It automatically imports RNA-seq counts,
metadata, and the program versions used.

This class contains raw read counts and length-scaled
transcripts per million (TPM) generated by \code{\link[tximport:tximport]{tximport::tximport()}}. Counts can
be loaded at gene or transcript level.
}
\note{
\code{bcbioRNASeq} extended \code{SummarizedExperiment} prior to v0.2.0, where we
migrated to \code{RangedSummarizedExperiment}.
}
\section{Genome build}{


Ensure that the organism and genome build used with bcio match correctly
here in the function call. In particular, for the legacy \emph{Homo sapiens}
GRCh37/hg19 genome build, ensure that \code{genomeBuild = "GRCh37"}. Otherwise,
the genomic ranges set in \code{\link[=rowRanges]{rowRanges()}} will mismatch. It is recommended
for current projects that GRCh38/hg38 is used in place of GRCh37/hg19
if possible.
}

\section{Metadata}{


The \code{\link[=metadata]{metadata()}} slot contains:
\itemize{
\item Sample quality control metrics.
\item Ensembl annotations.
\item Server run paths.
\item R session information (e.g. \code{\link[utils:sessionInfo]{utils::sessionInfo()}}).
}
}

\section{DESeq2}{


DESeq2 is run automatically when \code{bcbioRNASeq()} is called, and variance
stabilized counts are slotted into \code{\link[=assays]{assays()}}. If the number of samples is
bigger than the \code{transformationLimit} argument, \code{rlog} and \code{vst} counts will
not be slotted into \code{\link[=assays]{assays()}}. In this case, we recommend visualization
using TMM counts, which are automatically calculated using edgeR.
}

\section{Remote data}{


When working in RStudio, we recommend connecting to the bcbio run directory
as a remote connection over
\href{https://github.com/osxfuse/osxfuse/wiki/SSHFS}{sshfs}.
}

\examples{
upload_dir <- system.file("extdata/bcbio", package = "bcbioRNASeq")

# Gene level
x <- bcbioRNASeq(
    uploadDir = upload_dir,
    level = "genes",
    caller = "salmon",
    organism = "Mus musculus",
    ensemblRelease = 87L
)
show(x)
is(x, "RangedSummarizedExperiment")
validObject(x)

# Transcript level
x <- bcbioRNASeq(
    uploadDir = upload_dir,
    level = "transcripts",
    caller = "salmon",
    organism = "Mus musculus",
    ensemblRelease = 87L
)
show(x)
validObject(x)
}
\seealso{
\itemize{
\item \code{\link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment::SummarizedExperiment()}}.
\item \code{\link[methods:initialize]{methods::initialize()}}.
\item \code{\link[methods:validObject]{methods::validObject()}}.
\item \code{\link[BiocGenerics:updateObject]{BiocGenerics::updateObject()}}.
\item \code{.S4methods(class = "bcbioRNASeq")}.
}

Other S4 Object: \code{\link{coerce}},
  \code{\link{extract}}, \code{\link{show}},
  \code{\link{updateObject}}
}
\author{
Michael Steinbaugh, Lorena Pantano, Rory Kirchner, Victor Barrera
}
\concept{S4 Object}
