% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{bcbioRNASeqTestsURL}
\alias{bcbioRNASeqTestsURL}
\title{Cache URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
bcbioRNASeqTestsURL
}
\description{
Cache URL
}
\examples{
bcbioRNASeqTestsURL
}
\keyword{internal}
