% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-coerce.R
\name{coerce}
\alias{coerce}
\alias{coerce-bcbioRNASeq-DESeqDataSet}
\title{Coerce Object}
\value{
Object of new class.
}
\description{
Coerce Object
}
\examples{
# DESeqDataSet ====
dds <- as(bcb_small, "DESeqDataSet")
show(dds)

# RangedSummarizedExperiment ====
rse <- as(bcb_small, "RangedSummarizedExperiment")
show(rse)

# SummarizedExperiment ====
se <- as(bcb_small, "SummarizedExperiment")
show(se)
}
\author{
Michael Steinbaugh
}
