% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-coerce.R
\name{coerce}
\alias{coerce}
\alias{coerce-bcbioRNASeq-DESeqDataSet}
\alias{coerce-bcbioRNASeq-list}
\title{Coerce Object}
\value{
Object of new class.
}
\description{
Coerce Object
}
\examples{
# DESeqDataSet ====
dds <- as(bcb_small, "DESeqDataSet")
show(dds)

# RangedSummarizedExperiment ====
rse <- as(bcb_small, "RangedSummarizedExperiment")
slotNames(rse)
show(rse)

# SummarizedExperiment ====
# Coerced to RangedSummarizedExperiment first.
# Otherwise, rowData will be NULL.
se <- as(rse, "SummarizedExperiment")
slotNames(se)
show(se)

# list ====
list <- as(bcb_small, "list")
names(list)
}
\author{
Michael Steinbaugh
}
