% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-coerce.R
\name{coerce}
\alias{coerce}
\alias{as}
\alias{coerce-bcbioRNASeq-DESeqDataSet}
\title{Methods for Coercing an Object to a Class}
\value{
Object of new class.
}
\description{
Methods for Coercing an Object to a Class
}
\examples{
# DESeqDataSet ====
x <- as(bcb_small, "DESeqDataSet")
names(S4Vectors::mcols(x))
class(x)
show(x)

# RangedSummarizedExperiment ====
x <- as(bcb_small, "RangedSummarizedExperiment")
slotNames(x)
show(x)

# SummarizedExperiment ====
# Coerce to RangedSummarizedExperiment first.
x <- as(bcb_small, "RangedSummarizedExperiment")
x <- as(x, "SummarizedExperiment")
class(x)
slotNames(x)
show(x)
}
\seealso{
\itemize{
\item \code{\link[methods:as]{methods::as()}}.
\item \code{\link[methods:coerce]{methods::coerce()}}.
}

Other S4 Object: \code{\link{bcbioRNASeq}},
  \code{\link{show}}, \code{\link{subset}},
  \code{\link{updateObject}}
}
\author{
Michael Steinbaugh
}
