% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAs.R
\name{coerce}
\alias{coerce}
\alias{upgrade-bcbioRNASeq}
\alias{coerce-bcbioRNASeq-SummarizedExperiment}
\title{Coerce Object}
\arguments{
\item{from}{Class for which the coerce method will perform coercion.}
}
\description{
Coerce Object
}
\section{Upgrade \code{bcbioRNASeq} to current version}{

This method adds support for upgrading \link{bcbioRNADataSet} objects to the
latest \link{bcbioRNASeq} class version. This should be backwards compatible to
\link{bcbioRNASeq} version 0.0.26. Previous objects saved using \code{bcbioRnaseq}
(note case) will likely fail to load with newer versions of the package.
}

\section{\link{bcbioRNASeq} to \link{SummarizedExperiment}}{

Since \link{bcbioRNASeq} is an extension of \link{SummarizedExperiment}, this
coercion method is very simple. Here we're simply dropping our \code{@bcbio} slot,
which contains raw cellular barcodes and other bcbio-specific metadata.
}

\examples{
\dontrun{
# bcbioRNADataSet
newClass <- as(oldClass, "bcbioRNASeq")
}
}
\seealso{
\code{help(topic = "coerce", package = "methods")}.
}
\author{
Michael Steinbaugh
}
