% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-methods.R
\name{coerce}
\alias{coerce}
\alias{as}
\alias{coerce,bcbioRNASeq,DESeqDataSet-method}
\alias{coerce,bcbioRNASeq,DESeqTransform-method}
\alias{coerce,DESeqAnalysis,DESeqDataSet-method}
\alias{coerce,DESeqAnalysis,DESeqTransform-method}
\alias{coerce,DESeqAnalysis,DESeqResults-method}
\title{Methods for Coercing an Object to a Class}
\value{
Object of new class.
}
\description{
Methods for Coercing an Object to a Class
}
\examples{
# bcbioRNASeq to DESeqDataSet ====
x <- as(bcb_small, "DESeqDataSet")
names(S4Vectors::mcols(x))
class(x)
show(x)

# bcbioRNASeq to RangedSummarizedExperiment ====
x <- as(bcb_small, "RangedSummarizedExperiment")
slotNames(x)
show(x)

# bcbioRNASeq to SummarizedExperiment ====
# Coerce to RangedSummarizedExperiment first.
x <- as(bcb_small, "RangedSummarizedExperiment")
x <- as(x, "SummarizedExperiment")
class(x)
slotNames(x)
show(x)

# DESeqAnalysis ====
dds <- as(deseq_small, "DESeqDataSet")
print(dds)
dt <- as(deseq_small, "DESeqTransform")
print(dt)
# Pulls the first results slotted.
res <- as(deseq_small, "DESeqResults")
contrastName(res)
summary(res)
}
\seealso{
\itemize{
\item \code{\link[methods:as]{methods::as()}}.
\item \code{\link[methods:canCoerce]{methods::canCoerce()}}.
}

Other S4 Functions: \code{\link{extract}},
  \code{\link{show}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Functions}
