% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-coerce.R
\name{coerce}
\alias{coerce}
\alias{as}
\alias{coerce-bcbioRNASeq-DESeqDataSet}
\alias{coerce-bcbioRNASeq-list}
\title{Methods for Coercing an Object to a Class}
\value{
Object of new class.
}
\description{
Methods for Coercing an Object to a Class
}
\examples{
# DESeqDataSet ====
dds <- as(bcb_small, "DESeqDataSet")
class(dds)
show(dds)

# RangedSummarizedExperiment ====
rse <- as(bcb_small, "RangedSummarizedExperiment")
slotNames(rse)
show(rse)

# SummarizedExperiment ====
# Coerced to RangedSummarizedExperiment first.
# Otherwise, rowData will be NULL.
se <- as(rse, "SummarizedExperiment")
class(se)
slotNames(se)
show(se)

# list ====
list <- as(bcb_small, "list")
class(list)
names(list)
}
\seealso{
\itemize{
\item \code{\link[methods:as]{methods::as()}}.
\item \code{\link[methods:coerce]{methods::coerce()}}.
}

Other S4 Class Definition: \code{\link{bcbioRNASeq-class}},
  \code{\link{colData}}, \code{\link{rowData}},
  \code{\link{show}}, \code{\link{subset}},
  \code{\link{updateObject}}
}
\author{
Michael Steinbaugh
}
