% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAs.R
\name{coerce}
\alias{coerce}
\alias{coerce-bcbioRNASeq-SummarizedExperiment}
\title{Coerce Object}
\arguments{
\item{from}{Class for which the coerce method will perform coercion.}
}
\description{
Coerce Object
}
\section{bcbioRNASeq to SummarizedExperiment}{

Since \code{bcbioRNASeq} is an extension of \code{SummarizedExperiment}, this
coercion method is very simple. Here we're simply dropping the \code{bcbio} slot,
which contains raw cellular barcodes and other bcbio-specific metadata.
}

\examples{
load(system.file("extdata/bcb.rda", package = "bcbioRNASeq"))
se <- as(bcb, "SummarizedExperiment")
print(se)
}
\seealso{
\code{help(topic = "coerce", package = "methods")}.
}
\author{
Michael Steinbaugh
}
