% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-methods.R
\name{coerce}
\alias{coerce}
\alias{coerce,bcbioRNASeq,DESeqDataSet-method}
\alias{coerce,bcbioRNASeq,DESeqTransform-method}
\title{Methods for coercing an object to a class}
\description{
Force an object to belong to a class.
}
\section{bcbioRNASeq to DESeqDataSet}{

\enumerate{
\item Coerces to \code{RangedSummarizedExperiment}.
\item Rounds raw counts to \code{integer matrix}.
\item Subsets \code{colData} to include only clean factor columns.
See \code{\link[basejump:sampleData]{sampleData()}} for more details.
\item Simplifies \code{metadata} to include only relevant information and updates
\code{sessionInfo} slot.
}

Note that gene-level counts are required. Alternatively
\code{\link[tximport:summarizeToGene]{tximport::summarizeToGene()}} can be called to convert transcript-level
counts to gene-level. By default, we're using length-scaled TPM, so a
corresponding average transcript length matrix isn't necessary. The average
transcript length matrix is only necessary when raw counts matrix isn't
scaled during tximport call (see \code{countsFromAbundance} in tximport).
}

\section{bcbioRNASeq to DESeqTransform}{

\enumerate{
\item Coerces to \code{DESeqDataSet}.
\item Calls \code{\link[DESeq2:DESeq]{DESeq2::DESeq()}}.
\item Calls \code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}}.
}
}

\examples{
data(bcb)

## bcbioRNASeq to DESeqDataSet ====
x <- as(bcb, "DESeqDataSet")
names(S4Vectors::mcols(x))
class(x)
show(x)

## bcbioRNASeq to RangedSummarizedExperiment ====
x <- as(bcb, "RangedSummarizedExperiment")
slotNames(x)
show(x)

## bcbioRNASeq to SummarizedExperiment ====
## Coerce to RangedSummarizedExperiment first.
x <- as(bcb, "RangedSummarizedExperiment")
x <- as(x, "SummarizedExperiment")
class(x)
slotNames(x)
show(x)
}
\seealso{
\itemize{
\item \code{\link[tximport:tximport]{tximport::tximport()}}.
\item \code{\link[DESeq2:DESeqDataSetFromTximport]{DESeq2::DESeqDataSetFromTximport()}}.
}
}
\author{
Michael Steinbaugh
}
