% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-colData.R
\docType{methods}
\name{colData}
\alias{colData}
\alias{colData<-,bcbioRNASeq,DataFrame-method}
\title{Column Data}
\usage{
\S4method{colData}{bcbioRNASeq,DataFrame}(x, ...) <- value
}
\arguments{
\item{x}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{value}{Value to assign.}
}
\description{
Improved assignment method support for \link{bcbioRNASeq} object.

This method support will also update the \code{colData} inside the \code{bcbio} and
\code{assays} slots.
}
\examples{
bcb <- examples[["bcb"]]
cd <- colData(bcb)
cd[["age"]] <- factor(c(14, 30, 14, 30))
colData(bcb) <- cd
colData(bcb)

# These internal objects will also get updated
bcbio(bcb, "DESeqDataSet") \%>\% colData()
assays(bcb)[["rlog"]] \%>\% colData()
assays(bcb)[["vst"]] \%>\% colData()
}
\seealso{
\code{help("colData", "SummarizedExperiment")}
}
