% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-colData.R
\docType{methods}
\name{colData}
\alias{colData}
\alias{colData<-,bcbioRNASeq,DataFrame-method}
\title{Column Data}
\usage{
\S4method{colData}{bcbioRNASeq,DataFrame}(x, ...) <- value
}
\arguments{
\item{x}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{value}{Value to assign.}
}
\description{
Improved assignment method support for \link{bcbioRNASeq} object.

This method support will also update the \code{colData} inside the \code{bcbio} and
\code{assays} slots.
}
\examples{
load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioRNASeq"))

# Assignment support
colData <- colData(bcb)
colData[["age"]] <- factor(c(14, 30, 14, 30))
colData(bcb) <- colData
colData(bcb) \%>\% glimpse()

# These internal objects will also get updated
bcbio(bcb, "DESeqDataSet") \%>\% colData() \%>\% glimpse()
assays(bcb)[["rlog"]] \%>\% colData() \%>\% glimpse()
assays(bcb)[["vst"]] \%>\% colData() \%>\% glimpse()
}
\seealso{
\code{help("colData", "SummarizedExperiment")}
}
