% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-colData.R
\docType{methods}
\name{colData}
\alias{colData}
\alias{colData,bcbioRNASeq-method}
\alias{colData<-,bcbioRNASeq,data.frame-method}
\alias{colData<-,bcbioRNASeq,DataFrame-method}
\title{Column Data}
\usage{
\S4method{colData}{bcbioRNASeq}(x, return = c("DataFrame", "data.frame"))

\S4method{colData}{bcbioRNASeq,data.frame}(x, ...) <- value

\S4method{colData}{bcbioRNASeq,DataFrame}(x, ...) <- value
}
\arguments{
\item{x}{Object.}

\item{return}{Return as "\code{DataFrame}" or "\code{data.frame}".}

\item{...}{Additional arguments.}

\item{value}{Value to assign.}
}
\value{
Data describing the columns of the object.
}
\description{
Improved assignment method support for a \code{bcbioRNASeq} object.
}
\details{
This method support will also update the nested \code{\link[=colData]{colData()}} inside the
\code{assays} slot.
}
\examples{
load(system.file("extdata/bcb.rda", package = "bcbioRNASeq"))

# Return as data.frame
colData(bcb, return = "data.frame")

# Assignment support
colData <- colData(bcb)
# All columns will be coerced to factors
colData[["age"]] <- c(14L, 30L, 14L, 30L)
colData(bcb) <- colData
colData(bcb) \%>\% glimpse()

# These internal objects will also get updated
assays(bcb)[["rlog"]] \%>\% colData() \%>\% glimpse()
assays(bcb)[["vst"]] \%>\% colData() \%>\% glimpse()
}
\seealso{
\code{help("colData", "SummarizedExperiment")}
}
