% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-colData.R
\docType{methods}
\name{colData}
\alias{colData}
\alias{colData<-,bcbioRNASeq,data.frame-method}
\alias{colData<-,bcbioRNASeq,DataFrame-method}
\title{Column Data}
\usage{
\S4method{colData}{bcbioRNASeq,data.frame}(x, ...) <- value

\S4method{colData}{bcbioRNASeq,DataFrame}(x, ...) <- value
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}

\item{value}{Value to assign.}
}
\value{
\link{DataFrame}.
}
\description{
Improved assignment method support for \link{bcbioRNASeq} object.

This method support will also update the \code{colData} inside the \code{bcbio} and
\code{assays} slots.
}
\examples{
load(system.file("extdata/bcb.rda", package = "bcbioRNASeq"))

# Assignment support
colData <- colData(bcb)
# All columns will be coerced to factors
colData[["age"]] <- c(14L, 30L, 14L, 30L)
colData(bcb) <- colData
colData(bcb) \%>\% glimpse()

# These internal objects will also get updated
bcbio(bcb, "DESeqDataSet") \%>\% colData() \%>\% glimpse()
assays(bcb)[["rlog"]] \%>\% colData() \%>\% glimpse()
assays(bcb)[["vst"]] \%>\% colData() \%>\% glimpse()
}
\seealso{
\code{help("colData", "SummarizedExperiment")}
}
