% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-colData.R
\docType{methods}
\name{colData}
\alias{colData}
\alias{colData,bcbioRNASeq-method}
\alias{colData<-,bcbioRNASeq,data.frame-method}
\alias{colData<-,bcbioRNASeq,DataFrame-method}
\title{Column Data}
\usage{
\S4method{colData}{bcbioRNASeq}(x, return = c("data.frame", "DataFrame"))

\S4method{colData}{bcbioRNASeq,data.frame}(x) <- value

\S4method{colData}{bcbioRNASeq,DataFrame}(x) <- value
}
\arguments{
\item{x}{Object.}

\item{return}{Return as "\code{data.frame}" or "\code{DataFrame}".}

\item{value}{Value to assign.}
}
\value{
Data describing the columns of the object.
}
\description{
Improved assignment method support for a \code{bcbioRNASeq} object.
}
\details{
This method support will also update the nested \code{\link[=colData]{colData()}} inside the
\code{assays} slot.
}
\examples{
# bcbioRNASeq ====
colData(bcb_small) \%>\% glimpse()

# Assignment method support
# All columns will be coerced to factors
colData(bcb_small)[["age"]] <- c(14L, 30L, 14L, 30L, 14L, 30L)
colData(bcb_small) \%>\% glimpse()
}
\seealso{
\code{help("colData", "SummarizedExperiment")}

Other S4 Class Definition: \code{\link{bcbioRNASeq-class}},
  \code{\link{coerce}}, \code{\link{rowData}},
  \code{\link{show}}, \code{\link{subset}},
  \code{\link{updateObject}}
}
\author{
Michael Steinbaugh
}
