% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-contrastName.R
\docType{methods}
\name{contrastName}
\alias{contrastName}
\alias{contrastName}
\alias{contrastName,DESeqResults-method}
\title{Contrast Name}
\usage{
contrastName(object, ...)

\S4method{contrastName}{DESeqResults}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Contrast name \code{string}.
}
\description{
Contrast Name
}
\examples{
load(system.file("extdata/res.rda", package = "bcbioRNASeq"))

# DESeqResults
contrastName(res)
}
\author{
Michael Steinbaugh
}
