% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/contrastName-methods.R
\docType{methods}
\name{contrastName}
\alias{contrastName}
\alias{contrastName}
\alias{contrastName,DESeqResults-method}
\title{Contrast Name}
\usage{
contrastName(object, ...)

\S4method{contrastName}{DESeqResults}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Contrast name \code{string}.
}
\description{
Contrast Name
}
\examples{
# DESeqResults ====
contrastName(res_small)
}
\seealso{
Other Differential Expression Functions: \code{\link{alphaSummary}},
  \code{\link{plotDEGHeatmap}}, \code{\link{plotDEGPCA}},
  \code{\link{plotMeanAverage}}, \code{\link{plotVolcano}},
  \code{\link{resultsTables}}
}
\author{
Michael Steinbaugh
}
