% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_heatmap.R
\name{cor_heatmap}
\alias{cor_heatmap}
\title{Correlation matrix heatmap}
\usage{
cor_heatmap(counts, metadata, factor = "group", method = "pearson", ...)
}
\arguments{
\item{counts}{RNA-Seq counts}

\item{metadata}{Metadata data frame}

\item{factor}{Factor groups to use for clustering color headers}

\item{method}{Correlation coefficient (or covariance) to be computed}

\item{...}{Passthrough to \code{pheatmap()}}

\item{format}{Format of input count data}
}
\value{
Heatmap
}
\description{
Correlation matrix heatmap
}
\author{
Michael Steinbaugh
}
