% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_heatmap.R
\name{cor_heatmap}
\alias{cor_heatmap}
\title{Correlation matrix heatmap}
\usage{
cor_heatmap(bcbio, counts, method = "pearson", ...)
}
\arguments{
\item{bcbio}{bcbio list object}

\item{counts}{RNA-Seq counts}

\item{method}{Correlation coefficient (or covariance) to be computed}

\item{...}{Optional passthrough to \code{pheatmap()}}
}
\value{
Heatmap
}
\description{
Correlation matrix heatmap
}
\author{
Michael Steinbaugh
}
