% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_methods.R
\docType{methods}
\name{counts}
\alias{counts}
\title{Count matrix accessors}
\usage{
\S4method{counts}{bcbioRnaDataSet}(object, normalized = FALSE)
}
\description{
Count matrix accessors
}
\examples{
# Raw counts
counts(bcb, normalized = FALSE)

# TPM
counts(bcb, normalized = TRUE)
counts(bcb, normalized = "tpm")

# TMM
counts(bcb, normalized = "tmm")
}
\author{
Michael Steinbaugh
}
