% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_methods.R
\docType{methods}
\name{counts}
\alias{counts}
\alias{counts.bcbioRnaDataSet}
\alias{counts,bcbioRnaDataSet-method}
\alias{counts<-,bcbioRnaDataSet,matrix-method}
\alias{counts,bcbioRnaDataSet-method}
\alias{counts}
\title{Accessors for the count matrix of a bcbioRnaDataSet object.}
\usage{
counts.bcbioRnaDataSet(object)

\S4method{counts}{bcbioRnaDataSet}(object)

\S4method{counts}{bcbioRnaDataSet,matrix}(object) <- value
}
\arguments{
\item{object}{a \code{bcbioRnaDataSet} object}

\item{value}{an integer matrix}
}
\value{
\code{\link[base]{matrix}} with raw count data.
}
\description{
Accessors for the count matrix of a bcbioRnaDataSet object.
}
\author{
Lorena Pantano
}
