% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-counts.R
\docType{methods}
\name{counts,bcbioRNADataSet-method}
\alias{counts,bcbioRNADataSet-method}
\title{Count Matrix Accessors}
\usage{
\S4method{counts}{bcbioRNADataSet}(object, normalized = FALSE)
}
\arguments{
\item{object}{Object.}

\item{normalized}{Select raw counts (\code{FALSE}), DESeq2 normalized counts
(\code{TRUE}), or additional normalization methods:
\itemize{
\item \code{tpm}: Transcripts per million.
\item \code{tmm}: Trimmed mean of M-values (edgeR).
\item \code{rlog}: Regularized log transformation (\code{\link[DESeq2:rlog]{DESeq2::rlog()}}).
\item \code{vst}: Variance stabilizing transformation
(\code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}}).
}}
}
\value{
\link{matrix}.
}
\description{
By default, \code{\link[=counts]{counts()}} returns the raw counts. Normalized counts, including
transcripts per million (TPM) can be accessed using the \code{normalized}
argument.
}
\examples{
data(bcb)
# Raw counts
counts(bcb, normalized = FALSE) \%>\% glimpse

# DESeq2 normalized counts
counts(bcb, normalized = TRUE) \%>\% glimpse

# TPM
counts(bcb, normalized = "tpm") \%>\% glimpse

# TMM
counts(bcb, normalized = "tmm") \%>\% glimpse

# rlog
counts(bcb, normalized = "rlog") \%>\% glimpse

# VST
counts(bcb, normalized = "vst") \%>\% glimpse
}
\author{
Michael Steinbaugh
}
