% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-counts.R
\docType{methods}
\name{counts,bcbioRnaDataSet-method}
\alias{counts,bcbioRnaDataSet-method}
\title{Count matrix accessors}
\usage{
\S4method{counts}{bcbioRnaDataSet}(object, normalized = FALSE)
}
\arguments{
\item{object}{Object.}

\item{normalized}{Select normalized counts (\code{TRUE}), raw counts (\code{FALSE}),
or specifically request TMM-normalized counts used for QC functions (\code{tmm}).}
}
\description{
Count matrix accessors
}
\examples{
\dontrun{
# Raw counts
counts(bcb, normalized = FALSE)

# TPM
counts(bcb, normalized = TRUE)
counts(bcb, normalized = "tpm")

# TMM
counts(bcb, normalized = "tmm")
}
}
\author{
Michael Steinbaugh
}
