% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-counts.R
\docType{methods}
\name{counts}
\alias{counts}
\alias{counts,bcbioRNASeq-method}
\title{Count Matrix Accessors}
\usage{
\S4method{counts}{bcbioRNASeq}(object, normalized = FALSE)
}
\arguments{
\item{object}{Object.}

\item{normalized}{Select raw counts (\code{FALSE}), DESeq2 normalized counts
(\code{TRUE}), or additional normalization methods:
\itemize{
\item \code{tpm}: Transcripts per million.
\item \code{tmm}: Trimmed mean of M-values (edgeR).
\item \code{rlog}: Regularized log transformation (\code{\link[DESeq2:rlog]{DESeq2::rlog()}}).
\item \code{vst}: Variance stabilizing transformation
(\code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}}).
}}
}
\value{
\code{matrix}.
}
\description{
By default, \code{\link[=counts]{counts()}} returns the raw counts. Normalized counts, including
transcripts per million (TPM) can be accessed using the \code{normalized}
argument.
}
\examples{
# Raw counts (from tximport)
counts(bcb_small, normalized = FALSE) \%>\% summary()

# TPM (from tximport)
counts(bcb_small, normalized = "tpm") \%>\% summary()

# Normalized counts (from DESeq2)
counts(bcb_small, normalized = TRUE) \%>\% summary()

# rlog (from DESeq2)
counts(bcb_small, normalized = "rlog") \%>\% summary()

# VST (from DESeq2)
counts(bcb_small, normalized = "vst") \%>\% summary()

# TMM (from edgeR)
counts(bcb_small, normalized = "tmm") \%>\% summary()
}
\author{
Michael Steinbaugh
}
