% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-counts.R
\docType{methods}
\name{counts}
\alias{counts}
\alias{counts,bcbioRNASeq-method}
\title{Count Matrix Accessors}
\usage{
\S4method{counts}{bcbioRNASeq}(object, normalized = FALSE)
}
\arguments{
\item{object}{Object.}

\item{normalized}{Select raw counts (\code{FALSE}), DESeq2 normalized counts
(\code{TRUE}), or additional normalization methods:
\itemize{
\item \code{tpm}: Transcripts per million.
\item \code{tmm}: Trimmed mean of M-values (edgeR).
\item \code{rlog}: Regularized log transformation (\code{\link[DESeq2:rlog]{DESeq2::rlog()}}).
\item \code{vst}: Variance stabilizing transformation
(\code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}}).
}}
}
\value{
\link{matrix}.
}
\description{
By default, \code{\link[=counts]{counts()}} returns the raw counts. Normalized counts, including
transcripts per million (TPM) can be accessed using the \code{normalized}
argument.
}
\examples{
data(bcb)

# Raw counts
counts(bcb, normalized = FALSE) \%>\% summary()

# TPM
counts(bcb, normalized = "tpm") \%>\% summary()

# DESeq2 normalized counts
\dontrun{
counts(bcb, normalized = TRUE)
}

# rlog
\dontrun{
counts(bcb, normalized = "rlog")
}

# TMM
\dontrun{
counts(bcb, normalized = "tmm")
}

# VST
\dontrun{
counts(bcb, normalized = "vst")
}
}
\author{
Michael Steinbaugh
}
