% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-counts.R, R/reexports.R
\docType{methods}
\name{counts,bcbioRNADataSet-method}
\alias{counts,bcbioRNADataSet-method}
\alias{counts}
\title{Count matrix accessors}
\usage{
\S4method{counts}{bcbioRNADataSet}(object, normalized = FALSE)

counts(object, ...)
}
\arguments{
\item{object}{Object.}

\item{normalized}{Select normalized counts (\code{TRUE}), raw counts (\code{FALSE}),
or specifically request TMM-normalized counts used for QC functions (\code{tmm}).}
}
\description{
By default, \code{\link[=counts]{counts()}} returns the raw counts. This method will return
transcripts per million (TPM) by default when \code{normalized = TRUE}. This can
be overriden by requesting the normalization method directly.
}
\examples{
\dontrun{
# Raw counts
counts(bcb, normalized = FALSE)

# TPM
counts(bcb, normalized = TRUE)
counts(bcb, normalized = "tpm")

# TMM
counts(bcb, normalized = "tmm")
}
}
\author{
Michael Steinbaugh
}
