% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-counts.R
\docType{methods}
\name{counts}
\alias{counts}
\alias{counts,bcbioRNASeq-method}
\title{Count Matrix Accessors}
\usage{
\S4method{counts}{bcbioRNASeq}(object, normalized = FALSE)
}
\arguments{
\item{object}{Object.}

\item{normalized}{Logical or character indicating which normalization
method to apply:
\itemize{
\item \code{FALSE}: Raw counts (tximport).
\item \code{TRUE}: DESeq2 normalized counts. Calculated on the fly.
\item "\code{tpm}": Transcripts per million (tximport).
\item "\code{tmm}": edgeR trimmed mean of M-values. Calculated on the fly.
\item "\code{rlog}": DESeq2 \strong{log2} regularized log transformation.
\item "\code{vst}": DESeq2 \strong{log2} variance stabilizing transformation.
}}
}
\value{
\code{matrix}.
}
\description{
By default, \code{\link[=counts]{counts()}} returns the raw counts. Normalized counts, including
transcripts per million (TPM) can be accessed using the "\code{normalized}"
argument.
}
\examples{
load(system.file("extdata/bcb_small.rda", package = "bcbioRNASeq"))

# bcbioRNASeq ====
counts(bcb_small, normalized = FALSE) \%>\% summary()
counts(bcb_small, normalized = TRUE) \%>\% summary()
counts(bcb_small, normalized = "tpm") \%>\% summary()
counts(bcb_small, normalized = "tmm") \%>\% summary()

# log2 scale
counts(bcb_small, normalized = "rlog") \%>\% summary()
counts(bcb_small, normalized = "vst") \%>\% summary()
}
\seealso{
\itemize{
\item \code{\link[=tpm]{tpm()}}.
\item \code{\link[=tmm]{tmm()}}.
\item \code{\link[DESeq2:counts]{DESeq2::counts()}}.
\item \code{\link[DESeq2:rlog]{DESeq2::rlog()}}.
\item \code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}}.
}

Other Data Functions: \code{\link{colData}},
  \code{\link{rowData}}, \code{\link{selectSamples}},
  \code{\link{subset}}, \code{\link{tmm}},
  \code{\link{tpm}}
}
\author{
Michael Steinbaugh, Lorena Pantano
}
