% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts-methods.R
\docType{methods}
\name{counts}
\alias{counts}
\alias{counts,bcbioRNASeq-method}
\title{Count matrix accessors}
\usage{
counts(object, ...)

\S4method{counts}{bcbioRNASeq}(object, normalized = FALSE)
}
\arguments{
\item{object}{Object.}

\item{normalized}{\code{string} or \code{boolean}. Which normalization method to apply:
\itemize{
\item \code{FALSE}: Raw counts (tximport).
\item \code{TRUE}: DESeq2 normalized counts. Calculated on the fly.
\item "\code{tpm}": Transcripts per million (tximport).
\item "\code{vst}": DESeq2 \strong{log2} variance stabilizing transformation.
\item "\code{rlog}": DESeq2 \strong{log2} regularized log transformation.
\item "\code{tmm}": edgeR trimmed mean of M-values. Calculated on the fly.
\item "\code{rle}": Relative log expression transformation.
}}

\item{...}{Additional arguments.}
}
\value{
\code{matrix}.
}
\description{
By default, \code{\link[=counts]{counts()}} returns the raw counts. Normalized counts, including
transcripts per million (TPM) can be accessed using the "\code{normalized}"
argument.
}
\examples{
# bcbioRNASeq ====
counts(bcb_small, normalized = FALSE) \%>\% summary()
counts(bcb_small, normalized = TRUE) \%>\% summary()
counts(bcb_small, normalized = "tpm") \%>\% summary()
counts(bcb_small, normalized = "vst") \%>\% summary()
}
\seealso{
\itemize{
\item \code{\link[=tpm]{tpm()}}.
\item \code{\link[=tmm]{tmm()}}.
\item \code{\link[DESeq2:counts]{DESeq2::counts()}}.
\item \code{\link[DESeq2:rlog]{DESeq2::rlog()}}.
\item \code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}}.
}

Other Data Functions: \code{\link{metrics}},
  \code{\link{tmm}}
}
\author{
Michael Steinbaugh, Lorena Pantano
}
\concept{Data Functions}
