% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts-methods.R
\name{counts}
\alias{counts}
\alias{counts<-}
\alias{counts,bcbioRNASeq-method}
\title{Counts}
\usage{
\S4method{counts}{bcbioRNASeq}(object, normalized = FALSE)
}
\arguments{
\item{object}{Object.}

\item{normalized}{\code{character(1)} or \code{logical(1)}.
Normalization method to apply:
\itemize{
\item \code{FALSE}: Raw counts.
When using a \href{https://bioconductor.org/packages/tximport/}{tximport}-compatible caller, these are length scaled
by default (see \code{countsFromAbundance} argument).
When using a \href{http://bioinf.wehi.edu.au/featureCounts/}{featureCounts}-compatible caller, these are \code{integer}.
}

\href{https://bioconductor.org/packages/tximport/}{tximport} caller-specific normalizations:
\itemize{
\item \code{"tpm"}: \strong{T}ranscripts \strong{p}er \strong{m}illion.
}

Additional gene-level-specific normalizations:
\itemize{
\item \code{TRUE} / \code{"sf"}: \strong{S}ize \strong{f}actor (i.e. library size) normalized
counts.\cr
See \code{\link[DESeq2:sizeFactors]{DESeq2::sizeFactors()}} for details.
\item \code{"fpkm"}: \strong{F}ragments \strong{p}er \strong{k}ilobase per \strong{m}illion mapped
fragments.\cr
Requires \code{fpkm = TRUE} in \code{\link[=bcbioRNASeq]{bcbioRNASeq()}} call and gene annotations in
\code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{rowRanges()}} with defined
\code{\link[BiocGenerics:start]{width}}.\cr
See \code{\link[DESeq2:fpkm]{DESeq2::fpkm()}} for details.
\item \code{"vst"}: \strong{V}ariance-\strong{s}tabilizing \strong{t}ransformation (log2).\cr
Requires \code{vst = TRUE} to be set during \code{\link[=bcbioRNASeq]{bcbioRNASeq()}} call.\cr
See \code{DESeq2::varianceStabilizingTransformation} for more information.
\item \code{"rlog"}: \strong{R}egularized \strong{log} transformation (log2).\cr
Requires \code{rlog = TRUE} to be set during \code{\link[=bcbioRNASeq]{bcbioRNASeq()}} call.\cr
See \code{\link[DESeq2:rlog]{DESeq2::rlog()}} for details.
\item \code{"tmm"}: \strong{T}rimmed \strong{m}ean of \strong{M}-values.\cr
Calculated on the fly.\cr
See \code{\link[edgeR:calcNormFactors]{edgeR::calcNormFactors()}} for details.
\item \code{"rle"}: \strong{R}elative \strong{l}og \strong{e}xpression transformation.\cr
Calculated on the fly.\cr
See \code{\link[=relativeLogExpression]{relativeLogExpression()}} for details.

Note that \code{logical(1)} support only applies to \code{counts()}. Other
functions in the package require \code{character(1)} and use
\code{\link[base:match.arg]{match.arg()}} internally.
}}
}
\value{
\code{matrix}.
}
\description{
Count matrix.
}
\details{
By default, return the raw counts. Normalized counts in a variety of formats
can be accessed using the \code{normalized} argument.
}
\note{
Updated 2020-09-23.
}
\examples{
data(bcb)
summary(counts(bcb))
}
\references{
\itemize{
\item TMM: Robinson and Oshlack (2010).
\item RLE: Anders and Huber (2010).
}
}
\seealso{
\itemize{
\item \code{tximport::tximport()}.
\item \code{DESeq2::counts()}.
\item \code{DESeq2::sizeFactors()}.
\item \code{DESeq2::varianceStabilizingTransformation()}.
\item \code{DESeq2::rlog()}.
\item \code{DESeq2::fpkm()}.
\item \code{edgeR::calcNormFactors()}.
\item \code{tmm()}.
\item \code{relativeLogExpression()}.
}
}
\author{
Michael Steinbaugh, Lorena Pantano
}
