% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts-methods.R
\docType{methods}
\name{counts}
\alias{counts}
\alias{counts,bcbioRNASeq-method}
\title{Counts}
\usage{
\S4method{counts}{bcbioRNASeq}(object, normalized = FALSE)
}
\arguments{
\item{object}{Object.}

\item{normalized}{\code{string} or \code{boolean}. Which normalization method to apply:
\itemize{
\item \code{FALSE}: Raw counts.
\itemize{
\item When using a \href{https://bioconductor.org/packages/release/bioc/html/tximport.html}{tximport}-compatible caller, these are length scaled
by default (see \code{countsFromAbundance} argument).
\item When using a \href{http://bioinf.wehi.edu.au/featureCounts/}{featureCounts}-compatible caller, these are \code{integer}.
}
}

\href{https://bioconductor.org/packages/release/bioc/html/tximport.html}{tximport} caller-specific normalizations:
\itemize{
\item \code{"tpm"}: \strong{T}ranscripts \strong{p}er \strong{m}illion.
}

Additional gene-level-specific normalizations:
\itemize{
\item \code{TRUE}: Size factor-adjusted counts.\cr
See \code{DESeq2::sizeFactors()} for more information.
\item \code{"vst"}: \strong{V}ariance-\strong{s}tabilizing \strong{t}ransformation.\cr
Requires \code{vst = TRUE} to be set during \code{bcbioRNASeq()} call.\cr
See \code{DESeq2::varianceStabilizingTransformation()} for more information.
\item \code{"rlog"}: \strong{R}egularized \strong{log} transformation (log2).\cr
Requires \code{rlog = TRUE} to be set during \code{bcbioRNASeq()} call.\cr
See \code{DESeq2::rlog()} for more information.
\item \code{"tmm"}: \strong{T}rimmed \strong{m}ean of \strong{M}-values.\cr
See \code{edgeR::calcNormFactors()} for more information.
\item \code{"rle"}: \strong{R}elative \strong{l}og \strong{e}xpression transformation.
\item \code{"fpkm"}: \strong{F}ragments \strong{p}er \strong{k}ilobase per \strong{m}illion mapped
fragments.\cr
Requires annotations in \code{rowRanges()} with defined \code{width}; otherwise,
will be skipped during the \code{bcbioRNASeq()} load call.\cr
See \code{BiocGenerics::width()} and \code{DESeq2::fpkm()} for more information.
}}
}
\value{
\code{matrix}.
}
\description{
Count matrix.
}
\details{
By default, return the raw counts. Normalized counts in a variety of formats
can be accessed using the \code{normalized} argument.
}
\examples{
data(bcb)
counts(bcb) \%>\% summary()
}
\references{
\itemize{
\item TMM: Robinson and Oshlack (2010).
\item RLE: Anders and Huber (2010).
}
}
\seealso{
\itemize{
\item \code{tximport::tximport()}.
\item \code{DESeq2::counts()}.
\item \code{DESeq2::sizeFactors()}.
\item \code{DESeq2::varianceStabilizingTransformation()}.
\item \code{DESeq2::rlog()}.
\item \code{DESeq2::fpkm()}.
\item \code{edgeR::calcNormFactors()}.
\item \code{tmm()}.
\item \code{relativeLogExpression()}.
}
}
\author{
Michael Steinbaugh, Lorena Pantano
}
