% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_project.R
\name{create_new_project}
\alias{create_new_project}
\title{Create new project directory structure}
\usage{
create_new_project()
}
\description{
Create new project directory structure
}
\author{
Michael Steinbaugh
}
