% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-project_dir.R
\name{.data_versions}
\alias{.data_versions}
\title{Read data versions}
\usage{
.data_versions(project_dir)
}
\arguments{
\item{project_dir}{Project directory.}
}
\description{
Read data versions
}
\author{
Michael Steinbaugh
}
\keyword{internal}
