% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/david.R
\name{run_david}
\alias{run_david}
\alias{david_table}
\title{Query the DAVID website}
\usage{
run_david(res_tables, direction = "both", alpha = 0.05, count = 3,
  save = TRUE)

david_table(david)
}
\arguments{
\item{res_tables}{\code{res_tables()} list object}

\item{direction}{Gene direction: up, down, or both}

\item{alpha}{Alpha level cutoff}

\item{count}{Minimum hit count}

\item{save}{Save files to disk (\code{TRUE/FALSE})}

\item{david}{DAVID results list generated by \code{run_david}}
}
\value{
List of \code{RDAVIDWebService} report objects

Cutoff chart kable
}
\description{
Wrapper function that runs gene set enrichment analysis (GSEA)
  with the \code{RDAVIDWebService} package, using simplified input options.

Generate a DAVID results table for knit report
}
\author{
Michael Steinbaugh
}
