% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dds_small}
\alias{dds_small}
\title{\code{DESeqDataSet} example}
\format{An object of class \code{DESeqDataSet} with 500 rows and 6 columns.}
\usage{
dds_small
}
\description{
\code{DESeqDataSet} example
}
\seealso{
Other GSE65267: \code{\link{bcb_small}},
  \code{\link{res_small}}
}
\author{
Lorena Pantano, Michael Steinbaugh
}
\concept{GSE65267}
\keyword{datasets}
