% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.R
\name{plot_ma}
\alias{plot_ma}
\alias{plot_volcano}
\title{Differential expression plots}
\usage{
plot_ma(res, ylim = 2)

plot_volcano(run, res, lfc = 1, text_labels = 30)
}
\arguments{
\item{res}{\code{DESeqResults}}

\item{ylim}{Y-axis maximum (single integer)}

\item{run}{\code{bcbio-nextgen} run}

\item{lfc}{Log fold change ratio (base 2) cutoff for coloring}

\item{text_labels}{Number of text labels to plot}
}
\value{
MA plot

Volcano plot
}
\description{
Wrapper for \code{DESeq2::plotMA} that generates a title
  automatically

Volcano plot
}
\author{
Michael Steinbaugh
}
