% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-de.R
\name{plot_ma}
\alias{plot_ma}
\alias{plot_volcano}
\title{Differential expression plots}
\usage{
plot_ma(res, ylim = 2)

plot_volcano(bcb, res, lfc = 1, text_labels = 30, merge = TRUE,
  direction = "both", title = NULL, shade_color = "orange",
  shade_alpha = 0.25, point_color = "gray", point_alpha = 0.75,
  point_outline_color = "darkgray")
}
\arguments{
\item{res}{\link{DESeqResults}.}

\item{ylim}{Y-axis maximum (single integer).}

\item{bcb}{\link{bcbioRnaDataSet}.}

\item{lfc}{Log fold change ratio (base 2) cutoff for coloring.}

\item{text_labels}{Number of text labels to plot.}

\item{merge}{merge all plots into one.}

\item{direction}{Plot \code{up}, \code{down}, or \code{both} directions.}

\item{title}{Title for the figure.}

\item{shade_color}{Shading color for bounding box.}

\item{shade_alpha}{Shading transparency alpha.}

\item{point_color}{Point color.}

\item{point_alpha}{Point transparency alpha.}

\item{point_outline_color}{Point outline color.}
}
\value{
MA plot.

Volcano plot.
}
\description{
Wrapper for \code{\link[DESeq2:plotMA]{DESeq2::plotMA()}} that generates a title
automatically.

Volcano plot.
}
\author{
Michael Steinbaugh
}
