% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{defunct}
\alias{defunct}
\alias{plotCountDensity}
\alias{resultsTables}
\title{Defunct functions}
\usage{
plotCountDensity(...)

resultsTables(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
\code{.Defunct}.
}
\description{
Defunct functions
}
\keyword{internal}
