% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{defunct}
\alias{defunct}
\alias{plotGeneHeatmap}
\alias{txi}
\alias{plotCountDensity}
\alias{plotPCACovariates}
\alias{resultsTables}
\title{Defunct Functions}
\usage{
plotGeneHeatmap(...)

txi(...)

plotCountDensity(...)

plotPCACovariates(...)

resultsTables(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
\code{.Defunct()}.
}
\description{
Defunct Functions
}
\keyword{internal}
