% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{methods}
\name{deprecated}
\alias{deprecated}
\alias{download}
\alias{plotGeneDetectionSaturation}
\alias{plotDispersion}
\alias{loadRNASeqRun}
\alias{plotGeneHeatmap}
\alias{txi}
\alias{bcbio,bcbioRNASeq-method}
\alias{bcbio<-,bcbioRNASeq-method}
\alias{plot53Bias}
\title{Defunct or Deprecated Functions}
\usage{
download(...)

plotGeneDetectionSaturation(...)

plotDispersion(...)

loadRNASeqRun(...)

plotGeneHeatmap(...)

txi(...)

\S4method{bcbio}{bcbioRNASeq}(object, ...)

\S4method{bcbio}{bcbioRNASeq}(object, ...) <- value

plot53Bias(...)
}
\arguments{
\item{...}{Additional arguments.}

\item{object}{Object.}

\item{value}{Value to assign.}
}
\value{
No value.
}
\description{
Defunct or Deprecated Functions
}
\keyword{internal}
