% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{correlation_heatmap}
\alias{correlation_heatmap}
\alias{import_file}
\alias{import_metadata}
\alias{import_summary}
\alias{load_bcbio_run}
\alias{load_bcbio_run_from_yaml}
\alias{load_run_S4}
\alias{pool_counts}
\alias{pool_dds}
\alias{plot_deseq_pca}
\alias{pool_lane_split_counts}
\alias{pool_lane_split_dds}
\alias{res_tables}
\alias{tmm_normalize}
\alias{aggregate_replicates}
\alias{alpha_summary}
\alias{downloads}
\alias{load_run}
\alias{melt_log10}
\alias{metadata_table}
\alias{plot_5_3_bias}
\alias{plot_correlation_heatmap}
\alias{plot_count_density}
\alias{plot_counts_per_gene}
\alias{plot_deg_heatmap}
\alias{plot_dispersion}
\alias{plot_exonic_mapping_rate}
\alias{plot_gender_markers}
\alias{plot_gene}
\alias{plot_gene_detection_saturation}
\alias{plot_gene_heatmap}
\alias{plot_genes_detected}
\alias{plot_intronic_mapping_rate}
\alias{plot_mirna_counts}
\alias{plot_mapped_reads}
\alias{plot_mapping_rate}
\alias{plot_ma}
\alias{plot_mean_sd}
\alias{plot_pca_covariates}
\alias{plot_pattern}
\alias{plot_rrna_mapping_rate}
\alias{plot_pca}
\alias{plot_size_distribution}
\alias{plot_srna_clusters}
\alias{plot_total_reads}
\alias{plot_volcano}
\alias{results_tables}
\alias{sample_dirs}
\alias{top_tables}
\title{Deprecated Functions}
\usage{
correlation_heatmap()

import_file()

import_metadata()

import_summary()

load_bcbio_run()

load_bcbio_run_from_yaml()

load_run_S4()

pool_counts()

pool_dds()

plot_deseq_pca()

pool_lane_split_counts()

pool_lane_split_dds()

res_tables()

tmm_normalize()

aggregate_replicates()

alpha_summary()

downloads()

load_run()

melt_log10()

metadata_table()

plot_5_3_bias()

plot_correlation_heatmap()

plot_count_density()

plot_counts_per_gene()

plot_deg_heatmap()

plot_dispersion()

plot_exonic_mapping_rate()

plot_gender_markers()

plot_gene()

plot_gene_detection_saturation()

plot_gene_heatmap()

plot_genes_detected()

plot_intronic_mapping_rate()

plot_mirna_counts()

plot_mapped_reads()

plot_mapping_rate()

plot_ma()

plot_mean_sd()

plot_pca_covariates()

plot_pattern()

plot_rrna_mapping_rate()

plot_pca()

plot_size_distribution()

plot_srna_clusters()

plot_total_reads()

plot_volcano()

results_tables()

sample_dirs()

top_tables()
}
\arguments{
\item{...}{Additional parameters.}
}
\description{
Deprecated Functions
}
\keyword{internal}
