% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{methods}
\name{deprecated}
\alias{deprecated}
\alias{download}
\alias{plotGeneDetectionSaturation}
\alias{plotDispersion}
\alias{loadRNASeqRun}
\alias{plotGeneHeatmap}
\alias{txi}
\alias{meltLog10}
\alias{plot53Bias}
\alias{bcbio,bcbioRNASeq-method}
\alias{bcbio<-,bcbioRNASeq-method}
\alias{design,bcbioRNASeq-method}
\alias{design<-,bcbioRNASeq,formula-method}
\title{Defunct or Deprecated Functions}
\usage{
download(...)

plotGeneDetectionSaturation(...)

plotDispersion(...)

loadRNASeqRun(...)

plotGeneHeatmap(...)

txi(...)

meltLog10(object, ...)

plot53Bias(...)

\S4method{bcbio}{bcbioRNASeq}(object, ...)

\S4method{bcbio}{bcbioRNASeq}(object, ...) <- value

\S4method{design}{bcbioRNASeq}(object, ...)

\S4method{design}{bcbioRNASeq,formula}(object, ...) <- value
}
\arguments{
\item{...}{Additional arguments.}

\item{object}{Object.}

\item{value}{Value to assign.}
}
\value{
No value.
}
\description{
Defunct or Deprecated Functions
}
\keyword{internal}
