% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{download}
\alias{plotGeneDetectionSaturation}
\alias{plotDispersion}
\alias{loadRNASeqRun}
\title{Deprecated Functions}
\usage{
download()

plotGeneDetectionSaturation(...)

plotDispersion(...)

loadRNASeqRun(...)
}
\arguments{
\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
No value.
}
\description{
Deprecated Functions
}
\keyword{internal}
