% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{download}
\alias{plotGeneDetectionSaturation}
\alias{plotDispersion}
\alias{loadRNASeqRun}
\alias{plotGeneHeatmap}
\alias{txi}
\alias{annotable}
\title{Defunct or Deprecated Functions}
\usage{
download()

plotGeneDetectionSaturation(...)

plotDispersion(...)

loadRNASeqRun()

plotGeneHeatmap(...)

txi(object)

annotable(...)
}
\arguments{
\item{...}{Additional arguments.}

\item{object}{Object.}
}
\value{
No value.
}
\description{
Defunct or Deprecated Functions
}
\keyword{internal}
