% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{correlation_heatmap}
\alias{correlation_heatmap}
\alias{import_file}
\alias{import_metadata}
\alias{import_summary}
\alias{load_bcbio_run}
\alias{load_bcbio_run_from_yaml}
\alias{load_run_S4}
\alias{pool_counts}
\alias{pool_dds}
\alias{plot_deseq_pca}
\alias{pool_lane_split_counts}
\alias{pool_lane_split_dds}
\alias{res_tables}
\alias{tmm_normalize}
\title{Deprecated functions}
\usage{
correlation_heatmap(...)

import_file()

import_metadata()

import_summary()

load_bcbio_run(...)

load_bcbio_run_from_yaml(...)

load_run_S4(...)

pool_counts(...)

pool_dds(...)

plot_deseq_pca(...)

pool_lane_split_counts(...)

pool_lane_split_dds(...)

res_tables(...)

tmm_normalize(...)
}
\arguments{
\item{...}{Additional parameters.}
}
\description{
Deprecated functions
}
\keyword{internal}
