% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{correlation_heatmap}
\alias{import_file}
\title{Deprecated functions}
\usage{
correlation_heatmap(...)

import_file()
}
\arguments{
\item{...}{Passthrough parameters.}
}
\description{
Deprecated functions
}
\keyword{internal}
