% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{download}
\alias{plotGeneDetectionSaturation}
\alias{plotDispersion}
\alias{loadRNASeqRun}
\alias{plotGeneHeatmap}
\alias{txi}
\alias{plot53Bias}
\title{Defunct or Deprecated Functions}
\usage{
download(...)

plotGeneDetectionSaturation(...)

plotDispersion(...)

loadRNASeqRun(...)

plotGeneHeatmap(...)

txi(...)

plot53Bias(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
No value.
}
\description{
Defunct or Deprecated Functions
}
\keyword{internal}
