% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{download}
\alias{plotGeneDetectionSaturation}
\alias{plotDispersion}
\alias{loadRNASeqRun}
\alias{plotGeneHeatmap}
\alias{txi}
\title{Deprecated Functions}
\usage{
download()

plotGeneDetectionSaturation(...)

plotDispersion(...)

loadRNASeqRun()

plotGeneHeatmap(...)

txi(object)
}
\arguments{
\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{object}{Object.}
}
\value{
No value.
}
\description{
Deprecated Functions
}
\keyword{internal}
