% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{loadRNASeq}
\alias{prepareRNASeqTemplate}
\title{Deprecated functions}
\usage{
loadRNASeq(...)

prepareRNASeqTemplate(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
\code{.Deprecated}.
}
\description{
Deprecated functions
}
\keyword{internal}
