% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{methods}
\name{deprecated}
\alias{deprecated}
\alias{meltLog10}
\alias{plot53Bias}
\alias{design,bcbioRNASeq-method}
\alias{design<-,bcbioRNASeq,formula-method}
\alias{loadRNASeq}
\alias{plotMeanAverage}
\title{Defunct or Deprecated Functions}
\usage{
meltLog10(object, ...)

plot53Bias(...)

\S4method{design}{bcbioRNASeq}(object, ...)

\S4method{design}{bcbioRNASeq,formula}(object, ...) <- value

loadRNASeq(...)

plotMeanAverage(...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{Value to assign.}
}
\value{
No value.
}
\description{
Defunct or Deprecated Functions
}
\keyword{internal}
