% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{loadRNASeq}
\alias{prepareRNASeqTemplate}
\alias{plotCountsPerGene}
\alias{plotGenesDetected}
\title{Deprecated functions}
\usage{
loadRNASeq(...)

prepareRNASeqTemplate(...)

plotCountsPerGene(object, title = "Counts per gene", ...)

plotGenesDetected(object, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{object}{Object.}
}
\value{
\code{.Deprecated}.
}
\description{
Deprecated functions
}
\keyword{internal}
