% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{methods}
\name{deprecated}
\alias{deprecated}
\alias{plotGeneDetectionSaturation}
\alias{plotDispersion}
\alias{loadRNASeqRun}
\alias{meltLog10}
\alias{design,bcbioRNASeq-method}
\alias{design<-,bcbioRNASeq,formula-method}
\alias{loadRNASeq}
\alias{prepareRNASeqTemplate}
\title{Deprecated functions}
\usage{
plotGeneDetectionSaturation(...)

plotDispersion(...)

loadRNASeqRun(...)

meltLog10(object, ...)

\S4method{design}{bcbioRNASeq}(object, ...)

\S4method{design}{bcbioRNASeq,formula}(object, ...) <- value

loadRNASeq(...)

prepareRNASeqTemplate(...)
}
\arguments{
\item{...}{Additional arguments.}

\item{object}{Object.}
}
\value{
\code{.Deprecated}.
}
\description{
Deprecated functions
}
\keyword{internal}
