% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/deprecated-f1000v2.R
\name{plotDegHeatmap}
\alias{plotDegHeatmap}
\alias{topTables}
\alias{deprecated}
\alias{plotDegHeatmap,missing-method}
\alias{plotGenesDetected}
\alias{plotMeanAverage}
\alias{prepareRNASeqTemplate}
\alias{topTables,DFrameList-method}
\alias{writeCounts}
\title{Deprecated functions}
\usage{
plotDegHeatmap(object, ...)

topTables(object, ...)

\S4method{plotDegHeatmap}{missing}(object, results, counts, ...)

plotGenesDetected(object, ...)

plotMeanAverage(...)

prepareRNASeqTemplate(...)

\S4method{topTables}{DFrameList}(object, ...)

writeCounts(..., dir = getOption(x = "acid.export.dir", default = getwd()))
}
\arguments{
\item{object}{Object.}

\item{counts}{\code{matrix}.
Count matrix. Normalized counts are recommended.}

\item{...}{Additional arguments.}

\item{dir}{\code{character(1)}.
Directory path.}

\item{`results`}{\code{DESeqResults.}}

\item{`counts`}{\code{DESeqTransform}.}
}
\value{
\code{.Deprecated}.
}
\description{
Deprecated functions
}
\examples{
a <- function(...) {
    .Deprecated("b")
    b(...)
}

b <- function(x) {
    x + 1L
}

withCallingHandlers(
    expr = a(1L),
    warning = function(w) {
        print(w)
        invokeRestart("muffleWarning")
    }
)
}
\seealso{
\link[base:Deprecated]{Deprecated}.
}
\keyword{internal}
