% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_lane_pool.R
\name{deseq_lane_pool}
\alias{deseq_lane_pool}
\title{Run DESeq2 with pooled technical replicates}
\usage{
deseq_lane_pool(bcbio, dds)
}
\arguments{
\item{bcbio}{bcbio list object}

\item{dds}{DESeqDataSet object}
}
\value{
DESeqDataSet object using pooled technical replicates
}
\description{
Run DESeq2 with pooled technical replicates
}
\author{
Michael Steinbaugh
}
