% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DESeqAnalysis-class}
\alias{DESeqAnalysis-class}
\title{\code{DESeqAnalysis} Class}
\description{
Class containing all elements generated during differential expression
analysis with DESeq2. This class is essentially a \code{list} with validity checks
to ensure \code{DESeqTransform} and \code{DESeqResults} correspond to the
\code{DESeqDataSet}.
}
\section{Slots}{

\describe{
\item{\code{data}}{\code{DESeqDataSet}.}

\item{\code{transform}}{\code{DESeqTransform}.}

\item{\code{results}}{\code{list}. One or more unshrunken \code{DESeqResults}.}

\item{\code{lfcShrink}}{\code{list}. One or more shrunken \code{DESeqResults}.}
}}

\section{DESeqDataSet}{


We recommend generating the \code{DESeqDataSet} by coercion from \code{bcbioRNASeq}
object using \code{as(dds, "bcbioRNASeq")}. Don't use the \code{\link[DESeq2:DESeqDataSet]{DESeq2::DESeqDataSet()}}
or \code{\link[DESeq2:DESeqDataSetFromMatrix]{DESeq2::DESeqDataSetFromMatrix()}} constructors to generate the
\code{DESeqDataSet} object.
}

\section{DESeqResults}{


Don't modify any of the \code{DESeqResults} objects manually. This includes
rearranging the rows or dropping genes without adjusted P values. We'll take
care of this automatically in supported methods.
}

\seealso{
Other S4 Object: \code{\link{DESeqAnalysis}},
  \code{\link{DESeqResultsTables-class}},
  \code{\link{DESeqResultsTables}},
  \code{\link{bcbioRNASeq-class}}, \code{\link{coerce}},
  \code{\link{extract}}, \code{\link{show}},
  \code{\link{updateObject}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Object}
