% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DESeqAnalysis-class}
\alias{DESeqAnalysis-class}
\alias{DESeqAnalysis}
\title{DESeqAnalysis Class}
\value{
\code{DESeqAnalysis}.
}
\description{
Class containing all elements generated during differential expression
analysis with DESeq2.
}
\section{Slots}{

\describe{
\item{\code{data}}{\code{DESeqDataSet}.}

\item{\code{transform}}{\code{DESeqTransform}.}

\item{\code{results}}{\code{list}. One or more \code{DESeqResults}. We're using a list
here to support multiple pairwise contrasts.}

\item{\code{lfcShrink}}{\code{list}. One or more \code{DESeqResults} returned from
\code{\link[DESeq2:lfcShrink]{DESeq2::lfcShrink()}}.}
}}

\section{DESeqDataSet}{


We recommend generating the \code{DESeqDataSet} by coercion from \code{bcbioRNASeq}
object using \code{as(dds, "bcbioRNASeq")}. Don't use the \code{\link[DESeq2:DESeqDataSet]{DESeq2::DESeqDataSet()}}
or \code{\link[DESeq2:DESeqDataSetFromMatrix]{DESeq2::DESeqDataSetFromMatrix()}} constructors to generate the
\code{DESeqDataSet} object.
}

\section{DESeqResults}{


Don't modify any of the \code{DESeqResults} objects manually. This includes
rearranging the rows or dropping genes without adjusted P values. We'll take
care of this automatically in supported methods.
}

\examples{
library(DESeq2)
dds <- as(bcb_small, "DESeqDataSet")
design(dds) <- ~ treatment
dds <- DESeq(dds)
class(dds)
vst <- varianceStabilizingTransformation(dds)
class(vst)
resultsNames(dds)
res <- results(dds, name = resultsNames(dds)[[2L]])
class(res)
x <- DESeqAnalysis(
    data = dds,
    transform = vst,
    results = list(res),
    lfcShrink = list(lfcShrink(dds = dds, coef = 2L))
)
class(x)
slotNames(x)
}
\seealso{
Other S4 Object: \code{\link{DESeqResultsTables-class}},
  \code{\link{bcbioRNASeq-class}},
  \code{\link{bcbioRNASeq}}, \code{\link{coerce}},
  \code{\link{extract}}, \code{\link{show}},
  \code{\link{updateObject}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Object}
