% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/DESeqAnalysis-constructor.R
\docType{class}
\name{DESeqAnalysis-class}
\alias{DESeqAnalysis}
\title{DESeq2 Differential Expression Analysis}
\usage{
DESeqAnalysis(DESeqDataSet, DESeqTransform, DESeqResults, lfcShrink)
}
\arguments{
\item{DESeqDataSet}{\code{DESeqDataSet}.}

\item{DESeqTransform}{\code{DESeqTransform}.}

\item{DESeqResults}{\code{list}. One or more \code{DESeqResults}. We're using a list
here to support multiple pairwise contrasts.}

\item{lfcShrink}{\code{list}. One or more \code{DESeqResults} returned from
\code{\link[DESeq2:lfcShrink]{DESeq2::lfcShrink()}}.}
}
\value{
\code{DESeqAnalysis}.
}
\description{
Class containing all elements generated during differential expression
analysis with DESeq2.
}
\section{DESeqDataSet}{


We recommend generating the \code{DESeqDataSet} by coercion from \code{bcbioRNASeq}
object using \code{as(dds, "bcbioRNASeq")}. Don't use the \code{\link[DESeq2:DESeqDataSet]{DESeq2::DESeqDataSet()}}
or \code{\link[DESeq2:DESeqDataSetFromMatrix]{DESeq2::DESeqDataSetFromMatrix()}} constructors to generate the
\code{DESeqDataSet} object.
}

\section{DESeqResults}{


Don't modify any of the \code{DESeqResults} objects manually. This includes
rearranging the rows or dropping genes without adjusted P values. We'll take
care of this automatically in supported methods.
}

\examples{
library(DESeq2)
dds <- dds_small
class(dds)
vst <- varianceStabilizingTransformation(dds)
class(vst)
resultsNames(dds)
res <- results(dds, name = resultsNames(dds)[[2L]])
class(res)
x <- DESeqAnalysis(
    DESeqDataSet = dds_small,
    DESeqTransform = vst_small,
    DESeqResults = list(
        res_small
    ),
    lfcShrink = list(
        lfcShrink(dds = dds_small, coef = 2L)
    )
)
class(x)
slotNames(x)
}
\seealso{
Other S4 Object: \code{\link{bcbioRNASeq}},
  \code{\link{coerce}}, \code{\link{extract}},
  \code{\link{show}}, \code{\link{updateObject}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Object}
