% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/DESeqAnalysis-constructor.R
\docType{class}
\name{DESeqAnalysis-class}
\alias{DESeqAnalysis}
\title{DESeq2 Differential Expression Analysis}
\usage{
DESeqAnalysis(data, transform, results)
}
\arguments{
\item{data}{\code{DESeqDataSet}.}

\item{transform}{\code{DESeqTransform}.}

\item{results}{\code{list}. One or more \code{DESeqResults}. We're using a list here
to support multiple pairwise contrasts.}
}
\description{
Class containing all elements generated during differential expression
analysis with DESeq2.
}
\examples{
x <- DESeqAnalysis(
    data = dds_small,
    transform = vst_small,
    results = list(res_small)
)
class(x)
slotNames(x)
}
\seealso{
Other S4 Object: \code{\link{bcbioRNASeq}},
  \code{\link{coerce}}, \code{\link{extract}},
  \code{\link{show}}, \code{\link{updateObject}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Object}
