% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeqAnalysis.R
\name{DESeqAnalysis}
\alias{DESeqAnalysis}
\title{DESeq2 Analysis Container}
\usage{
DESeqAnalysis(data, transform, results, lfcShrink)
}
\arguments{
\item{data}{\code{DESeqDataSet}.}

\item{transform}{\code{DESeqTransform}.}

\item{results}{\code{list}. One or more unshrunken \code{DESeqResults}. Assign the
\code{\link[DESeq2:results]{DESeq2::results()}} return here.}

\item{lfcShrink}{\code{list}. One or more shrunken \code{DESeqResults}. Assign the
\code{\link[DESeq2:lfcShrink]{DESeq2::lfcShrink()}} return here.}
}
\value{
\code{DESeqAnalysis}.
}
\description{
Class containing all elements generated during differential expression
analysis with DESeq2. This class is essentially a \code{list} with validity checks
to ensure \code{DESeqTransform} and \code{DESeqResults} correspond to the
\code{DESeqDataSet}.
}
\section{DESeqDataSet}{


We recommend generating the \code{DESeqDataSet} by coercion from \code{bcbioRNASeq}
object using \code{as(dds, "bcbioRNASeq")}. Don't use the \code{\link[DESeq2:DESeqDataSet]{DESeq2::DESeqDataSet()}}
or \code{\link[DESeq2:DESeqDataSetFromMatrix]{DESeq2::DESeqDataSetFromMatrix()}} constructors to generate the
\code{DESeqDataSet} object.
}

\section{DESeqResults}{


Don't modify any of the \code{DESeqResults} objects manually. This includes
rearranging the rows or dropping genes without adjusted P values. We'll take
care of this automatically in supported methods.
}

\examples{
library(DESeq2)
data(bcb_small)

dds <- as(bcb_small, "DESeqDataSet")
design(dds) <- ~ treatment
dds <- DESeq(dds)
class(dds)

vst <- varianceStabilizingTransformation(dds)
class(vst)

resultsNames(dds)
res <- results(dds, name = resultsNames(dds)[[2L]])
class(res)

x <- DESeqAnalysis(
    data = dds,
    transform = vst,
    results = list(res),
    lfcShrink = list(lfcShrink(dds = dds, coef = 2L))
)
print(x)
}
\seealso{
\code{\link[=DESeqAnalysis]{DESeqAnalysis()}}.

Other S4 classes: \code{\link{DESeqResultsTables-class}},
  \code{\link{bcbioRNASeq-class}}
}
\author{
Michael Steinbaugh
}
