% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{DESeqAnalysis}
\alias{DESeqAnalysis}
\title{\code{DESeqAnalysis} Generator}
\usage{
DESeqAnalysis(data, transform, results, lfcShrink)
}
\arguments{
\item{data}{\code{DESeqDataSet}.}

\item{transform}{\code{DESeqTransform}.}

\item{results}{\code{list}. One or more unshrunken \code{DESeqResults}. Assign the
\code{\link[DESeq2:results]{DESeq2::results()}} return here.}

\item{lfcShrink}{\code{list}. One or more shrunken \code{DESeqResults}. Assign the
\code{\link[DESeq2:lfcShrink]{DESeq2::lfcShrink()}} return here.}
}
\value{
\code{DESeqAnalysis}.
}
\description{
\code{DESeqAnalysis} Generator
}
\examples{
library(DESeq2)
dds <- as(bcb_small, "DESeqDataSet")
design(dds) <- ~ treatment
dds <- DESeq(dds)
class(dds)
vst <- varianceStabilizingTransformation(dds)
class(vst)
resultsNames(dds)
res <- results(dds, name = resultsNames(dds)[[2L]])
class(res)
x <- DESeqAnalysis(
    data = dds,
    transform = vst,
    results = list(res),
    lfcShrink = list(lfcShrink(dds = dds, coef = 2L))
)
print(x)
}
\seealso{
Other S4 Object: \code{\link{DESeqAnalysis-class}},
  \code{\link{DESeqResultsTables-class}},
  \code{\link{DESeqResultsTables}},
  \code{\link{bcbioRNASeq-class}},
  \code{\link{bcbioRNASeq}}, \code{\link{coerce}},
  \code{\link{extract}}, \code{\link{show}},
  \code{\link{updateObject}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Object}
