% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DESeqResultsTables-class}
\alias{DESeqResultsTables-class}
\title{\code{DESeqResultsTables} Class}
\value{
\code{DESeqResultsTables}.
}
\description{
\code{DESeqResultsTables} Class
}
\section{Slots}{

\describe{
\item{\code{all}}{\code{DESeqResults}. Original unmodified \code{DESeqResults}. Should contain
all genes, including those with \code{NA} adjusted \emph{P} values.}

\item{\code{deg}}{\code{DataFrame}. Subset containing genes that pass adjusted \emph{P} value
and log2 fold change cutoffs.}

\item{\code{degUp}}{\code{DataFrame}. Directional subset containing only upregulated
genes.}

\item{\code{degDown}}{\code{DataFrame}. Directional subset containing only downregulated
genes.}
}}

\examples{
object <- deseq_small@results[[1L]]
x <- resultsTables(object)
print(x)
}
\seealso{
Other S4 Object: \code{\link{DESeqAnalysis-class}},
  \code{\link{bcbioRNASeq-class}},
  \code{\link{bcbioRNASeq}}, \code{\link{coerce}},
  \code{\link{extract}}, \code{\link{show}},
  \code{\link{updateObject}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Object}
