% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{DESeqResultsTables}
\alias{DESeqResultsTables}
\title{\code{DESeqResultsTables} Generator}
\usage{
DESeqResultsTables(object)
}
\arguments{
\item{object}{\code{DESeqResults}.}
}
\value{
\code{DESeqResultsTables}.
}
\description{
\code{DESeqResultsTables} Generator
}
\examples{
object <- deseq_small@results[[1L]]
class(object)
x <- DESeqResultsTables(object)
print(x)
}
\seealso{
Other S4 Object: \code{\link{DESeqAnalysis-class}},
  \code{\link{DESeqAnalysis}},
  \code{\link{DESeqResultsTables-class}},
  \code{\link{bcbioRNASeq-class}},
  \code{\link{bcbioRNASeq}}, \code{\link{coerce}},
  \code{\link{extract}}, \code{\link{show}},
  \code{\link{updateObject}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Object}
