% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/DESeqResultsTables-methods.R
\docType{methods}
\name{DESeqResultsTables}
\alias{DESeqResultsTables}
\alias{DESeqResultsTables,DESeqResults-method}
\alias{DESeqResultsTables,DESeqAnalysis-method}
\title{\code{DESeqResultsTables} Generator}
\usage{
DESeqResultsTables(object, ...)

\S4method{DESeqResultsTables}{DESeqResults}(object)

\S4method{DESeqResultsTables}{DESeqAnalysis}(object, results = 1L,
  lfcShrink = TRUE, rowData = TRUE, counts = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{results}{\code{scalar}. Position or name of \code{DESeqResults}.}

\item{lfcShrink}{\code{boolean}. Use shrunken log2 fold change (LFC) values.}

\item{rowData}{\code{boolean}. Include gene annotations.}

\item{counts}{\code{boolean}. Include DESeq2 normalized counts.}
}
\value{
\code{DESeqResultsTables}.
}
\description{
\code{DESeqResultsTables} Generator
}
\note{
Log fold change cutoff threshold ("\code{lfcThreshold}") does not apply to
statistical hypothesis testing, only gene filtering in the results tables.
See \code{\link[DESeq2:results]{DESeq2::results()}} for additional information about using
\code{lfcThreshold} and \code{altHypothesis} to set an alternative hypothesis based
on expected fold changes.
}
\examples{
data(deseq_small)

# DESeqAnalysis ====
# This is the recommended default method.
x <- DESeqResultsTables(deseq_small)
print(x)

# DESeqResults ====
res <- as(deseq_small, "DESeqResults")
x <- DESeqResultsTables(res)
print(x)
}
\seealso{
\itemize{
\item \code{\link[=markdown]{markdown()}}.
\item \code{\link[=write]{write()}}.
}

Other S4 Generators: \code{\link{DESeqAnalysis}},
  \code{\link{bcbioRNASeq}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Generators}
