% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/DESeqResultsTables-methods.R
\docType{methods}
\name{DESeqResultsTables}
\alias{DESeqResultsTables}
\alias{DESeqResultsTables,DESeqResults-method}
\alias{DESeqResultsTables,DESeqAnalysis-method}
\title{DESeq2 Differential Expression Results Tables}
\usage{
DESeqResultsTables(object, ...)

\S4method{DESeqResultsTables}{DESeqResults}(object)

\S4method{DESeqResultsTables}{DESeqAnalysis}(object, results = 1L,
  lfcShrink = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{results}{\code{scalar}. Position or name of \code{DESeqResults}.}

\item{lfcShrink}{\code{boolean}. Use shrunken log2 fold change (LFC) values.}
}
\value{
\code{DESeqResultsTables}.
}
\description{
\code{DESeqResults} object with \code{DataFrame} subsets and file path information.
}
\section{Obtaining results from DESeq2}{


It is recommended to specify the \code{contrast} argument as \code{character}:
\enumerate{
\item Design matrix factor of interest.
\item Numerator for LFC (expt).
\item Denominator for LFC (control).
}

For example, to get the relative change in gene expression between mutant
and wildtype samples, here's how to set the contrast vector:\preformatted{c(
    factor = "genotype",
    numerator = "mutant",
    denominator = "wildtype"
)
}
}

\section{Log fold change threshold cutoffs}{


Refer to \code{\link[DESeq2:results]{DESeq2::results()}} for additional information about using
\code{lfcThreshold} and \code{altHypothesis} to set an alternative hypothesis based on
expected fold changes. In addition, the "Hypothesis tests with thresholds on
effect size" section in the DESeq2 paper provides additional explanation.

Don't apply \emph{post hoc} LFC threshold filtering to obtain results tables, as
this will destroy the meaning of the adjusted \emph{P} values computed. If you are
expecting a biological effect size past a particular threshold, set
\code{lfcThreshold}, but be conservative.

This \href{https://support.bioconductor.org/p/101504/}{thread} on the Bioconductor forums explains how \code{\link[DESeq2:results]{DESeq2::results()}}
should be called with regard to LFC cutoffs in nice detail. In particular,
refer to Mike Love:

"A common procedure is to disregard genes whose estimated LFC \emph{beta ir} is
below some threshold, \emph{beta ir ≤ theta}. However, this approach loses the
benefit of an easily interpretable FDR, as the reported \emph{P} value and
adjusted \emph{P} value still correspond to the test of \emph{zero} LFC. It is
therefore desirable to include the threshold in the statistical testing
procedure directly, i.e., not to filter post hoc on a reported fold-change
\emph{estimate}, but rather to evaluate statistically directly whether there is
sufficient evidence that the LFC is above the chosen threshold."
}

\examples{
data(deseq)

## DESeqAnalysis ====
## This is the recommended default method.
x <- DESeqResultsTables(deseq)
print(x)

## DESeqResults ====
res <- as(deseq, "DESeqResults")
x <- DESeqResultsTables(res)
print(x)
}
\seealso{
\itemize{
\item \code{\link[DESeq2:results]{DESeq2::results()}}.
\item \code{\link[=markdown]{markdown()}}, \code{\link[=write]{write()}}.
}
}
\author{
Michael Steinbaugh
}
