% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-design.R
\docType{methods}
\name{design}
\alias{design}
\alias{design,bcbioRNASeq-method}
\alias{design<-,bcbioRNASeq,formula-method}
\title{Design Formula Accessor}
\usage{
\S4method{design}{bcbioRNASeq}(object)

\S4method{design}{bcbioRNASeq,formula}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{a \code{formula} used for estimating dispersion
and fitting Negative Binomial GLMs}
}
\value{
\link{formula}.
}
\description{
Design Formula Accessor
}
\examples{
load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioRNASeq"))

# bcbioRNASeq
design(bcb) \%>\% print(showEnv = FALSE)

# Assignment support
design(bcb) <- formula(~group)
design(bcb) \%>\% print(showEnv = FALSE)
}
\seealso{
\itemize{
\item \link[BiocGenerics:design]{BiocGenerics::design}.
\item \link[DESeq2:design]{DESeq2::design}.
}
}
\author{
Michael Steinbaugh
}
