% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{detect_organism}
\alias{detect_organism}
\title{Detect the organism from the genome build name}
\usage{
detect_organism(genome_build)
}
\arguments{
\item{genome_build}{Genome build}
}
\value{
Organism string
}
\description{
Detect the organism from the genome build name
}
\author{
Michael Steinbaugh
}
\keyword{internal}
