% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-detect.R
\name{.detect_organism}
\alias{.detect_organism}
\title{Detect the Organism from the Genome Build Name}
\usage{
.detect_organism(genome_build)
}
\arguments{
\item{genome_build}{Genome build.}
}
\value{
Organism string.
}
\description{
Detect the Organism from the Genome Build Name
}
\author{
Michael Steinbaugh
}
\keyword{internal}
