% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\title{Download Dependency File}
\usage{
download(file = NULL)
}
\arguments{
\item{file}{\emph{Optional}. File name. If \code{NULL} (default), download the default
dependency files for a new experiment.}
}
\description{
If the required dependency file isn't present, download latest version from
the \href{http://bioinformatics.sph.harvard.edu/}{HBC website}.
}
\author{
Michael Steinbaugh
}
