% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{downloads}
\alias{downloads}
\title{File downloads}
\usage{
downloads(file = NULL)
}
\arguments{
\item{file}{File name. If \code{NULL} (default), download the default dependency
files for a consult.}
}
\description{
If file isn't present, download latest version from the
\href{http://bioinformatics.sph.harvard.edu/}{HBC website}.
}
\details{
File download utility functions for RMarkdown knit reports.
}
\author{
Michael Steinbaugh
}
