% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{hbc_download}
\alias{hbc_download}
\alias{download_rmarkdown_files}
\title{File downloads}
\usage{
hbc_download(file)

download_rmarkdown_files()
}
\arguments{
\item{file}{File name.}
}
\description{
If file isn't present, download latest version from the
\href{http://bioinformatics.sph.harvard.edu/}{HBC website}.
}
\details{
File download utility functions for RMarkdown knit reports.
}
\author{
Michael Steinbaugh
}
\keyword{internal}
