% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{downloads}
\alias{downloads}
\title{File Downloads}
\usage{
downloads(file = NULL)
}
\arguments{
\item{file}{\emph{Optional}. File name. If \code{NULL} (default), download the default
dependency files for a new experiment.}
}
\value{
No return.
}
\description{
If the required dependency file isn't present, download latest version from
the \href{http://bioinformatics.sph.harvard.edu/}{HBC website}.
}
\author{
Michael Steinbaugh
}
