% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl.R
\name{gene_level_annotations}
\alias{gene_level_annotations}
\alias{tx2gene}
\title{Ensembl annotations}
\usage{
gene_level_annotations(run)

tx2gene(run)
}
\arguments{
\item{run}{bcbio-nextgen run.}
}
\value{
Tibble arranged by \code{ensembl_gene_id}.

Tibble with \code{ensembl_transcript_id} and \code{ensembl_gene_id}.
}
\description{
Gene-level annotations with \code{description} and \code{gene_biotype},
and \code{broad_class}.

Transcript-to-gene annotations, to be used with \code{\link[=tximport]{tximport()}}.
}
\details{
Transcript-level annotations are downloaded from
\href{http://www.ensembl.org/}{Ensembl} and saved as a tibbled grouped by
\code{ensembl_gene_id} with nested \code{ensembl_transcript_id} in the bcbio run
object. We also define broad classes, which are used in quality control
analysis.
}
\author{
Michael Steinbaugh
}
