% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_annotations.R
\name{ensembl_annotations}
\alias{ensembl_annotations}
\title{Get gene annotations from Ensembl}
\usage{
ensembl_annotations(organism)
}
\arguments{
\item{organism}{Organism identifier}
}
\value{
Data frame
}
\description{
Get gene annotations from Ensembl
}
\examples{
\dontrun{
ensembl_annotations("mmusculus")
}
}
\author{
Michael Steinbaugh
}
\keyword{internal}
