% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_annotations.R
\name{ensembl_annotations}
\alias{ensembl_annotations}
\title{\href{http://www.ensembl.org/}{Ensembl} annotations}
\usage{
ensembl_annotations(run, attributes = NULL, filters = "ensembl_gene_id",
  values = NULL)
}
\arguments{
\item{run}{bcbio-nextgen run.}

\item{attributes}{Ensembl attributes.}

\item{filters}{biomaRt filters.}

\item{values}{Ensembl gene identifier values. Optional but will run faster if
specified.}
}
\value{
Data frame.
}
\description{
Download the most recent gene annotations from Ensembl.
}
\examples{
\dontrun{
ensembl_annotations(run, values = "ENSMUSG00000000001")
}

}
\seealso{
\code{\link[biomaRt]{listAttributes}}, \code{\link[biomaRt]{listFilters}}.
}
\author{
Michael Steinbaugh
}
