% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_annotations.R
\name{ensembl_annotations}
\alias{ensembl_annotations}
\title{Ensembl annotations}
\usage{
ensembl_annotations(run)
}
\arguments{
\item{run}{bcbio-nextgen run.}
}
\value{
Data frame.
}
\description{
Download transcript-to-gene annotations from
\href{http://www.ensembl.org/}{Ensembl}. This function also defines broad classes,
which are used in quality control analysis.
}
\author{
Rory Kirchner

Michael Steinbaugh
}
