% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_annotations.R
\name{ensembl_annotations}
\alias{ensembl_annotations}
\title{Get gene annotations from Ensembl}
\usage{
ensembl_annotations(bcbio, attributes = NULL, filters = "ensembl_gene_id",
  values = NULL)
}
\arguments{
\item{bcbio}{bcbio list object}

\item{attributes}{Ensembl attributes. See \code{biomaRt::listAttributes()}.}

\item{filters}{biomaRt filters. See \code{biomaRt::listFilters()}.}

\item{values}{Ensembl gene identifier values. Optional but will run faster if
specified.}
}
\value{
Data frame
}
\description{
Get gene annotations from Ensembl
}
\examples{
\dontrun{
ensembl_annotations(bcbio, values = "ENSMUSG00000000001")
}
}
\author{
Michael Steinbaugh
}
\keyword{internal}
