% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-methods.R
\docType{methods}
\name{export}
\alias{export}
\alias{export,DESeqResultsTables-method}
\title{Export}
\usage{
\S4method{export}{DESeqResultsTables}(x, dir = ".", dropboxDir = NULL,
  rdsToken = NULL)
}
\arguments{
\item{x}{Object.}

\item{dir}{\code{string}. Local directory path.}

\item{dropboxDir}{\code{string} or \code{NULL}. Dropbox directory path where to archive
the results tables for permanent storage (e.g. Stem Cell Commons). When
this option is enabled, unique links per file are generated internally with
the rdrop2 package. Note that local files are written to \code{\link[base:tempdir]{base::tempdir()}}
and the \code{dir} argument is ignored, if this is enabled.}

\item{rdsToken}{\code{string} or \code{NULL}. RDS file token to use for Dropbox
authentication. If set \code{NULL} and \code{dropboxDir} is defined, then an
interactive prompt will appear requesting authorization.}
}
\value{
\itemize{
\item \code{DESeqResultsTables}: CSV files.
}
}
\description{
Export data out of R and write to disk.
}
\details{
This is a wrapper for \code{\link[rio:export]{rio::export()}} that adds support for additional S4
classes in Bioconductor.
}
\examples{
x <- resultsTables(deseq_small)
export(x, dir = "~")
}
\seealso{
\code{\link[rio:export]{rio::export()}}.
}
