% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-methods.R
\docType{methods}
\name{extract}
\alias{extract}
\alias{[,bcbioRNASeq,ANY,ANY,ANY-method}
\title{Extract or replace parts of an object}
\usage{
\S4method{[}{bcbioRNASeq,ANY,ANY,ANY}(x, i, j, drop = FALSE,
  transform = TRUE)
}
\arguments{
\item{x}{
    object from which to extract element(s) or in which to replace element(s).
  }

\item{i}{
    indices specifying elements to extract or replace.  Indices are
    \code{numeric} or \code{character} vectors or empty (missing) or
    \code{NULL}.  Numeric values are coerced to integer as by
    \code{\link{as.integer}} (and hence truncated towards zero).
    Character vectors will be matched to the \code{\link{names}} of the
    object (or for matrices/arrays, the \code{\link{dimnames}}):
    see \sQuote{Character indices} below for further details.

    For \code{[}-indexing only: \code{i}, \code{j}, \code{\dots} can be
    logical vectors, indicating elements/slices to select.  Such vectors
    are recycled if necessary to match the corresponding extent.
    \code{i}, \code{j}, \code{\dots} can also be negative integers,
    indicating elements/slices to leave out of the selection.

    When indexing arrays by \code{[} a single argument \code{i} can be a
    matrix with as many columns as there are dimensions of \code{x}; the
    result is then a vector with elements corresponding to the sets of
    indices in each row of \code{i}.

    An index value of \code{NULL} is treated as if it were \code{integer(0)}.
  }

\item{j}{
    indices specifying elements to extract or replace.  Indices are
    \code{numeric} or \code{character} vectors or empty (missing) or
    \code{NULL}.  Numeric values are coerced to integer as by
    \code{\link{as.integer}} (and hence truncated towards zero).
    Character vectors will be matched to the \code{\link{names}} of the
    object (or for matrices/arrays, the \code{\link{dimnames}}):
    see \sQuote{Character indices} below for further details.

    For \code{[}-indexing only: \code{i}, \code{j}, \code{\dots} can be
    logical vectors, indicating elements/slices to select.  Such vectors
    are recycled if necessary to match the corresponding extent.
    \code{i}, \code{j}, \code{\dots} can also be negative integers,
    indicating elements/slices to leave out of the selection.

    When indexing arrays by \code{[} a single argument \code{i} can be a
    matrix with as many columns as there are dimensions of \code{x}; the
    result is then a vector with elements corresponding to the sets of
    indices in each row of \code{i}.

    An index value of \code{NULL} is treated as if it were \code{integer(0)}.
  }

\item{drop}{For matrices and arrays.  If \code{TRUE} the result is
    coerced to the lowest possible dimension (see the examples).  This
    only works for extracting elements, not for the replacement.  See
    \code{\link{drop}} for further details.
  }

\item{transform}{\code{boolean}. Recalculate slotted DESeq2 transformations.
Recommended by default but can take a while for large datasets.}
}
\value{
\code{bcbioRNASeq}.
}
\description{
Extract genes by row and samples by column from a \code{bcbioRNASeq} object.
Internal count transformations are rescaled automatically, if defined.
}
\note{
DESeq2 transformations will only be calculated when \code{transform = TRUE}
and either \code{rlog} or \code{vst} counts are defined in \code{\link[=assays]{assays()}}.
}
\examples{
object <- bcb_small

# Minimum of 100 genes, 2 samples
genes <- head(rownames(object), 100L)
head(genes)
samples <- head(colnames(object), 2L)
head(samples)

# Extract by sample name
object[, samples]

# Extract by gene list
object[genes, ]

# Extract by both genes and samples
x <- object[genes, samples]
print(x)
assayNames(x)

# Fast subsetting, by skipping DESeq2 transformations
# Note that rlog and vst counts will not be defined
x <- object[, samples, transform = FALSE]
print(x)
names(assays(x))
}
\seealso{
\code{help("[", "base")}.

Other S4 Object: \code{\link{bcbioRNASeq}},
  \code{\link{coerce}}, \code{\link{show}},
  \code{\link{updateObject}}
}
\author{
Lorena Pantano, Michael Steinbaugh
}
\concept{S4 Object}
