% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-methods.R
\docType{methods}
\name{extract}
\alias{extract}
\alias{[,bcbioRNASeq,ANY,ANY,ANY-method}
\title{Extract or Replace Parts of an Object}
\usage{
\S4method{[}{bcbioRNASeq,ANY,ANY,ANY}(x, i, j, drop = FALSE,
  recalculate = TRUE)
}
\arguments{
\item{x}{
    object from which to extract element(s) or in which to replace element(s).
  }

\item{i}{
    indices specifying elements to extract or replace.  Indices are
    \code{numeric} or \code{character} vectors or empty (missing) or
    \code{NULL}.  Numeric values are coerced to integer as by
    \code{\link{as.integer}} (and hence truncated towards zero).
    Character vectors will be matched to the \code{\link{names}} of the
    object (or for matrices/arrays, the \code{\link{dimnames}}):
    see \sQuote{Character indices} below for further details.

    For \code{[}-indexing only: \code{i}, \code{j}, \code{\dots} can be
    logical vectors, indicating elements/slices to select.  Such vectors
    are recycled if necessary to match the corresponding extent.
    \code{i}, \code{j}, \code{\dots} can also be negative integers,
    indicating elements/slices to leave out of the selection.

    When indexing arrays by \code{[} a single argument \code{i} can be a
    matrix with as many columns as there are dimensions of \code{x}; the
    result is then a vector with elements corresponding to the sets of
    indices in each row of \code{i}.

    An index value of \code{NULL} is treated as if it were \code{integer(0)}.
  }

\item{j}{
    indices specifying elements to extract or replace.  Indices are
    \code{numeric} or \code{character} vectors or empty (missing) or
    \code{NULL}.  Numeric values are coerced to integer as by
    \code{\link{as.integer}} (and hence truncated towards zero).
    Character vectors will be matched to the \code{\link{names}} of the
    object (or for matrices/arrays, the \code{\link{dimnames}}):
    see \sQuote{Character indices} below for further details.

    For \code{[}-indexing only: \code{i}, \code{j}, \code{\dots} can be
    logical vectors, indicating elements/slices to select.  Such vectors
    are recycled if necessary to match the corresponding extent.
    \code{i}, \code{j}, \code{\dots} can also be negative integers,
    indicating elements/slices to leave out of the selection.

    When indexing arrays by \code{[} a single argument \code{i} can be a
    matrix with as many columns as there are dimensions of \code{x}; the
    result is then a vector with elements corresponding to the sets of
    indices in each row of \code{i}.

    An index value of \code{NULL} is treated as if it were \code{integer(0)}.
  }

\item{drop}{For matrices and arrays.  If \code{TRUE} the result is
    coerced to the lowest possible dimension (see the examples).  This
    only works for extracting elements, not for the replacement.  See
    \code{\link{drop}} for further details.
  }

\item{recalculate}{\code{logical(1)}. Recalculate DESeq2 normalized counts and
variance-stabilizing transformations defined in \code{assays}. Recommended by
default, but can take a long time for large datasets.}
}
\value{
\code{bcbioRNASeq}.
}
\description{
Extract genes by row and samples by column.
}
\details{
Internal count transformations are rescaled automatically, if defined. DESeq2
transformations will only be updated when \code{recalculate = TRUE} and either
\code{rlog} or \code{vst} counts are defined in \code{assays}.
}
\examples{
data(bcb)
object <- bcb

## Minimum of 100 genes, 2 samples.
genes <- head(rownames(object), 100L)
head(genes)
samples <- head(colnames(object), 2L)
head(samples)

## Extract by sample name.
object[, samples]

## Extract by gene list.
object[genes, ]

## Extract by both genes and samples.
x <- object[genes, samples]
print(x)
assayNames(x)

## Fast subsetting, by skipping DESeq2 recalculations.
## Note that `normalized`, `rlog`, and `vst` assays will not be defined.
x <- object[, samples, recalculate = FALSE]
print(x)
names(assays(x))
}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988)
  \emph{The New S Language}.
  Wadsworth & Brooks/Cole.
}
\author{
Michael Steinbaugh, Lorena Pantano
}
