% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gender_markers}
\alias{gender_markers}
\title{Gender marker genes}
\format{An object of class \code{list} of length 2.}
\usage{
gender_markers
}
\description{
Gender marker genes
}
\author{
Michael Steinbaugh
}
\keyword{internal}
