% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-gene2symbol.R
\docType{methods}
\name{gene2symbol}
\alias{gene2symbol}
\alias{gene2symbol,bcbioRnaDataSet-method}
\title{Gene identifier to symbol conversion}
\usage{
gene2symbol(object)

\S4method{gene2symbol}{bcbioRnaDataSet}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\link{data.frame}.
}
\description{
Gene identifier to symbol conversion
}
\author{
Michael Steinbaugh
}
