% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basejump.R, R/methods-gene2symbol.R
\docType{methods}
\name{gene2symbol}
\alias{gene2symbol}
\alias{gene2symbol,bcbioRNADataSet-method}
\title{Gene Identifier to Symbol Conversion}
\usage{
gene2symbol(object)

\S4method{gene2symbol}{bcbioRNADataSet}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\link{data.frame}.
}
\description{
Gene Identifier to Symbol Conversion
}
\author{
Michael Steinbaugh
}
