% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-gene2symbol.R
\docType{methods}
\name{gene2symbol}
\alias{gene2symbol}
\alias{gene2symbol,bcbioRNASeq-method}
\title{Gene to Symbol Mappings}
\usage{
\S4method{gene2symbol}{bcbioRNASeq}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\link{data.frame} containing Ensembl gene identifier (\code{ensgene}) and
symbol (\code{symbol}) mappings.
}
\description{
Gene to Symbol Mappings
}
\examples{
load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioRNASeq"))

# bcbioRNASeq
gene2symbol(bcb) \%>\% head()
}
\author{
Michael Steinbaugh
}
