% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-gene2symbol.R
\docType{methods}
\name{gene2symbol,bcbioRNADataSet-method}
\alias{gene2symbol,bcbioRNADataSet-method}
\title{Gene Identifier to Symbol Conversion}
\usage{
\S4method{gene2symbol}{bcbioRNADataSet}(object)
}
\value{
\link{data.frame}.
}
\description{
Gene Identifier to Symbol Conversion
}
\author{
Michael Steinbaugh
}
