% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-gene2symbol.R
\docType{methods}
\name{gene2symbol,bcbioRNADataSet-method}
\alias{gene2symbol,bcbioRNADataSet-method}
\title{Gene identifier to symbol conversion}
\usage{
\S4method{gene2symbol}{bcbioRNADataSet}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\link{data.frame}.
}
\description{
Gene identifier to symbol conversion
}
\author{
Michael Steinbaugh
}
