% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_heatmap.R
\name{gene_heatmap}
\alias{gene_heatmap}
\title{Gene heatmap}
\usage{
gene_heatmap(counts, deg = NULL, metadata, factor = "intgroup")
}
\arguments{
\item{counts}{Counts matrix}

\item{deg}{Optional data frame of differential expressed genes (DEG) produced
by DESeq2. If \code{NULL}, will output a heatmap of all genes.}

\item{metadata}{Metadata data frame}

\item{factor}{Factor list}
}
\description{
Gene heatmap
}
\author{
Michael Steinbaugh
}
