% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-general.R
\name{general}
\alias{general}
\title{General Arguments}
\arguments{
\item{object}{Object.}

\item{i}{An integer or numeric scalar.}

\item{value}{Value to assign.}

\item{x}{Object.}

\item{...}{Additional arguments.}

\item{alpha}{Adjusted P value ("alpha") cutoff.}

\item{colData}{\code{data.frame} describing the columns of the object. Must
contain \code{rownames} identical to the \code{colnames} of the object.}

\item{color}{Desired ggplot color scale. Must supply discrete values. When
set to \code{NULL}, the default ggplot2 color palette will be used. If manual
color definitions are desired, we recommend using
\code{\link[ggplot2:scale_colour_manual]{ggplot2::scale_colour_manual()}}.}

\item{counts}{Object containing a normalized counts matrix.}

\item{dir}{Local directory path.}

\item{direction}{Plot "\code{both}", "\code{up}", or "\code{down}" directions.}

\item{fill}{Desired ggplot fill scale. Must supply discrete values. When set
to \code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.}

\item{flip}{Flip x and y axes.}

\item{gene2symbol}{\code{data.frame} containing gene-to-symbol mappings. Must
contain the columns \code{geneID} and \code{geneName}.}

\item{genes}{Character vector of genes to include. These must match the
rownames of the object. It is best practice to use the stable gene
identifiers from Ensembl (e.g. "ENSG00000000003") and not the gene symbols.}

\item{headerLevel}{R Markdown header level.}

\item{interestingGroups}{Character vector denoting groups of interest that
define the samples. If left unset, defaults to \code{sampleName}.}

\item{lfcThreshold}{Log fold change ratio (base 2) cutoff threshold.}

\item{limit}{Threshold to denote on the plot, using a dashed line.}

\item{minCounts}{Minimum number of counts per gene in the counts matrix.}

\item{normalized}{Character indicating which normalization method to apply:
\itemize{
\item "\code{tpm}": Transcripts per million (tximport).
\item "\code{tmm}": edgeR trimmed mean of M-values. Calculated on the fly.
\item "\code{rlog}": DESeq2 \strong{log2} regularized log transformation.
\item "\code{vst}": DESeq2 \strong{log2} variance stabilizing transformation.
}}

\item{ntop}{Number of top genes to label.}

\item{passLimit}{Threshold to plot pass color marker.}

\item{pointColor}{Default point color for the plot.}

\item{results}{\code{DESeqResults} object.}

\item{return}{Object class to return. Uses \code{\link[=match.arg]{match.arg()}} internally and picks
the first item in the vector by default.}

\item{samples}{Character vector of samples to include.}

\item{sigPointColor}{\code{character} vector containing color names for labeling
upregulated and downregulated genes. Also supports a character string for
labeling DEGs with the same color, regardless of direction.}

\item{subtitle}{Subtitle of plot.}

\item{title}{Title of plot.}

\item{warnLimit}{Threshold to plot warning color marker.}

\item{withDimnames}{A \code{logical}, indicating whether dimnames should be
applied to extracted assay elements.}
}
\value{
No value.
}
\description{
General Arguments
}
\keyword{internal}
