% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_counts.R, R/import_file.R,
%   R/import_metadata.R, R/import_summary.R
\name{import_counts}
\alias{import_counts}
\alias{import_file}
\alias{import_metadata}
\alias{import_summary}
\title{Import data from bcbio-nextgen run}
\usage{
import_counts(run, type = "salmon", grep = NULL, samples = NULL)

import_file(run, file, row_names = NULL, ...)

import_metadata(run, grep = NULL, save = FALSE, pool = FALSE)

import_summary(run, save = FALSE)
}
\arguments{
\item{run}{\code{bcbio-nextgen} run}

\item{type}{The type of software used to generate the abundances. Follows the
conventions of \code{tximport()}.}

\item{grep}{Apply grep pattern matching to samples}

\item{samples}{Specify the names of samples in bcbio final directory to
input. If \code{NULL} (default), all samples will be imported.}

\item{file}{File name}

\item{row_names}{Column identifier to use for row names}

\item{...}{Optional parameters for \code{readr}}

\item{save}{Save data}

\item{pool}{Pool lane split samples}
}
\value{
txi \code{tximport} list object

bcbio run data

Metadata data frame

Summary data frame
}
\description{
Import RNA-Seq counts using \code{tximport}

Import a project data file

Import metadata

Import project summary statistics
}
\author{
Michael Steinbaugh

Rory Kirchner
}
