% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_counts.R
\name{import_counts}
\alias{import_counts}
\title{Import counts}
\usage{
import_counts(bcbio, type = "salmon", save_tpm = TRUE)
}
\arguments{
\item{bcbio}{bcbio run object}

\item{type}{The type of software used to generate the abundances. Follows the
conventions of \code{tximport()}.}

\item{save_tpm}{Whether to save transcripts per million data frame}
}
\value{
txi \code{tximport} list object
}
\description{
Import \code{bcbio-rnaseq} count data run using \code{tximport}
}
\examples{
\dontrun{
import_counts(bcbio, type = "salmon")
}
}
\author{
Michael Steinbaugh

Rory Kirchner
}
