% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_file.R
\name{import_file}
\alias{import_file}
\title{Import a bcbio project data file}
\usage{
import_file(bcbio, filename, input = "tsv", output = "data.frame",
  rownames = NULL)
}
\arguments{
\item{bcbio}{bcbio run object}

\item{filename}{Filename}

\item{input}{Input format}

\item{output}{Desired output format}

\item{rownames}{Column identifier to use for rownames}
}
\value{
bcbio data
}
\description{
Import a bcbio project data file
}
\examples{
\dontrun{
import_file(bcbio, "combined.counts",
            output = "matrix", rownames = "id")
}
}
\author{
Michael Steinbaugh
}
