% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_metadata.R
\name{import_metadata}
\alias{import_metadata}
\title{Metadata for bcbio run}
\usage{
import_metadata(bcbio, save = FALSE)
}
\arguments{
\item{bcbio}{bcbio run object}

\item{save}{Save data frame}
}
\value{
Metadata data frame
}
\description{
Metadata for bcbio run
}
\examples{
\dontrun{
import_metadata(bcbio)
}
}
\author{
Michael Steinbaugh
}
