% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_metadata.R
\name{import_metadata}
\alias{import_metadata}
\title{Metadata for bcbio run}
\usage{
import_metadata(bcbio, intgroup, lane_split = FALSE)
}
\arguments{
\item{bcbio}{bcbio run object}

\item{intgroup}{Internal sample groups used for plot coloring.}

\item{lane_split}{Whether samples were split across flow cell lanes.}
}
\value{
Metadata data frame
}
\description{
Metadata for bcbio run
}
\examples{
\dontrun{
import_metadata(bcbio, intgroup = "treatment")
}
}
\author{
Michael Steinbaugh
}
