% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_counts.R
\name{import_sailfish}
\alias{import_sailfish}
\title{Import sailfish counts}
\usage{
import_sailfish(...)
}
\arguments{
\item{...}{Passthrough to \code{import_counts()}}
}
\description{
Import sailfish counts
}
\keyword{internal}
