% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_counts.R
\name{import_salmon}
\alias{import_salmon}
\title{Import salmon counts}
\usage{
import_salmon(...)
}
\arguments{
\item{...}{Passthrough to \code{import_counts()}}
}
\description{
Import salmon counts
}
\keyword{internal}
