% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_integrity_checks.R
\name{check_run}
\alias{check_run}
\alias{check_dds}
\alias{check_dt}
\alias{check_res}
\title{Perform object integrity checks.}
\usage{
check_run(run)

check_dds(dds, stop = TRUE)

check_dt(dt, stop = TRUE)

check_res(res, stop = TRUE)
}
\arguments{
\item{run}{bcbio-nextgen run}

\item{dds}{\linkS4class{DESeqDataSet}.}

\item{stop}{Stop if class doesn't match.}

\item{dt}{\linkS4class{DESeqTransform}.}

\item{res}{\linkS4class{DESeqResults}.}
}
\description{
Check bcbio-nextgen run. Always stop on failure.

Check \linkS4class{DESeqDataSet}.

Check \linkS4class{DESeqTransform}.

Check \linkS4class{DESeqResults}.
}
\author{
Michael Steinbaugh
}
\keyword{internal}
