% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrity_checks.R
\name{check_run}
\alias{check_run}
\alias{check_dds}
\alias{check_dt}
\alias{check_res}
\title{Perform object integrity checks}
\usage{
check_run(run)

check_dds(dds, stop = TRUE)

check_dt(dt, stop = TRUE)

check_res(res, stop = TRUE)
}
\arguments{
\item{run}{\code{bcbio-nextgen} run}

\item{dds}{\code{DESeqDataSet}}

\item{stop}{Stop if class doesn't match}

\item{dt}{\code{DESeqTransform}}

\item{res}{\code{DESeqResults}}
}
\description{
Check \code{bcbio-nextgen} run

Check \code{DESeqDataSet}

Check \code{DESeqTransform}

Check \code{DESeqResults}
}
\author{
Michael Steinbaugh
}
