% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_integrity_checks.R
\name{check_run}
\alias{check_run}
\alias{check_dds}
\alias{check_dt}
\alias{check_res}
\title{Perform object integrity checks}
\usage{
check_run(run)

check_dds(dds, stop = TRUE)

check_dt(dt, stop = TRUE)

check_res(res, stop = TRUE)
}
\arguments{
\item{run}{bcbio-nextgen run}

\item{dds}{DESeqDataSet}

\item{stop}{Stop if class doesn't match}

\item{dt}{DESeqTransform}

\item{res}{DESeqResults}
}
\description{
Check bcbio-nextgen run. Always stop on failure.

Check DESeqDataSet

Check DESeqTransform

Check DESeqResults
}
\author{
Michael Steinbaugh
}
