% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-interestingGroups.R
\docType{methods}
\name{interestingGroups}
\alias{interestingGroups}
\alias{interestingGroups,bcbioRNASeqANY-method}
\title{Interesting Groups}
\usage{
\S4method{interestingGroups}{bcbioRNASeqANY}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Character vector.
}
\description{
Interesting Groups
}
