% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_bcbio_run.R
\name{load_bcbio_run}
\alias{load_bcbio_run}
\title{Load bcbio run}
\usage{
load_bcbio_run(final_dir = "final", intgroup = "description",
  organism = NULL)
}
\arguments{
\item{final_dir}{Path to final output directory. This path is set when
running \code{bcbio_nextgen -w template}.}

\item{intgroup}{Character vector of interesting groups. First entry is used
for plot colors during quality control (QC) analysis. Entire vector is used
for PCA and heatmap QC functions.}

\item{organism}{Organism name, following Ensembl/Biomart conventions. Must be
lowercase and one word (e.g. hsapiens). This will be detected automatically
for common reference genomes.}
}
\value{
\code{bcbio-nextgen} run object
}
\description{
We recommend loading the \code{bcbio-nextgen} run as a remote connection over
\code{sshfs}. This requires setting \code{parent_dir} in the function.
}
\author{
Michael Steinbaugh
}
