% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_run.R
\name{load_run}
\alias{load_run}
\title{Load bcbio-nextgen run}
\usage{
load_run(upload_dir = "final", analysis = "rnaseq",
  groups_of_interest = "description", design = NULL, contrast = NULL,
  custom_metadata_file = NULL)
}
\arguments{
\item{upload_dir}{Path to final upload directory. This path is set when
running \code{bcbio_nextgen -w template}.}

\item{analysis}{Analysis type (e.g. \code{rnaseq} or \code{srnaseq}).}

\item{groups_of_interest}{Character vector of interesting groups. First entry
is used for plot colors during quality control (QC) analysis. Entire vector
is used for PCA and heatmap QC functions.}

\item{design}{(\emph{Optional}). Design formula for DESeq2.See
\code{\link[DESeq2:design]{DESeq2::design()}} for more information.}

\item{contrast}{(\emph{Optional}). Contrast vector for DESeq2. See
\code{\link[DESeq2:results]{DESeq2::results()}} for more information.}

\item{custom_metadata_file}{(\emph{Optional}). Custom metadata file containing
sample information. Otherwise defaults to sample metadata saved in the YAML
file.}
}
\value{
\link{bcbioRnaDataSet}.
}
\description{
Simply point to the final upload directory output by
\href{https://bcbio-nextgen.readthedocs.io/}{bcbio-nextgen}, and this function
will take care of the rest. It automatically imports RNA-seq counts,
metadata, and program versions used.
}
\details{
When working in RStudio, we recommend connecting to the bcbio-nextgen run
directory as a remote connection over
\href{https://github.com/osxfuse/osxfuse/wiki/SSHFS}{sshfs}.
}
\author{
Michael Steinbaugh

Lorena Patano
}
