% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R, R/load.R
\name{load_run_S4}
\alias{load_run_S4}
\alias{load_run}
\title{Load bcbio-nextgen run}
\usage{
load_run_S4(upload_dir = "final", analysis = "rnaseq",
  intgroup = "description", metadata_file = NULL, organism = NULL, ...)

load_run(upload_dir = "final", analysis = "rnaseq", alpha = 0.05,
  lfc = 1, intgroup = "description", metadata_file = NULL,
  organism = NULL)
}
\arguments{
\item{upload_dir}{Path to final upload directory. This path is set when
running \code{bcbio_nextgen -w template}.}

\item{analysis}{Analysis type (e.g. \code{rnaseq} or \code{srnaseq}).}

\item{intgroup}{Character vector of interesting groups. First entry is used
for plot colors during quality control (QC) analysis. Entire vector is used
for PCA and heatmap QC functions.}

\item{metadata_file}{Custom metadata file to import. Otherwise defaults to
sample metadata saved in the YAML file.}

\item{organism}{Organism name, following Ensembl/Biomart conventions. Must be
lowercase and one word (e.g. hsapiens). This will be detected automatically
for common reference genomes and normally does not need to be set.}

\item{...}{Arguments provided to \link{SummarizedExperiment}. Remove this once
we migrate the S4 loading variant over as main method.}

\item{alpha}{Alpha (false discovery rate) cutoff.}

\item{lfc}{Log2 fold change ratio cutoff.}
}
\value{
\link{bcbioRnaDataSet} containing counts and metadata.
}
\description{
Simply point to the final upload directory output by
\href{https://bcbio-nextgen.readthedocs.io/}{bcbio-nextgen}, and this function
will take care of the rest. It automatically imports RNA-seq counts,
metadata, and program versions used.
}
\details{
When working in RStudio, we recommend connecting to the bcbio-nextgen run
directory as a remote connection over
\href{https://github.com/osxfuse/osxfuse/wiki/SSHFS}{sshfs}.
}
\author{
Michael Steinbaugh

Lorena Patano
}
\keyword{internal}
