% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_run.R
\name{load_run}
\alias{load_run}
\title{Load bcbio-nextgen run.}
\usage{
load_run(upload_dir = "final", analysis = "rnaseq",
  intgroup = "description", metadata_file = NULL, organism = NULL,
  read_counts = TRUE)
}
\arguments{
\item{upload_dir}{Path to final upload directory. This path is set when
running \code{bcbio_nextgen -w template}.}

\item{analysis}{Analysis type (e.g. \code{rnaseq} or \code{srnaseq})}

\item{intgroup}{Character vector of interesting groups. First entry is used
for plot colors during quality control (QC) analysis. Entire vector is used
for PCA and heatmap QC functions.}

\item{metadata_file}{Custom metadata file to import. Otherwise defaults to
sample metadata saved in the YAML file.}

\item{organism}{Organism name, following Ensembl/Biomart conventions. Must be
lowercase and one word (e.g. hsapiens). This will be detected automatically
for common reference genomes and normally does not need to be set.}

\item{read_counts}{Automatically read in the count data using
\code{read_bcbio_counts()}}
}
\value{
bcbio-nextgen run list object, containing counts and metadata.
}
\description{
Simply point to the YAML file generated by bcbio-nextgen in the final
upload directory of the run, and this function will take care of the rest. It
imports data and program versions, and imports the RNA-seq counts
automatically. When working in RStudio, we recommend connecting to the
bcbio-nextgen run directory as a remote connection over \code{sshfs}.
}
\author{
Michael Steinbaugh
}
