% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_run.R
\name{load_run}
\alias{load_run}
\title{Load bcbio run}
\usage{
load_run(yaml_file = "project-summary.yaml", analysis = "rnaseq",
  intgroup = "description", metadata_file = NULL, organism = NULL,
  read_counts = TRUE)
}
\arguments{
\item{yaml_file}{YAML file saved during final upload from
\code{bcbio-nextgen}}

\item{analysis}{Analysis type (e.g. \code{rnaseq} or \code{srnaseq})}

\item{intgroup}{Character vector of interesting groups. First entry is used
for plot colors during quality control (QC) analysis. Entire vector is used
for PCA and heatmap QC functions.}

\item{metadata_file}{Custom metadata file to import. Otherwise defaults to
sample metadata saved in the YAML file.}

\item{organism}{Organism name, following Ensembl/Biomart conventions. Must be
lowercase and one word (e.g. hsapiens). This will be detected automatically
for common reference genomes and normally does not need to be set.}

\item{read_counts}{Automatically read in the count data using
\code{read_bcbio_counts()}}
}
\value{
\code{bcbio-nextgen} run object
}
\description{
We recommend loading the \code{bcbio-nextgen} run as a remote connection over
\code{sshfs}. This requires setting \code{parent_dir} in the function.
}
\author{
Michael Steinbaugh
}
