% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRNASeqRun.R
\name{loadRNASeqRun}
\alias{loadRNASeqRun}
\title{Load bcbio RNA-Seq Run}
\usage{
loadRNASeqRun(uploadDir, interestingGroups = "sampleName",
  sampleMetadataFile = NULL, maxSamples = 50L, ensemblVersion = "current",
  ...)
}
\arguments{
\item{uploadDir}{Path to final upload directory. This path is set when
running \code{bcbio_nextgen -w template}.}

\item{interestingGroups}{Character vector of interesting groups. First entry
is used for plot colors during quality control (QC) analysis. Entire vector
is used for PCA and heatmap QC functions.}

\item{sampleMetadataFile}{\emph{Optional}. Custom metadata file containing
sample information. Otherwise defaults to sample metadata saved in the YAML
file.}

\item{maxSamples}{Maximum number of samples to calculate \code{\link[DESeq2:rlog]{DESeq2::rlog()}} and
\code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}} matrix. See Details.}

\item{ensemblVersion}{Ensembl release version. Defaults to current, and does
not typically need to be user-defined. This parameter can be useful for
matching Ensembl annotations against an outdated bcbio annotation build.}

\item{...}{Additional arguments, slotted into the \code{\link[=metadata]{metadata()}} accessor.}
}
\value{
\link{bcbioRNASeq}.
}
\description{
Simply point to the final upload directory output by
\href{https://bcbio-nextgen.readthedocs.io/}{bcbio}, and this function will take
care of the rest. It automatically imports RNA-seq counts, metadata, and
program versions used.
}
\details{
When number of samples is bigger than \code{maxSamples}, \code{rlog} and \code{vst} slot
in \link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment::SummarizedExperiment} will be the output of
\link{edgeR} normalization method.
}
\note{
When working in RStudio, we recommend connecting to the bcbio-nextgen
run directory as a remote connection over
\href{https://github.com/osxfuse/osxfuse/wiki/SSHFS}{sshfs}.
}
\examples{
uploadDir <- system.file("extdata/bcbio", package = "bcbioRNASeq")
bcb <- loadRNASeqRun(uploadDir, interestingGroups = "group")
}
\author{
Michael Steinbaugh, Lorena Pantano
}
