% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown-methods.R
\docType{methods}
\name{markdown}
\alias{markdown}
\alias{markdown,DESeqResultsTables-method}
\title{Markdown}
\usage{
\S4method{markdown}{DESeqResultsTables}(object, headerLevel = 2L)
}
\arguments{
\item{object}{Object.}

\item{headerLevel}{\code{scalar integer} (\code{1}-\code{7}). R Markdown header level.}
}
\value{
Markdown output.
}
\description{
Markdown
}
\examples{
object <- resultsTables(deseq_small)
markdown(object)
}
\author{
Michael Steinbaugh
}
