% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown-methods.R
\docType{methods}
\name{markdown}
\alias{markdown}
\alias{markdown,bcbioRNASeq-method}
\title{Markdown}
\usage{
\S4method{markdown}{bcbioRNASeq}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Markdown output.
}
\description{
Markdown
}
\section{Methods (by class)}{
\itemize{
\item \code{bcbioRNASeq}: Sample metadata table.
}}

\examples{
data(deseq)
x <- DESeqResultsTables(deseq)
markdown(x)
}
\author{
Michael Steinbaugh
}
