% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{md_res_tables}
\alias{md_res_tables}
\title{Markdown list of results files}
\usage{
md_res_tables(res_tbl)
}
\arguments{
\item{res_tbl}{List of results tables generated by \code{\link[=res_tables]{res_tables()}}.}
}
\value{
Unordered Markdown list.
}
\description{
Enables looping of results contrast file links for RMarkdown.
}
\author{
Michael Steinbaugh
}
\keyword{internal}
