% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-rmarkdown.R
\name{.md_results_tables}
\alias{.md_results_tables}
\title{Markdown List of Results Files}
\usage{
.md_results_tables(res_tbl, dir)
}
\arguments{
\item{res_tbl}{List of results tables generated by \code{\link[=results_tables]{results_tables()}}.}

\item{dir}{Output directory.}
}
\value{
\code{\link[=writeLines]{writeLines()}}.
}
\description{
Enables looping of results contrast file links for RMarkdown.
}
\author{
Michael Steinbaugh
}
\keyword{internal}
