% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt_log10.R
\name{melt_log10}
\alias{melt_log10}
\title{Melt RNA-Seq count data to long format and log10 transform}
\usage{
melt_log10(counts, metadata)
}
\arguments{
\item{counts}{Counts matrix}

\item{metadata}{Metadata data frame}
}
\value{
log10 melted data frame
}
\description{
Melt RNA-Seq count data to long format and log10 transform
}
\author{
Michael Steinbaugh
}
