% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt_log10.R
\name{melt_log10}
\alias{melt_log10}
\title{Melt RNA-Seq count data to long format and log10 transform}
\usage{
melt_log10(bcbio, counts)
}
\arguments{
\item{bcbio}{bcbio list object}

\item{counts}{Counts matrix}
}
\value{
log10 melted data frame
}
\description{
Melt RNA-Seq count data to long format and log10 transform
}
\author{
Michael Steinbaugh
}
