% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-melt_log10.R
\docType{methods}
\name{melt_log10}
\alias{melt_log10}
\alias{.melt_log10}
\alias{.join_melt}
\alias{melt_log10,bcbioRNADataSet-method}
\alias{melt_log10,DESeqDataSet-method}
\alias{melt_log10,DESeqTransform-method}
\title{Melt count matrix to long format and log10 transform}
\usage{
\S4method{melt_log10}{bcbioRNADataSet}(object, normalized = TRUE,
  interesting_groups = NULL)

\S4method{melt_log10}{DESeqDataSet}(object, normalized = TRUE,
  interesting_groups = NULL)

\S4method{melt_log10}{DESeqTransform}(object, interesting_groups = NULL)
}
\arguments{
\item{object}{Object containing counts matrix.}

\item{...}{Optional parameters.}

\item{normalized}{Select normalized counts (\code{TRUE}), raw counts (\code{FALSE}),
or specifically request TMM-normalized counts (\code{tmm}).}

\item{interesting_groups}{\emph{Optional}. Interesting groups.}
}
\value{
log10 melted \link{data.frame}.
}
\description{
Melt count matrix to long format and log10 transform
}
\seealso{
\code{\link[reshape2:melt]{reshape2::melt()}}.
}
\author{
Michael Steinbaugh
}
