% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-meltLog10.R
\docType{methods}
\name{meltLog10}
\alias{meltLog10}
\alias{meltLog10}
\alias{meltLog10,bcbioRNASeq-method}
\alias{meltLog10,DESeqDataSet-method}
\alias{meltLog10,DESeqTransform-method}
\title{Melt Count Matrix to Long Format and log10 Transform}
\usage{
meltLog10(object, ...)

\S4method{meltLog10}{bcbioRNASeq}(object, normalized = TRUE)

\S4method{meltLog10}{DESeqDataSet}(object, normalized = TRUE)

\S4method{meltLog10}{DESeqTransform}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{normalized}{Select normalized counts (\code{TRUE}), raw counts (\code{FALSE}),
or specifically request TMM-normalized counts (\code{tmm}).}
}
\value{
log10 melted \link{data.frame}.
}
\description{
Melt Count Matrix to Long Format and log10 Transform
}
\examples{
load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioRNASeq"))
load(system.file(
    file.path("extdata", "dds.rda"),
    package = "bcbioRNASeq"))
load(system.file(
    file.path("extdata", "rld.rda"),
    package = "bcbioRNASeq"))

# bcbioRNASeq
meltLog10(bcb) \%>\% glimpse()

# DESeqDataSet
meltLog10(dds) \%>\% glimpse()

# DESeqTransform
meltLog10(rld) \%>\% glimpse()
}
\seealso{
\code{\link[reshape2:melt]{reshape2::melt()}}.
}
\author{
Michael Steinbaugh
}
