% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-meltLog10.R
\docType{methods}
\name{meltLog10}
\alias{meltLog10}
\alias{meltLog10,bcbioRNADataSet-method}
\alias{meltLog10,DESeqDataSet-method}
\alias{meltLog10,DESeqTransform-method}
\title{Melt Count Matrix to Long Format and log10 Transform}
\usage{
meltLog10(object, ...)

\S4method{meltLog10}{bcbioRNADataSet}(object, normalized = TRUE)

\S4method{meltLog10}{DESeqDataSet}(object, normalized = TRUE)

\S4method{meltLog10}{DESeqTransform}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{normalized}{Select normalized counts (\code{TRUE}), raw counts (\code{FALSE}),
or specifically request TMM-normalized counts (\code{tmm}).}
}
\value{
log10 melted \link{data.frame}.
}
\description{
Melt Count Matrix to Long Format and log10 Transform
}
\examples{
data(bcb, dds, rld)

# bcbioRNADataSet
meltLog10(bcb) \%>\% str

# DESeqDataSet
meltLog10(dds) \%>\% str

# DESeqTransform
meltLog10(rld) \%>\% str
}
\seealso{
\code{\link[reshape2:melt]{reshape2::melt()}}.
}
\author{
Michael Steinbaugh
}
