% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-metadata_table.R
\docType{methods}
\name{metadata_table}
\alias{metadata_table}
\alias{.metadata_table}
\alias{metadata_table,bcbioRNADataSet-method}
\alias{metadata_table,DESeqDataSet-method}
\alias{metadata_table,DESeqTransform-method}
\title{Metadata Table}
\usage{
metadata_table(object, ...)

\S4method{metadata_table}{bcbioRNADataSet}(object, ...)

\S4method{metadata_table}{DESeqDataSet}(object, ...)

\S4method{metadata_table}{DESeqTransform}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional parameters, passed to \code{\link[=kable]{kable()}}.}
}
\value{
\link{data.frame} containing only the columns of interest.
}
\description{
Returns a subset of metadata columns of interest used for knit reports. These
"interesting group" columns are defined as \code{interesting_groups} in the
\link{bcbioRNADataSet} object.
}
\author{
Michael Steinbaugh
}
