% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_table.R
\name{metadata_table}
\alias{metadata_table}
\title{Metadata table}
\usage{
metadata_table(run)
}
\arguments{
\item{run}{bcbio-nextgen run}
}
\value{
Data frame containing only the columns of interest
}
\description{
Returns a subset of metadata columns of interest used for knit reports. These
"interesting group" columns are defined as \code{intgroup} in the
bcbio-nextgen run object.
}
\author{
Michael Steinbaugh
}
