% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/all_methods.R
\docType{methods}
\name{metadata_table}
\alias{metadata_table}
\alias{metadata_table,bcbioRnaDataSet-method}
\alias{metadata_table,DESeqDataSet-method}
\alias{metadata_table,DESeqTransform-method}
\title{Metadata table}
\usage{
metadata_table(object, ...)

\S4method{metadata_table}{bcbioRnaDataSet}(object, ...)

\S4method{metadata_table}{DESeqDataSet}(object, ...)

\S4method{metadata_table}{DESeqTransform}(object, ...)
}
\arguments{
\item{...}{\code{\link[=kable]{kable()}} passthrough parameters.}
}
\value{
Data frame containing only the columns of interest.
}
\description{
Returns a subset of metadata columns of interest used for knit reports. These
"interesting group" columns are defined as \code{interesting_groups} in the
\link{bcbioRnaDataSet} object.
}
