% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-methods.R
\docType{methods}
\name{metrics}
\alias{metrics}
\alias{metrics,bcbioRNASeq-method}
\title{Sample Metrics}
\usage{
\S4method{metrics}{bcbioRNASeq}(object, interestingGroups = NULL)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}. Groups of interest that define the
samples. If left unset, defaults to \code{sampleName}.}
}
\value{
\code{data.frame}.
}
\description{
Sample Metrics
}
\examples{
metrics(bcb_small) \%>\% glimpse()
}
\seealso{
Other Data Functions: \code{\link{counts}},
  \code{\link{sampleData}}, \code{\link{tmm}},
  \code{\link{tpm}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
