% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metrics.R
\docType{methods}
\name{metrics,bcbioRNADataSet-method}
\alias{metrics,bcbioRNADataSet-method}
\title{Sample metrics}
\usage{
\S4method{metrics}{bcbioRNADataSet}(object)
}
\arguments{
\item{object}{\link{bcbioRNADataSet} object.}
}
\value{
\link{tibble}.
}
\description{
Sample metrics
}
