% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metrics.R
\docType{methods}
\name{metrics}
\alias{metrics}
\alias{metrics,bcbioRNADataSet-method}
\title{Sample Metrics}
\usage{
\S4method{metrics}{bcbioRNADataSet}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\link{data.frame}.
}
\description{
Sample Metrics
}
\examples{
data(bcb)
metrics(bcb) \%>\% glimpse
}
