% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metrics.R
\docType{methods}
\name{metrics}
\alias{metrics}
\alias{metrics,bcbioRNASeq-method}
\title{Sample Metrics}
\usage{
\S4method{metrics}{bcbioRNASeq}(object, interestingGroups)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{Character vector denoting groups of interest that
define the samples. If left unset, defaults to \code{sampleName}.}
}
\value{
\code{data.frame}.
}
\description{
Sample Metrics
}
\examples{
metrics(bcb_small) \%>\% glimpse()
}
\seealso{
Other Data Functions: \code{\link{counts}},
  \code{\link{tmm}}, \code{\link{tpm}}
}
\author{
Michael Steinbaugh
}
