% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metrics.R
\docType{methods}
\name{metrics}
\alias{metrics}
\alias{metrics,bcbioRNASeq-method}
\title{Sample Metrics}
\usage{
\S4method{metrics}{bcbioRNASeq}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{data.frame}.
}
\description{
Sample Metrics
}
\examples{
metrics(bcb_small) \%>\% glimpse()
}
