% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-methods.R
\docType{methods}
\name{metrics}
\alias{metrics}
\alias{metrics,bcbioRNASeq-method}
\title{Quality control metrics}
\usage{
metrics(object, ...)

\S4method{metrics}{bcbioRNASeq}(object, interestingGroups)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}. Groups of interest that define the
samples. If left unset, defaults to \code{sampleName}.}

\item{...}{Additional arguments.}
}
\value{
\code{data.frame}.
}
\description{
Quality control metrics
}
\examples{
metrics(bcb_small) \%>\% tibble::glimpse(.)
}
\seealso{
Other Data Functions: \code{\link{counts}},
  \code{\link{tmm}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
