% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Shared parameter documentation}
\arguments{
\item{normalized}{\code{character(1)}.
Which normalization method to apply:
\itemize{
\item "\code{tpm}": Transcripts per million (tximport).
\item "\code{tmm}": edgeR trimmed mean of M-values. Calculated on the fly.
\item "\code{rlog}": DESeq2 log2 regularized log transformation.
\item "\code{vst}": DESeq2 log2 variance stabilizing transformation.
}}

\item{object}{Object.}

\item{title}{\code{character(1)} or \code{NULL}.
Plot title.}

\item{...}{Additional arguments.}
}
\value{
No value.
}
\description{
Shared parameter documentation
}
\keyword{internal}
