% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_samples_yaml.R
\name{parse_samples_yaml}
\alias{parse_samples_yaml}
\title{Read sample}
\usage{
parse_samples_yaml(run, nested_operators)
}
\arguments{
\item{run}{\code{bcbio-nextgen} run}

\item{nested_operators}{Nested operators work as an ordered character vector,
recursing a level down for each entry}
}
\description{
Read sample
}
\author{
Michael Steinbaugh
}
