% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plot53Bias.R
\docType{methods}
\name{plot53Bias}
\alias{plot53Bias}
\alias{plot53Bias}
\alias{plot53Bias,bcbioRNADataSet-method}
\alias{plot53Bias,data.frame-method}
\title{Plot 5'->3' Bias}
\usage{
plot53Bias(object, ...)

\S4method{plot53Bias}{bcbioRNADataSet}(object, ...)

\S4method{plot53Bias}{data.frame}(object, interestingGroup = "sampleName",
  warnLimit = 2L, flip = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{interestingGroup}{\emph{Optional}. Category to use to group samples (color
and shape). If unset, this is automatically determined by the metadata set
inside the \link{bcbioRNADataSet}.}

\item{warnLimit}{Threshold to plot warning color marker.}

\item{flip}{Flip X and Y axes.}
}
\value{
\link{ggplot}.
}
\description{
Plot 5'->3' Bias
}
\examples{
data(bcb)

# bcbioRNADataSet
plot53Bias(bcb)

# data.frame
metrics <- metrics(bcb)
plot53Bias(metrics)
}
\seealso{
Other Quality Control Plots: \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGeneDetectionSaturation}},
  \code{\link{plotGenesDetected}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh
}
