% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot5Prime3PrimeBias-methods.R
\docType{methods}
\name{plot5Prime3PrimeBias}
\alias{plot5Prime3PrimeBias}
\alias{plot5Prime3PrimeBias}
\alias{plot5Prime3PrimeBias,bcbioRNASeq-method}
\title{Plot 5'->3' Bias}
\usage{
plot5Prime3PrimeBias(object, ...)

\S4method{plot5Prime3PrimeBias}{bcbioRNASeq}(object, interestingGroups,
  limit = 2L, fill = NULL, flip = TRUE, title = "5'->3' bias")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{interestingGroups}{Character vector denoting groups of interest that
define the samples. If left unset, defaults to \code{sampleName}.}

\item{limit}{Threshold to denote on the plot, using a dashed line.}

\item{fill}{Desired ggplot fill scale. Must supply discrete values. When set
to \code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.}

\item{flip}{Flip x and y axes.}

\item{title}{Title of plot.}
}
\value{
\code{ggplot}.
}
\description{
Plot 5'->3' Bias
}
\examples{
plot5Prime3PrimeBias(bcb_small)
}
\seealso{
Other Quality Control Functions: \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotDispEsts}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotGenesDetected}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}}, \code{\link{plotMeanSD}},
  \code{\link{plotPCACovariates}}, \code{\link{plotPCA}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh
}
