% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot5Prime3PrimeBias-methods.R
\docType{methods}
\name{plot5Prime3PrimeBias}
\alias{plot5Prime3PrimeBias}
\alias{plot5Prime3PrimeBias,bcbioRNASeq-method}
\title{Plot 5'->3' Bias}
\usage{
\S4method{plot5Prime3PrimeBias}{bcbioRNASeq}(object,
  interestingGroups = NULL, limit = 0L,
  fill = getOption("basejump.fill.discrete", NULL),
  flip = getOption("basejump.flip", TRUE), title = "5'->3' bias")
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character} or \code{NULL}. Groups of interest that
define the samples. If left unset, defaults to \code{sampleName}.}

\item{limit}{\code{scalar numeric}. Threshold to denote on the plot, using a
dashed line.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 fill scale.
Must supply discrete values. When set to \code{NULL}, the default ggplot2 color
palette will be used. If manual color definitions are desired, we recommend
using \code{ggplot2::scale_fill_manual()}. To set the discrete fill palette
globally, use \code{options(bcbio.discrete.fill = scale_fill_viridis_d())}.}

\item{flip}{\code{boolean}. Flip x and y axes. Recommended for plots containing
many samples.}

\item{title}{\code{string} or \code{NULL}. Plot title.}
}
\value{
\code{ggplot}.
}
\description{
RNA-seq data can have specific biases at either the 5’ or 3’ end of sequenced
fragments.
}
\details{
It is common to see a small amount of bias, especially if polyA
enrichment was performed, or if there is any sample degradation. If a large
amount of bias is observed here, be sure to analyze the samples with a
Bioanalyzer and check the RIN scores.

5' (3') bias is generally calculated as the median of the following ratio:\preformatted{[mean expression of 5' (3')] / [mean expression of whole transcript]
}

For example:
\itemize{
\item Mean expression for 5' (3') is calculated as mean coverage of first (last)
100 bases.
\item Mean expression of transcript is the mean coverage of all bases in that
transcript.
\item Median is calculated for the representative set of 1000 transcripts.
}
}
\examples{
data(bcb)
plot5Prime3PrimeBias(bcb)
}
\author{
Michael Steinbaugh
}
