% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot5Prime3PrimeBias-methods.R
\docType{methods}
\name{plot5Prime3PrimeBias}
\alias{plot5Prime3PrimeBias}
\alias{plot5Prime3PrimeBias,bcbioRNASeq-method}
\title{Plot 5' to 3' bias}
\usage{
\S4method{plot5Prime3PrimeBias}{bcbioRNASeq}(object,
  interestingGroups = NULL, color = getOption(x =
  "acid.color.discrete", default =
  acidplots::scale_colour_synesthesia_d()), flip = getOption(x =
  "acid.flip", default = TRUE), title = "5'->3' bias")
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.

To set the discrete color palette globally, use:\preformatted{options(acid.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{flip}{\code{logical(1)}.
Flip x and y axes. Recommended for plots containing many samples.}

\item{title}{\code{character(1)}.
Plot title.}
}
\value{
\code{ggplot}.
}
\description{
RNA-seq data can have specific biases at either the 5’ or 3’ end of sequenced
fragments.
}
\details{
It is common to see a small amount of bias, especially if polyA enrichment
was performed, or if there is any sample degradation. If a large amount of
bias is observed here, be sure to analyze the samples with a Bioanalyzer and
check the RIN scores.

5' (3') bias is generally calculated as the median of the following ratio:\preformatted{[mean expression of 5' (3')] /
    [mean expression of whole transcript]
}

For example:
\itemize{
\item Mean expression for 5' (3') is calculated as mean coverage of first (last)
100 bases.
\item Mean expression of transcript is the mean coverage of all bases in that
transcript.
\item Median is calculated for the representative set of 1000 transcripts.
}
}
\examples{
data(bcb)
plot5Prime3PrimeBias(bcb)
}
\author{
Michael Steinbaugh
}
