% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-qc.R
\name{plot_correlation_heatmap}
\alias{plot_correlation_heatmap}
\title{Correlation matrix heatmap}
\usage{
plot_correlation_heatmap(run, dt, method = "pearson")
}
\arguments{
\item{run}{bcbio-nextgen run.}

\item{dt}{\linkS4class{DESeqTransform} generated from \code{\link[DESeq2:rlog]{DESeq2::rlog()}} or
\code{\link[DESeq2:vst]{DESeq2::vst()}} on a \linkS4class{DESeqDataSet}.}

\item{method}{Correlation coefficient (or covariance) to be computed.
Defaults to \code{pearson} but \code{spearman} can also be used.}
}
\value{
Correlation heatmap.
}
\description{
Correlation matrix heatmap
}
\author{
Michael Steinbaugh
}
