% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlation_heatmap.R
\name{plot_correlation_heatmap}
\alias{plot_correlation_heatmap}
\title{Correlation matrix heatmap}
\usage{
plot_correlation_heatmap(run, dt, method = "pearson")
}
\arguments{
\item{run}{bcbio-nextgen run object}

\item{dt}{\code{DESeqTransform} generated from \code{rlog()} or \code{vst()}
on a \code{DESeqDataSet} object}

\item{method}{Correlation coefficient (or covariance) to be computed.
Defaults to \code{pearson} but \code{spearman} can also be used.}
}
\value{
Heatmap
}
\description{
Correlation matrix heatmap
}
\author{
Michael Steinbaugh
}
