% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_count_density.R
\name{plot_count_density}
\alias{plot_count_density}
\title{Count density plot}
\usage{
plot_count_density(bcbio, counts)
}
\arguments{
\item{bcbio}{bcbio list object}

\item{counts}{Counts matrix}
}
\description{
Count density plot
}
\examples{
\dontrun{
plot_count_density(bcbio, counts)
}
}
\author{
Michael Steinbaugh
}
