% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-plot_count_density.R
\docType{methods}
\name{plot_count_density}
\alias{plot_count_density}
\alias{.plot_count_density}
\alias{plot_count_density,bcbioRNADataSet-method}
\title{Plot Count Density}
\usage{
plot_count_density(object, ...)

.plot_count_density(melted, interesting_group = "sample_name")

\S4method{plot_count_density}{bcbioRNADataSet}(object, normalized = "tmm")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{melted}{Melted counts generated by \code{\link[=melt_log10]{melt_log10()}}.}

\item{interesting_group}{\emph{Optional}. Category to use to group samples (color
and shape). If unset, this is automatically determined by the metadata set
inside the \link{bcbioRNADataSet}.}

\item{normalized}{Count normalization method. See \code{\link[=counts]{counts()}} documentation
for more information.}
}
\value{
\link{ggplot}.
}
\description{
Plot Count Density
}
\examples{
data(bcb)
plot_count_density(bcb)
}
\seealso{
Other Quality Control Plots: \code{\link{plot_counts_per_gene}},
  \code{\link{plot_exonic_mapping_rate}},
  \code{\link{plot_gene_detection_saturation}},
  \code{\link{plot_genes_detected}},
  \code{\link{plot_intronic_mapping_rate}},
  \code{\link{plot_mapped_reads}},
  \code{\link{plot_mapping_rate}},
  \code{\link{plot_rrna_mapping_rate}},
  \code{\link{plot_total_reads}}
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
