% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R,
%   R/methods-plot_counts_per_gene.R
\docType{methods}
\name{plot_counts_per_gene}
\alias{plot_counts_per_gene}
\alias{.plot_counts_per_gene}
\alias{plot_counts_per_gene,bcbioRNADataSet-method}
\title{Plot Counts Per Gene}
\usage{
plot_counts_per_gene(object, ...)

.plot_counts_per_gene(melted, interesting_group, flip = TRUE)

\S4method{plot_counts_per_gene}{bcbioRNADataSet}(object, normalized = "tmm",
  ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{interesting_group}{\emph{Optional}. Category to use to group samples (color
and shape). If unset, this is automatically determined by the metadata set
inside the \link{bcbioRNADataSet}.}

\item{normalized}{Count normalization method. See \code{\link[=counts]{counts()}} documentation
for more information.}
}
\value{
\link{ggplot}.
}
\description{
Plot Counts Per Gene
}
\examples{
data(bcb)
plot_counts_per_gene(bcb)
}
\seealso{
Other Quality Control Plots: \code{\link{plot_count_density}},
  \code{\link{plot_exonic_mapping_rate}},
  \code{\link{plot_gene_detection_saturation}},
  \code{\link{plot_genes_detected}},
  \code{\link{plot_mapped_reads}},
  \code{\link{plot_mapping_rate}},
  \code{\link{plot_total_reads}}
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
