% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_counts_per_gene.R
\name{plot_counts_per_gene}
\alias{plot_counts_per_gene}
\title{Counts per gene plot}
\usage{
plot_counts_per_gene(bcbio, counts, flip = TRUE)
}
\arguments{
\item{bcbio}{bcbio list object}

\item{counts}{Counts matrix}

\item{flip}{Flip x and y axes}
}
\value{
Boxplot
}
\description{
Counts per gene plot
}
\examples{
\dontrun{
plot_counts_per_gene(bcbio, counts)
}
}
\author{
Michael Steinbaugh
}
