% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-de.R
\name{plot_deg_heatmap}
\alias{plot_deg_heatmap}
\title{Differentially expressed gene heatmap}
\usage{
plot_deg_heatmap(run, res, dt, lfc = 0)
}
\arguments{
\item{run}{bcbio-nextgen run.}

\item{res}{\linkS4class{DESeqResults}.}

\item{dt}{\linkS4class{DESeqTransform}. We recommend passing in \code{\link[=rlog]{rlog()}}
counts in as the transform object by default.}

\item{lfc}{Log2 fold change ratio cutoff.}
}
\value{
Gene heatmap.
}
\description{
Differentially expressed gene heatmap
}
\author{
Michael Steinbaugh
}
