% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-de.R
\name{plot_deg_heatmap}
\alias{plot_deg_heatmap}
\title{Differentially expressed gene heatmap}
\usage{
plot_deg_heatmap(run, res, dt, lfc = 0, intgroup = NULL)
}
\arguments{
\item{run}{\link{bcbioRnaDataSet}.}

\item{res}{\link{DESeqResults}.}

\item{dt}{\link{DESeqTransform}. Transformed counts generated by \code{\link[=rlog]{rlog()}}
(\strong{recommended}) or \code{\link[=vst]{vst()}} are supported.}

\item{lfc}{Log2 fold change ratio cutoff.}

\item{intgroup}{Interesting groups to label with bars above heatmap. If
\code{NULL}, defaults to \code{intgroup} defined in \link{bcbioRnaDataSet}.}
}
\value{
Gene heatmap.
}
\description{
Differentially expressed gene heatmap
}
\author{
Michael Steinbaugh
}
