% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot_deg_heatmap.R
\docType{methods}
\name{plot_deg_heatmap,bcbioRNAResults,missingOrNULL-method}
\alias{plot_deg_heatmap,bcbioRNAResults,missingOrNULL-method}
\alias{plot_deg_heatmap,DESeqResults,DESeqTransform-method}
\title{Differentially expressed gene heatmap}
\usage{
\S4method{plot_deg_heatmap}{bcbioRNAResults,missingOrNULL}(object, counts, lfc,
  title = NULL, ...)

\S4method{plot_deg_heatmap}{DESeqResults,DESeqTransform}(object, counts)
}
\arguments{
\item{object}{Primary object containing a list of DEGs.}

\item{counts}{Secondary object containing normalized counts.}

\item{lfc}{\link{log2} fold change ratio cutoff.}

\item{alpha}{Alpha level cutoff.}
}
\description{
This function is a simplified version of \code{\link[=plot_gene_heatmap]{plot_gene_heatmap()}} that is
optimized for handling a \link{DESeqResults} object rather a gene vector. All of
the optional parameters for \code{\link[=plot_gene_heatmap]{plot_gene_heatmap()}} are also available to this
function.
}
\seealso{
Other Heatmaps: \code{\link{plot_correlation_heatmap,bcbioRNADataSet-method}},
  \code{\link{plot_gene_heatmap,bcbioRNADataSet-method}}
}
\author{
Michael Steinbaugh
}
