% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_deg_heatmap.R
\name{plot_deg_heatmap}
\alias{plot_deg_heatmap}
\title{Differentially expressed gene (DEG) heatmap}
\usage{
plot_deg_heatmap(run, dds, contrast, alpha = 0.1, lfc = 0)
}
\arguments{
\item{run}{\code{bcbio-nextgen} run object}

\item{dds}{\code{DESeqDataSet}}

\item{contrast}{Contrast}

\item{alpha}{Alpha level cutoff}

\item{lfc}{Log fold change ratio (base 2) cutoff}
}
\value{
Gene heatmap
}
\description{
Differentially expressed gene (DEG) heatmap
}
\author{
Michael Steinbaugh
}
