% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_deseq_pca.R
\name{plot_deseq_pca}
\alias{plot_deseq_pca}
\title{Quickly plot principal component analysis (PCA)}
\usage{
plot_deseq_pca(dt, factor = c("group"), ...)
}
\arguments{
\item{dt}{\code{DESeqTransform} object generated from \code{rlog()} or
\code{vst()} on a \code{DESeqDataSet} object. We prefer to pipe the
transformed data in to this function because the operation is CPU
intensive, and is slow when applied to multiple PCA plot operations.}

\item{factor}{PCA factor groups for \code{DESeq2::plotPCA()}. Shapes are
defined by the first value.}

\item{...}{Passthrough to \code{DESeq2::plotPCA()}.}
}
\value{
PCA plot
}
\description{
Quickly plot principal component analysis (PCA)
}
\examples{
\dontrun{
plot_deseq_pca(rlog)
}
}
\author{
Michael Steinbaugh
}
