% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-plot_dispersion.R
\docType{methods}
\name{plot_dispersion}
\alias{plot_dispersion}
\alias{plot_dispersion,bcbioRNADataSet-method}
\title{Plot Dispersion Estimates}
\usage{
plot_dispersion(object)

\S4method{plot_dispersion}{bcbioRNADataSet}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Dispersion plot \link{ggplot}.
}
\description{
\code{\link[DESeq2:plotDispEsts]{DESeq2::plotDispEsts()}} wrapper supporting a \link{bcbioRNADataSet}.
}
\examples{
data(bcb)
plot_dispersion(bcb)
}
\seealso{
\code{\link[DESeq2:plotDispEsts]{DESeq2::plotDispEsts()}}.

Other DESeq2 Utilities: \code{\link{plot_mean_sd}}
}
\author{
Michael Steinbaugh
}
