% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot_dispersion.R
\docType{methods}
\name{plot_dispersion,bcbioRNADataSet-method}
\alias{plot_dispersion,bcbioRNADataSet-method}
\title{Plot dispersion estimates}
\usage{
\S4method{plot_dispersion}{bcbioRNADataSet}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Dispersion plot \link{ggplot}.
}
\description{
\code{\link[DESeq2:plotDispEsts]{DESeq2::plotDispEsts()}} wrapper supporting a \link{bcbioRNADataSet}.
}
\examples{
data(bcb)
plot_dispersion(bcb)
}
\seealso{
Other DESeq2 utilities: \code{\link{plot_mean_sd,bcbioRNADataSet-method}}
}
\author{
Michael Steinbaugh
}
