% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-deseq.R
\name{plot_dispersion}
\alias{plot_dispersion}
\title{Plot dispersion estimates}
\usage{
plot_dispersion(bcb)
}
\arguments{
\item{bcb}{\link{bcbioRNADataSet}.}
}
\value{
\link{ggplot}.
}
\description{
\code{\link[DESeq2:plotDispEsts]{DESeq2::plotDispEsts()}} wrapper supporting a \link{bcbioRNADataSet}.
}
\author{
Michael Steinbaugh
}
