% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exonic_mapping_rate.R
\name{plot_exonic_mapping_rate}
\alias{plot_exonic_mapping_rate}
\title{Exonic mapping rate plot}
\usage{
plot_exonic_mapping_rate(summary)
}
\arguments{
\item{summary}{\code{bcbio-rnaseq} summary report}
}
\description{
Exonic mapping rate plot
}
\examples{
\dontrun{
plot_exonic_mapping_rate(summary)
}
}
\author{
Michael Steinbaugh
}
