% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R,
%   R/methods-plot_exonic_mapping_rate.R
\docType{methods}
\name{plot_exonic_mapping_rate}
\alias{plot_exonic_mapping_rate}
\alias{.plot_exonic_mapping_rate}
\alias{plot_exonic_mapping_rate,bcbioRNADataSet-method}
\alias{plot_exonic_mapping_rate,data.frame-method}
\title{Plot Exonic Mapping Rate}
\usage{
plot_exonic_mapping_rate(object, ...)

.plot_exonic_mapping_rate(metrics, interesting_group = "sample_name",
  pass_limit = 60L, flip = TRUE)

\S4method{plot_exonic_mapping_rate}{bcbioRNADataSet}(object, ...)

\S4method{plot_exonic_mapping_rate}{data.frame}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{metrics}{Metrics \link{data.frame} generated by \code{\link[=metrics]{metrics()}}.}

\item{interesting_group}{\emph{Optional}. Category to use to group samples (color
and shape). If unset, this is automatically determined by the metadata set
inside the \link{bcbioRNADataSet}.}

\item{pass_limit}{Threshold to plot pass color marker.}

\item{flip}{Flip X and Y axes.}
}
\value{
\link{ggplot}.
}
\description{
Plot Exonic Mapping Rate
}
\examples{
data(bcb)
plot_exonic_mapping_rate(bcb)
}
\seealso{
Other Quality Control Plots: \code{\link{plot_5_3_bias}},
  \code{\link{plot_count_density}},
  \code{\link{plot_counts_per_gene}},
  \code{\link{plot_gene_detection_saturation}},
  \code{\link{plot_genes_detected}},
  \code{\link{plot_intronic_mapping_rate}},
  \code{\link{plot_mapped_reads}},
  \code{\link{plot_mapping_rate}},
  \code{\link{plot_rrna_mapping_rate}},
  \code{\link{plot_total_reads}}
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
