% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-plot_gender_markers.R
\docType{methods}
\name{plot_gender_markers}
\alias{plot_gender_markers}
\alias{plot_gender_markers,bcbioRNADataSet-method}
\title{Plot Sexually Dimorphic Gender Markers}
\usage{
plot_gender_markers(object, ...)

\S4method{plot_gender_markers}{bcbioRNADataSet}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\link{ggplot}.
}
\description{
Plot Sexually Dimorphic Gender Markers
}
\examples{
data(bcb)
plot_gender_markers(bcb)
}
\author{
Michael Steinbaugh
}
