% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gender_markers.R
\name{plot_gender_markers}
\alias{plot_gender_markers}
\title{Graph sexually dimorphic marker genes}
\usage{
plot_gender_markers(bcbio, counts)
}
\arguments{
\item{bcbio}{bcbio list object}

\item{counts}{Counts matrix. Transcripts per million (TPM) or normalized
counts are preferred.}
}
\value{
Scatterplot
}
\description{
Graph sexually dimorphic marker genes
}
\examples{
\dontrun{
plot_gender_markers(bcbio, tpm)
}
}
\author{
Michael Steinbaugh
}
