% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot_gender_markers.R
\docType{methods}
\name{plot_gender_markers,bcbioRNADataSet-method}
\alias{plot_gender_markers,bcbioRNADataSet-method}
\title{Plot sexually dimorphic gender markers}
\usage{
\S4method{plot_gender_markers}{bcbioRNADataSet}(object)
}
\arguments{
\item{object}{Object.}
}
\description{
Plot sexually dimorphic gender markers
}
\author{
Michael Steinbaugh
}
