% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-gene.R
\name{plot_gender_markers}
\alias{plot_gender_markers}
\title{Plot sexually dimorphic gender markers}
\usage{
plot_gender_markers(bcb)
}
\arguments{
\item{bcb}{\link{bcbioRNADataSet}.}
}
\description{
Plot sexually dimorphic gender markers
}
