% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot_gene.R
\docType{methods}
\name{plot_gene,bcbioRNADataSet,character-method}
\alias{plot_gene,bcbioRNADataSet,character-method}
\title{Plot an individual gene}
\usage{
\S4method{plot_gene}{bcbioRNADataSet,character}(object, keys,
  format = "symbol")
}
\arguments{
\item{object}{Object.}

\item{format}{Ensembl identifier format. Defaults to the gene symbol (a.k.a.
\code{external_gene_name}).}

\item{gene}{Gene identifier. Can input multiple genes as a character vector.}
}
\description{
Plot an individual gene
}
\author{
Michael Steinbaugh
}
