% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene.R
\name{plot_gene}
\alias{plot_gene}
\title{Plot an individual gene}
\usage{
plot_gene(bcbio, counts, gene)
}
\arguments{
\item{bcbio}{bcbio list object}

\item{counts}{Counts matrix}

\item{gene}{Gene identifier}
}
\description{
Plot an individual gene
}
\author{
Michael Steinbaugh
}
