% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-gene.R
\name{plot_gene}
\alias{plot_gene}
\title{Plot an individual gene}
\usage{
plot_gene(bcb, gene, format = "symbol")
}
\arguments{
\item{bcb}{\link{bcbioRnaDataSet}.}

\item{gene}{Gene identifier. Can input multiple genes as a character vector.}

\item{format}{Ensembl identifier format. Defaults to the gene symbol (a.k.a.
\code{external_gene_name}).}
}
\description{
Plot an individual gene
}
\author{
Michael Steinbaugh
}
