% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene.R
\name{plot_gene}
\alias{plot_gene}
\title{Plot an individual gene}
\usage{
plot_gene(gene, counts, metadata)
}
\arguments{
\item{gene}{Gene identifier}

\item{counts}{Counts matrix}

\item{metadata}{Metadata data frame}
}
\description{
Plot an individual gene
}
\author{
Michael Steinbaugh
}
