% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene.R
\name{plot_gene}
\alias{plot_gene}
\title{Plot an individual gene.}
\usage{
plot_gene(run, normalized_counts, gene, format = "external_gene_name")
}
\arguments{
\item{run}{bcbio-nextgen run.}

\item{normalized_counts}{Normalized counts matrix. Can be obtained from
DESeqDataSet by running \code{counts(normalized = TRUE)}. Transcripts per
million (TPM) are also acceptable.}

\item{gene}{Gene identifier. Can input multiple genes as a character vector.}

\item{format}{Ensembl identifier format. Defaults to the gene name (symbol).}
}
\description{
Plot an individual gene.
}
\author{
Michael Steinbaugh
}
