% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R,
%   R/methods-plot_gene_detection_saturation.R
\docType{methods}
\name{plot_gene_detection_saturation}
\alias{plot_gene_detection_saturation}
\alias{.plot_gene_detection_saturation}
\alias{plot_gene_detection_saturation,bcbioRNADataSet-method}
\title{Plot Gene Detection Saturation}
\usage{
plot_gene_detection_saturation(object, ...)

.plot_gene_detection_saturation(metrics, counts,
  interesting_group = "sample_name", min_counts = 0L)

\S4method{plot_gene_detection_saturation}{bcbioRNADataSet}(object,
  normalized = "tmm", ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{metrics}{Metrics \link{data.frame} generated by \code{\link[=metrics]{metrics()}}.}

\item{counts}{Counts matrix.}

\item{interesting_group}{\emph{Optional}. Category to use to group samples (color
and shape). If unset, this is automatically determined by the metadata set
inside the \link{bcbioRNADataSet}.}

\item{min_counts}{Numeric value for filtering the counts matrix before
plotting.}

\item{normalized}{Count normalization method. See \code{\link[=counts]{counts()}} documentation
for more information.}
}
\value{
\link{ggplot}.
}
\description{
Plot Gene Detection Saturation
}
\examples{
data(bcb)
plot_gene_detection_saturation(bcb)
}
\seealso{
Other Quality Control Plots: \code{\link{.plot_rrna_mapping_rate}},
  \code{\link{plot_count_density}},
  \code{\link{plot_counts_per_gene}},
  \code{\link{plot_exonic_mapping_rate}},
  \code{\link{plot_genes_detected}},
  \code{\link{plot_intronic_mapping_rate}},
  \code{\link{plot_mapped_reads}},
  \code{\link{plot_mapping_rate}},
  \code{\link{plot_total_reads}}
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
