% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_detection_saturation.R
\name{plot_gene_detection_saturation}
\alias{plot_gene_detection_saturation}
\title{Genes detection saturation plot}
\usage{
plot_gene_detection_saturation(counts, summary)
}
\arguments{
\item{counts}{Counts matrix}

\item{summary}{\code{bcbio-rnaseq} summary report}
}
\description{
Genes detection saturation plot
}
\examples{
\dontrun{
plot_gene_detection_saturation(counts, summary)
}
}
\author{
Michael Steinbaugh
}
