% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-heatmap.R
\name{plot_gene_heatmap}
\alias{plot_gene_heatmap}
\alias{plot_deg_heatmap}
\title{Gene heatmap}
\usage{
plot_gene_heatmap(bcb, dt = NULL, genes, clustering_method = "ward.D2",
  cluster_rows = TRUE, cluster_cols = TRUE, scale = "row",
  interesting_groups = NULL, annotation = NULL, title = NULL, ...)

plot_deg_heatmap(bcb, dt, res, lfc, title = NULL, ...)
}
\arguments{
\item{bcb}{\link{bcbioRNADataSet}.}

\item{dt}{\link{DESeqTransform}. Transformed counts generated by \code{\link[=rlog]{rlog()}}
(\strong{recommended}) or \code{\link[=vst]{vst()}} on a \link{DESeqDataSet} are supported.}

\item{genes}{Character vector of specific gene identifiers to plot.}

\item{clustering_method}{Hierarchical clustering method. Accepts the same
values as \code{\link[stats:hclust]{stats::hclust()}}.}

\item{cluster_rows}{Use hierarchical clustering to arrange rows.}

\item{cluster_cols}{Use hierarchical clustering to arrange columns.}

\item{scale}{Character indicating if the values should be centered and scaled
in either the \code{row} direction, \code{column} direction, or \code{none}.}

\item{interesting_groups}{\emph{Optional}. Interesting groups to label with bars
above heatmap. If \code{NULL}, defaults to \code{interesting_groups} defined in the
\link{bcbioRNADataSet}.}

\item{annotation}{\emph{Optional}. Alternative annotation to use. Useful when
plotting more than one column.}

\item{title}{\emph{Optional}. Text to include in plot title.}

\item{...}{Additional arguments.}

\item{res}{\link{DESeqResults}.}

\item{lfc}{\link{log2} fold change ratio.}
}
\description{
These functions facilitate heatmap plotting of a specified set of genes. By
default, row- and column-wise hierarchical clustering is performed using the
Ward method, but this behavior can be overrided by setting \code{cluster_rows} or
\code{cluster_cols} to \code{FALSE}. When column clustering is disabled, the columns
are sorted by the interesting groups (\code{interesting_groups}) specified in the
\link{bcbioRNADataSet} and then the sample names.
}
\details{
\code{\link[=plot_deg_heatmap]{plot_deg_heatmap()}} is a simplified version of
\code{\link[=plot_gene_heatmap]{plot_gene_heatmap()}} that is optimized for handling a \link{DESeqResults}
object rather a gene vector. All of the optional parameters for
\code{\link[=plot_gene_heatmap]{plot_gene_heatmap()}} are also available to this function.
}
\seealso{
\code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}.
}
\author{
Michael Steinbaugh
}
