% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genes_detected.R
\name{plot_genes_detected}
\alias{plot_genes_detected}
\title{Genes detected plot}
\usage{
plot_genes_detected(summary, counts)
}
\arguments{
\item{summary}{\code{bcbio-rnaseq} summary report}

\item{counts}{Raw counts matrix}
}
\description{
Genes detected plot
}
\examples{
\dontrun{
plot_genes_detected(summary, raw_counts)
}
}
\author{
Michael Steinbaugh
}
