% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-plot_genes_detected.R
\docType{methods}
\name{plot_genes_detected}
\alias{plot_genes_detected}
\alias{.plot_genes_detected}
\alias{plot_genes_detected,bcbioRNADataSet-method}
\title{Plot Genes Detected}
\usage{
plot_genes_detected(object, ...)

.plot_genes_detected(metrics, counts, interesting_group = "sample_name",
  pass_limit = 20000L, warn_limit = 15000L, min_counts = 0L,
  flip = TRUE)

\S4method{plot_genes_detected}{bcbioRNADataSet}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{metrics}{Metrics \link{data.frame} generated by \code{\link[=metrics]{metrics()}}.}

\item{counts}{Counts matrix.}

\item{interesting_group}{\emph{Optional}. Category to use to group samples (color
and shape). If unset, this is automatically determined by the metadata set
inside the \link{bcbioRNADataSet}.}

\item{pass_limit}{Threshold to plot pass color marker.}

\item{warn_limit}{Threshold to plot warning color marker.}

\item{min_counts}{Numeric value for filtering the counts matrix before
plotting.}

\item{flip}{Flip X and Y axes.}
}
\value{
\link{ggplot}.
}
\description{
Plot Genes Detected
}
\examples{
data(bcb)
plot_genes_detected(bcb, pass_limit = NULL)
}
\seealso{
Other Quality Control Plots: \code{\link{plot_count_density}},
  \code{\link{plot_counts_per_gene}},
  \code{\link{plot_exonic_mapping_rate}},
  \code{\link{plot_gene_detection_saturation}},
  \code{\link{plot_intronic_mapping_rate}},
  \code{\link{plot_mapped_reads}},
  \code{\link{plot_mapping_rate}},
  \code{\link{plot_rrna_mapping_rate}},
  \code{\link{plot_total_reads}}
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
