% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_intronic_mapping_rate.R
\name{plot_intronic_mapping_rate}
\alias{plot_intronic_mapping_rate}
\title{Intronic mapping rate plot}
\usage{
plot_intronic_mapping_rate(bcbio)
}
\arguments{
\item{bcbio}{bcbio list object}
}
\description{
Intronic mapping rate plot
}
\examples{
\dontrun{
plot_intronic_mapping_rate(bcbio)
}
}
\author{
Michael Steinbaugh
}
