% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-de.R
\name{plot_ma}
\alias{plot_ma}
\title{MA-plot from base means and log fold changes}
\usage{
plot_ma(res, ylim = 2L)
}
\arguments{
\item{res}{\link{DESeqResults}.}

\item{ylim}{Y-axis maximum (single integer).}
}
\value{
Mean average (MA) \link{ggplot}.
}
\description{
\code{\link[DESeq2:plotMA]{DESeq2::plotMA()}} wrapper that generates a title automatically.
}
\author{
Michael Steinbaugh
}
