% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-de.R
\name{plot_ma}
\alias{plot_ma}
\title{MA-plot from base means and log fold changes}
\usage{
plot_ma(res, ylim = 2L, title = NULL)
}
\arguments{
\item{res}{\link{DESeqResults}.}

\item{ylim}{Y-axis maximum (single integer).}

\item{title}{\emph{Optional}. Plot title.}
}
\value{
Mean average (MA) \link{ggplot}.
}
\description{
\code{\link[DESeq2:plotMA]{DESeq2::plotMA()}} wrapper that generates a title from the contrast
automatically.
}
\author{
Michael Steinbaugh
}
