% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-plot_ma.R
\docType{methods}
\name{plot_ma}
\alias{plot_ma}
\alias{.plot_ma}
\alias{plot_ma,DESeqResults-method}
\alias{plot_ma,data.frame-method}
\title{Plot Mean Average}
\usage{
plot_ma(object, ...)

\S4method{plot_ma}{DESeqResults}(object, label_points = NULL)

\S4method{plot_ma}{data.frame}(object, title = NULL, label_points = NULL,
  label_column = "symbol")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{label_points}{\emph{Optional}. Label these particular points.}

\item{label_column}{Match \code{label_points} argument to this column in the
results.}

\item{title}{\emph{Optional}. Plot title.}
}
\value{
\link{ggplot}.
}
\description{
Plot Mean Average
}
\examples{
data(bcb)
dds <- DESeqDataSetFromTximport(
    txi = txi(bcb),
    colData = colData(bcb),
    design = formula(~group)) \%>\%
    DESeq
res <- results(dds)
plot_ma(res)
}
\seealso{
Other Differential Expression Plots: \code{\link{plot_volcano}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
