% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-de.R
\name{plot_ma}
\alias{plot_ma}
\title{MA-plot from base means and log fold changes}
\usage{
plot_ma(res, title = NULL, label_points = NULL, label_column = "symbol")
}
\arguments{
\item{res}{\link{DESeqResults} or a data frame of DESeqResults.}

\item{title}{\emph{Optional}. Plot title.}

\item{label_points}{Optionally label these particular points}

\item{label_column}{Match label_points to this column in the results}
}
\value{
MA-plot \link{ggplot}.
}
\description{
MA-plot
}
\examples{
data(dummy)
library(DESeq2)
dds = DESeqDataSetFromTximport(bcbio(bcb, "tximport"),
colData(bcb),design=~group)
dds=DESeq(dds)
res = results(dds)
plot_ma(res)
}
\author{
Michael Steinbaugh

Rory Kirchner
}
