% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-de.R
\name{plot_ma}
\alias{plot_ma}
\title{MA-plot from base means and log fold changes}
\usage{
plot_ma(res, title = NULL, label_points = NULL, label_column = "symbol")
}
\arguments{
\item{res}{\link{DESeqResults} or a modified \link{data.frame} derived from
\link{DESeqResults}.}

\item{title}{\emph{Optional}. Plot title.}

\item{label_points}{\emph{Optional}. Label these particular points.}

\item{label_column}{Match label_points to this column in the results.}
}
\value{
\link{ggplot}.
}
\description{
MA-plot from base means and log fold changes
}
\examples{
data(dummy)
dds <- DESeqDataSetFromTximport(
    txi = txi(bcb),
    colData = colData(bcb),
    design = formula(~group))
dds <- DESeq(dds)
res <- results(dds)
plot_ma(res)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
