% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot_ma.R
\docType{methods}
\name{.plot_ma}
\alias{.plot_ma}
\alias{plot_ma,DESeqResults-method}
\alias{plot_ma,data.frame-method}
\title{MA-plot from base means and log fold changes}
\usage{
\S4method{plot_ma}{DESeqResults}(object, title = NULL, label_points = NULL,
  label_column = "symbol")

\S4method{plot_ma}{data.frame}(object, title = NULL, label_points = NULL,
  label_column = "symbol")
}
\arguments{
\item{title}{\emph{Optional}. Plot title.}

\item{label_points}{\emph{Optional}. Label these particular points.}

\item{label_column}{Match \code{label_points} argument to this column in the
results.}

\item{object}{Object.}
}
\value{
\link{ggplot}.
}
\description{
MA-plot from base means and log fold changes
}
\examples{
data(bcb)
dds <- DESeqDataSetFromTximport(
    txi = txi(bcb),
    colData = colData(bcb),
    design = formula(~group)) \%>\%
    DESeq
res <- results(dds)
plot_ma(res)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
