% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-plot_ma.R
\docType{methods}
\name{plot_ma}
\alias{plot_ma}
\alias{.plot_ma}
\alias{plot_ma,DESeqResults-method}
\alias{plot_ma,data.frame-method}
\title{Plot Mean Average}
\usage{
plot_ma(object, ...)

\S4method{plot_ma}{DESeqResults}(object, ...)

\S4method{plot_ma}{data.frame}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{label_points}{\emph{Optional}. Label these particular points.}

\item{label_column}{Match \code{label_points} argument to this column in the
results.}

\item{point_color_scale}{Point color scale. See \link[ggplot2:color_manual]{ggplot2::color_manual} for
more information.}

\item{label_color}{Text label color.}

\item{title}{\emph{Optional}. Plot title.}
}
\value{
\link{ggplot}.
}
\description{
Plot Mean Average
}
\examples{
data(bcb)
dds <- DESeqDataSetFromTximport(
    txi = txi(bcb),
    colData = colData(bcb),
    design = formula(~group)) \%>\%
    DESeq
res <- results(dds)

genes <- c("ENSMUSG00000104523", "ENSMUSG00000016918")

# DESeqResults
plot_ma(res, label_points = genes)

# data.frame
res_df <- as.data.frame(res)
plot_ma(res_df, label_points = genes)
}
\seealso{
Other Differential Expression Plots: \code{\link{plot_volcano}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
