% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mapped_reads.R
\name{plot_mapped_reads}
\alias{plot_mapped_reads}
\title{Mapped reads plot}
\usage{
plot_mapped_reads(summary)
}
\arguments{
\item{summary}{\code{bcbio-rnaseq} summary report}
}
\description{
Mapped reads plot
}
\examples{
\dontrun{
plot_mapped_reads(summary)
}
}
\author{
Michael Steinbaugh
}
