% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot_mean_sd.R
\docType{methods}
\name{plot_mean_sd,bcbioRNADataSet-method}
\alias{plot_mean_sd,bcbioRNADataSet-method}
\title{Plot row standard deviations versus row means}
\usage{
\S4method{plot_mean_sd}{bcbioRNADataSet}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\link{ggplot} grid.
}
\description{
\code{\link[vsn:meanSdPlot]{vsn::meanSdPlot()}} wrapper that plots \code{\link[=log2]{log2()}}, \code{\link[=rlog]{rlog()}}, and
\code{\link[=varianceStabilizingTransformation]{varianceStabilizingTransformation()}} normalized counts.
}
\examples{
data(bcb)
plot_mean_sd(bcb)
}
\seealso{
Other DESeq2 utilities: \code{\link{plot_dispersion,bcbioRNADataSet-method}}
}
\author{
Michael Steinbaugh, Lorena Patano
}
