% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-deseq.R
\name{plot_mean_sd}
\alias{plot_mean_sd}
\title{Plot row standard deviations versus row means}
\usage{
plot_mean_sd(bcb)
}
\arguments{
\item{bcb}{\link{bcbioRNADataSet}.}
}
\value{
ggplot grid.
}
\description{
\code{\link[vsn:meanSdPlot]{vsn::meanSdPlot()}} wrapper that plots \code{\link[=log2]{log2()}}, \code{\link[=rlog]{rlog()}}, and
\code{\link[=varianceStabilizingTransformation]{varianceStabilizingTransformation()}} normalized counts.
}
\author{
Michael Steinbaugh, Lorena Patano
}
