% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-DESeq2.R
\name{plot_mean_sd}
\alias{plot_mean_sd}
\title{Plot row standard deviations versus row means}
\usage{
plot_mean_sd(dds, rld = NULL, vsd = NULL)
}
\arguments{
\item{dds}{\link{DESeqDataSet}.}

\item{rld}{\link{DESeqTransform} resulting from \code{\link[=rlog]{rlog()}} applied to
\link{DESeqDataSet}.}

\item{vsd}{\link{DESeqTransform} resulting from
\code{\link[=varianceStabilizingTransformation]{varianceStabilizingTransformation()}} applied to \link{DESeqDataSet}.}
}
\description{
Modified wrapper function of \code{\link[vsn:meanSdPlot]{vsn::meanSdPlot()}} that plots \code{\link[=log2]{log2()}},
\code{\link[=rlog]{rlog()}}, and \code{\link[=varianceStabilizingTransformation]{varianceStabilizingTransformation()}} count data.
}
