% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot_mirna_counts.R
\docType{methods}
\name{plot_mirna_counts,bcbioRNADataSet-method}
\alias{plot_mirna_counts,bcbioRNADataSet-method}
\title{Plot of total miRNA counts}
\usage{
\S4method{plot_mirna_counts}{bcbioRNADataSet}(object)
}
\value{
\link{ggplot}.
}
\description{
Plot of total miRNA counts
}
\author{
Lorena Pantano
}
