% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-de.R
\name{plot_pattern}
\alias{plot_pattern}
\title{\code{\link[DEGreport:degPatterns]{DEGreport::degPatterns()}} wrapper}
\usage{
plot_pattern(bcb, res, fdr = 0.1, n = NULL, ...)
}
\arguments{
\item{bcb}{\link{bcbioRNADataSet} object.}

\item{res}{Table with padj as column and gene names as \link{rownames}.}

\item{fdr}{Float cutoff to consider genes significant.}

\item{n}{Integer maximum number of genes to consider.}

\item{...}{Additional parameters.}
}
\description{
Adds support for a \link{bcbioRNADataSet} object.
}
\author{
Lorena Pantano
}
