% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot_pattern.R
\docType{methods}
\name{plot_pattern,bcbioRNADataSet-method}
\alias{plot_pattern,bcbioRNADataSet-method}
\title{Make groups of genes using expression profile}
\usage{
\S4method{plot_pattern}{bcbioRNADataSet}(object, res, fdr = 0.1, n = NULL,
  ...)
}
\arguments{
\item{object}{Object.}

\item{res}{Table with padj as column and gene names as \link{rownames}.}

\item{fdr}{Float cutoff to consider genes significant.}

\item{n}{Integer maximum number of genes to consider.}

\item{...}{Additional parameters, passed to \code{\link[DEGreport:degPatterns]{DEGreport::degPatterns()}}.}
}
\value{
\link{ggplot}.
}
\description{
\code{\link[DEGreport:degPatterns]{DEGreport::degPatterns()}} wrapper supporting a \link{bcbioRNADataSet}.
}
\author{
Lorena Pantano, Michael Steinbaugh
}
