% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-de.R
\name{plot_pattern}
\alias{plot_pattern}
\title{Wrap degPatter function around bcbio object}
\usage{
plot_pattern(bcb, res, fdr = 0.1, n = NULL, ...)
}
\arguments{
\item{bcb}{\link{bcbioRnaseq} object}

\item{res}{table with padj as column and gene names as row.names}

\item{fdr}{float cutoff to consider genes significant}

\item{n}{integer maximum number of genes to consider}

\item{...}{extra parameters for degPatterns function}
}
\description{
It runs \link[DEGreport:degPatterns]{DEGreport::degPatterns} function from a bcbio object
}
\author{
Lorena Pantano
}
