% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot PCA}
\usage{
plot_pca(run, dt, label = TRUE)
}
\arguments{
\item{run}{bcbio-nextgen run}

\item{dt}{\code{DESeqTransform} generated from \code{rlog()} or \code{vst()}
on a \code{DESeqDataSet} object}

\item{label}{Superimpose sample text labels on the plot}
}
\value{
PCA plot
}
\description{
Wrapper for \code{DESeq2::plotPCA()} that improves PCA sample coloring and
labeling
}
