% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot principal component analysis (PCA)}
\usage{
plot_pca(run, dt, label = TRUE, ntop = 500)
}
\arguments{
\item{run}{\link{bcbioRnaDataSet}.}

\item{dt}{\link{DESeqTransform} generated from \code{\link[=rlog]{rlog()}} (\strong{recommended}) or
\code{\link[=vst]{vst()}} on a \link{DESeqDataSet}.}

\item{label}{Superimpose sample text labels on the plot.}

\item{ntop}{Number of genes to use for PCA. To plot all gene variation, use
\code{ntop = nrow(.)}. Note that using all genes for PCA is generally \emph{not
recommended}.}
}
\value{
\link{ggplot}.
}
\description{
Wrapper for \code{\link[DESeq2:plotPCA]{DESeq2::plotPCA()}} that improves PCA sample coloring and
labeling.
}
\seealso{
\itemize{
\item \code{\link[DESeq2:plotPCA]{DESeq2::plotPCA()}}.
\item \href{https://support.bioconductor.org/p/51270/}{Bioconductor thread on ntop usage}.
}
}
