% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-pca.R
\name{plot_pca}
\alias{plot_pca}
\title{Sample PCA plot for transformed data}
\usage{
plot_pca(bcb, dt, genes = NULL, interesting_groups = NULL, shape = FALSE,
  label = TRUE)
}
\arguments{
\item{bcb}{\link{bcbioRNADataSet}.}

\item{dt}{\link{DESeqTransform} generated from \code{\link[=rlog]{rlog()}} (\strong{recommended}) or
\code{\link[=vst]{vst()}} on a \link{DESeqDataSet}.}

\item{genes}{\emph{Optional}. Character vector of gene identifiers to use.}

\item{interesting_groups}{\emph{Optional}. Interesting groups to use for point
appearance. If \code{NULL}, color defaults to all \code{interesting_groups}
parameters set in \link{bcbioRNADataSet}.}

\item{shape}{\emph{Optional}. Make points easier to inspect with differing shapes.}

\item{label}{\emph{Optional}. Superimpose sample text labels on the plot.}
}
\value{
\link{ggplot}.
}
\description{
Wrapper for \code{\link[DESeq2:plotPCA]{DESeq2::plotPCA()}} that improves principal component analysis
(PCA) sample coloring and labeling.
}
\seealso{
\code{\link[DESeq2:plotPCA]{DESeq2::plotPCA()}}.
}
