% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-plot_pca.R
\docType{methods}
\name{plot_pca}
\alias{plot_pca}
\alias{plot_pca,bcbioRNADataSet-method}
\title{Sample PCA Plot for Transformed Data}
\usage{
plot_pca(object, ...)

\S4method{plot_pca}{bcbioRNADataSet}(object, transform = "rlog",
  genes = NULL, censor_samples = NULL, interesting_groups = NULL,
  shape = FALSE, label = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{transform}{String specifying \link{rlog} (\strong{recommended}) or \link{vst}
\link{DESeqTransform} slotted inside the \link{bcbioRNADataSet}.}

\item{genes}{\emph{Optional}. Character vector of gene identifiers to use.}

\item{censor_samples}{\emph{Optional}. Censors to exclude from PCA plot.}

\item{interesting_groups}{\emph{Optional}. Interesting groups to use for point
appearance. If \code{NULL}, color defaults to all \code{interesting_groups}
parameters set in \link{bcbioRNADataSet}.}

\item{shape}{\emph{Optional}. Make points easier to inspect with differing shapes.}

\item{label}{\emph{Optional}. Superimpose sample text labels on the plot.}
}
\value{
\link{ggplot}.
}
\description{
Wrapper for \code{\link[DESeq2:plotPCA]{DESeq2::plotPCA()}} that improves principal component analysis
(PCA) sample coloring and labeling.
}
\examples{
data(bcb)
plot_pca(bcb)
}
\seealso{
\code{\link[DESeq2:plotPCA]{DESeq2::plotPCA()}}.
}
\author{
Michael Steinbaugh
}
