% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot_pca_covariates.R
\docType{methods}
\name{plot_pca_covariates,bcbioRNADataSet-method}
\alias{plot_pca_covariates,bcbioRNADataSet-method}
\title{Find correlation between principal components (PCs) and covariates.}
\usage{
\S4method{plot_pca_covariates}{bcbioRNADataSet}(object, transform = "rlog",
  use = NULL, ...)
}
\arguments{
\item{object}{Object.}

\item{transform}{String specifying \link{rlog} (\strong{recommended}) or \link{vst}
\link{DESeqTransform} slotted inside the \link{bcbioRNADataSet}.}

\item{use}{\emph{Optional}. Character vector of columns to use.}

\item{...}{Passthrough arguments to \code{\link[DEGreport:degCovariates]{DEGreport::degCovariates()}}.}
}
\value{
\link{ggplot}.
}
\description{
\code{\link[DEGreport:degCovariates]{DEGreport::degCovariates()}} wrapper supporting a \link{bcbioRNADataSet}.
}
\author{
Lorena Pantano, Michael Steinbaugh
}
