% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-pca.R
\name{plot_pca_covariates}
\alias{plot_pca_covariates}
\title{Find correlation between principal components (PCs) and covariates.}
\usage{
plot_pca_covariates(bcb, transform = "rlog", use = NULL, ...)
}
\arguments{
\item{bcb}{\link{bcbioRNADataSet}.}

\item{transform}{String specifying \link{rlog} (\strong{recommended}) or \link{vst}
\link{DESeqTransform} slotted inside the \link{bcbioRNADataSet}.}

\item{use}{\emph{Optional}. Character vector of columns to use.}

\item{...}{Passthrough arguments to \code{\link[DEGreport:degCovariates]{DEGreport::degCovariates()}}.}
}
\value{
\link{ggplot}.
}
\description{
\code{\link[DEGreport:degCovariates]{DEGreport::degCovariates()}} wrapper supporting a \link{bcbioRNADataSet}.
}
\author{
Lorena Pantano
}
