% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-pca.R
\name{plot_pca_covariates}
\alias{plot_pca_covariates}
\title{Find correlation between principal components (PCs) and covariates.}
\usage{
plot_pca_covariates(bcb, dt, use = NULL, ...)
}
\arguments{
\item{bcb}{\link{bcbioRNADataSet}.}

\item{dt}{\link{DESeqTransform}. \code{\link[=rlog]{rlog()}}-transformed counts are recommended.}

\item{use}{\emph{Optional}. Character vector of columns to use.}

\item{...}{Passthrough arguments to \code{\link[DEGreport:degCovariates]{DEGreport::degCovariates()}}.}
}
\description{
Wrapper for \code{\link[DEGreport:degCovariates]{DEGreport::degCovariates()}} supporting a \link{bcbioRNADataSet}.
}
\author{
Lorena Pantano
}
