% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-qc.R
\name{plot_pca_covariates}
\alias{plot_pca_covariates}
\title{Find correlation between PCs and covariates}
\usage{
plot_pca_covariates(run, dt, ...)
}
\arguments{
\item{run}{\link{bcbioRnaDataSet}.}

\item{dt}{\link{DESeqTransform}. \code{\link[=rlog]{rlog()}}-transformed counts are recommended.}

\item{...}{Passthrough parmeters to \code{\link[DEGreport:degCovariates]{DEGreport::degCovariates()}}.}
}
\description{
Find correlation between PCs and covariates
}
\author{
Lorena Pantano
}
