% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-pca.R
\name{plot_pca_covariates}
\alias{plot_pca_covariates}
\title{Find correlation between PCs and covariates}
\usage{
plot_pca_covariates(bcb, dt, use = NULL, ...)
}
\arguments{
\item{bcb}{\link{bcbioRnaDataSet}.}

\item{dt}{\link{DESeqTransform}. \code{\link[=rlog]{rlog()}}-transformed counts are recommended.}

\item{use}{Character vector. List of columns to use in degCovariates.}

\item{...}{Passthrough parmeters to \code{\link[DEGreport:degCovariates]{DEGreport::degCovariates()}}.}
}
\description{
Find correlation between PCs and covariates
}
\author{
Lorena Pantano
}
