% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-plot_pca_covariates.R
\docType{methods}
\name{plot_pca_covariates}
\alias{plot_pca_covariates}
\alias{plot_pca_covariates,bcbioRNADataSet-method}
\title{Find Correlation Between Principal Components (PCs) and Covariates}
\usage{
plot_pca_covariates(object, ...)

\S4method{plot_pca_covariates}{bcbioRNADataSet}(object, transform = "rlog",
  metrics = TRUE, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments, passed to \code{\link[=degCovariates]{degCovariates()}}.}

\item{transform}{String specifying \link{DESeqTransform} slotted inside the
\link{bcbioRNADataSet}:
\itemize{
\item \code{rlog} (\strong{recommended}).
\item \code{vst}: variance stabilizing transformation.
}}

\item{metrics}{Includesample summary metrics as covariates.}
}
\value{
\link{ggplot}.
}
\description{
\code{\link[DEGreport:degCovariates]{DEGreport::degCovariates()}} wrapper supporting a \link{bcbioRNADataSet}.
}
\examples{
data(bcb)
plot_pca_covariates(bcb)
}
\seealso{
\itemize{
\item \code{\link[DEGreport:degCovariates]{DEGreport::degCovariates()}}.
\item \code{\link[DESeq2:rlog]{DESeq2::rlog()}}.
\item \code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}}.
}
}
\author{
Lorena Pantano, Michael Steinbaugh
}
