% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rrna_mapping_rate.R
\name{plot_rrna_mapping_rate}
\alias{plot_rrna_mapping_rate}
\title{rRNA contamination mapping rate}
\usage{
plot_rrna_mapping_rate(summary)
}
\arguments{
\item{summary}{\code{bcbio-rnaseq} summary report}
}
\description{
rRNA contamination mapping rate
}
\examples{
\dontrun{
plot_rrna_mapping_rate(summary)
}
}
\author{
Michael Steinbaugh
}
