% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot_rrna_mapping_rate.R
\docType{methods}
\name{.plot_rrna_mapping_rate}
\alias{.plot_rrna_mapping_rate}
\alias{plot_rrna_mapping_rate,bcbioRNADataSet-method}
\alias{plot_rrna_mapping_rate,data.frame-method}
\title{Plot rRNA Mapping Rate}
\usage{
.plot_rrna_mapping_rate(metrics, interesting_group = "sample_name",
  warn_limit = 10L, flip = TRUE)

\S4method{plot_rrna_mapping_rate}{bcbioRNADataSet}(object, ...)

\S4method{plot_rrna_mapping_rate}{data.frame}(object, ...)
}
\arguments{
\item{metrics}{Metrics \link{data.frame} generated by \code{\link[=metrics]{metrics()}}.}

\item{interesting_group}{\emph{Optional}. Category to use to group samples (color
and shape). If unset, this is automatically determined by the metadata set
inside the \link{bcbioRNADataSet}.}

\item{warn_limit}{Threshold to plot warning color marker.}

\item{flip}{Flip X and Y axes.}
}
\value{
\link{ggplot}.
}
\description{
Plot rRNA Mapping Rate
}
\examples{
data(bcb)

# bcbioRNADataSet
plot_rrna_mapping_rate(bcb)

# data.frame
metrics <- metrics(bcb)
plot_rrna_mapping_rate(metrics)
}
\seealso{
Other Quality Control Plots: \code{\link{plot_count_density}},
  \code{\link{plot_counts_per_gene}},
  \code{\link{plot_exonic_mapping_rate}},
  \code{\link{plot_gene_detection_saturation}},
  \code{\link{plot_genes_detected}},
  \code{\link{plot_intronic_mapping_rate}},
  \code{\link{plot_mapped_reads}},
  \code{\link{plot_mapping_rate}},
  \code{\link{plot_total_reads}}
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
