% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot_size_distribution.R
\docType{methods}
\name{plot_size_distribution,bcbioRNADataSet-method}
\alias{plot_size_distribution,bcbioRNADataSet-method}
\title{Plot size distribution of small RNA-seq data}
\usage{
\S4method{plot_size_distribution}{bcbioRNADataSet}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\link{ggplot}.
}
\description{
Plot size distribution of small RNA-seq data
}
\seealso{
Other Small RNA-seq utilities: \code{\link{plot_srna_clusters}}
}
\author{
Lorena Pantano, Michael Steinbaugh
}
