% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{plot_size_distribution}
\alias{plot_size_distribution}
\title{Plot size distribution of small RNAseq data}
\usage{
plot_size_distribution(run)
}
\arguments{
\item{run}{\code{bcbio-nextgen} run}
}
\value{
ggplot figure
}
\description{
Plot size distribution of small RNAseq data
}
\author{
Lorena Pantano
}
