% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-plot_srna_clusters.R
\docType{methods}
\name{plot_srna_clusters}
\alias{plot_srna_clusters}
\alias{plot_srna_clusters,bcbioRNADataSet-method}
\title{Cluster Small RNA Samples}
\usage{
plot_srna_clusters(object, ...)

\S4method{plot_srna_clusters}{bcbioRNADataSet}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\link{ggplot}.
}
\description{
Cluster Small RNA Samples
}
\examples{
data(bcb)
plot_srna_clusters(bcb)
}
\seealso{
Other Small RNA-Seq Utilities: \code{\link{plot_size_distribution}}
}
\author{
Lorena Pantano, Michael Steinbaugh
}
