% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot_srna_clusters.R
\name{plot_srna_clusters}
\alias{plot_srna_clusters}
\title{Cluster Small RNA samples}
\usage{
plot_srna_clusters(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\link{ggplot}.
}
\description{
Cluster Small RNA samples
}
\seealso{
Other Small RNA-Seq Utilities: \code{\link{plot_size_distribution}}
}
\author{
Lorena Pantano, Michael Steinbaugh
}
