% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_total_reads.R
\name{plot_total_reads}
\alias{plot_total_reads}
\title{Total reads plot}
\usage{
plot_total_reads(bcbio)
}
\arguments{
\item{bcbio}{list object}
}
\description{
Total reads plot
}
\examples{
\dontrun{
plot_total_reads(bcbio)
}
}
\author{
Michael Steinbaugh
}
