% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano plot}
\usage{
plot_volcano(bcbio, res, lfc = 1, text_labels = 20)
}
\arguments{
\item{bcbio}{bcbio list object}

\item{res}{\code{DESeqResults} object}

\item{lfc}{Log fold change ratio (base 2) cutoff for coloring}
}
\value{
Volcano plot
}
\description{
Volcano plot
}
\examples{
\dontrun{
plot_volcano(bcbio,
             res,
             alpha = 0.05,
             lfc = 1)
}
}
\author{
Michael Steinbaugh
}
