% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-de.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano plot.}
\usage{
plot_volcano(bcb, res, lfc = 1L, text_labels = 30L, merge_plots = TRUE,
  direction = "both", title = NULL, shade_color = "orange",
  shade_alpha = 0.25, point_color = "gray", point_alpha = 0.75,
  point_outline_color = "darkgray")
}
\arguments{
\item{bcb}{\link{bcbioRNADataSet}.}

\item{res}{\link{DESeqResults}.}

\item{lfc}{Log fold change ratio (base 2) cutoff for coloring.}

\item{text_labels}{Number of text labels to plot.}

\item{merge_plots}{Merge all plots into one.}

\item{direction}{Plot \code{up}, \code{down}, or \code{both} directions.}

\item{title}{(\emph{Optional}). Custom title.}

\item{shade_color}{Shading color for bounding box.}

\item{shade_alpha}{Shading transparency alpha.}

\item{point_color}{Point color.}

\item{point_alpha}{Point transparency alpha.}

\item{point_outline_color}{Point outline color.}
}
\value{
Volcano plot arranged as \code{ggrid} (\code{merge_plots = TRUE}), or \code{\link[=show]{show()}}
individual \link{ggplot}s (\code{merge_plots = FALSE}).
}
\description{
Volcano plot.
}
\seealso{
This function is an updated variant of
\code{CHBUtils::volcano_density_plot()}.
}
\author{
Michael Steinbaugh, Lorena Pantano (based on John Hutchinson's work)
}
