% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-plot_volcano.R
\docType{methods}
\name{plot_volcano}
\alias{plot_volcano}
\alias{plot_volcano,DESeqResults-method}
\alias{.plot_volcano}
\title{Plot Volcano}
\usage{
plot_volcano(object, ...)

\S4method{plot_volcano}{DESeqResults}(object, ...)

.plot_volcano(res_df, alpha = 0.05, padj = TRUE, lfc = 1L,
  text_labels = 10L, direction = "both", shade_color = "orange",
  shade_alpha = 0.25, point_color = "gray", point_alpha = 0.75,
  point_outline_color = "darkgray", ggdraw = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{padj}{Use P values adjusted for multiple comparisions.}

\item{lfc}{Log fold change ratio (base 2) cutoff for coloring.}

\item{direction}{Plot \code{up}, \code{down}, or \code{both} (\strong{default}) directions.}

\item{shade_color}{Shading color for bounding box.}

\item{shade_alpha}{Shading transparency alpha.}

\item{point_color}{Point color.}

\item{point_alpha}{Point transparency alpha.}

\item{point_outline_color}{Point outline color.}

\item{ggdraw}{Merge all plots using \code{\link[=ggdraw]{ggdraw()}}.}

\item{ntop}{Number of top genes to label.}

\item{title}{\emph{Optional}. Plot title.}
}
\value{
Volcano plot arranged as \code{ggrid} (\code{ggdraw = TRUE}), or \code{\link[=show]{show()}}
individual \link{ggplot} (\code{ggdraw = FALSE}).
}
\description{
Plot Volcano
}
\examples{
data(bcb)
dds <- DESeqDataSetFromTximport(
    txi = txi(bcb),
    colData = colData(bcb),
    design = formula(~group)) \%>\%
    DESeq
res <- results(dds)
plot_volcano(res)
}
\seealso{
This function is an updated variant of
\code{CHBUtils::volcano_density_plot()}.

Other Differential Expression Plots: \code{\link{plot_ma}}
}
\author{
John Hutchinson, Michael Steinbaugh, Lorena Pantano
}
