% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotCorrelationHeatmap.R
\docType{methods}
\name{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap,bcbioRNASeq-method}
\title{Plot Correlation Heatmap}
\usage{
\S4method{plotCorrelationHeatmap}{bcbioRNASeq}(object, normalized = c("rlog",
  "vst", "tmm", "tpm"), ...)
}
\arguments{
\item{object}{Object.}

\item{normalized}{Character indicating which normalization method to apply:
\itemize{
\item "\code{tpm}": Transcripts per million (tximport).
\item "\code{tmm}": edgeR trimmed mean of M-values. Calculated on the fly.
\item "\code{rlog}": DESeq2 \strong{log2} regularized log transformation.
\item "\code{vst}": DESeq2 \strong{log2} variance stabilizing transformation.
}}

\item{...}{Passthrough arguments to \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}. The names of the
arguments must be formatted in camel case, not snake case.}
}
\value{
Show heatmap and invisibly return a \code{list} of the components.
}
\description{
This function calculates a correlation matrix based on feature expression per
sample.
}
\examples{
# bcbioRNASeq ====
# Pearson correlation
plotCorrelationHeatmap(bcb_small, method = "pearson")

# Spearman correlation
plotCorrelationHeatmap(bcb_small, method = "spearman")

# Inferno palette
plotCorrelationHeatmap(
    bcb_small,
    color = inferno,
    legendColor = inferno
)

# Default pheatmap palette
plotCorrelationHeatmap(
    bcb_small,
    color = NULL,
    legendColor = NULL
)
}
\seealso{
Other Quality Control Functions: \code{\link{plot5Prime3PrimeBias}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotDispEsts}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotGenesDetected}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}}, \code{\link{plotMeanSD}},
  \code{\link{plotPCACovariates}}, \code{\link{plotPCA}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh
}
