% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-plotCorrelationHeatmap.R
\docType{methods}
\name{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap,bcbioRNADataSet-method}
\title{Plot Correlation Heatmap}
\usage{
plotCorrelationHeatmap(object, ...)

\S4method{plotCorrelationHeatmap}{bcbioRNADataSet}(object, transform = "rlog",
  method = "pearson", clusteringMethod = "ward.D2", genes = NULL,
  samples = NULL, interestingGroups = NULL, annotationCol = NULL,
  title = NULL, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments, passed to \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}.}

\item{transform}{String specifying \code{rlog} (\strong{recommended}) or \code{vst}
(\code{varianceStabilizingTransformation}) \link{DESeqTransform} object slotted
inside the \link{bcbioRNADataSet}.}

\item{method}{Correlation coefficient (or covariance) to be computed.
Defaults to \code{pearson} but \code{spearman} can also be used. Consult the
\code{\link[stats:cor]{stats::cor()}} documentation for more information.}

\item{clusteringMethod}{Hierarchical clustering method. Accepts the same
values as \code{\link[stats:hclust]{stats::hclust()}}.}

\item{genes}{\emph{Optional}. Character vector of specific gene identifiers to
plot.}

\item{samples}{\emph{Optional}. Character vector of specific samples.}

\item{interestingGroups}{\emph{Optional}. Interesting groups to label with bars
above heatmap. If \code{NULL}, defaults to \code{interestingGroups} defined in the
\link{bcbioRNADataSet}.}

\item{annotationCol}{\emph{Optional}. Alternative annotation to use. Useful when
plotting more than one column.}

\item{title}{\emph{Optional}. Text to include in plot title.}
}
\value{
\code{\link[=pheatmap]{pheatmap()}}.
}
\description{
This function calculates a correlation matrix based on gene expression per
sample. By default, this function processes all gene counts per sample to
calculate the corrlation matrix. This behavior can be overrided with the
input of \code{gene} identifier vector. In this case, only the expression of the
desired genes will be used to calculate the correlation matrix.
}
\examples{
data(bcb)
plotCorrelationHeatmap(bcb)
}
\seealso{
\itemize{
\item \code{\link[stats:cor]{stats::cor()}}.
\item \code{\link[stats:hclust]{stats::hclust()}}.
\item \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}.
}

Other Heatmaps: \code{\link{plotDEGHeatmap}}
}
\author{
Michael Steinbaugh
}
