% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-plotCorrelationHeatmap.R
\docType{methods}
\name{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap,bcbioRNASeq-method}
\title{Plot Correlation Heatmap}
\usage{
plotCorrelationHeatmap(object, ...)

\S4method{plotCorrelationHeatmap}{bcbioRNASeq}(object, transform = "rlog",
  method = "pearson", interestingGroups, genes = NULL, samples = NULL,
  title = NULL, color = viridis::viridis(256),
  legendColor = viridis::viridis)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}.}

\item{transform}{String specifying \code{rlog} (\strong{recommended}) or \code{vst}
(\code{varianceStabilizingTransformation}) \link{DESeqTransform} object slotted
inside the \link{bcbioRNASeq} object.}

\item{method}{Correlation coefficient (or covariance) method to be computed.
Defaults to \code{pearson} but \code{spearman} can also be used. Consult the
\code{\link[stats:cor]{stats::cor()}} documentation for more information.}

\item{interestingGroups}{Category to use to group samples. In the plotting
functions, this will define color and shape, where applicable. If unset,
this is automatically determined by the metadata set inside the
\link{bcbioRNASeq} object. When set to \code{NULL}, this will default to
\code{sampleName}.}

\item{genes}{\emph{Optional}. Character vector of specific gene identifiers to
plot.}

\item{samples}{\emph{Optional}. Character vector of specific samples.}

\item{title}{\emph{Optional}. Plot title.}

\item{color}{Colors to use for plot. Defaults to \code{\link[=inferno]{inferno()}} palette.}

\item{legendColor}{Colors to use for legend labels. Defaults to \code{\link[=viridis]{viridis()}}
palette.}
}
\value{
Graphical output only.
}
\description{
This function calculates a correlation matrix based on gene expression per
sample. By default, this function processes all gene counts per sample to
calculate the corrlation matrix. This behavior can be overrided with the
input of \code{gene} identifier vector. In this case, only the expression of the
desired genes will be used to calculate the correlation matrix.
}
\examples{
bcb <- examples[["bcb"]]

# Pearson correlation (default)
plotCorrelationHeatmap(bcb)

# Spearman correlation
plotCorrelationHeatmap(bcb, method = "spearman")

# Flip the palettes used for plot and legend
plotCorrelationHeatmap(
    bcb,
    color = viridis(256),
    legendColor = inferno)

# Default pheatmap palette
plotCorrelationHeatmap(
    bcb,
    color = NULL,
    legendColor = NULL)
}
\seealso{
\itemize{
\item \code{\link[stats:cor]{stats::cor()}}.
\item \code{\link[stats:hclust]{stats::hclust()}}.
}

Other Heatmaps: \code{\link{plotDEGHeatmap}},
  \code{\link{plotHeatmap}}
}
\author{
Michael Steinbaugh
}
