% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-plotCorrelationHeatmap.R
\docType{methods}
\name{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap,bcbioRNASeq-method}
\title{Plot Correlation Heatmap}
\usage{
plotCorrelationHeatmap(object, ...)

\S4method{plotCorrelationHeatmap}{bcbioRNASeq}(object, normalized = "rlog",
  method = "pearson", interestingGroups, genes = NULL, samples = NULL,
  color = viridis::viridis(256L), legendColor = viridis::viridis,
  title = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{normalized}{Select raw counts (\code{FALSE}), DESeq2 normalized counts
(\code{TRUE}), or additional normalization methods:
\itemize{
\item \code{tpm}: Transcripts per million.
\item \code{tmm}: Trimmed mean of M-values (edgeR).
\item \code{rlog}: Regularized log transformation (\code{\link[DESeq2:rlog]{DESeq2::rlog()}}).
\item \code{vst}: Variance stabilizing transformation
(\code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}}).
}}

\item{method}{Correlation coefficient (or covariance) method to be computed.
Defaults to \code{pearson} but \code{spearman} can also be used. Consult the
\code{\link[stats:cor]{stats::cor()}} documentation for more information.}

\item{interestingGroups}{Category to use to group samples. In the plotting
functions, this will define color and shape, where applicable. If unset,
this is automatically determined by the metadata set inside the
\link{bcbioRNASeq} object. When set to \code{NULL}, this will default to
\code{sampleName}.}

\item{genes}{\emph{Optional}. Character vector of specific gene identifiers to
plot.}

\item{samples}{\emph{Optional}. Character vector of specific samples.}

\item{color}{Desired ggplot color scale. Defaults to
\code{\link[viridis:scale_color_viridis]{viridis::scale_color_viridis()}}. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using
\code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}

\item{legendColor}{Colors to use for legend labels. Defaults to the
\code{\link[viridis:viridis]{viridis::viridis()}}.}

\item{title}{Include plot title.}
}
\description{
This function calculates a correlation matrix based on gene expression per
sample. By default, this function processes all gene counts per sample to
calculate the corrlation matrix. This behavior can be overrided with the
input of \code{gene} identifier vector. In this case, only the expression of the
desired genes will be used to calculate the correlation matrix.
}
\details{
When column clustering is disabled, the columns are sorted by the interesting
groups (\code{interestingGroups}) specified in the \link{bcbioRNASeq} and then the
sample names.
}
\examples{
load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioRNASeq"))

# Pearson correlation
plotCorrelationHeatmap(bcb, method = "pearson")

# Spearman correlation
plotCorrelationHeatmap(bcb, method = "spearman")

# Inferno palette
plotCorrelationHeatmap(
    bcb,
    color = viridis::inferno(256),
    legendColor = viridis::inferno)

# Default pheatmap palette
plotCorrelationHeatmap(
    bcb,
    color = NULL,
    legendColor = NULL)
}
\seealso{
\itemize{
\item \code{\link[stats:cor]{stats::cor()}}.
\item \code{\link[stats:hclust]{stats::hclust()}}.
}

Other Heatmaps: \code{\link{plotDEGHeatmap}},
  \code{\link{plotHeatmap}}
}
\author{
Michael Steinbaugh
}
