% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-plotCorrelationHeatmap.R
\docType{methods}
\name{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap,bcbioRNADataSet-method}
\title{Plot Correlation Heatmap}
\usage{
plotCorrelationHeatmap(object, ...)

\S4method{plotCorrelationHeatmap}{bcbioRNADataSet}(object, transform = "rlog",
  method = "pearson", genes = NULL, samples = NULL, title = NULL)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{transform}{String specifying \code{rlog} (\strong{recommended}) or \code{vst}
(\code{varianceStabilizingTransformation}) \link{DESeqTransform} object slotted
inside the \link{bcbioRNADataSet}.}

\item{method}{Correlation coefficient (or covariance) method to be computed.
Defaults to \code{pearson} but \code{spearman} can also be used. Consult the
\code{\link[stats:cor]{stats::cor()}} documentation for more information.}

\item{genes}{\emph{Optional}. Character vector of specific gene identifiers to
plot.}

\item{samples}{\emph{Optional}. Character vector of specific samples.}

\item{title}{\emph{Optional}. Plot title.}
}
\value{
Graphical output only.
}
\description{
This function calculates a correlation matrix based on gene expression per
sample. By default, this function processes all gene counts per sample to
calculate the corrlation matrix. This behavior can be overrided with the
input of \code{gene} identifier vector. In this case, only the expression of the
desired genes will be used to calculate the correlation matrix.
}
\examples{
data(bcb)
plotCorrelationHeatmap(bcb)
plotCorrelationHeatmap(bcb, method = "spearman")
}
\seealso{
\itemize{
\item \code{\link[stats:cor]{stats::cor()}}.
\item \code{\link[stats:hclust]{stats::hclust()}}.
}

Other Heatmaps: \code{\link{plotDEGHeatmap}},
  \code{\link{plotGeneHeatmap}}
}
\author{
Michael Steinbaugh
}
\concept{Heatmaps}
