% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotCorrelationHeatmap.R
\docType{methods}
\name{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap,bcbioRNASeq-method}
\title{Plot Correlation Heatmap}
\usage{
\S4method{plotCorrelationHeatmap}{bcbioRNASeq}(object, normalized = c("rlog",
  "vst", "tmm", "tpm"), method = c("pearson", "spearman"), genes = NULL,
  samples = NULL, interestingGroups, color = viridis,
  legendColor = viridis, title = TRUE, ...)
}
\arguments{
\item{object}{Object.}

\item{normalized}{Logical or character indicating which normalization
method to apply:
\itemize{
\item \code{FALSE}: Raw counts (tximport).
\item \code{TRUE}: DESeq2 normalized counts. Calculated on the fly.
\item "\code{tpm}": Transcripts per million (tximport).
\item "\code{tmm}": edgeR trimmed mean of M-values. Calculated on the fly.
\item "\code{rlog}": DESeq2 \strong{log2} regularized log transformation.
\item "\code{vst}": DESeq2 \strong{log2} variance stabilizing transformation.
}}

\item{method}{Correlation coefficient (or covariance) method to be computed.
Defaults to "\code{pearson}" but "\code{spearman}" can also be used. Consult the
\code{\link[stats:cor]{stats::cor()}} documentation for more information.}

\item{genes}{\emph{Optional.} Character vector of genes to include.}

\item{samples}{\emph{Optional.} Character vector of samples to include.}

\item{interestingGroups}{Character vector denoting groups of interest that
define the samples. If left unset, defaults to \code{sampleName}.}

\item{color}{Colors to use for plot. Defaults to the \code{\link[=viridis]{viridis()}}
palette.}

\item{legendColor}{Colors to use for legend labels. Defaults to the
\code{\link[=viridis]{viridis()}} palette.}

\item{title}{\emph{Optional.} Plot title.}

\item{...}{Passthrough arguments to \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}.}
}
\value{
Show heatmap and return \code{list}, containing \code{gtable}.
}
\description{
This function calculates a correlation matrix based on gene expression per
sample. By default, this function processes all gene counts per sample to
calculate the corrlation matrix. This behavior can be overrided with the
input of \code{gene} identifier vector. In this case, only the expression of the
desired genes will be used to calculate the correlation matrix.
}
\examples{
# Pearson correlation
plotCorrelationHeatmap(bcb_small, method = "pearson")

# Spearman correlation
plotCorrelationHeatmap(bcb_small, method = "spearman")

# Inferno palette
plotCorrelationHeatmap(
    bcb_small,
    color = inferno,
    legendColor = inferno
)

# Default pheatmap palette
plotCorrelationHeatmap(
    bcb_small,
    color = NULL,
    legendColor = NULL
)
}
\seealso{
Other Quality Control Functions: \code{\link{metrics}},
  \code{\link{plot5Prime3PrimeBias}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotDispEsts}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotGenesDetected}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}}, \code{\link{plotMeanSD}},
  \code{\link{plotPCACovariates}}, \code{\link{plotPCA}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh
}
