% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelationHeatmap-methods.R
\docType{methods}
\name{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap,bcbioRNASeq-method}
\title{Plot Correlation Heatmap}
\usage{
\S4method{plotCorrelationHeatmap}{bcbioRNASeq}(object,
  normalized = c("vst", "rlog", "tmm", "tpm", "rle"),
  interestingGroups = NULL, method = c("pearson", "spearman"),
  clusteringMethod = "ward.D2", showRownames = TRUE,
  showColnames = TRUE, treeheightRow = 0L, treeheightCol = 50L,
  color = viridis::viridis, legendColor = NULL, borderColor = NULL,
  title = TRUE, ...)
}
\arguments{
\item{object}{Object.}

\item{normalized}{\code{string}. Which normalization method to apply:
- "\code{tpm}": Transcripts per million (tximport).
- "\code{tmm}": edgeR trimmed mean of M-values. Calculated on the fly.
- "\code{rlog}": DESeq2 \strong{log2} regularized log transformation.
- "\code{vst}": DESeq2 \strong{log2} variance stabilizing transformation.}

\item{interestingGroups}{\code{character}. Groups of interest.}

\item{method}{\code{string}. Correlation coefficient (or covariance) method to be
computed. Defaults to "\code{pearson}" but "\code{spearman}" can also be used.
Consult the \code{\link[stats:cor]{stats::cor()}} documentation for more information.}

\item{clusteringMethod}{\code{string}. Clustering method. Accepts the same values
as \code{\link[stats:hclust]{stats::hclust()}}.}

\item{showRownames}{\code{boolean}. Show row or column names.}

\item{showColnames}{\code{boolean}. Show row or column names.}

\item{treeheightRow}{\code{scalar integer}. Size of the row and
column dendrograms. Use \code{0} to disable.}

\item{treeheightCol}{\code{scalar integer}. Size of the row and
column dendrograms. Use \code{0} to disable.}

\item{color}{\code{function}, \code{character}, or \code{NULL}. Hexadecimal color function
or values to use for plot. We generally recommend these hexadecimal
functions from the viridis package:
\itemize{
\item \code{\link[viridis:viridis]{viridis::viridis()}} (\emph{default}).
\item \code{\link[viridis:inferno]{viridis::inferno()}}.
\item \code{\link[viridis:magma]{viridis::magma()}}.
\item \code{\link[viridis:plasma]{viridis::plasma()}}.
Alternatively, colors can be defined manually using hexadecimal values
(e.g. \code{c("#FF0000", "#0000FF")}), but this is not generally recommended.
Refer to the RColorBrewer package for hexadecimal color palettes that may
be suitable. If set \code{NULL}, will use the default pheatmap colors.
}}

\item{legendColor}{\code{function} or \code{NULL}. Hexadecimal color function to use
for legend labels. Note that hexadecimal values are not supported. If set
\code{NULL}, will use the default pheatmap colors.}

\item{borderColor}{\code{string} or \code{NULL}. Border color. Disabled by default for
improved aesthetics.}

\item{title}{\code{string} or \code{NULL}. Plot title.}

\item{...}{Passthrough arguments to \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}. The names of the
arguments should be formatted in camel case, not snake case.}
}
\value{
Show heatmap and invisibly return a \code{list} of the components.
}
\description{
This function calculates a correlation matrix based on feature expression per
sample.
}
\examples{
plotCorrelationHeatmap(bcb_small, method = "pearson")
plotCorrelationHeatmap(bcb_small, method = "spearman")
}
\seealso{
\itemize{
\item \code{help("plotCorrelationHeatmap", "basejump")}.
\item \code{findMethod("plotCorrelationHeatmap", "SummarizedExperiment")}.
}

Other Quality Control Functions: \code{\link{plot5Prime3PrimeBias}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotDispEsts}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotGenesDetected}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}}, \code{\link{plotMeanSD}},
  \code{\link{plotPCACovariates}}, \code{\link{plotPCA}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalCounts}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh
}
\concept{Quality Control Functions}
