% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelationHeatmap-methods.R
\docType{methods}
\name{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap,bcbioRNASeq-method}
\alias{plotCorrelationHeatmap,DESeqTransform-method}
\title{Correlation heatmap}
\usage{
plotCorrelationHeatmap(object, ...)

\S4method{plotCorrelationHeatmap}{bcbioRNASeq}(object,
  normalized = c("vst", "rlog", "tmm", "tpm", "rle"), ...)

\S4method{plotCorrelationHeatmap}{DESeqTransform}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{normalized}{\code{string}. Which normalization method to apply:
\itemize{
\item "\code{tpm}": Transcripts per million (tximport).
\item "\code{tmm}": edgeR trimmed mean of M-values. Calculated on the fly.
\item "\code{rlog}": DESeq2 \strong{log2} regularized log transformation.
\item "\code{vst}": DESeq2 \strong{log2} variance stabilizing transformation.
}}

\item{...}{Passthrough arguments to \code{SummarizedExperiment} method.}
}
\value{
Plot.
}
\description{
Correlation heatmap
}
\examples{
# bcbioRNASeq ====
# Pearson correlation
plotCorrelationHeatmap(bcb_small, method = "pearson")

# Spearman correlation
plotCorrelationHeatmap(bcb_small, method = "spearman")
}
\seealso{
\itemize{
\item \code{help("plotCorrelationHeatmap", "bcbioBase")}.
\item \code{findMethod("plotCorrelationHeatmap", "SummarizedExperiment")}.
}

Other Quality Control Functions: \code{\link{plot5Prime3PrimeBias}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotDispEsts}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotGenesDetected}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}}, \code{\link{plotMeanSD}},
  \code{\link{plotPCACovariates}}, \code{\link{plotPCA}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh
}
\concept{Quality Control Functions}
