% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/plotCorrelationHeatmap-methods.R
\name{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap,bcbioRNASeq-method}
\title{Correlation heatmap}
\usage{
\S4method{plotCorrelationHeatmap}{bcbioRNASeq}(object, normalized = c("vst", "rlog"), ...)
}
\arguments{
\item{object}{Object.}

\item{normalized}{\code{character(1)} or \code{logical(1)}.
Normalization method to apply:
\itemize{
\item \code{FALSE}: Raw counts.
When using a \href{https://bioconductor.org/packages/tximport/}{tximport}-compatible caller, these are length scaled
by default (see \code{countsFromAbundance} argument).
When using a \href{http://bioinf.wehi.edu.au/featureCounts/}{featureCounts}-compatible caller, these are \code{integer}.
}

\href{https://bioconductor.org/packages/tximport/}{tximport} caller-specific normalizations:
\itemize{
\item \code{"tpm"}: \strong{T}ranscripts \strong{p}er \strong{m}illion.
}

Additional gene-level-specific normalizations:
\itemize{
\item \code{TRUE} / \code{"sf"}: \strong{S}ize \strong{f}actor (i.e. library size) normalized
counts.\cr
See \code{\link[DESeq2:sizeFactors]{DESeq2::sizeFactors()}} for details.
\item \code{"fpkm"}: \strong{F}ragments \strong{p}er \strong{k}ilobase per \strong{m}illion mapped
fragments.\cr
Requires \code{fast = FALSE} in \code{\link[=bcbioRNASeq]{bcbioRNASeq()}} call and gene annotations in
\code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{rowRanges()}} with defined
\code{\link[BiocGenerics:start]{width}}.\cr
See \code{\link[DESeq2:fpkm]{DESeq2::fpkm()}} for details.
\item \code{"vst"}: \strong{V}ariance-\strong{s}tabilizing \strong{t}ransformation (log2).\cr
Requires \code{fast = FALSE} to be set during \code{\link[=bcbioRNASeq]{bcbioRNASeq()}} call.\cr
See \code{DESeq2::varianceStabilizingTransformation} for more information.
\item \code{"tmm"}: \strong{T}rimmed \strong{m}ean of \strong{M}-values.\cr
Calculated on the fly.\cr
See \code{\link[edgeR:calcNormFactors]{edgeR::calcNormFactors()}} for details.
\item \code{"rle"}: \strong{R}elative \strong{l}og \strong{e}xpression transformation.\cr
Calculated on the fly.\cr
See \code{\link[=relativeLogExpression]{relativeLogExpression()}} for details.
\item \code{"rlog"}: \emph{Deprecated}.
\strong{R}egularized \strong{log} transformation (log2).\cr
No longer calculated automatically during \code{\link[=bcbioRNASeq]{bcbioRNASeq()}} call, but may
be defined in legacy objects.\cr
See \code{\link[DESeq2:rlog]{DESeq2::rlog()}} for details.\cr
Note that VST is more performant and now recommended by default instead.

Note that \code{logical(1)} support only applies to \code{counts()}. Other
functions in the package require \code{character(1)} and use
\code{\link[base:match.arg]{match.arg()}} internally.
}}

\item{...}{Passthrough to \code{SummarizedExperiment} method defined in AcidPlots.
See \code{\link[AcidPlots:plotCorrelationHeatmap]{AcidPlots::plotCorrelationHeatmap()}} for details.}
}
\value{
Plot.
}
\description{
Correlation heatmap
}
\note{
Updated 2019-09-15.
}
\examples{
data(bcb)
plotCorrelationHeatmap(bcb, method = "pearson")
plotCorrelationHeatmap(bcb, method = "spearman")
}
\author{
Michael Steinbaugh
}
