% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelationHeatmap-methods.R
\docType{methods}
\name{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap}
\alias{plotCorrelationHeatmap,bcbioRNASeq-method}
\title{Correlation heatmap}
\usage{
\S4method{plotCorrelationHeatmap}{bcbioRNASeq}(object,
  normalized = c("vst", "rlog", "tmm", "rle", "tpm", "fpkm"),
  interestingGroups = NULL, method = c("pearson", "kendall",
  "spearman"), clusteringMethod = "ward.D2", showRownames = TRUE,
  showColnames = TRUE, treeheightRow = 0L, treeheightCol = 50L,
  color = viridis::viridis, legendColor = viridis::viridis,
  borderColor = NULL, title = TRUE, ...)
}
\arguments{
\item{object}{Object.}

\item{normalized}{\code{character(1)}.
Which normalization method to apply:\preformatted{- "`tpm`": Transcripts per million (tximport).
- "`tmm`": edgeR trimmed mean of M-values. Calculated on the fly.
- "`rlog`": DESeq2 **log2** regularized log transformation.
- "`vst`": DESeq2 **log2** variance stabilizing transformation.
}}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{method}{\code{character(1)}.
Correlation coefficient (or covariance) method to be computed. Defaults to
"\code{pearson}" but "\code{spearman}" or "\code{kendall}" can also be used. Refer to the
\code{\link[stats:cor]{stats::cor()}} documentation for details.}

\item{clusteringMethod}{\code{character(1)}.
Clustering method. Accepts the same values as \code{\link[stats:hclust]{stats::hclust()}}.}

\item{showRownames}{\code{logical(1)}.
Show row or column names.}

\item{showColnames}{\code{logical(1)}.
Show row or column names.}

\item{treeheightRow}{\code{integer(1)}.
Size of the row and column dendrograms. Use \code{0} to disable.}

\item{treeheightCol}{\code{integer(1)}.
Size of the row and column dendrograms. Use \code{0} to disable.}

\item{color}{\code{function}, \code{character}, or \code{NULL}.
Hexadecimal color function or values to use for plot.

We generally recommend these hexadecimal functions from the viridis
package:
\itemize{
\item \code{viridis::viridis()} (\emph{default}).
\item \code{viridis::inferno()}.
\item \code{viridis::magma()}.
\item \code{viridis::plasma()}.
}

Alternatively, colors can be defined manually using hexadecimal values
(e.g. \code{c("#FF0000", "#0000FF")}), but this is not generally recommended.
Refer to the RColorBrewer package for hexadecimal color palettes that may
be suitable. If set \code{NULL}, will use the default pheatmap colors.}

\item{legendColor}{\code{function} or \code{NULL}.
Hexadecimal color function to use for legend labels. Note that hexadecimal
values are not supported. If set \code{NULL}, will use the default pheatmap
colors.}

\item{borderColor}{\code{character(1)} or \code{NULL}.
Border color. Disabled by default for improved aesthetics.}

\item{title}{\code{character(1)} or \code{NULL}.
Plot title.}

\item{...}{Additional arguments.}
}
\value{
\code{pheatmap}.
}
\description{
Correlation heatmap
}
\examples{
data(bcb)
plotCorrelationHeatmap(bcb, method = "pearson")
plotCorrelationHeatmap(bcb, method = "spearman")
}
\author{
Michael Steinbaugh
}
