% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotCountDensity.R
\docType{methods}
\name{plotCountDensity}
\alias{plotCountDensity}
\alias{plotCountDensity,bcbioRNASeqANY-method}
\alias{plotCountDensity,data.frame-method}
\title{Plot Count Density}
\usage{
plotCountDensity(object, ...)

\S4method{plotCountDensity}{bcbioRNASeqANY}(object, interestingGroups,
  normalized = "tmm", style = "color")

\S4method{plotCountDensity}{data.frame}(object,
  interestingGroups = "sampleName", style = "color")
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{interestingGroups}{Category to use to group samples (color and shape).
If unset, this is automatically determined by the metadata set inside the
\link{bcbioRNASeq} object.}

\item{normalized}{Count normalization method. See \code{\link[=counts]{counts()}} documentation
for more information.}

\item{style}{Desired plot style (\code{color} or \code{fill}).}
}
\value{
\link{ggplot}.
}
\description{
Plot Count Density
}
\examples{
data(bcb)

# bcbioRNASeq
plotCountDensity(bcb)

\dontrun{
plotCountDensity(bcb, interestingGroups = "group")

# data.frame
meltLog10(bcb, normalized = "tmm") \%>\%
    plotCountDensity()
}
}
\seealso{
Other Quality Control Plots: \code{\link{plot53Bias}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotGenesDetected}}, \code{\link{plotGene}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
\concept{Quality Control Plots}
