% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCounts-methods.R
\docType{methods}
\name{plotCounts}
\alias{plotCounts}
\alias{plotCounts,SummarizedExperiment-method}
\alias{plotCounts,bcbioRNASeq-method}
\alias{plotCounts,DESeqDataSet-method}
\alias{plotCounts,DESeqTransform-method}
\title{Plot counts}
\usage{
plotCounts(object, ...)

\S4method{plotCounts}{SummarizedExperiment}(object, genes,
  interestingGroups, countsAxisLabel = "counts", medianLine = TRUE,
  color = getOption("bcbio.discrete.color", NULL),
  legend = getOption("bcbio.legend", TRUE), headerLevel = 2L,
  return = c("facet", "wide"))

\S4method{plotCounts}{bcbioRNASeq}(object, normalized = c("vst", "rlog",
  "tmm", "tpm", "rle"), ...)

\S4method{plotCounts}{DESeqDataSet}(object, ...)

\S4method{plotCounts}{DESeqTransform}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{genes}{\code{character}. Genes to include. These must match the rownames of
the object. It is best practice to use the stable gene identifiers from
Ensembl (e.g. "ENSG00000000003") and not the gene symbols.}

\item{interestingGroups}{\code{character}. Groups of interest that define the
samples. If left unset, defaults to \code{sampleName}.}

\item{countsAxisLabel}{\code{string}. Label to use for the counts axis.}

\item{medianLine}{\code{boolean}. Include median line for each group. Disabled if
samples are colored by sample name.}

\item{color}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.
To set the discrete color palette globally, use
\code{options(bcbio.discrete.color = scale_color_viridis_d())}.}

\item{legend}{\code{boolean}. Show plot legend.}

\item{headerLevel}{\code{scalar integer} (\code{1}-\code{7}). R Markdown header level.}

\item{return}{\code{string}. Object class to return. Uses \code{\link[=match.arg]{match.arg()}} internally
and picks the first item in the vector by default.}

\item{normalized}{\code{string}. Which normalization method to apply:
\itemize{
\item "\code{tpm}": Transcripts per million (tximport).
\item "\code{tmm}": edgeR trimmed mean of M-values. Calculated on the fly.
\item "\code{rlog}": DESeq2 \strong{log2} regularized log transformation.
\item "\code{vst}": DESeq2 \strong{log2} variance stabilizing transformation.
}}

\item{...}{Additional arguments.}
}
\value{
\itemize{
\item "\code{facet}": \code{ggplot} grouped by \code{sampleName}, with \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}
applied to panel the samples.
\item "\code{wide}": Show \code{ggplot} in wide format, with genes on the x-axis.
}
}
\description{
Plot counts
}
\examples{
# Gene identifiers
genes <- head(rownames(bcb_small), 4L)
glimpse(genes)

# bcbioRNASeq ====
plotCounts(bcb_small, genes = genes, normalized = "vst", return = "facet")
plotCounts(bcb_small, genes = genes, normalized = "vst", return = "wide")

# DESeqTransform ====
vst_small <- DESeq2::varianceStabilizingTransformation(dds_small)
plotCounts(vst_small, genes = genes)
}
\seealso{
\code{\link[DESeq2:plotCounts]{DESeq2::plotCounts()}}.
}
\author{
Michael Steinbaugh
}
\concept{Gene Expression Functions}
