% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCountsPerBiotype-methods.R
\docType{methods}
\name{plotCountsPerBiotype}
\alias{plotCountsPerBiotype}
\alias{plotCountsPerBroadClass}
\alias{plotCountsPerBiotype,bcbioRNASeq-method}
\alias{plotCountsPerBroadClass,bcbioRNASeq-method}
\title{Plot counts per biotype}
\usage{
plotCountsPerBiotype(object, ...)

plotCountsPerBroadClass(object, ...)

\S4method{plotCountsPerBiotype}{bcbioRNASeq}(object,
  normalized = c("tpm", "vst", "rlog", "tmm", "rle", "fpkm"), n = 9L,
  interestingGroups = NULL, trans = c("identity", "log2", "log10"),
  fill = getOption(x = "acid.fill.discrete", default =
  acidplots::scale_fill_synesthesia_d()), countsAxisLabel = "counts",
  title = "Counts per biotype")

\S4method{plotCountsPerBroadClass}{bcbioRNASeq}(object,
  normalized = c("tpm", "vst", "rlog", "tmm", "rle", "fpkm"),
  trans = c("log2", "log10"), interestingGroups = NULL,
  fill = getOption(x = "acid.fill.discrete", default =
  acidplots::scale_fill_synesthesia_d()), countsAxisLabel = "counts",
  title = "Counts per broad class")
}
\arguments{
\item{object}{Object.}

\item{normalized}{\code{character(1)}.
Which normalization method to apply:
\itemize{
\item "\code{tpm}": Transcripts per million (tximport).
\item "\code{tmm}": edgeR trimmed mean of M-values. Calculated on the fly.
\item "\code{rlog}": DESeq2 log2 regularized log transformation.
\item "\code{vst}": DESeq2 log2 variance stabilizing transformation.
}}

\item{n}{\code{integer(1)}.
Number to include.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.

To set the discrete fill palette globally, use:\preformatted{options(acid.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{countsAxisLabel}{\code{character(1)}.
Counts axis label.}

\item{title}{\code{character(1)}.
Plot title.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot counts per biotype
}
\examples{
data(bcb)
plotCountsPerBiotype(bcb)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
