% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCountsPerBiotype-methods.R
\docType{methods}
\name{plotCountsPerBiotype}
\alias{plotCountsPerBiotype}
\alias{plotCountsPerBiotype,bcbioRNASeq-method}
\alias{plotCountsPerBroadClass,bcbioRNASeq-method}
\title{Plot Counts per Biotype}
\usage{
\S4method{plotCountsPerBiotype}{bcbioRNASeq}(object,
  normalized = c("tpm", "vst", "rlog", "tmm", "rle", "fpkm"), n = 9L,
  interestingGroups = NULL, trans = "log2",
  countsAxisLabel = "counts")

\S4method{plotCountsPerBroadClass}{bcbioRNASeq}(object,
  normalized = c("tpm", "vst", "rlog", "tmm", "rle", "fpkm"),
  interestingGroups = NULL, trans = "log2",
  countsAxisLabel = "counts")
}
\arguments{
\item{object}{Object.}

\item{normalized}{\code{string}. Which normalization method to apply:
\itemize{
\item "\code{tpm}": Transcripts per million (tximport).
\item "\code{tmm}": edgeR trimmed mean of M-values. Calculated on the fly.
\item "\code{rlog}": DESeq2 \strong{log2} regularized log transformation.
\item "\code{vst}": DESeq2 \strong{log2} variance stabilizing transformation.
}}

\item{n}{\code{scalar integer}. Number to include.}

\item{interestingGroups}{\code{character}. Groups of interest that
define the samples. If left unset, defaults to \code{sampleName}.}

\item{trans}{\code{string}. Name of the axis scale transformation to apply. See
\code{help("scale_x_continuous", "ggplot2")} for more information.}

\item{countsAxisLabel}{\code{string}. Counts axis label.}
}
\value{
\code{ggplot}.
}
\description{
Plot Counts per Biotype
}
\examples{
data(bcb)
plotCountsPerBiotype(bcb)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
