% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCountsPerFeature-methods.R
\docType{methods}
\name{plotCountsPerFeature}
\alias{plotCountsPerFeature}
\alias{plotCountsPerFeature,bcbioRNASeq-method}
\title{Plot counts per feature}
\usage{
\S4method{plotCountsPerFeature}{bcbioRNASeq}(object,
  normalized = c("tmm", "vst", "rlog", "rle", "tpm", "fpkm"),
  trans = c("log2", "log10"), minCounts = 1L,
  minCountsMethod = c("perFeature", "absolute"),
  interestingGroups = NULL, geom = c("boxplot", "density", "violin"),
  color = getOption(x = "acid.color.discrete", default =
  acidplots::scale_colour_synesthesia_d()), fill = getOption(x =
  "acid.fill.discrete", default = acidplots::scale_fill_synesthesia_d()),
  flip = getOption(x = "acid.flip", default = TRUE),
  title = "Counts per feature")
}
\arguments{
\item{object}{Object.}

\item{normalized}{\code{character(1)}.
Which normalization method to apply:
\itemize{
\item "\code{tpm}": Transcripts per million (tximport).
\item "\code{tmm}": edgeR trimmed mean of M-values. Calculated on the fly.
\item "\code{rlog}": DESeq2 log2 regularized log transformation.
\item "\code{vst}": DESeq2 log2 variance stabilizing transformation.
}}

\item{trans}{\code{character(1)}.
Logarithmic transformation to apply. Note that \code{vst} and \code{rlog} counts are
already log2.}

\item{minCounts}{\code{integer(1)}.
Minimum number of counts per gene in the count matrix.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{geom}{\code{character(1)}.
Plot type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to
the first argument in the \code{character} vector.}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.

To set the discrete color palette globally, use:\preformatted{options(acid.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.

To set the discrete fill palette globally, use:\preformatted{options(acid.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{flip}{\code{logical(1)}.
Flip x and y axes. Recommended for plots containing many samples.}

\item{title}{\code{character(1)}.
Plot title.}
}
\value{
\code{ggplot}.
}
\description{
Generally, we expect similar count spreads for all genes between samples
unless the library sizes or total RNA expression are different.
}
\section{Trimmed Mean of M-Values}{


We recommend visualizing counts normalized with the \strong{T}rimmed \strong{M}ean of
\strong{M}-Values (TMM) method here. TMM normalization equates the overall
expression levels of genes between samples under the assumption that the
majority of them are not differentially expressed. Therefore, by normalizing
for total RNA expression by sample, we expect the spread of the
TMM-normalized counts per gene to be similar for every sample.
}

\examples{
data(bcb)
plotCountsPerFeature(bcb)
}
\references{
TMM: Robinson, et al., 2010.
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
