% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCountsPerGene-methods.R
\docType{methods}
\name{plotCountsPerGene}
\alias{plotCountsPerGene}
\alias{plotCountsPerGene,bcbioRNASeq-method}
\title{Plot Counts Per Gene}
\usage{
\S4method{plotCountsPerGene}{bcbioRNASeq}(object, normalized = c("tmm",
  "vst", "rlog", "tpm", "rle"), interestingGroups = NULL,
  geom = c("density", "boxplot", "violin"), trans = c("identity",
  "log2", "log10"), color = getOption("basejump.discrete.color", NULL),
  fill = getOption("basejump.discrete.fill", NULL),
  flip = getOption("basejump.flip", TRUE), countsAxisLabel = "counts",
  title = "counts per gene")
}
\arguments{
\item{object}{Object.}

\item{normalized}{\code{string}. Which normalization method to apply:
- "\code{tpm}": Transcripts per million (tximport).
- "\code{tmm}": edgeR trimmed mean of M-values. Calculated on the fly.
- "\code{rlog}": DESeq2 \strong{log2} regularized log transformation.
- "\code{vst}": DESeq2 \strong{log2} variance stabilizing transformation.}

\item{interestingGroups}{\code{character}. Groups of interest.}

\item{geom}{\code{string}. Type of ggplot2 geometric object to use.}

\item{trans}{\code{string}. Apply a log transformation (e.g. \code{log2(x + 1L)}) to
the count matrix prior to melting, if desired. Use \code{"identity"} to return
unmodified (default).}

\item{color}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}. To set the discrete color
palette globally, use \code{options(basejump.discrete.color = ggplot2::scale_color_viridis_d())}.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 fill scale.
Must supply discrete values. When set to \code{NULL}, the default ggplot2 color
palette will be used. If manual color definitions are desired, we recommend
using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}. To set the discrete fill palette
globally, use \code{options(bcbio.discrete.fill = scale_fill_viridis_d())}.}

\item{flip}{\code{boolean}. Flip x and y axes. Recommended for quality control
plots containing many samples.}

\item{countsAxisLabel}{\code{string}. Counts axis label.}

\item{title}{\code{string}. Plot title.}
}
\value{
\code{ggplot}.
}
\description{
Generally, we expect similar count spreads for all genes between samples
unless the library sizes or total RNA expression are different.
}
\section{TMM}{

We recommend visualizing counts normalized with the Trimmed Mean of M-Values
(TMM) method here (Robinson, et al., 2010). TMM normalization equates the
overall expression levels of genes between samples under the assumption that
the majority of them are not differentially expressed. Therefore, by
normalizing for total RNA expression by sample, we expect the spread of the
TMM-normalized counts per gene to be similar for every sample.
}

\examples{
plotCountsPerGene(bcb_small)
}
\seealso{
Other Quality Control Functions: \code{\link{plot5Prime3PrimeBias}},
  \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotDispEsts}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}}, \code{\link{plotMeanSD}},
  \code{\link{plotPCA}}, \code{\link{plotQC}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
\concept{Quality Control Functions}
