% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotDEGenePCA.R
\docType{methods}
\name{plotDEGenePCA}
\alias{plotDEGenePCA}
\alias{plotDEGenePCA}
\alias{plotDEGenePCA,DESeqResults,DESeqTransform-method}
\alias{plotDEGPCA}
\title{Plot DEG PCA}
\usage{
plotDEGenePCA(object, counts, ...)

\S4method{plotDEGenePCA}{DESeqResults,DESeqTransform}(object, counts,
  interestingGroups = "sampleName", lfc = 0L,
  color = scale_color_viridis(discrete = TRUE), label = FALSE,
  returnData = FALSE)

plotDEGPCA(...)
}
\arguments{
\item{object}{Object.}

\item{counts}{Secondary object containing a normalized counts matrix.}

\item{...}{Additional arguments.}

\item{interestingGroups}{\emph{Optional.} Interesting groups to use for point
appearance. If missing, color defaults to all \code{interestingGroups}
parameters set in the \link{bcbioRNASeq} object.}

\item{lfc}{log2 fold change ratio cutoff.}

\item{color}{Desired ggplot color scale. Defaults to
\code{\link[=scale_color_viridis]{scale_color_viridis()}}. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[=scale_color_manual]{scale_color_manual()}}.}

\item{label}{\emph{Optional.} Superimpose sample text labels on the plot.}

\item{returnData}{Return PCA loadings data instead of plotting.}
}
\description{
Plot DEG PCA
}
\examples{
load(system.file(
    file.path("extdata", "res.rda"),
    package = "bcbioRNASeq"))
load(system.file(
    file.path("extdata", "rld.rda"),
    package = "bcbioRNASeq"))

# DESeqResults + rlog counts
plotDEGenePCA(res, counts = rld)
}
\author{
Michael Steinbaugh
}
