% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/aliases.R,
%   R/methods-plotDEGenePCA.R
\docType{methods}
\name{plotDEGenePCA}
\alias{plotDEGenePCA}
\alias{plotDEGPCA}
\alias{plotDEGenePCA}
\alias{plotDEGenePCA,DESeqResults,DESeqTransform-method}
\title{Plot DEG PCA}
\usage{
plotDEGenePCA(object, counts, ...)

plotDEGPCA(...)

\S4method{plotDEGenePCA}{DESeqResults,DESeqTransform}(object, counts, alpha,
  lfc = 0L, gene2symbol = TRUE, color = viridis::viridis(256L))
}
\arguments{
\item{object}{Object.}

\item{counts}{Secondary object containing a normalized counts matrix.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{alpha}{\emph{Optional} Alpha level cutoff. If missing, the function will
use the alpha level defined in the object.}

\item{lfc}{log2 fold change ratio cutoff.}

\item{gene2symbol}{Apply gene identifier to symbol mappings. If set \code{TRUE},
the function will attempt to automatically map gene identifiers to symbols
from Ensembl using \code{\link[=annotable]{annotable()}}. If set \code{FALSE}/\code{NULL}, then
gene2symbol mapping will be disabled. This is useful when working with a
poorly annotated genome. Alternatively, a gene2symbol \link{data.frame} can be
passed in, and must contain the columns \code{ensgene} and \code{symbol}. then the
Ensembl gene identifiers will be labeled in place of gene symbols.}

\item{color}{Desired ggplot color scale. Defaults to
\code{\link[viridis:scale_color_viridis]{viridis::scale_color_viridis()}}. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using
\code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}
}
\description{
Plot DEG PCA
}
\examples{
load(system.file(
    file.path("extdata", "res.rda"),
    package = "bcbioRNASeq"))
load(system.file(
    file.path("extdata", "rld.rda"),
    package = "bcbioRNASeq"))

# DESeqResults + rlog counts
plotDEGenePCA(res, counts = rld)
}
\author{
Michael Steinbaugh
}
