% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotDEGHeatmap-methods.R
\docType{methods}
\name{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\alias{plotDEGHeatmap,DESeqResults,SummarizedExperiment-method}
\alias{plotDEGHeatmap,DESeqResults,bcbioRNASeq-method}
\alias{plotDEGHeatmap,DESeqResults,DESeqDataSet-method}
\alias{plotDEGHeatmap,DESeqResults,DESeqTransform-method}
\title{Differentially Expressed Gene Heatmap}
\usage{
plotDEGHeatmap(results, counts, ...)

\S4method{plotDEGHeatmap}{DESeqResults,SummarizedExperiment}(results,
  counts, alpha, lfcThreshold = 0L, title = TRUE, ...)

\S4method{plotDEGHeatmap}{DESeqResults,bcbioRNASeq}(results, counts,
  normalized = c("vst", "rlog", "tmm", "tpm", "rle"), ...)

\S4method{plotDEGHeatmap}{DESeqResults,DESeqDataSet}(results, counts, ...)

\S4method{plotDEGHeatmap}{DESeqResults,DESeqTransform}(results, counts,
  alpha, lfcThreshold = 0L, title = TRUE, ...)
}
\arguments{
\item{results}{\code{DESeqResults}.}

\item{counts}{\code{matrix}. Normalized counts.}

\item{...}{Passthrough arguments to \code{\link[=plotHeatmap]{plotHeatmap()}}.}

\item{alpha}{\code{scalar numeric}. Adjusted P value ("alpha") cutoff.}

\item{lfcThreshold}{\code{scalar numeric}. Log fold change ratio (base 2) cutoff
threshold.}

\item{title}{\code{character(1)} or \code{NULL}.
Plot title.}

\item{normalized}{\code{string}. Which normalization method to apply:
\itemize{
\item "\code{tpm}": Transcripts per million (tximport).
\item "\code{tmm}": edgeR trimmed mean of M-values. Calculated on the fly.
\item "\code{rlog}": DESeq2 \strong{log2} regularized log transformation.
\item "\code{vst}": DESeq2 \strong{log2} variance stabilizing transformation.
}}
}
\value{
\code{pheatmap}.
}
\description{
This function is a simplified version of \code{\link[=plotHeatmap]{plotHeatmap()}} that is
optimized for handling a \code{DESeqResults} object rather a gene vector. All of
the optional parameters for \code{\link[=plotHeatmap]{plotHeatmap()}} are also available to this
function.
}
\details{
To adjust the annotation columns, modify the \code{colData} of the \code{counts}
argument, which must contain a \code{SummarizedExperiment} (e.g. \code{DESeqTransform},
\code{DESeqDataSet}).
}
\section{Scaling}{



Here we're scaling simply by calculating the standard score (z-score).
\itemize{
\item mu: mean.
\item sigma: standard deviation.
\item x: raw score (e.g. count matrix).
\item z: standard score (z-score).
}\preformatted{z = (x - mu) / sigma
}

See also:
\itemize{
\item \code{pheatmap:::scale_rows()}.
\item \code{scale()} for additional scaling approaches.
}

}

\section{Hierarchical clustering}{



Row- and column-wise hierarchical clustering is performed when \code{clusterRows}
and/or \code{clusterCols} are set to \code{TRUE}. Internally, this calls
\code{\link[stats:hclust]{stats::hclust()}}, and defaults to the Ward method.

Automatic hierarchical clustering of rows and/or columns can error for some
datasets. When this occurs, you'll likely see this error:\preformatted{Error in hclust(d, method = method) :
NA/NaN/Inf in foreign function call
}

In this case, either set \code{clusterRows} and/or \code{clusterCols} to \code{FALSE}, or
you can attempt to pass an \code{hclust} object to these arguments. This is
recommended as an alternate approach to be used with \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}},
which is called internally by our plotting code. Here's how this can be
accomplished:\preformatted{mat <- assay(mat)
dist <- dist(mat)
hclust <- hclust(dist, method = "ward.D2")
}

}

\examples{
# DESeqResults, bcbioRNASeq ====
plotDEGHeatmap(
    results = res_small,
    counts = bcb_small,
    normalized = "vst"
)

# DESeqResults, DESeqTransform ====
vst_small <- DESeq2::varianceStabilizingTransformation(dds_small)
plotDEGHeatmap(
    results = res_small,
    counts = vst_small
)

# DESeqResults, DESeqDataSet ====
# This always uses normalized counts
# Using default ggplot2 colors
plotDEGHeatmap(
    results = res_small,
    counts = dds_small,
    color = NULL,
    legendColor = NULL
)
}
\seealso{
\itemize{
\item \code{help("plotHeatmap", "bcbioBase")}.
\item \code{findMethod("plotHeatmap", "SummarizedExperiment")}.
}

Other Differential Expression Functions: \code{\link{alphaSummary}},
  \code{\link{contrastName}}, \code{\link{plotDEGPCA}},
  \code{\link{plotMeanAverage}}, \code{\link{plotVolcano}},
  \code{\link{resultsTables}}
}
\author{
Michael Steinbaugh
}
\concept{Differential Expression Functions}
