% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEGHeatmap-methods.R
\docType{methods}
\name{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\alias{plotDEGHeatmap,DESeqResults-method}
\alias{plotDEGHeatmap,DESeqAnalysis-method}
\title{Differentially Expressed Gene Heatmap}
\usage{
\S4method{plotDEGHeatmap}{DESeqResults}(object, counts,
  direction = c("both", "up", "down"), clusteringMethod = "ward.D2",
  scale = "row", title = TRUE, interestingGroups = NULL,
  clusterRows = TRUE, clusterCols = TRUE, showRownames = FALSE,
  showColnames = TRUE, treeheightRow = 0L, treeheightCol = 50L,
  color = viridis::viridis, legendColor = viridis::viridis,
  borderColor = NULL, ...)

\S4method{plotDEGHeatmap}{DESeqAnalysis}(object, results, counts,
  direction = c("both", "up", "down"), clusteringMethod = "ward.D2",
  scale = "row", title = TRUE, interestingGroups = NULL,
  clusterRows = TRUE, clusterCols = TRUE, showRownames = FALSE,
  showColnames = TRUE, treeheightRow = 0L, treeheightCol = 50L,
  color = viridis::viridis, legendColor = viridis::viridis,
  borderColor = NULL, ...)
}
\arguments{
\item{object}{Object.}

\item{counts}{\code{DESeqTransform}.}

\item{direction}{\code{string}. Plot "\code{both}", "\code{up}", or "\code{down}" directions.}

\item{title}{\code{string} or \code{NULL}. Plot title.}

\item{interestingGroups}{\code{character}. Groups of interest that define the
samples. If left unset, defaults to \code{sampleName}.}

\item{color}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.
To set the discrete color palette globally, use
\code{options(bcbio.discrete.color = scale_color_viridis_d())}.}

\item{...}{Additional arguments.}

\item{results}{\code{scalar}. Position or name of \code{DESeqResults}.}
}
\value{
\code{pheatmap}.
}
\description{
Construct a simple heatmap.
}
\details{
This function is a simplified version of \code{basejump.plots::plotHeatmap()} that
is optimized for automatic handling differentially expressed genes, rather
than requiring manual input of a gene vector. All of the optional parameters
for \code{plotHeatmap} are also available to this function.

To adjust the annotation columns, modify the \code{colData} of the \code{counts}
argument, which must contain/extend a \code{SummarizedExperiment}.
}
\examples{
data(deseq)

## DESeqAnalysis ====
plotDEGHeatmap(deseq)

## DESeqResults ====
plotDEGHeatmap(
    object = as(deseq, "DESeqResults"),
    counts = as(deseq, "DESeqTransform")
)
}
\author{
Michael Steinbaugh
}
