% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotDEGHeatmap.R
\docType{methods}
\name{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\alias{plotDEGHeatmap,DESeqResults,DESeqTransform-method}
\alias{plotDEGHeatmap,DESeqResults,DESeqDataSet-method}
\title{Differentially Expressed Gene Heatmap}
\usage{
plotDEGHeatmap(object, counts, ...)

\S4method{plotDEGHeatmap}{DESeqResults,DESeqTransform}(object, counts,
  lfc = 0, ...)

\S4method{plotDEGHeatmap}{DESeqResults,DESeqDataSet}(object, counts, lfc = 0,
  title = TRUE, ...)
}
\arguments{
\item{object}{Object.}

\item{counts}{Secondary object containing a normalized count matrix.}

\item{...}{Passthrough arguments to \code{\link[=plotHeatmap]{plotHeatmap()}}.}

\item{lfc}{log2 fold change ratio cutoff.}

\item{title}{\emph{Optional}. Plot title.}
}
\value{
Graphical output only.
}
\description{
This function is a simplified version of \code{\link[=plotHeatmap]{plotHeatmap()}} that is
optimized for handling a \link{DESeqResults} object rather a gene vector. All of
the optional parameters for \code{\link[=plotHeatmap]{plotHeatmap()}} are also available to this
function.
}
\examples{
# DESeqResults, DESeqTransform
res <- examples[["res"]]
rld <- examples[["rld"]]
plotDEGHeatmap(res, counts = rld)

# DESeqResults, DESeqDataSet
# Using default ggplot2 colors
dds <- examples[["dds"]]
plotDEGHeatmap(
    res,
    counts = dds,
    color = NULL,
    legendColor = NULL)
}
\seealso{
Other Heatmaps: \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotHeatmap}}
}
\author{
Michael Steinbaugh
}
