% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotDEGHeatmap.R
\docType{methods}
\name{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\alias{plotDEGHeatmap,DESeqResults-method}
\title{Differentially Expressed Gene Heatmap}
\usage{
plotDEGHeatmap(object, ...)

\S4method{plotDEGHeatmap}{DESeqResults}(object, counts, lfc = 0L,
  gene2symbol = NULL, title = TRUE, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[=plotHeatmap]{plotHeatmap()}} \code{SummarizedExperiment}
method.}

\item{counts}{Object containing a normalized counts matrix.}

\item{lfc}{Log fold change ratio (base 2) cutoff.}

\item{gene2symbol}{\code{data.frame} containing gene-to-symbol mappings. Must
contain the columns \code{geneID} and \code{geneName}.}

\item{title}{\emph{Optional.} Plot title.}
}
\value{
Show heatmap and invisibly return a \code{list} of the components.
}
\description{
This function is a simplified version of \code{\link[=plotHeatmap]{plotHeatmap()}} that is
optimized for handling a \code{DESeqResults} object rather a gene vector. All of
the optional parameters for \code{\link[=plotHeatmap]{plotHeatmap()}} are also available to this
function.
}
\details{
To adjust the annotation columns, modify the \code{colData} of the \code{counts}
argument, which must contain a \code{SummarizedExperiment} (e.g. \code{DESeqTransform},
\code{DESeqDataSet}).
}
\examples{
# Use our stashed gene2symbol
gene2symbol <- gene2symbol(bcb_small)

# DESeqResults, DESeqTransform ====
plotDEGHeatmap(
    object = res_small,
    counts = rld_small,
    gene2symbol = gene2symbol
)

# DESeqResults, DESeqDataSet ====
# Using default ggplot2 colors
plotDEGHeatmap(
    object = res_small,
    counts = dds_small,
    gene2symbol = gene2symbol,
    color = NULL,
    legendColor = NULL
)
}
\seealso{
\itemize{
\item \code{help("plotHeatmap", "bcbioBase")}.
\item \code{findMethod("plotHeatmap", "SummarizedExperiment")}.
}

Other Differential Expression Functions: \code{\link{alphaSummary}},
  \code{\link{contrastName}}, \code{\link{plotDEGPCA}},
  \code{\link{plotMA}}, \code{\link{plotVolcano}},
  \code{\link{resultsTables}}
}
\author{
Michael Steinbaugh
}
