% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotDEGHeatmap.R
\docType{methods}
\name{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\alias{plotDEGHeatmap,DESeqResults-method}
\title{Differentially Expressed Gene Heatmap}
\usage{
plotDEGHeatmap(object, ...)

\S4method{plotDEGHeatmap}{DESeqResults}(object, counts, lfc = 0L,
  gene2symbol = NULL, showColnames = TRUE, showRownames = TRUE,
  title = TRUE, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}. The names of the
arguments must be formatted in camel case, not snake case.}

\item{counts}{Object containing a normalized counts matrix.}

\item{lfc}{Log fold change ratio (base 2) cutoff.}

\item{gene2symbol}{\code{data.frame} containing gene-to-symbol mappings. Must
contain the columns \code{geneID} and \code{geneName}.}

\item{showColnames}{Show column names.}

\item{showRownames}{Show row names.}

\item{title}{\emph{Optional.} Plot title.}
}
\value{
Show heatmap and invisibly return a \code{list} of the components.
}
\description{
This function is a simplified version of \code{\link[=plotHeatmap]{plotHeatmap()}} that is
optimized for handling a \code{DESeqResults} object rather a gene vector. All of
the optional parameters for \code{\link[=plotHeatmap]{plotHeatmap()}} are also available to this
function.
}
\examples{
# Use our stashed gene2symbol
gene2symbol <- gene2symbol(bcb_small)
annotationCol <- colData(bcb_small) \%>\%
    .[, interestingGroups(bcb_small), drop = FALSE]

# DESeqResults, DESeqTransform ====
plotDEGHeatmap(
    object = res_small,
    counts = rld_small,
    gene2symbol = gene2symbol,
    annotationCol = annotationCol
)

# DESeqResults, DESeqDataSet ====
# Using default ggplot2 colors
plotDEGHeatmap(
    object = res_small,
    counts = dds_small,
    gene2symbol = gene2symbol,
    color = NULL,
    legendColor = NULL
)
}
\seealso{
Other Differential Expression Functions: \code{\link{alphaSummary}},
  \code{\link{contrastName}}, \code{\link{plotDEGPCA}},
  \code{\link{plotMA}}, \code{\link{plotVolcano}},
  \code{\link{resultsTables}}
}
\author{
Michael Steinbaugh
}
