% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotDEGHeatmap.R
\docType{methods}
\name{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\alias{plotDEGHeatmap,DESeqResults,DESeqDataSet-method}
\alias{plotDEGHeatmap,DESeqResults,DESeqTransform-method}
\alias{plotDEGHeatmap,DESeqResults,matrix-method}
\title{Differentially Expressed Gene Heatmap}
\usage{
plotDEGHeatmap(object, counts, ...)

\S4method{plotDEGHeatmap}{DESeqResults,DESeqDataSet}(object, counts, lfc = 0,
  title = TRUE, ...)

\S4method{plotDEGHeatmap}{DESeqResults,DESeqTransform}(object, counts,
  lfc = 0, title = TRUE, ...)

\S4method{plotDEGHeatmap}{DESeqResults,matrix}(object, counts, lfc = 0,
  title = TRUE, ...)
}
\arguments{
\item{object}{Object.}

\item{counts}{Secondary object containing a normalized counts matrix.}

\item{...}{Passthrough arguments to \code{\link[=plotHeatmap]{plotHeatmap()}}.}

\item{lfc}{log2 fold change ratio cutoff.}

\item{title}{\emph{Optional}. Plot title.}
}
\value{
\code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}} ggplot return.
}
\description{
This function is a simplified version of \code{\link[=plotHeatmap]{plotHeatmap()}} that is
optimized for handling a \link{DESeqResults} object rather a gene vector. All of
the optional parameters for \code{\link[=plotHeatmap]{plotHeatmap()}} are also available to this
function.
}
\details{
When column clustering is disabled, the columns are sorted by the interesting
groups (\code{interestingGroups}) specified in the \link{bcbioRNASeq} and then the
sample names.
}
\examples{
load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioRNASeq"))
load(system.file(
    file.path("extdata", "dds.rda"),
    package = "bcbioRNASeq"))
load(system.file(
    file.path("extdata", "res.rda"),
    package = "bcbioRNASeq"))
load(system.file(
    file.path("extdata", "rld.rda"),
    package = "bcbioRNASeq"))

# Use our stashed gene2symbol for better speed
gene2symbol <- gene2symbol(bcb)
annotationCol <- sampleMetadata(bcb) \%>\%
    .[, interestingGroups(bcb), drop = FALSE]

# DESeqResults, DESeqTransform
plotDEGHeatmap(
    res,
    counts = rld,
    gene2symbol = gene2symbol,
    annotationCol = annotationCol)

# DESeqResults, DESeqDataSet
# Using default ggplot2 colors
plotDEGHeatmap(
    res,
    counts = dds,
    gene2symbol = gene2symbol,
    color = NULL,
    legendColor = NULL)
}
\seealso{
Other Heatmaps: \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotHeatmap}}
}
\author{
Michael Steinbaugh
}
