% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotDEGHeatmap.R
\docType{methods}
\name{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\alias{plotDEGHeatmap,DESeqResults,DESeqTransform-method}
\alias{plotDEGHeatmap,data.frame,ANY-method}
\title{Differentially Expressed Gene Heatmap}
\usage{
plotDEGHeatmap(object, counts, ...)

\S4method{plotDEGHeatmap}{DESeqResults,DESeqTransform}(object, counts, ...)

\S4method{plotDEGHeatmap}{data.frame,ANY}(object, counts, alpha = 0.05,
  lfc = 0L, ...)
}
\arguments{
\item{object}{Object.}

\item{counts}{Secondary object containing normalized counts.}

\item{...}{Additional arguments.}

\item{alpha}{Alpha level cutoff.}

\item{lfc}{\link{log2} fold change ratio cutoff.}
}
\value{
Graphical output only.
}
\description{
This function is a simplified version of \code{\link[=plotGeneHeatmap]{plotGeneHeatmap()}} that is
optimized for handling a \link{DESeqResults} object rather a gene vector. All of
the optional parameters for \code{\link[=plotGeneHeatmap]{plotGeneHeatmap()}} are also available to this
function.
}
\examples{
data(bcb)
dds <- DESeqDataSetFromTximport(
    txi = txi(bcb),
    colData = colData(bcb),
    design = formula(~group)) \%>\%
    DESeq
res <- results(dds)
rld <- rlog(dds)
plotDEGHeatmap(res, rld)
}
\seealso{
Other Heatmaps: \code{\link{plotCorrelationHeatmap}}
}
\author{
Michael Steinbaugh
}
