% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotDEGHeatmap.R
\docType{methods}
\name{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\alias{plotDEGHeatmap,DESeqResults-method}
\title{Differentially Expressed Gene Heatmap}
\usage{
plotDEGHeatmap(object, ...)

\S4method{plotDEGHeatmap}{DESeqResults}(object, counts, lfc = 0L,
  gene2symbol = NULL, title = TRUE, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[=plotHeatmap]{plotHeatmap()}}.}

\item{counts}{Object containing a normalized counts matrix.}

\item{lfc}{Log fold change ratio (base 2) cutoff.}

\item{gene2symbol}{\code{data.frame} containing gene-to-symbol mappings. Must
contain the columns \code{geneID} and \code{geneName}.}

\item{title}{\emph{Optional.} Plot title.}
}
\value{
Show heatmap and return \code{list}, containing \code{gtable}.
}
\description{
This function is a simplified version of \code{\link[=plotHeatmap]{plotHeatmap()}} that is
optimized for handling a \code{DESeqResults} object rather a gene vector. All of
the optional parameters for \code{\link[=plotHeatmap]{plotHeatmap()}} are also available to this
function.
}
\details{
When column clustering is disabled, the columns are sorted by the interesting
groups (\code{interestingGroups}) specified in the \link{bcbioRNASeq} and then the
sample names.
}
\examples{
load(system.file("extdata/bcb_small.rda", package = "bcbioRNASeq"))
load(system.file("extdata/dds_small.rda", package = "bcbioRNASeq"))
load(system.file("extdata/res_small.rda", package = "bcbioRNASeq"))
load(system.file("extdata/rld_small.rda", package = "bcbioRNASeq"))

# Use our stashed gene2symbol
gene2symbol <- gene2symbol(bcb_small)
annotationCol <- colData(bcb_small) \%>\%
    .[, interestingGroups(bcb_small), drop = FALSE]

# DESeqResults, DESeqTransform ====
plotDEGHeatmap(
    object = res_small,
    counts = rld_small,
    gene2symbol = gene2symbol,
    annotationCol = annotationCol
)

# DESeqResults, DESeqDataSet ====
# Using default ggplot2 colors
plotDEGHeatmap(
    object = res_small,
    counts = dds_small,
    gene2symbol = gene2symbol,
    color = NULL,
    legendColor = NULL
)
}
\seealso{
Other Differential Expression Functions: \code{\link{alphaSummary}},
  \code{\link{contrastName}}, \code{\link{plotDEGPCA}},
  \code{\link{plotMA}}, \code{\link{plotVolcano}},
  \code{\link{resultsTables}}
}
\author{
Michael Steinbaugh
}
