% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotDEGHeatmap.R
\docType{methods}
\name{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\alias{plotDEGHeatmap,DESeqResults,DESeqDataSet-method}
\alias{plotDEGHeatmap,DESeqResults,DESeqTransform-method}
\alias{plotDEGHeatmap,DESeqResults,matrix-method}
\title{Differentially Expressed Gene Heatmap}
\usage{
plotDEGHeatmap(object, counts, ...)

\S4method{plotDEGHeatmap}{DESeqResults,DESeqDataSet}(object, counts, alpha,
  lfc = 0, gene2symbol = TRUE, annotationCol = NULL, scale = "row",
  color = viridis::viridis(256), legendColor = viridis::viridis,
  title = TRUE, ...)

\S4method{plotDEGHeatmap}{DESeqResults,DESeqTransform}(object, counts, alpha,
  lfc = 0, gene2symbol = TRUE, annotationCol = NULL, scale = "row",
  color = viridis::viridis(256), legendColor = viridis::viridis,
  title = TRUE, ...)

\S4method{plotDEGHeatmap}{DESeqResults,matrix}(object, counts, alpha, lfc = 0,
  gene2symbol = TRUE, annotationCol = NULL, scale = "row",
  color = viridis::viridis(256), legendColor = viridis::viridis,
  title = TRUE, ...)
}
\arguments{
\item{object}{Object.}

\item{counts}{Secondary object containing a normalized counts matrix.}

\item{...}{Passthrough arguments to \code{\link[=plotHeatmap]{plotHeatmap()}}.}

\item{alpha}{\emph{Optional} Alpha level cutoff. If missing, the function will
use the alpha level defined in the object.}

\item{lfc}{log2 fold change ratio cutoff.}

\item{gene2symbol}{Apply gene identifier to symbol mappings. If set \code{TRUE},
the function will attempt to automatically map gene identifiers to symbols
from Ensembl using \code{\link[basejump:annotable]{basejump::annotable()}}. If set \code{FALSE}/\code{NULL}, then
gene2symbol mapping will be disabled. This is useful when working with a
poorly annotated genome. Alternatively, a gene2symbol \link{data.frame} can be
passed in, and must contain the columns \code{ensgene} and \code{symbol}. then the
Ensembl gene identifiers will be labeled in place of gene symbols.}

\item{annotationCol}{\emph{Optional}. \link{data.frame} that defines annotation
mappings for the columns.}

\item{scale}{Character indicating if the values should be centered and scaled
in either the row direction or the column direction, or none. Corresponding
values are "row", "column" and "none".}

\item{color}{Colors to use for plot. Defaults to the \code{\link[viridis:viridis]{viridis::viridis()}}
palette.}

\item{legendColor}{Colors to use for legend labels. Defaults to the
\code{\link[viridis:viridis]{viridis::viridis()}} palette.}

\item{title}{\emph{Optional}. Plot title.}
}
\value{
\code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}} return \link{list}, containing \code{gtable}.
}
\description{
This function is a simplified version of \code{\link[=plotHeatmap]{plotHeatmap()}} that is
optimized for handling a \link{DESeqResults} object rather a gene vector. All of
the optional parameters for \code{\link[=plotHeatmap]{plotHeatmap()}} are also available to this
function.
}
\details{
When column clustering is disabled, the columns are sorted by the interesting
groups (\code{interestingGroups}) specified in the \link{bcbioRNASeq} and then the
sample names.
}
\examples{
load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioRNASeq"))
load(system.file(
    file.path("extdata", "dds.rda"),
    package = "bcbioRNASeq"))
load(system.file(
    file.path("extdata", "res.rda"),
    package = "bcbioRNASeq"))
load(system.file(
    file.path("extdata", "rld.rda"),
    package = "bcbioRNASeq"))

# Use our stashed gene2symbol for better speed
gene2symbol <- gene2symbol(bcb)
annotationCol <- sampleMetadata(bcb) \%>\%
    .[, interestingGroups(bcb), drop = FALSE]

# DESeqResults, DESeqTransform
plotDEGHeatmap(
    res,
    counts = rld,
    gene2symbol = gene2symbol,
    annotationCol = annotationCol)

# DESeqResults, DESeqDataSet
# Using default ggplot2 colors
plotDEGHeatmap(
    res,
    counts = dds,
    gene2symbol = gene2symbol,
    color = NULL,
    legendColor = NULL)
}
\seealso{
Other Heatmaps: \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotHeatmap}}
}
\author{
Michael Steinbaugh
}
