% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotDEGHeatmap.R
\docType{methods}
\name{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\alias{plotDEGHeatmap,DESeqResults,SummarizedExperiment-method}
\alias{plotDEGHeatmap,DESeqResults,DESeqDataSet-method}
\alias{plotDEGHeatmap,DESeqResults,bcbioRNASeq-method}
\title{Differentially Expressed Gene Heatmap}
\usage{
plotDEGHeatmap(results, counts, ...)

\S4method{plotDEGHeatmap}{DESeqResults,SummarizedExperiment}(results, counts,
  alpha, lfcThreshold = 0L, title = TRUE, ...)

\S4method{plotDEGHeatmap}{DESeqResults,DESeqDataSet}(results, counts, ...)

\S4method{plotDEGHeatmap}{DESeqResults,bcbioRNASeq}(results, counts,
  normalized = c("vst", "rlog", "tmm", "tpm", "rle"), ...)
}
\arguments{
\item{results}{\code{DESeqResults} object.}

\item{counts}{Object containing a normalized counts matrix.}

\item{...}{Passthrough arguments to \code{\link[=plotHeatmap]{plotHeatmap()}}.}

\item{alpha}{Adjusted P value ("alpha") cutoff.}

\item{lfcThreshold}{Log fold change ratio (base 2) cutoff threshold.}

\item{title}{\emph{Optional.} Plot title.}

\item{normalized}{Character indicating which normalization method to apply:
\itemize{
\item "\code{tpm}": Transcripts per million (tximport).
\item "\code{tmm}": edgeR trimmed mean of M-values. Calculated on the fly.
\item "\code{rlog}": DESeq2 \strong{log2} regularized log transformation.
\item "\code{vst}": DESeq2 \strong{log2} variance stabilizing transformation.
}}
}
\value{
Show heatmap and invisibly return a \code{list} of the components.
}
\description{
This function is a simplified version of \code{\link[=plotHeatmap]{plotHeatmap()}} that is
optimized for handling a \code{DESeqResults} object rather a gene vector. All of
the optional parameters for \code{\link[=plotHeatmap]{plotHeatmap()}} are also available to this
function.
}
\details{
To adjust the annotation columns, modify the \code{colData} of the \code{counts}
argument, which must contain a \code{SummarizedExperiment} (e.g. \code{DESeqTransform},
\code{DESeqDataSet}).
}
\examples{
# DESeqResults, SummarizedExperiment ====
plotDEGHeatmap(
    results = res_small,
    counts = rld_small
)

# DESeqResults, DESeqDataSet ====
# This always uses normalized counts
# Using default ggplot2 colors
plotDEGHeatmap(
    results = res_small,
    counts = dds_small,
    color = NULL,
    legendColor = NULL
)

# DESeqResults, bcbioRNASeq ====
plotDEGHeatmap(
    results = res_small,
    counts = bcb_small
)
}
\seealso{
\itemize{
\item \code{help("plotHeatmap", "bcbioBase")}.
\item \code{findMethod("plotHeatmap", "SummarizedExperiment")}.
}

Other Differential Expression Functions: \code{\link{alphaSummary}},
  \code{\link{contrastName}}, \code{\link{plotDEGPCA}},
  \code{\link{plotMeanAverage}}, \code{\link{plotVolcano}},
  \code{\link{resultsTables}}
}
\author{
Michael Steinbaugh
}
