% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotDEGHeatmap-methods.R
\docType{methods}
\name{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\alias{plotDEGHeatmap,DESeqResults,SummarizedExperiment-method}
\alias{plotDEGHeatmap,DESeqResults,DESeqTransform-method}
\alias{plotDEGHeatmap,DESeqResults,DESeqDataSet-method}
\alias{plotDEGHeatmap,DESeqResults,bcbioRNASeq-method}
\title{Differentially Expressed Gene Heatmap}
\usage{
plotDEGHeatmap(results, counts, ...)

\S4method{plotDEGHeatmap}{DESeqResults,SummarizedExperiment}(results,
  counts, alpha = NULL, lfcThreshold = 0L, scale = "row",
  title = TRUE, interestingGroups, clusterRows = TRUE,
  clusterCols = TRUE, showRownames = FALSE, showColnames = TRUE,
  treeheightRow = 0L, treeheightCol = 50L, color = viridis::viridis,
  legendColor = NULL, borderColor = NULL, ...)

\S4method{plotDEGHeatmap}{DESeqResults,DESeqTransform}(results, counts,
  alpha = NULL, lfcThreshold = 0L, scale = "row", title = TRUE,
  interestingGroups, clusterRows = TRUE, clusterCols = TRUE,
  showRownames = FALSE, showColnames = TRUE, treeheightRow = 0L,
  treeheightCol = 50L, color = viridis::viridis, legendColor = NULL,
  borderColor = NULL, ...)

\S4method{plotDEGHeatmap}{DESeqResults,DESeqDataSet}(results, counts, ...)

\S4method{plotDEGHeatmap}{DESeqResults,bcbioRNASeq}(results, counts,
  normalized = c("vst", "rlog", "tmm", "tpm", "rle"), ...)
}
\arguments{
\item{results}{\code{DESeqResults}.}

\item{counts}{\code{matrix}. Normalized counts.}

\item{...}{Passthrough arguments to \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}. The names of the
arguments should be formatted in camel case, not snake case.}

\item{alpha}{\code{scalar numeric}. Adjusted P value ("alpha") cutoff.}

\item{lfcThreshold}{\code{scalar numeric}. Log fold change ratio (base 2) cutoff
threshold.}

\item{scale}{\code{string}. Whether the values should be centered and scaled in
either the row or column direction ("\code{row}", "\code{column}"), or remain
unscaled ("\code{none}").}

\item{title}{\code{string} or \code{NULL}. Plot title.}

\item{interestingGroups}{\code{character}. Groups of interest.}

\item{clusterRows}{\code{boolean}. Arrange with hierarchical
clustering.}

\item{clusterCols}{\code{boolean}. Arrange with hierarchical
clustering.}

\item{showRownames}{\code{boolean}. Show row or column names.}

\item{showColnames}{\code{boolean}. Show row or column names.}

\item{treeheightRow}{\code{scalar integer}. Size of the row and
column dendrograms. Use \code{0} to disable.}

\item{treeheightCol}{\code{scalar integer}. Size of the row and
column dendrograms. Use \code{0} to disable.}

\item{color}{\code{function}, \code{character}, or \code{NULL}. Hexadecimal color function
or values to use for plot. We generally recommend these hexadecimal
functions from the viridis package:
\itemize{
\item \code{\link[viridis:viridis]{viridis::viridis()}} (\emph{default}).
\item \code{\link[viridis:inferno]{viridis::inferno()}}.
\item \code{\link[viridis:magma]{viridis::magma()}}.
\item \code{\link[viridis:plasma]{viridis::plasma()}}.
Alternatively, colors can be defined manually using hexadecimal values
(e.g. \code{c("#FF0000", "#0000FF")}), but this is not generally recommended.
Refer to the RColorBrewer package for hexadecimal color palettes that may
be suitable. If set \code{NULL}, will use the default pheatmap colors.
}}

\item{legendColor}{\code{function} or \code{NULL}. Hexadecimal color function to use
for legend labels. Note that hexadecimal values are not supported. If set
\code{NULL}, will use the default pheatmap colors.}

\item{borderColor}{\code{string} or \code{NULL}. Border color. Disabled by default for
improved aesthetics.}

\item{normalized}{\code{string}. Which normalization method to apply:
\itemize{
\item "\code{tpm}": Transcripts per million (tximport).
\item "\code{tmm}": edgeR trimmed mean of M-values. Calculated on the fly.
\item "\code{rlog}": DESeq2 \strong{log2} regularized log transformation.
\item "\code{vst}": DESeq2 \strong{log2} variance stabilizing transformation.
}}
}
\value{
Show heatmap and invisibly return a \code{list} of the components.
}
\description{
This function is a simplified version of \code{\link[=plotHeatmap]{plotHeatmap()}} that is
optimized for handling a \code{DESeqResults} object rather a gene vector. All of
the optional parameters for \code{\link[=plotHeatmap]{plotHeatmap()}} are also available to this
function.
}
\details{
To adjust the annotation columns, modify the \code{colData} of the \code{counts}
argument, which must contain a \code{SummarizedExperiment} (e.g. \code{DESeqTransform},
\code{DESeqDataSet}).
}
\examples{
# DESeqResults, bcbioRNASeq ====
plotDEGHeatmap(
    results = res_small,
    counts = bcb_small,
    normalized = "vst"
)

# DESeqResults, DESeqTransform ====
vst_small <- DESeq2::varianceStabilizingTransformation(dds_small)
plotDEGHeatmap(
    results = res_small,
    counts = vst_small
)

# DESeqResults, DESeqDataSet ====
# This always uses normalized counts
# Using default ggplot2 colors
plotDEGHeatmap(
    results = res_small,
    counts = dds_small,
    color = NULL,
    legendColor = NULL
)
}
\seealso{
\itemize{
\item \code{help("plotHeatmap", "basejump")}.
\item \code{findMethod("plotHeatmap", "SummarizedExperiment")}.
}

Other Differential Expression Functions: \code{\link{alphaSummary}},
  \code{\link{contrastName}}, \code{\link{plotDEGPCA}},
  \code{\link{plotMeanAverage}}, \code{\link{plotVolcano}},
  \code{\link{resultsTables}}
}
\author{
Michael Steinbaugh
}
\concept{Differential Expression Functions}
