% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotDEGHeatmap.R
\docType{methods}
\name{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\alias{plotDEGHeatmap,DESeqResults,DESeqTransform-method}
\alias{plotDEGHeatmap,DESeqResults,DESeqDataSet-method}
\title{Differentially Expressed Gene Heatmap}
\usage{
plotDEGHeatmap(object, counts, ...)

\S4method{plotDEGHeatmap}{DESeqResults,DESeqTransform}(object, counts,
  lfc = 0L)

\S4method{plotDEGHeatmap}{DESeqResults,DESeqDataSet}(object, counts, lfc = 0L)
}
\arguments{
\item{object}{Object.}

\item{counts}{Secondary object containing normalized counts.}

\item{...}{Additional arguments.}

\item{lfc}{\link{log2} fold change ratio cutoff.}

\item{alpha}{Alpha level cutoff.}

\item{title}{Plot title.}
}
\value{
Graphical output only.
}
\description{
This function is a simplified version of \code{\link[=plotGeneHeatmap]{plotGeneHeatmap()}} that is
optimized for handling a \link{DESeqResults} object rather a gene vector. All of
the optional parameters for \code{\link[=plotGeneHeatmap]{plotGeneHeatmap()}} are also available to this
function.
}
\examples{
data(dds, res, rld)

# DESeqResults, DESeqTransform
plotDEGHeatmap(res, rld)

# DESeqResults, DESeqDataSet
plotDEGHeatmap(res, dds)
}
\seealso{
Other Heatmaps: \code{\link{plotCorrelationHeatmap}}
}
\author{
Michael Steinbaugh
}
