% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotDEGHeatmap.R
\docType{methods}
\name{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\alias{plotDEGHeatmap}
\alias{plotDEGHeatmap,DESeqResults,DESeqDataSet-method}
\alias{plotDEGHeatmap,DESeqResults,DESeqTransform-method}
\alias{plotDEGHeatmap,DESeqResults,matrix-method}
\title{Differentially Expressed Gene Heatmap}
\usage{
plotDEGHeatmap(object, counts, ...)

\S4method{plotDEGHeatmap}{DESeqResults,DESeqDataSet}(object, counts, lfc = 0L,
  gene2symbol = NULL, title = TRUE, ...)

\S4method{plotDEGHeatmap}{DESeqResults,DESeqTransform}(object, counts,
  lfc = 0L, gene2symbol = NULL, title = TRUE, ...)

\S4method{plotDEGHeatmap}{DESeqResults,matrix}(object, counts, lfc = 0L,
  gene2symbol = NULL, title = TRUE, ...)
}
\arguments{
\item{object}{Object.}

\item{counts}{Secondary object containing a normalized counts matrix.}

\item{...}{Passthrough arguments to \code{\link[=plotHeatmap]{plotHeatmap()}}.}

\item{lfc}{log2 fold change ratio cutoff.}

\item{gene2symbol}{Apply gene identifier to symbol mappings. A gene2symbol
\link{data.frame} can be passed in, and must contain the columns \code{ensgene} and
\code{symbol}. then the Ensembl gene identifiers will be labeled in place of
gene symbols.}

\item{title}{Title.}

\item{alpha}{\emph{Optional} Alpha level cutoff. If missing, the function will
use the alpha level defined in the object.}
}
\description{
This function is a simplified version of \code{\link[=plotHeatmap]{plotHeatmap()}} that is
optimized for handling a \link{DESeqResults} object rather a gene vector. All of
the optional parameters for \code{\link[=plotHeatmap]{plotHeatmap()}} are also available to this
function.
}
\details{
When column clustering is disabled, the columns are sorted by the interesting
groups (\code{interestingGroups}) specified in the \link{bcbioRNASeq} and then the
sample names.
}
\examples{
load(system.file("extdata/bcb.rda", package = "bcbioRNASeq"))
load(system.file("extdata/dds.rda", package = "bcbioRNASeq"))
load(system.file("extdata/res.rda", package = "bcbioRNASeq"))
load(system.file("extdata/rld.rda", package = "bcbioRNASeq"))

# Use our stashed gene2symbol
gene2symbol <- gene2symbol(bcb)
annotationCol <- sampleMetadata(bcb) \%>\%
    .[, interestingGroups(bcb), drop = FALSE]

# DESeqResults, DESeqTransform ====
plotDEGHeatmap(
    object = res,
    counts = rld,
    gene2symbol = gene2symbol,
    annotationCol = annotationCol)

# DESeqResults, DESeqDataSet ====
# Using default ggplot2 colors
plotDEGHeatmap(
    object = res,
    counts = dds,
    gene2symbol = gene2symbol,
    color = NULL,
    legendColor = NULL)
}
\seealso{
Other Heatmaps: \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotHeatmap}}
}
\author{
Michael Steinbaugh
}
