% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotDEGPCA.R
\docType{methods}
\name{plotDEGPCA}
\alias{plotDEGPCA}
\alias{plotDEGPCA}
\alias{plotDEGPCA,DESeqResults-method}
\title{Plot DEG PCA}
\usage{
plotDEGPCA(object, ...)

\S4method{plotDEGPCA}{DESeqResults}(object, counts,
  interestingGroups = "sampleName", lfc = 0L,
  color = scale_color_viridis(discrete = TRUE), label = FALSE,
  title = "deg pca", return = c("ggplot", "data.frame"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{counts}{Object containing a normalized counts matrix.}

\item{interestingGroups}{\emph{Optional.} Interesting groups to use for point
appearance. If missing, color defaults to all \code{interestingGroups}
parameters set in the \link{bcbioRNASeq} object.}

\item{lfc}{log2 fold change ratio cutoff.}

\item{color}{Desired ggplot color scale. Defaults to
\code{\link[=scale_color_viridis]{scale_color_viridis()}}. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[=scale_color_manual]{scale_color_manual()}}.}

\item{label}{\emph{Optional.} Superimpose sample text labels on the plot.}

\item{title}{Title.}

\item{return}{Return \code{ggplot} or \code{data.frame}.}
}
\description{
Plot DEG PCA
}
\examples{
# DESeqResults, DESeqTransform ====
plotDEGPCA(
    object = res_small,
    counts = rld_small,
    interestingGroups = interestingGroups(bcb_small),
    label = TRUE
)
}
\seealso{
Other Differential Expression Functions: \code{\link{alphaSummary}},
  \code{\link{contrastName}}, \code{\link{plotDEGHeatmap}},
  \code{\link{plotMA}}, \code{\link{plotVolcano}},
  \code{\link{resultsTables}}
}
\author{
Michael Steinbaugh
}
