% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotDEGPCA.R
\docType{methods}
\name{plotDEGPCA}
\alias{plotDEGPCA}
\alias{plotDEGPCA}
\alias{plotDEGPCA,DESeqResults,DESeqTransform-method}
\title{Plot DEG PCA}
\usage{
plotDEGPCA(object, counts, ...)

\S4method{plotDEGPCA}{DESeqResults,DESeqTransform}(object, counts,
  interestingGroups = "sampleName", lfc = 0L,
  color = scale_color_viridis(discrete = TRUE), label = FALSE,
  returnData = FALSE)
}
\arguments{
\item{object}{Object.}

\item{counts}{Secondary object containing a normalized counts matrix.}

\item{...}{Additional arguments.}

\item{interestingGroups}{\emph{Optional.} Interesting groups to use for point
appearance. If missing, color defaults to all \code{interestingGroups}
parameters set in the \link{bcbioRNASeq} object.}

\item{lfc}{log2 fold change ratio cutoff.}

\item{color}{Desired ggplot color scale. Defaults to
\code{\link[=scale_color_viridis]{scale_color_viridis()}}. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[=scale_color_manual]{scale_color_manual()}}.}

\item{label}{\emph{Optional.} Superimpose sample text labels on the plot.}

\item{returnData}{Return PCA loadings data instead of plotting.}
}
\description{
Plot DEG PCA
}
\examples{
# DESeqResults, DESeqTransform ====
plotDEGPCA(
    object = res_small,
    counts = rld_small,
    interestingGroups = interestingGroups(bcb_small)
)
}
\author{
Michael Steinbaugh
}
