% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotDEGPCA-methods.R
\docType{methods}
\name{plotDEGPCA}
\alias{plotDEGPCA}
\alias{plotDEGPCA,DESeqAnalysis-method}
\alias{plotDEGPCA,DESeqResults-method}
\title{Plot DEG PCA}
\usage{
plotDEGPCA(object, ...)

\S4method{plotDEGPCA}{DESeqAnalysis}(object, counts = NULL, results,
  direction = c("both", "up", "down"), interestingGroups = NULL,
  color = getOption("bcbio.discrete.color", NULL),
  label = getOption("bcbio.label", TRUE), return = c("ggplot",
  "DataFrame"))

\S4method{plotDEGPCA}{DESeqResults}(object, counts, direction = c("both",
  "up", "down"), interestingGroups = NULL,
  color = getOption("bcbio.discrete.color", NULL),
  label = getOption("bcbio.label", TRUE), return = c("ggplot",
  "DataFrame"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{counts}{\code{DESeqTransform}.}

\item{results}{\code{scalar}. Position or name of \code{DESeqResults}.}

\item{direction}{\code{string}. Plot "\code{both}", "\code{up}", or "\code{down}" directions.}

\item{interestingGroups}{\code{character}. Groups of interest that define the
samples. If left unset, defaults to \code{sampleName}.}

\item{color}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.
To set the discrete color palette globally, use
\code{options(bcbio.discrete.color = scale_color_viridis_d())}.}

\item{label}{\code{boolean}. Superimpose sample text labels on the plot.}

\item{return}{\code{string}. Object class to return. Uses \code{\link[=match.arg]{match.arg()}} internally
and picks the first item in the vector by default.}
}
\value{
\code{ggplot} or \code{DataFrame}.
}
\description{
Wrapper for \code{\link[DESeq2:plotPCA]{DESeq2::plotPCA()}} that improves principal component analysis
(PCA) sample coloring and labeling.
}
\details{
PCA (Jolliffe, et al., 2002) is a multivariate technique that allows us to
summarize the systematic patterns of variations in the data. PCA takes the
expression levels for genes and transforms it in principal component space,
reducing each sample into one point. Thereby, we can separate samples by
expression variation, and identify potential sample outliers. The PCA plot is
a way to look at how samples are clustering.
}
\examples{
# DESeqAnalysis ====
plotDEGPCA(deseq_small)

# DESeqResults ====
plotDEGPCA(
    object = as(deseq_small, "DESeqResults"),
    counts = as(deseq_small, "DESeqTransform")
)
}
\seealso{
Other Differential Expression Functions: \code{\link{alphaSummary}},
  \code{\link{contrastName}}, \code{\link{plotDEGHeatmap}},
  \code{\link{plotMA}}, \code{\link{plotVolcano}}
}
\author{
Michael Steinbaugh
}
\concept{Differential Expression Functions}
