% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotDEGPCA.R
\docType{methods}
\name{plotDEGPCA}
\alias{plotDEGPCA}
\alias{plotDEGPCA}
\alias{plotDEGPCA,DESeqResults,SummarizedExperiment-method}
\alias{plotDEGPCA,DESeqResults,DESeqDataSet-method}
\alias{plotDEGPCA,DESeqResults,bcbioRNASeq-method}
\title{Plot DEG PCA}
\usage{
plotDEGPCA(results, counts, ...)

\S4method{plotDEGPCA}{DESeqResults,SummarizedExperiment}(results, counts,
  interestingGroups, alpha, lfcThreshold = 0L, color = scale_color_hue(),
  label = FALSE, title = "deg pca", return = c("ggplot", "data.frame"))

\S4method{plotDEGPCA}{DESeqResults,DESeqDataSet}(results, counts, ...)

\S4method{plotDEGPCA}{DESeqResults,bcbioRNASeq}(results, counts,
  normalized = c("rlog", "vst", "tmm", "tpm"), ...)
}
\arguments{
\item{results}{\code{DESeqResults} object.}

\item{counts}{Object containing a normalized counts matrix.}

\item{...}{Additional arguments.}

\item{interestingGroups}{Character vector denoting groups of interest that
define the samples. If left unset, defaults to \code{sampleName}.}

\item{alpha}{Adjusted P value ("alpha") cutoff.}

\item{lfcThreshold}{Log fold change ratio (base 2) cutoff threshold.}

\item{color}{Desired ggplot color scale. Must supply discrete values. When
set to \code{NULL}, the default ggplot2 color palette will be used. If manual
color definitions are desired, we recommend using
\code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}

\item{label}{Superimpose sample text labels on the plot.}

\item{title}{Title of plot.}

\item{return}{Object class to return. Uses \code{\link[=match.arg]{match.arg()}} internally and picks
the first item in the vector by default.}

\item{normalized}{Character indicating which normalization method to apply:
\itemize{
\item "\code{tpm}": Transcripts per million (tximport).
\item "\code{tmm}": edgeR trimmed mean of M-values. Calculated on the fly.
\item "\code{rlog}": DESeq2 \strong{log2} regularized log transformation.
\item "\code{vst}": DESeq2 \strong{log2} variance stabilizing transformation.
}}
}
\value{
\code{ggplot} or \code{data.frame}.
}
\description{
Plot DEG PCA
}
\examples{
# DESeqResults, SummarizedExperiment ====
plotDEGPCA(
    results = res_small,
    counts = rld_small,
    label = TRUE
)

# DESeqResults, bcbioRNASeq ====
plotDEGPCA(
    results = res_small,
    counts = bcb_small,
    label = TRUE
)
}
\seealso{
Other Differential Expression Functions: \code{\link{alphaSummary}},
  \code{\link{contrastName}}, \code{\link{plotDEGHeatmap}},
  \code{\link{plotMeanAverage}}, \code{\link{plotVolcano}},
  \code{\link{resultsTables}}
}
\author{
Michael Steinbaugh
}
