% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEGPCA-methods.R
\docType{methods}
\name{plotDEGPCA}
\alias{plotDEGPCA}
\alias{plotDEGPCA,DESeqResults,SummarizedExperiment-method}
\alias{plotDEGPCA,DESeqResults,bcbioRNASeq-method}
\alias{plotDEGPCA,DESeqResults,DESeqDataSet-method}
\alias{plotDEGPCA,DESeqResults,DESeqTransform-method}
\title{Plot DEG PCA}
\usage{
plotDEGPCA(results, counts, ...)

\S4method{plotDEGPCA}{DESeqResults,SummarizedExperiment}(results, counts,
  interestingGroups, alpha, lfcThreshold = 0L, direction = c("both",
  "up", "down"), color = getOption("bcbio.discrete.color", NULL),
  label = getOption("bcbio.label", FALSE), return = c("ggplot",
  "data.frame"))

\S4method{plotDEGPCA}{DESeqResults,bcbioRNASeq}(results, counts,
  normalized = c("vst", "rlog", "tmm", "tpm", "rle"), ...)

\S4method{plotDEGPCA}{DESeqResults,DESeqDataSet}(results, counts, ...)

\S4method{plotDEGPCA}{DESeqResults,DESeqTransform}(results, counts,
  interestingGroups, alpha, lfcThreshold = 0L, direction = c("both",
  "up", "down"), color = getOption("bcbio.discrete.color", NULL),
  label = getOption("bcbio.label", FALSE), return = c("ggplot",
  "data.frame"))
}
\arguments{
\item{results}{\code{DESeqResults}.}

\item{counts}{\code{matrix}. Normalized counts.}

\item{...}{Passthrough arguments to \code{SummarizedExperiment} method, defined
in \code{\link[acidplots:plotPCA]{acidplots::plotPCA()}}.}

\item{interestingGroups}{\code{character}. Groups of interest that define the
samples. If left unset, defaults to \code{sampleName}.}

\item{alpha}{\code{scalar numeric}. Adjusted P value ("alpha") cutoff.}

\item{lfcThreshold}{\code{scalar numeric}. Log fold change ratio (base 2) cutoff
threshold.}

\item{direction}{\code{string}. Plot "\code{both}", "\code{up}", or "\code{down}" directions.}

\item{color}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.
To set the discrete color palette globally, use
\code{options(bcbio.discrete.color = scale_color_viridis_d())}.}

\item{label}{\code{boolean}. Superimpose sample text labels on the plot.}

\item{return}{\code{string}. Object class to return. Uses \code{\link[=match.arg]{match.arg()}} internally
and picks the first item in the vector by default.}

\item{normalized}{\code{string}. Which normalization method to apply:
\itemize{
\item "\code{tpm}": Transcripts per million (tximport).
\item "\code{tmm}": edgeR trimmed mean of M-values. Calculated on the fly.
\item "\code{rlog}": DESeq2 \strong{log2} regularized log transformation.
\item "\code{vst}": DESeq2 \strong{log2} variance stabilizing transformation.
}}
}
\value{
\code{ggplot}.
}
\description{
Wrapper for \code{\link[DESeq2:plotPCA]{DESeq2::plotPCA()}} that improves principal component analysis
(PCA) sample coloring and labeling.
}
\details{
PCA (Jolliffe, et al., 2002) is a multivariate technique that allows us to
summarize the systematic patterns of variations in the data. PCA takes the
expression levels for genes and transforms it in principal component space,
reducing each sample into one point. Thereby, we can separate samples by
expression variation, and identify potential sample outliers. The PCA plot is
a way to look at how samples are clustering.
}
\examples{
# DESeqResults, DESeqTransform ====
vst_small <- DESeq2::varianceStabilizingTransformation(dds_small)
plotDEGPCA(
    results = res_small,
    counts = vst_small,
    label = TRUE
)

# DESeqResults, bcbioRNASeq ====
plotDEGPCA(
    results = res_small,
    counts = bcb_small,
    normalized = "vst",
    label = TRUE
)
}
\seealso{
Other Differential Expression Functions: \code{\link{alphaSummary}},
  \code{\link{contrastName}}, \code{\link{plotDEGHeatmap}},
  \code{\link{plotMA}}, \code{\link{plotVolcano}},
  \code{\link{resultsTables}}
}
\author{
Michael Steinbaugh
}
\concept{Differential Expression Functions}
