% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotDispEsts.R
\docType{methods}
\name{plotDispEsts}
\alias{plotDispEsts}
\alias{plotDispEsts,bcbioRNASeq-method}
\title{Plot Dispersion Estimates}
\usage{
\S4method{plotDispEsts}{bcbioRNASeq}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[DESeq2:plotDispEsts]{DESeq2::plotDispEsts()}}.}
}
\value{
\code{ggplot}.
}
\description{
Method support for plotting the dispersion of counts stored in a
\code{bcbioRNASeq} object. Here we're generating a \code{DESeqDataSet} object on the
fly, which already has method support for plotting dispersion, provided by
the DESeq2 package.
}
\examples{
# bcbioRNASeq ====
plotDispEsts(bcb_small)

# Custom colors, using DESeq2 parameters
plotDispEsts(
    bcb_small,
    genecol = "gray",
    fitcol = "purple",
    finalcol = "orange"
)
}
\seealso{
\itemize{
\item \code{\link[DESeq2:plotDispEsts]{DESeq2::plotDispEsts()}}.
\item \code{getMethod("plotDispEsts", "DESeqDataSet")}.
}

Other Quality Control Functions: \code{\link{metrics}},
  \code{\link{plot5Prime3PrimeBias}},
  \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotGenesDetected}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}}, \code{\link{plotMeanSD}},
  \code{\link{plotPCACovariates}}, \code{\link{plotPCA}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh
}
