% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDispEsts-methods.R
\docType{methods}
\name{plotDispEsts}
\alias{plotDispEsts}
\alias{plotDispEsts,bcbioRNASeq-method}
\title{Plot Dispersion Estimates}
\usage{
\S4method{plotDispEsts}{bcbioRNASeq}(object, ymin, CV = FALSE,
  genecol = "black", fitcol = "red", finalcol = "dodgerblue",
  legend = TRUE, xlab, ylab, log = "xy", cex = 0.45, ...)
}
\arguments{
\item{object}{Object.}

\item{ymin}{the lower bound for points on the plot, points beyond this
are drawn as triangles at ymin}

\item{CV}{logical, whether to plot the asymptotic or biological
coefficient of variation (the square root of dispersion) on the y-axis.
As the mean grows to infinity, the square root of dispersion gives
the coefficient of variation for the counts. Default is \code{FALSE},
plotting dispersion.}

\item{genecol}{the color for gene-wise dispersion estimates}

\item{fitcol}{the color of the fitted estimates}

\item{finalcol}{the color of the final estimates used for testing}

\item{legend}{logical, whether to draw a legend}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{log}{log}

\item{cex}{cex}

\item{...}{further arguments to \code{plot}}
}
\value{
\code{ggplot}.
}
\description{
This plot shows the dispersion by mean of normalized counts. We expect the
dispersion to decrease as the mean of normalized counts increases.
}
\note{
Here we're generating a \code{DESeqDataSet} object on the fly, which already
has method support for plotting dispersion, provided by the DESeq2 package.
}
\examples{
plotDispEsts(bcb_small)

# Custom colors, using DESeq2 parameters.
plotDispEsts(
    object = bcb_small,
    genecol = "gray",
    fitcol = "purple",
    finalcol = "orange"
)
}
\seealso{
\itemize{
\item \code{\link[DESeq2:plotDispEsts]{DESeq2::plotDispEsts()}}.
\item \code{getMethod("plotDispEsts", "DESeqDataSet")}.
}

Other Quality Control Functions: \code{\link{plot5Prime3PrimeBias}},
  \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}}, \code{\link{plotMeanSD}},
  \code{\link{plotPCACovariates}}, \code{\link{plotPCA}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh
}
\concept{Quality Control Functions}
