% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotDispEsts.R
\docType{methods}
\name{plotDispEsts}
\alias{plotDispEsts}
\alias{plotDispEsts,bcbioRNASeqANY-method}
\title{Plot Dispersion Estimates}
\usage{
\S4method{plotDispEsts}{bcbioRNASeqANY}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[DESeq2:plotDispEsts]{DESeq2::plotDispEsts()}}.}
}
\value{
\link{ggplot}.
}
\description{
Plot Dispersion Estimates
}
\examples{
data(bcb)

# bcbioRNASeq
plotDispEsts(bcb)
plotDispEsts(bcb, genecol = "gray", fitcol = "purple", finalcol = "orange")
}
\seealso{
\code{\link[DESeq2:plotDispEsts]{DESeq2::plotDispEsts()}}.

Other Differential Expression Utilities: \code{\link{alphaSummary}},
  \code{\link{plotMeanSD}}
}
\author{
Michael Steinbaugh
}
\concept{Differential Expression Utilities}
