% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotDispEsts.R
\docType{methods}
\name{plotDispEsts}
\alias{plotDispEsts}
\alias{plotDispEsts,bcbioRNASeq-method}
\title{Plot Dispersion Estimates}
\usage{
\S4method{plotDispEsts}{bcbioRNASeq}(object, ...)
}
\arguments{
\item{object}{\link{bcbioRNASeq}.}

\item{...}{Passthrough arguments to \code{\link[DESeq2:plotDispEsts]{DESeq2::plotDispEsts()}}.}
}
\value{
\link{ggplot}.
}
\description{
Method support for plotting the dispersion of counts stored in a
\link{bcbioRNASeq} object. Here we're using the internally stored \link{DESeqDataSet},
which already has method support for plotting dispersion, provided by
the DESeq2 package.
}
\examples{
plotDispEsts(bcb)

\dontrun{
plotDispEsts(
    bcb,
    genecol = "gray",
    fitcol = "purple",
    finalcol = "orange")
}
}
\seealso{
\itemize{
\item \code{\link[DESeq2:plotDispEsts]{DESeq2::plotDispEsts()}}.
\item \code{getMethod("plotDispEsts", "DESeqDataSet")}.
}

Other Differential Expression Utilities: \code{\link{alphaSummary}},
  \code{\link{plotMeanSD}}
}
\author{
Michael Steinbaugh
}
\concept{Differential Expression Utilities}
