% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotDispEsts-methods.R
\name{plotDispEsts}
\alias{plotDispEsts}
\alias{plotDispEsts,bcbioRNASeq-method}
\title{Plot dispersion estimates}
\usage{
plotDispEsts(object, ...)

\S4method{plotDispEsts}{bcbioRNASeq}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough to \code{DESeqDataSet} method defined in DESeq2.
See \code{\link[DESeq2:plotDispEsts]{DESeq2::plotDispEsts()}} for details.}
}
\value{
\code{ggplot}.
}
\description{
A simple helper function that plots the per-gene dispersion
estimates together with the fitted mean-dispersion relationship.
}
\details{
This plot shows the dispersion by mean of normalized counts. We expect the
dispersion to decrease as the mean of normalized counts increases.

Here we're generating a \code{DESeqDataSet} object on the fly, which already has
method support for plotting dispersion, provided by the DESeq2 package.
}
\note{
Updated 2019-09-15.
}
\examples{
data(bcb)
plotDispEsts(bcb)

## Custom colors, using DESeq2 parameters.
plotDispEsts(
    object = bcb,
    genecol = "gray",
    fitcol = "purple",
    finalcol = "orange"
)
}
\seealso{
\code{\link[DESeq2:plotDispEsts]{DESeq2::plotDispEsts()}}.
}
\author{
Michael Steinbaugh
}
