% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-plotExonicMappingRate.R
\docType{methods}
\name{plotExonicMappingRate}
\alias{plotExonicMappingRate}
\alias{plotExonicMappingRate}
\alias{plotExonicMappingRate,bcbioRNASeq-method}
\alias{plotExonicMappingRate,data.frame-method}
\title{Plot Exonic Mapping Rate}
\usage{
plotExonicMappingRate(object, ...)

\S4method{plotExonicMappingRate}{bcbioRNASeq}(object, interestingGroups,
  passLimit = 60, fill = viridis::scale_fill_viridis(discrete = TRUE),
  flip = TRUE, title = TRUE)

\S4method{plotExonicMappingRate}{data.frame}(object,
  interestingGroups = "sampleName", passLimit = 60,
  fill = viridis::scale_fill_viridis(discrete = TRUE), flip = TRUE,
  title = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{interestingGroups}{Category to use to group samples. In the plotting
functions, this will define color and shape, where applicable. If unset,
this is automatically determined by the metadata set inside the
\link{bcbioRNASeq} object. When set to \code{NULL}, this will default to
\code{sampleName}.}

\item{passLimit}{Threshold to plot pass color marker.}

\item{fill}{Desired ggplot fill scale. Defaults to
\code{\link[viridis:scale_fill_viridis]{viridis::scale_fill_viridis()}}. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.}

\item{flip}{Flip x and y axes.}

\item{title}{Include plot title.}
}
\value{
\link{ggplot}.
}
\description{
Plot Exonic Mapping Rate
}
\examples{
load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioRNASeq"))

# bcbioRNASeq
plotExonicMappingRate(bcb)
plotExonicMappingRate(
    bcb,
    interestingGroups = "sampleName",
    fill = NULL)

# data.frame
df <- metrics(bcb)
plotExonicMappingRate(df)
}
\seealso{
Other Quality Control Plots: \code{\link{plot53Bias}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotGenesDetected}}, \code{\link{plotGene}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
