% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/plotExonicMappingRate-methods.R
\docType{methods}
\name{plotExonicMappingRate}
\alias{plotExonicMappingRate}
\alias{plotExonicMappingRate,bcbioRNASeq-method}
\title{Plot Exonic Mapping Rate}
\usage{
plotExonicMappingRate(object, ...)

\S4method{plotExonicMappingRate}{bcbioRNASeq}(object,
  interestingGroups = NULL, limit = 0.6,
  fill = getOption("basejump.discrete.fill", NULL),
  flip = getOption("basejump.flip", TRUE),
  title = "exonic mapping rate")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{interestingGroups}{\code{character}. Groups of interest that define the
samples. If left unset, defaults to \code{sampleName}.}

\item{limit}{\code{scalar numeric}. Threshold to denote on the plot, using a
dashed line.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 fill scale.
Must supply discrete values. When set to \code{NULL}, the default ggplot2 color
palette will be used. If manual color definitions are desired, we recommend
using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.
To set the discrete fill palette globally, use
\code{options(bcbio.discrete.fill = scale_fill_viridis_d())}.}

\item{flip}{\code{boolean}. Flip x and y axes. Recommended for quality control
plots containing many samples.}

\item{title}{\code{string} or \code{NULL}. Plot title.}
}
\value{
\code{ggplot}.
}
\description{
Ideally, at least 60 percent of total reads should map to exons.
}
\examples{
data(bcb_small)
plotExonicMappingRate(bcb_small)
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
