% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExonicMappingRate-methods.R
\docType{methods}
\name{plotExonicMappingRate}
\alias{plotExonicMappingRate}
\alias{plotExonicMappingRate,bcbioRNASeq-method}
\title{Plot exonic mapping rate}
\usage{
\S4method{plotExonicMappingRate}{bcbioRNASeq}(object,
  interestingGroups = NULL, limit = 0.6,
  fill = getOption("acid.fill.discrete", default = NULL),
  flip = getOption("acid.flip", default = TRUE),
  title = "exonic mapping rate")
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}
}
\value{
\code{ggplot}.
}
\description{
Ideally, at least 60 percent of total reads should map to exons for RNA-seq.
}
\examples{
data(bcb)
plotExonicMappingRate(bcb)
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
