% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExonicMappingRate-methods.R
\docType{methods}
\name{plotExonicMappingRate}
\alias{plotExonicMappingRate}
\alias{plotExonicMappingRate,bcbioRNASeq-method}
\title{Plot exonic mapping rate}
\usage{
plotExonicMappingRate(object, ...)

\S4method{plotExonicMappingRate}{bcbioRNASeq}(object,
  interestingGroups = NULL, limit = 0.6, fill = getOption(x =
  "acid.fill.discrete", default = acidplots::scale_fill_synesthesia_d()),
  flip = getOption(x = "acid.flip", default = TRUE),
  title = "Exonic mapping rate")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Ideally, at least 60 percent of total reads should map to exons for RNA-seq.
}
\note{
Updated 2019-08-20.
}
\examples{
data(bcb)
plotExonicMappingRate(bcb)
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
