% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExonicMappingRate-methods.R
\docType{methods}
\name{plotExonicMappingRate}
\alias{plotExonicMappingRate}
\alias{plotExonicMappingRate,bcbioRNASeq-method}
\title{Plot exonic mapping rate}
\usage{
\S4method{plotExonicMappingRate}{bcbioRNASeq}(object,
  interestingGroups = NULL, limit = 0.6,
  fill = getOption("acid.fill.discrete", default = NULL),
  flip = getOption("acid.flip", default = TRUE),
  title = "exonic mapping rate")
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{limit}{\code{numeric(1)}.
Threshold to denote on the plot, using a dashed line.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.

To set the discrete fill palette globally, use:\preformatted{options(acid.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{flip}{\code{logical(1)}.
Flip x and y axes. Recommended for plots containing many samples.}

\item{title}{\code{character(1)}.
Plot title.}
}
\value{
\code{ggplot}.
}
\description{
Ideally, at least 60 percent of total reads should map to exons for RNA-seq.
}
\examples{
data(bcb)
plotExonicMappingRate(bcb)
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
