% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/plotExonicMappingRate-methods.R
\name{plotExonicMappingRate}
\alias{plotExonicMappingRate}
\alias{plotExonicMappingRate,bcbioRNASeq-method}
\title{Plot exonic mapping rate}
\usage{
plotExonicMappingRate(object, ...)

\S4method{plotExonicMappingRate}{bcbioRNASeq}(
  object,
  interestingGroups = NULL,
  limit = 0.6,
  labels = list(title = "Exonic mapping rate", subtitle = NULL, sampleAxis = NULL,
    metricAxis = "exonic mapping rate (\%)"),
  flip = getOption(x = "acid.flip", default = TRUE)
)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{limit}{\code{numeric(1)}.
Threshold limit.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{flip}{\code{logical(1)}.
Flip x and y axes. Recommended for plots containing many samples.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Ideally, at least 60 percent of total reads should map to exons for RNA-seq.
}
\note{
Updated 2021-07-21.
}
\examples{
## bcbioRNASeq ====
data(bcb)
plotExonicMappingRate(bcb)
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
