% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGenderMarkers-methods.R
\docType{methods}
\name{plotGenderMarkers}
\alias{plotGenderMarkers}
\alias{plotGenderMarkers,bcbioRNASeq-method}
\title{Plot Sexually Dimorphic Gender Marker Genes}
\usage{
\S4method{plotGenderMarkers}{bcbioRNASeq}(object, normalized = c("vst",
  "rlog", "tmm", "tpm", "rle"), ...)
}
\arguments{
\item{object}{Object.}

\item{normalized}{\code{string}. Which normalization method to apply:
- "\code{tpm}": Transcripts per million (tximport).
- "\code{tmm}": edgeR trimmed mean of M-values. Calculated on the fly.
- "\code{rlog}": DESeq2 \strong{log2} regularized log transformation.
- "\code{vst}": DESeq2 \strong{log2} variance stabilizing transformation.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
This is a convenience function that wraps \code{\link[=plotGene]{plotGene()}} to quickly plot known
sexually dimorphic genes, and overall gene experssion for the X and Y
chromosomes. Currently only \emph{Homo sapiens} and \emph{Mus musculus}
genomes are supported.
}
\examples{
plotGenderMarkers(bcb_small, normalized = "vst")
}
\seealso{
Other Quality Control Functions: \code{\link{plot5Prime3PrimeBias}},
  \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotDispEsts}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}}, \code{\link{plotMeanSD}},
  \code{\link{plotPCA}}, \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh
}
\concept{Quality Control Functions}
