% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotGenderMarkers.R
\docType{methods}
\name{plotGenderMarkers}
\alias{plotGenderMarkers}
\alias{plotGenderMarkers}
\alias{plotGenderMarkers,bcbioRNADataSet-method}
\title{Plot Sexually Dimorphic Gender Markers}
\usage{
plotGenderMarkers(object, ...)

\S4method{plotGenderMarkers}{bcbioRNADataSet}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\link{ggplot}.
}
\description{
Plot Sexually Dimorphic Gender Markers
}
\examples{
data(bcb)
plotGenderMarkers(bcb)
}
\author{
Michael Steinbaugh
}
