% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotGenderMarkers.R
\docType{methods}
\name{plotGenderMarkers}
\alias{plotGenderMarkers}
\alias{plotGenderMarkers}
\alias{plotGenderMarkers,bcbioRNASeq-method}
\alias{plotGenderMarkers,DESeqDataSet-method}
\alias{plotGenderMarkers,matrix-method}
\title{Plot Sexually Dimorphic Gender Markers}
\usage{
plotGenderMarkers(object, ...)

\S4method{plotGenderMarkers}{bcbioRNASeq}(object, interestingGroups,
  color = scale_color_viridis(discrete = TRUE))

\S4method{plotGenderMarkers}{DESeqDataSet}(object,
  interestingGroups = "sampleName", organism,
  color = scale_color_viridis(discrete = TRUE))

\S4method{plotGenderMarkers}{matrix}(object, interestingGroups = "sampleName",
  organism, metadata, countsAxisLabel = "counts",
  color = scale_color_viridis(discrete = TRUE))
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{interestingGroups}{Category to use to group samples. In the plotting
functions, this will define color and shape, where applicable. If unset,
this is automatically determined by the metadata set inside the
\link{bcbioRNASeq} object. When set to \code{NULL}, this will default to
\code{sampleName}.}

\item{color}{Desired ggplot color scale. Defaults to
\code{\link[viridis:scale_color_viridis]{viridis::scale_color_viridis()}}. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using
\code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}

\item{organism}{\emph{Optional}. Organism name. Should be detected automatically,
unless a spike-in FASTA sequence is provided containing a gene identifier
that is first alphabetically in the count matrix rownames.}

\item{countsAxisLabel}{Text label of counts axis.}
}
\value{
\link{ggplot}.
}
\description{
Plot Sexually Dimorphic Gender Markers
}
\examples{
# bcbioRNASeq
# Use F1000 workflow example dataset
# The minimal example inside the package doesn't have dimorphic genes
loadRemoteData(
    file.path(
        "http://bcbiornaseq.seq.cloud",
        "f1000v1",
        "data",
        "bcb.rda"),
    quiet = TRUE)
plotGenderMarkers(bcb)
plotGenderMarkers(
    bcb,
    interestingGroups = "sampleName",
    color = NULL)

# DESeqDataSet
\dontrun{
dds <- bcbio(bcb, "DESeqDataSet")
plotGenderMarkers(dds)
}
}
\seealso{
Other Quality Control Plots: \code{\link{plot53Bias}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotGenesDetected}}, \code{\link{plotGene}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh
}
