% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotGenderMarkers.R
\docType{methods}
\name{plotGenderMarkers}
\alias{plotGenderMarkers}
\alias{plotGenderMarkers}
\alias{plotGenderMarkers,bcbioRNASeq-method}
\alias{plotGenderMarkers,DESeqDataSet-method}
\alias{plotGenderMarkers,DESeqTransform-method}
\alias{plotGenderMarkers,matrix-method}
\title{Plot Sexually Dimorphic Gender Markers}
\usage{
plotGenderMarkers(object, ...)

\S4method{plotGenderMarkers}{bcbioRNASeq}(object, interestingGroups,
  normalized = "tpm", color = scale_color_viridis(discrete = TRUE),
  title = TRUE)

\S4method{plotGenderMarkers}{DESeqDataSet}(object,
  interestingGroups = "sampleName", organism,
  color = scale_color_viridis(discrete = TRUE), title = TRUE)

\S4method{plotGenderMarkers}{DESeqTransform}(object,
  interestingGroups = "sampleName", organism,
  color = scale_color_viridis(discrete = TRUE), title = TRUE)

\S4method{plotGenderMarkers}{matrix}(object, metadata,
  interestingGroups = "sampleName", organism, countsAxisLabel = "counts",
  color = scale_color_viridis(discrete = TRUE), title = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{interestingGroups}{Category to use to group samples. In the plotting
functions, this will define color and shape, where applicable. If unset,
this is automatically determined by the metadata set inside the
\link{bcbioRNASeq} object. When set to \code{NULL}, this will default to
\code{sampleName}.}

\item{normalized}{Normalization method. Supports \code{tpm} (\strong{default}), \code{tmm},
\code{rlog}, or \code{vst}.}

\item{color}{Desired ggplot color scale. Defaults to
\code{\link[=scale_color_viridis]{scale_color_viridis()}}. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[=scale_color_manual]{scale_color_manual()}}.}

\item{title}{Title.}

\item{organism}{\emph{Optional.} Organism name. Should be detected automatically,
unless a spike-in FASTA sequence is provided containing a gene identifier
that is first alphabetically in the count matrix rownames.}

\item{metadata}{Sample metadata \link{data.frame}.}

\item{countsAxisLabel}{Text label of counts axis.}
}
\value{
\link{ggplot}.
}
\description{
Plot Sexually Dimorphic Gender Markers
}
\examples{
dir <- "http://bcbiornaseq.seq.cloud/f1000v1/data"
loadRemoteData(
    c(
        file.path(dir, "bcb.rda"),
        file.path(dir, "rld.rda"),
        file.path(dir, "vst.rda")
    ),
    quiet = TRUE)

# bcbioRNASeq
plotGenderMarkers(bcb)
plotGenderMarkers(
    bcb,
    interestingGroups = "sampleName",
    color = NULL)

# DESeqDataSet
dds <- bcbio(bcb, "DESeqDataSet")
plotGenderMarkers(dds, interestingGroups = "group")

# DESeqTransform
plotGenderMarkers(rld, interestingGroups = "group")
plotGenderMarkers(vst, interestingGroups = "group")
}
\seealso{
Other Quality Control Plots: \code{\link{plot53Bias}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotGenesDetected}}, \code{\link{plotGene}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh
}
