% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotGenderMarkers.R
\docType{methods}
\name{plotGenderMarkers}
\alias{plotGenderMarkers}
\alias{plotGenderMarkers}
\alias{plotGenderMarkers,matrix-method}
\alias{plotGenderMarkers,bcbioRNASeq-method}
\alias{plotGenderMarkers,DESeqDataSet-method}
\alias{plotGenderMarkers,DESeqTransform-method}
\title{Plot Sexually Dimorphic Gender Markers}
\usage{
plotGenderMarkers(object, ...)

\S4method{plotGenderMarkers}{matrix}(object, colData,
  interestingGroups = "sampleName", organism, countsAxisLabel = "counts",
  medianLine = TRUE, color = scale_color_viridis(discrete = TRUE),
  title = TRUE)

\S4method{plotGenderMarkers}{bcbioRNASeq}(object, interestingGroups,
  normalized = "rlog", color = scale_color_viridis(discrete = TRUE),
  title = TRUE)

\S4method{plotGenderMarkers}{DESeqDataSet}(object,
  interestingGroups = "sampleName", organism,
  color = scale_color_viridis(discrete = TRUE), title = TRUE)

\S4method{plotGenderMarkers}{DESeqTransform}(object,
  interestingGroups = "sampleName", organism,
  color = scale_color_viridis(discrete = TRUE), title = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{colData}{Sample metadata \code{data.frame}.}

\item{interestingGroups}{Character vector denoting groups of interest that
define the samples. If left unset, defaults to \code{sampleName}.}

\item{organism}{\emph{Optional.} Organism name. Should be detected automatically,
unless a spike-in FASTA sequence is provided containing a gene identifier
that is first alphabetically in the count matrix rownames.}

\item{countsAxisLabel}{Text label of counts axis.}

\item{medianLine}{Include median line for each group. Disabled when samples
are grouped by \code{sampleName}.}

\item{color}{Desired ggplot color scale. Defaults to
\code{\link[=scale_color_viridis]{scale_color_viridis()}}. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[=scale_color_manual]{scale_color_manual()}}.}

\item{title}{Title.}

\item{normalized}{Normalization method. Supports \code{tpm} (\strong{default}), \code{tmm},
\code{rlog}, or \code{vst}.}
}
\value{
\code{ggplot}.
}
\description{
Plot Sexually Dimorphic Gender Markers
}
\examples{
# bcbioRNASeq ====
plotGenderMarkers(bcb_small)

# DESeqDataSet ====
plotGenderMarkers(dds_small)

# DESeqTransform ====
plotGenderMarkers(rld_small)
}
\seealso{
Other Gene Expression Functions: \code{\link{plotGene}},
  \code{\link{plotHeatmap}}
}
\author{
Michael Steinbaugh
}
