% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotGenderMarkers.R
\docType{methods}
\name{plotGenderMarkers}
\alias{plotGenderMarkers}
\alias{plotGenderMarkers,bcbioRNASeq-method}
\alias{plotGenderMarkers,DESeqDataSet-method}
\alias{plotGenderMarkers,matrix-method}
\title{Plot Sexually Dimorphic Gender Markers}
\usage{
plotGenderMarkers(object, ...)

\S4method{plotGenderMarkers}{bcbioRNASeq}(object,
  color = scale_color_viridis(discrete = TRUE))

\S4method{plotGenderMarkers}{DESeqDataSet}(object, organism,
  color = scale_color_viridis(discrete = TRUE))

\S4method{plotGenderMarkers}{matrix}(object, organism, ylab = "counts",
  color = scale_color_viridis(discrete = TRUE))
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{color}{Desired ggplot color scale. Defaults to
\code{\link[viridis:scale_color_viridis]{viridis::scale_color_viridis()}}. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using
\code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}

\item{organism}{\emph{Optional}. Organism name. Should be detected automatically,
unless a spike-in FASTA sequence is provided containing a gene identifier
that is first alphabetically in the count matrix rownames.}

\item{ylab}{Y-axis label.}
}
\value{
\link{ggplot}.
}
\description{
Plot Sexually Dimorphic Gender Markers
}
\examples{
# Use F1000 workflow example dataset
# The minimal example inside the package doesn't have dimorphic genes
\dontrun{
loadRemoteData(
    file.path(
        "https://github.com",
        "hbc",
        "bcbioRNASeq",
        "raw",
        "f1000v1",
        "data",
        "bcb.rda"
    )
)
plotGenderMarkers(bcb)
}
}
\seealso{
Other Quality Control Plots: \code{\link{plot53Bias}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotGenesDetected}}, \code{\link{plotGene}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh
}
\concept{Quality Control Plots}
