% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotGenderMarkers.R
\docType{methods}
\name{plotGenderMarkers}
\alias{plotGenderMarkers}
\alias{plotGenderMarkers}
\alias{plotGenderMarkers,SummarizedExperiment-method}
\alias{plotGenderMarkers,bcbioRNASeq-method}
\alias{plotGenderMarkers,DESeqDataSet-method}
\alias{plotGenderMarkers,DESeqTransform-method}
\title{Plot Sexually Dimorphic Gender Markers}
\usage{
plotGenderMarkers(object, ...)

\S4method{plotGenderMarkers}{SummarizedExperiment}(object, ...)

\S4method{plotGenderMarkers}{bcbioRNASeq}(object, normalized = c("vst",
  "rlog", "tmm", "tpm", "rle"), ...)

\S4method{plotGenderMarkers}{DESeqDataSet}(object, ...)

\S4method{plotGenderMarkers}{DESeqTransform}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{normalized}{Character indicating which normalization method to apply:
\itemize{
\item "\code{tpm}": Transcripts per million (tximport).
\item "\code{tmm}": edgeR trimmed mean of M-values. Calculated on the fly.
\item "\code{rlog}": DESeq2 \strong{log2} regularized log transformation.
\item "\code{vst}": DESeq2 \strong{log2} variance stabilizing transformation.
}}
}
\value{
\code{ggplot}.
}
\description{
This is a convenience function that wraps \code{\link[=plotGene]{plotGene()}} to quickly plot known
sexually dimorphic genes. Currently only \emph{Homo sapiens} and \emph{Mus musculus}
genomes are supported.
}
\examples{
# bcbioRNASeq ====
plotGenderMarkers(bcb_small)

# DESeqDataSet ====
plotGenderMarkers(dds_small)

# DESeqTransform ====
plotGenderMarkers(rld_small)
}
\seealso{
\code{\link[=plotGene]{plotGene()}}.

Other Quality Control Functions: \code{\link{plot5Prime3PrimeBias}},
  \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotDispEsts}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotGenesDetected}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}}, \code{\link{plotMeanSD}},
  \code{\link{plotPCACovariates}}, \code{\link{plotPCA}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh
}
