% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGenderMarkers-methods.R
\docType{methods}
\name{plotGenderMarkers}
\alias{plotGenderMarkers}
\alias{plotGenderMarkers,bcbioRNASeq-method}
\title{Plot sexually dimorphic gender marker genes}
\usage{
plotGenderMarkers(object, ...)

\S4method{plotGenderMarkers}{bcbioRNASeq}(object, normalized = c("vst",
  "rlog", "tmm", "rle", "tpm", "fpkm"), interestingGroups = NULL,
  trans = c("identity", "log2", "log10"), medianLine = TRUE,
  color = getOption(x = "acid.color.discrete", default =
  acidplots::scale_colour_synesthesia_d()), legend = getOption(x =
  "acid.legend", default = TRUE), style = "wide")
}
\arguments{
\item{object}{Object.}

\item{normalized}{\code{character(1)}.
Type of normalized counts to visualize:
\itemize{
\item \code{tpm} Transcripts per million.
\item \code{tmm}: edgeR trimmed mean of M-values.
\item \code{rlog}: DESeq2 log2 regularized log transformation.
\item \code{vst}: DESeq2 log2 variance stabilizing transformation.
}}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}}

\item{medianLine}{\code{logical(1)}.
Include median line for each group. Disabled if samples are colored by
sample name.}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.

To set the discrete color palette globally, use:\preformatted{options(acid.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{legend}{\code{logical(1)}.
Include plot legend.}

\item{style}{\code{character(1)}.
Plot style.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
This is a convenience function that quickly plots known sexually dimorphic
genes, and overall gene experssion for the X and Y chromosomes.
}
\note{
Updated 2019-08-07.
}
\examples{
data(bcb)
plotGenderMarkers(bcb)
}
\seealso{
\code{plotCounts()}.
}
\author{
Michael Steinbaugh
}
