% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGene-methods.R
\docType{methods}
\name{plotGene}
\alias{plotGene}
\alias{plotGene,bcbioRNASeq-method}
\alias{plotGene,DESeqAnalysis-method}
\title{Plot Gene Expression}
\usage{
\S4method{plotGene}{bcbioRNASeq}(object, genes, normalized = c("vst",
  "rlog", "tmm", "rle", "tpm", "fpkm"), interestingGroups = NULL,
  medianLine = TRUE, color = getOption("basejump.discrete.color",
  NULL), legend = getOption("basejump.legend", TRUE),
  style = c("facet", "wide"))

\S4method{plotGene}{DESeqAnalysis}(object, genes,
  interestingGroups = NULL, medianLine = TRUE,
  color = getOption("basejump.discrete.color", NULL),
  legend = getOption("basejump.legend", TRUE), style = c("facet",
  "wide"))
}
\arguments{
\item{object}{Object.}

\item{genes}{\code{character}. Genes to include. These must match the rownames of
the object. It is best practice to use the stable gene identifiers from
Ensembl (e.g. "ENSG00000000003") and not the gene symbols.}

\item{normalized}{\code{string}. Which normalization method to apply:
- "\code{tpm}": Transcripts per million (tximport).
- "\code{tmm}": edgeR trimmed mean of M-values. Calculated on the fly.
- "\code{rlog}": DESeq2 \strong{log2} regularized log transformation.
- "\code{vst}": DESeq2 \strong{log2} variance stabilizing transformation.}

\item{interestingGroups}{\code{character}. Groups of interest that define the
samples. If left unset, defaults to \code{sampleName}.}

\item{color}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.
To set the discrete color palette globally, use
\code{options(bcbio.discrete.color = scale_color_viridis_d())}.}

\item{legend}{\code{boolean}. Show plot legend.}
}
\value{
\code{ggplot}.
}
\description{
Plot Gene Expression
}
\examples{
data(bcb, deseq)

## bcbioRNASeq ====
object <- bcb
g2s <- Gene2Symbol(object)
geneIDs <- head(g2s[["geneID"]])
print(geneIDs)
geneNames <- head(g2s[["geneName"]])
print(geneNames)

plotGene(
    object = object,
    genes = geneIDs,
    normalized = "vst",
    style = "facet"
)
plotGene(
    object = object,
    genes = geneNames,
    normalized = "vst",
    style = "wide"
)

## DESeqAnalysis ====
object <- deseq
plotGene(object, genes = geneIDs, style = "facet")
plotGene(object, genes = geneNames, style = "wide")
}
\author{
Michael Steinbaugh
}
