% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotGene.R
\docType{methods}
\name{plotGene}
\alias{plotGene}
\alias{plotGene,bcbioRNASeq-method}
\alias{plotGene,DESeqDataSet-method}
\alias{plotGene,DESeqTransform-method}
\title{Plot Individual Genes}
\usage{
\S4method{plotGene}{bcbioRNASeq}(object, genes, normalized = c("rlog", "vst",
  "tpm"), interestingGroups, medianLine = TRUE,
  color = scale_color_viridis(discrete = TRUE), headerLevel = 2L,
  return = c("grid", "wide", "list", "markdown"))

\S4method{plotGene}{DESeqDataSet}(object, genes, gene2symbol = NULL,
  interestingGroups = "sampleName", medianLine = TRUE,
  color = scale_color_viridis(discrete = TRUE), headerLevel = 2L,
  return = c("grid", "wide", "list", "markdown"))

\S4method{plotGene}{DESeqTransform}(object, genes, gene2symbol = NULL,
  interestingGroups = "sampleName", medianLine = TRUE,
  color = scale_color_viridis(discrete = TRUE), headerLevel = 2L,
  return = c("grid", "wide", "list", "markdown"))
}
\arguments{
\item{object}{Object.}

\item{genes}{Character vector of genes to include. These must match the
rownames of the object. It is best practice to use the stable gene
identifiers from Ensembl (e.g. "ENSG00000000003") and not the gene symbols.}

\item{normalized}{Character indicating which normalization method to apply:
\itemize{
\item "\code{tpm}": Transcripts per million (tximport).
\item "\code{tmm}": edgeR trimmed mean of M-values. Calculated on the fly.
\item "\code{rlog}": DESeq2 \strong{log2} regularized log transformation.
\item "\code{vst}": DESeq2 \strong{log2} variance stabilizing transformation.
}}

\item{interestingGroups}{Character vector denoting groups of interest that
define the samples. If left unset, defaults to \code{sampleName}.}

\item{medianLine}{Include median line for each group. Disabled when samples
are grouped by \code{sampleName}.}

\item{color}{Desired ggplot color scale. Must supply discrete values. When
set to \code{NULL}, the default ggplot2 color palette will be used. If manual
color definitions are desired, we recommend using
\code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}

\item{headerLevel}{R Markdown header level.}

\item{return}{Object class to return. Uses \code{\link[=match.arg]{match.arg()}} internally and picks
the first item in the vector by default.}

\item{gene2symbol}{\code{data.frame} containing gene-to-symbol mappings. Must
contain the columns \code{geneID} and \code{geneName}.}
}
\value{
\itemize{
\item "\code{grid}": Show \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}}, paneled per gene.
\item "\code{wide}": Show \code{ggplot} in wide format, with genes on the x-axis.
\item "\code{list}": \code{list}, containing per gene \code{ggplot} objects.
\item "\code{markdown}": Show tabset R Markdown output, tabbed per gene.
}
}
\description{
Plot Individual Genes
}
\examples{
load(system.file("extdata/bcb_small.rda", package = "bcbioRNASeq"))
load(system.file("extdata/dds_small.rda", package = "bcbioRNASeq"))
load(system.file("extdata/rld_small.rda", package = "bcbioRNASeq"))

# Gene identifiers
genes <- head(rownames(bcb_small), 8L)

# bcbioRNASeq ====
plotGene(bcb_small, genes = genes, return = "grid")
plotGene(bcb_small, genes = genes, return = "wide")

# DESeqDataSet ====
plotGene(dds_small, genes = genes)

# DESeqTransform ====
plotGene(rld_small, genes = genes)
}
\seealso{
\code{\link[DESeq2:plotCounts]{DESeq2::plotCounts()}}.

Other Gene Expression Functions: \code{\link{plotGenderMarkers}},
  \code{\link{plotHeatmap}}
}
\author{
Michael Steinbaugh
}
