% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotGene.R
\docType{methods}
\name{plotGene}
\alias{plotGene}
\alias{plotGene,bcbioRNADataSet-method}
\title{Plot Individual Genes}
\usage{
plotGene(object, ...)

\S4method{plotGene}{bcbioRNADataSet}(object, normalized = "tpm", gene,
  format = "symbol")
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{normalized}{Normalization method. Supports \code{tpm} (\strong{default}), \code{tmm},
\code{rlog}, or \code{vst}.}

\item{gene}{Gene identifier. Can input multiple genes as a character vector.}

\item{format}{Ensembl identifier format. Supports \code{symbol} (\strong{default}) or
\code{ensgene}.}
}
\value{
\link{ggplot}.
}
\description{
Plot Individual Genes
}
\examples{
data(bcb)
plotGene(bcb, gene = c("Sulf1", "Phf3"))
}
\seealso{
\code{\link[DESeq2:plotCounts]{DESeq2::plotCounts()}}.

Other Quality Control Plots: \code{\link{plot53Bias}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotGenesDetected}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh
}
\concept{Quality Control Plots}
