% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotGene.R
\docType{methods}
\name{plotGene}
\alias{plotGene}
\alias{plotGene,bcbioRNASeq-method}
\title{Plot Individual Genes}
\usage{
\S4method{plotGene}{bcbioRNASeq}(object, interestingGroups,
  normalized = "tpm", gene, format = "symbol",
  color = scale_color_viridis(discrete = TRUE))
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{Category to use to group samples (color and shape).
If unset, this is automatically determined by the metadata set inside the
\link{bcbioRNASeq} object.}

\item{normalized}{Normalization method. Supports \code{tpm} (\strong{default}), \code{tmm},
\code{rlog}, or \code{vst}.}

\item{gene}{Gene identifier. Can input multiple genes as a character vector.}

\item{format}{Ensembl identifier format. Supports \code{symbol} (\strong{default}) or
\code{ensgene}.}

\item{color}{Desired ggplot color scale. Defaults to
\code{\link[viridis:scale_color_viridis]{viridis::scale_color_viridis()}}. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using
\code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}
}
\value{
\itemize{
\item \code{return = FALSE}: \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}} graphical output.
\item \code{return = TRUE}: \link{list} of per gene \link{ggplot} objects.
}
}
\description{
Plot Individual Genes
}
\examples{
# Gene symbols
symbol <- rowData(bcb)[["symbol"]][1:3]
print(symbol)
plotGene(bcb, gene = symbol, format = "symbol")

# Gene identifiers
\dontrun{
ensgene <- rownames(bcb)[1:3]
print(ensgene)
plotGene(bcb, gene = ensgene, format = "ensgene")
}
}
\seealso{
\code{\link[DESeq2:plotCounts]{DESeq2::plotCounts()}}.

Other Quality Control Plots: \code{\link{plot53Bias}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotGenesDetected}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh
}
\concept{Quality Control Plots}
