% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotGene.R
\docType{methods}
\name{plotGene}
\alias{plotGene}
\alias{plotGene,matrix-method}
\alias{plotGene,bcbioRNASeq-method}
\alias{plotGene,DESeqDataSet-method}
\alias{plotGene,DESeqTransform-method}
\title{Plot Individual Genes}
\usage{
\S4method{plotGene}{matrix}(object, genes, gene2symbol = NULL, colData,
  interestingGroups = "sampleName", countsAxisLabel = "counts",
  medianLine = TRUE, color = scale_color_viridis(discrete = TRUE),
  headerLevel = 2L, return = c("grid", "wide", "list", "markdown"))

\S4method{plotGene}{bcbioRNASeq}(object, genes, normalized = "rlog",
  interestingGroups, medianLine = TRUE, color = scale_color_viridis(discrete
  = TRUE), headerLevel = 2L, return = c("grid", "wide", "list", "markdown"))

\S4method{plotGene}{DESeqDataSet}(object, genes, gene2symbol = NULL,
  interestingGroups = "sampleName", medianLine = TRUE,
  color = scale_color_viridis(discrete = TRUE), headerLevel = 2L,
  return = c("grid", "wide", "list", "markdown"))

\S4method{plotGene}{DESeqTransform}(object, genes, gene2symbol = NULL,
  interestingGroups = "sampleName", medianLine = TRUE,
  color = scale_color_viridis(discrete = TRUE), headerLevel = 2L,
  return = c("grid", "wide", "list", "markdown"))
}
\arguments{
\item{object}{Object.}

\item{genes}{Gene identifiers (rownames) to plot. These must be the stable
identifiers (e.g. "ENSG00000000003") used on Ensembl and not the gene
symbols.}

\item{gene2symbol}{Apply gene identifier to symbol mappings. A gene2symbol
\code{data.frame} can be passed in, and must contain the columns \code{geneID} and
\code{geneName}.}

\item{colData}{Sample metadata \code{data.frame}.}

\item{interestingGroups}{Character vector denoting groups of interest that
define the samples. If left unset, defaults to \code{sampleName}.}

\item{countsAxisLabel}{Text label of counts axis.}

\item{medianLine}{Include median line for each group. Disabled when samples
are grouped by \code{sampleName}.}

\item{color}{Desired ggplot color scale. Defaults to
\code{\link[=scale_color_viridis]{scale_color_viridis()}}. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[=scale_color_manual]{scale_color_manual()}}.}

\item{headerLevel}{R Markdown header level.}

\item{return}{Object class to return. Uses \code{\link[=match.arg]{match.arg()}} internally and picks
the first item in the vector by default.}

\item{normalized}{Normalization method. Supports \code{tpm} (\strong{default}), \code{tmm},
\code{rlog}, or \code{vst}.}
}
\value{
\itemize{
\item "\code{grid}": Show \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}}, paneled per gene.
\item "\code{wide}": Show \code{ggplot} in wide format, with genes on the x-axis.
\item "\code{list}": \code{list}, containing per gene \code{ggplot} objects.
\item "\code{markdown}": Show tabset R Markdown output, tabbed per gene.
}
}
\description{
Plot Individual Genes
}
\examples{
# Gene identifiers
genes <- head(rownames(bcb_small), 8L)

# bcbioRNASeq ====
plotGene(bcb_small, genes = genes, return = "grid")
plotGene(bcb_small, genes = genes, return = "wide")

# DESeqDataSet ====
plotGene(dds_small, genes = genes)

# DESeqTransform ====
plotGene(rld_small, genes = genes)
}
\seealso{
\code{\link[DESeq2:plotCounts]{DESeq2::plotCounts()}}.

Other Gene Expression Functions: \code{\link{plotGenderMarkers}},
  \code{\link{plotHeatmap}}
}
\author{
Michael Steinbaugh
}
