% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotGene.R
\docType{methods}
\name{plotGene}
\alias{plotGene}
\alias{plotGene,SummarizedExperiment-method}
\alias{plotGene,bcbioRNASeq-method}
\alias{plotGene,DESeqDataSet-method}
\alias{plotGene,DESeqTransform-method}
\title{Plot Individual Genes}
\usage{
\S4method{plotGene}{SummarizedExperiment}(object, genes, interestingGroups,
  countsAxisLabel = "counts", medianLine = TRUE,
  color = scale_color_viridis(discrete = TRUE), headerLevel = 2L,
  return = c("facet", "wide", "grid", "markdown", "list"))

\S4method{plotGene}{bcbioRNASeq}(object, normalized = c("rlog", "vst", "tpm"),
  ...)

\S4method{plotGene}{DESeqDataSet}(object, ...)

\S4method{plotGene}{DESeqTransform}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{genes}{Character vector of genes to include. These must match the
rownames of the object. It is best practice to use the stable gene
identifiers from Ensembl (e.g. "ENSG00000000003") and not the gene symbols.}

\item{interestingGroups}{Character vector denoting groups of interest that
define the samples. If left unset, defaults to \code{sampleName}.}

\item{medianLine}{Include median line for each group. Disabled if samples
are colored by sample name.}

\item{color}{Desired ggplot color scale. Must supply discrete values. When
set to \code{NULL}, the default ggplot2 color palette will be used. If manual
color definitions are desired, we recommend using
\code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}

\item{headerLevel}{R Markdown header level.}

\item{return}{Object class to return. Uses \code{\link[=match.arg]{match.arg()}} internally and picks
the first item in the vector by default.}

\item{normalized}{Character indicating which normalization method to apply:
\itemize{
\item "\code{tpm}": Transcripts per million (tximport).
\item "\code{tmm}": edgeR trimmed mean of M-values. Calculated on the fly.
\item "\code{rlog}": DESeq2 \strong{log2} regularized log transformation.
\item "\code{vst}": DESeq2 \strong{log2} variance stabilizing transformation.
}}

\item{...}{Additional arguments.}
}
\value{
\itemize{
\item "\code{grid}": Show \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}}, paneled per gene.
\item "\code{wide}": Show \code{ggplot} in wide format, with genes on the x-axis.
\item "\code{list}": \code{list}, containing per gene \code{ggplot} objects.
\item "\code{markdown}": Show tabset R Markdown output, tabbed per gene.
}
}
\description{
Plot Individual Genes
}
\examples{
# Gene identifiers
genes <- head(rownames(bcb_small), 8L)

# bcbioRNASeq ====
plotGene(bcb_small, genes = genes, return = "grid")
plotGene(bcb_small, genes = genes, return = "wide")

# DESeqDataSet ====
plotGene(dds_small, genes = genes)

# DESeqTransform ====
plotGene(rld_small, genes = genes)
}
\seealso{
\code{\link[DESeq2:plotCounts]{DESeq2::plotCounts()}}.
}
\author{
Michael Steinbaugh
}
