% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGene-methods.R
\docType{methods}
\name{plotGene}
\alias{plotGene}
\alias{plotGene,bcbioRNASeq-method}
\alias{plotGene,DESeqAnalysis-method}
\title{Plot Gene Expression}
\usage{
\S4method{plotGene}{bcbioRNASeq}(object, genes, normalized = c("vst",
  "rlog", "tmm", "tpm", "rle"), interestingGroups = NULL,
  medianLine = TRUE, color = getOption("basejump.discrete.color",
  NULL), legend = getOption("basejump.legend", TRUE),
  style = c("facet", "wide"))

\S4method{plotGene}{DESeqAnalysis}(object, genes,
  interestingGroups = NULL, medianLine = TRUE,
  color = getOption("basejump.discrete.color", NULL),
  legend = getOption("basejump.legend", TRUE), style = c("facet",
  "wide"))
}
\arguments{
\item{object}{Object.}

\item{genes}{\code{character}. Gene identifiers.}

\item{normalized}{\code{string}. Which normalization method to apply:
- "\code{tpm}": Transcripts per million (tximport).
- "\code{tmm}": edgeR trimmed mean of M-values. Calculated on the fly.
- "\code{rlog}": DESeq2 \strong{log2} regularized log transformation.
- "\code{vst}": DESeq2 \strong{log2} variance stabilizing transformation.}

\item{interestingGroups}{\code{character}. Groups of interest.}

\item{medianLine}{\code{boolean}. Include median line for each group. Disabled if
samples are colored by sample name.}

\item{color}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}. To set the discrete color
palette globally, use \code{options(basejump.discrete.color = ggplot2::scale_color_viridis_d())}.}

\item{legend}{\code{boolean}. Show plot legend.}

\item{style}{\code{string}. Plot style.}
}
\value{
\itemize{
\item \code{style = "facet"}: \code{ggplot} grouped by \code{sampleName}, with
\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} applied to panel the samples.
\item \code{style = "wide"}: \code{ggplot} in wide format, with genes on the x-axis.
}
}
\description{
Plot Gene Expression
}
\examples{
g2s <- gene2symbol(bcb_small)
geneIDs <- head(g2s[["geneID"]])
print(geneIDs)
geneNames <- head(g2s[["geneName"]])
print(geneNames)

# bcbioRNASeq ====
object <- bcb_small
plotGene(
    object = object,
    genes = geneIDs,
    normalized = "vst",
    style = "facet"
)
plotGene(
    object = object,
    genes = geneNames,
    normalized = "vst",
    style = "wide"
)

# DESeqAnalysis ====
object <- deseq_small
plotGene(object, genes = geneIDs, style = "facet")
plotGene(object, genes = geneNames, style = "wide")
}
\seealso{
Other Plot Functions: \code{\link{plotCountsPerBiotype}},
  \code{\link{plotHeatmap}}, \code{\link{plotQC}},
  \code{\link{plotZerosVsDepth}}
}
\author{
Michael Steinbaugh
}
\concept{Gene Expression Functions}
