% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotGene.R
\docType{methods}
\name{plotGene}
\alias{plotGene}
\alias{plotGene,bcbioRNASeq-method}
\alias{plotGene,matrix-method}
\title{Plot Individual Genes}
\usage{
\S4method{plotGene}{bcbioRNASeq}(object, genes, interestingGroups,
  normalized = "tpm", format = "ensgene",
  color = viridis::scale_color_viridis(discrete = TRUE), returnList = FALSE)

\S4method{plotGene}{matrix}(object, genes, metadata,
  interestingGroups = "sampleName",
  color = viridis::scale_color_viridis(discrete = TRUE),
  countsAxisLabel = "counts", returnList = FALSE)
}
\arguments{
\item{object}{Object.}

\item{genes}{Gene identifier(s). Can input multiple genes as a character
vector.}

\item{interestingGroups}{Category to use to group samples. In the plotting
functions, this will define color and shape, where applicable. If unset,
this is automatically determined by the metadata set inside the
\link{bcbioRNASeq} object. When set to \code{NULL}, this will default to
\code{sampleName}.}

\item{normalized}{Normalization method. Supports \code{tpm} (\strong{default}), \code{tmm},
\code{rlog}, or \code{vst}.}

\item{format}{Ensembl identifier format. Supports \code{ensgene} (\strong{recommended})
or \code{symbol}.}

\item{color}{Desired ggplot color scale. Defaults to
\code{\link[viridis:scale_color_viridis]{viridis::scale_color_viridis()}}. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using
\code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}

\item{returnList}{Return the plotlist used to generate the paneled,
multi-gene plot with \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}}.}

\item{metadata}{Sample metadata \link{data.frame}.}

\item{countsAxisLabel}{Text label of counts axis.}
}
\value{
\itemize{
\item \code{returnList = FALSE}: \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}} graphical output.
\item \code{returnList = TRUE}: \link{list} of per gene \link{ggplot} objects.
}
}
\description{
Plot Individual Genes
}
\examples{
load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioRNASeq"))

# Gene identifiers
ensgene <- rownames(bcb)[1:4]
print(ensgene)
plotGene(
    bcb,
    genes = ensgene,
    format = "ensgene")

# Gene symbols
symbol <- rowData(bcb)[["symbol"]][1:4]
print(symbol)
plotGene(
    bcb,
    genes = symbol,
    format = "symbol")

# Default ggplot2 color palette
plotGene(
    bcb,
    genes = ensgene,
    format = "ensgene",
    interestingGroups = "sampleName",
    color = NULL)
}
\seealso{
\code{\link[DESeq2:plotCounts]{DESeq2::plotCounts()}}.

Other Quality Control Plots: \code{\link{plot53Bias}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotGenesDetected}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh
}
