% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGene-methods.R
\docType{methods}
\name{plotGene}
\alias{plotGene}
\alias{plotGene,bcbioRNASeq-method}
\title{Plot gene expression}
\usage{
\S4method{plotGene}{bcbioRNASeq}(object, genes, normalized = c("vst",
  "rlog", "tmm", "rle", "tpm", "fpkm"), interestingGroups = NULL,
  medianLine = TRUE, color = getOption("basejump.discrete.color",
  NULL), legend = getOption("basejump.legend", TRUE),
  style = c("facet", "wide"))
}
\arguments{
\item{object}{Object.}

\item{genes}{\code{character}.
Gene identifiers. It is considered better practice to input the stable gene
identifiers from Ensembl (e.g. "ENSG00000000003") and not the (HGNC) gene
symbols (e.g. "TSPN6"), if possible.}

\item{normalized}{\code{character(1)}.
Which normalization method to apply:\preformatted{- "`tpm`": Transcripts per million (tximport).
- "`tmm`": edgeR trimmed mean of M-values. Calculated on the fly.
- "`rlog`": DESeq2 **log2** regularized log transformation.
- "`vst`": DESeq2 **log2** variance stabilizing transformation.
}}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{medianLine}{\code{logical(1)}.
Include median line for each group. Disabled if samples are colored by
sample name.}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.

To set the discrete color palette globally, use:\preformatted{options(basejump.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{legend}{\code{logical(1)}.
Show plot legend.}

\item{style}{\code{character(1)}.
Plot style.}
}
\value{
\itemize{
\item \code{style = "facet"}:
\code{ggplot} grouped by \code{sampleName}, with
\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} applied to panel the samples.
\item \code{style = "wide"}:
\code{ggplot} in wide format, with genes on the x-axis.
}
}
\description{
Plot gene expression
}
\examples{
data(bcb)

g2s <- basejump::Gene2Symbol(bcb)
geneIDs <- head(g2s[["geneID"]])
print(geneIDs)
geneNames <- head(g2s[["geneName"]])
print(geneNames)

plotGene(
    object = bcb,
    genes = geneIDs,
    normalized = "vst",
    style = "facet"
)
plotGene(
    object = bcb,
    genes = geneNames,
    normalized = "vst",
    style = "wide"
)
}
\author{
Michael Steinbaugh
}
