% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotGeneHeatmap.R
\docType{methods}
\name{plotGeneHeatmap}
\alias{plotGeneHeatmap}
\alias{plotGeneHeatmap}
\alias{plotGeneHeatmap,bcbioRNADataSet-method}
\alias{plotGeneHeatmap,DESeqDataSet-method}
\alias{plotGeneHeatmap,DESeqTransform-method}
\alias{plotGeneHeatmap,matrix-method}
\title{Gene Heatmap}
\usage{
plotGeneHeatmap(object, ...)

\S4method{plotGeneHeatmap}{bcbioRNADataSet}(object, genes = NULL,
  title = NULL)

\S4method{plotGeneHeatmap}{DESeqDataSet}(object, genes = NULL,
  annotationCol = NULL, title = NULL)

\S4method{plotGeneHeatmap}{DESeqTransform}(object, genes = NULL,
  annotationCol = NULL, title = NULL)

\S4method{plotGeneHeatmap}{matrix}(object, genes = NULL,
  annotationCol = NULL, title = NULL)
}
\arguments{
\item{object}{Counts matrix.}

\item{...}{Additional arguments.}

\item{genes}{Character vector of specific gene identifiers to plot.}

\item{title}{Plot title.}

\item{annotationCol}{\link{data.frame} that specifies the annotations shown on the
right side of the heatmap. Each row of this \link{data.frame} defines the
features of the heatmap columns.}
}
\value{
Graphical output only.
}
\description{
These functions facilitate heatmap plotting of a specified set of genes. By
default, row- and column-wise hierarchical clustering is performed using the
Ward method, but this behavior can be overrided by setting \code{cluster_rows} or
\code{cluster_cols} to \code{FALSE}. When column clustering is disabled, the columns
are sorted by the interesting groups (\code{interestingGroups}) specified in the
\link{bcbioRNADataSet} and then the sample names.
}
\examples{
data(bcb, dds, rld)

# bcbioRNADataSet
plotGeneHeatmap(bcb)

# Genes as Ensembl identifiers
genes <- counts(bcb)[1L:20L, ] \%>\% rownames
plotGeneHeatmap(bcb, genes = genes)

# DESeqDataSet
plotGeneHeatmap(dds)

# DESeqTransform
plotGeneHeatmap(rld)
}
\seealso{
\code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}.
}
\author{
Michael Steinbaugh
}
