% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotGeneSaturation.R
\docType{methods}
\name{plotGeneSaturation}
\alias{plotGeneSaturation}
\alias{plotGeneSaturation}
\alias{plotGeneSaturation,bcbioRNASeq-method}
\title{Plot Gene Detection Saturation}
\usage{
plotGeneSaturation(object, ...)

\S4method{plotGeneSaturation}{bcbioRNASeq}(object, normalized = c("rlog",
  "vst", "tmm", "tpm"), interestingGroups, minCounts = 0L, trendline = TRUE,
  color = scale_color_viridis(discrete = TRUE), title = "gene saturation")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{normalized}{Count normalization method. See \code{\link[=counts]{counts()}} documentation
for more information.}

\item{interestingGroups}{Category to use to group samples. In the plotting
functions, this will define color and shape, where applicable. If unset,
this is automatically determined by the metadata set inside the
\link{bcbioRNASeq} object. When set to \code{NULL}, this will default to
\code{sampleName}.}

\item{minCounts}{Numeric value for filtering the counts matrix before
plotting.}

\item{trendline}{Include a trendline for each group.}

\item{color}{Desired ggplot color scale. Defaults to
\code{\link[=scale_color_viridis]{scale_color_viridis()}}. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[=scale_color_manual]{scale_color_manual()}}.}

\item{title}{Title.}
}
\value{
\code{ggplot}.
}
\description{
Plot Gene Detection Saturation
}
\examples{
# Minimal exampe distorts the y-axis
plotGeneSaturation(bcb_small, interestingGroups = "sampleName")
}
\seealso{
Other Quality Control Functions: \code{\link{metrics}},
  \code{\link{plot5Prime3PrimeBias}},
  \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotDispEsts}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGenesDetected}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}}, \code{\link{plotMeanSD}},
  \code{\link{plotPCACovariates}}, \code{\link{plotPCA}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
