% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotGeneSaturation-methods.R
\docType{methods}
\name{plotGeneSaturation}
\alias{plotGeneSaturation}
\alias{plotGeneSaturation,bcbioRNASeq-method}
\title{Plot Gene Detection Saturation}
\usage{
plotGeneSaturation(object, ...)

\S4method{plotGeneSaturation}{bcbioRNASeq}(object, interestingGroups,
  minCounts = 1L, trendline = FALSE, label = getOption("bcbio.label",
  FALSE), color = getOption("bcbio.color", NULL),
  title = "gene saturation")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{interestingGroups}{Character vector denoting groups of interest that
define the samples. If left unset, defaults to \code{sampleName}.}

\item{minCounts}{Minimum number of counts per gene in the counts matrix.}

\item{trendline}{Include a trendline for each group.}

\item{label}{Superimpose sample text labels on the plot.}

\item{color}{Desired ggplot color scale. Must supply discrete values. When
set to \code{NULL}, the default ggplot2 color palette will be used. If manual
color definitions are desired, we recommend using
\code{\link[ggplot2:scale_colour_manual]{ggplot2::scale_colour_manual()}}.}

\item{title}{Title of plot.}
}
\value{
\code{ggplot}.
}
\description{
We should observe a linear trend in the number of genes detected with the
number of mapped reads, which indicates that the sample input was not
overloaded.
}
\examples{
plotGeneSaturation(bcb_small, label = FALSE)
plotGeneSaturation(bcb_small, label = TRUE)
}
\seealso{
Other Quality Control Functions: \code{\link{plot5Prime3PrimeBias}},
  \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotDispEsts}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGenesDetected}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}}, \code{\link{plotMeanSD}},
  \code{\link{plotPCACovariates}}, \code{\link{plotPCA}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
\concept{Quality Control Functions}
