% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-plotGeneDetectionSaturation.R
\docType{methods}
\name{plotGeneSaturation}
\alias{plotGeneSaturation}
\alias{plotGeneSaturation}
\alias{plotGeneSaturation,bcbioRNADataSet,missing-method}
\alias{plotGeneSaturation,data.frame,matrix-method}
\title{Plot Gene Detection Saturation}
\usage{
plotGeneSaturation(object, counts, ...)

\S4method{plotGeneSaturation}{bcbioRNADataSet,missing}(object,
  normalized = "tmm", ...)

\S4method{plotGeneSaturation}{data.frame,matrix}(object, counts,
  interestingGroup = "sampleName", minCounts = 0L)
}
\arguments{
\item{object}{Object.}

\item{counts}{Counts matrix.}

\item{...}{Additional arguments.}

\item{normalized}{Count normalization method. See \code{\link[=counts]{counts()}} documentation
for more information.}

\item{interestingGroup}{\emph{Optional}. Category to use to group samples (color
and shape). If unset, this is automatically determined by the metadata set
inside the \link{bcbioRNADataSet}.}

\item{minCounts}{Numeric value for filtering the counts matrix before
plotting.}
}
\value{
\link{ggplot}.
}
\description{
Plot Gene Detection Saturation
}
\examples{
data(bcb, rld)

# bcbioRNADataSet
plotGeneSaturation(bcb)

# data.frame + matrix
plotGeneSaturation(metrics(bcb), assay(rld))
}
\seealso{
Other Quality Control Plots: \code{\link{plot53Bias}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGenesDetected}}, \code{\link{plotGene}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
