% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneSaturation-methods.R
\docType{methods}
\name{plotGeneSaturation}
\alias{plotGeneSaturation}
\alias{plotGeneSaturation,bcbioRNASeq-method}
\title{Plot gene detection saturation}
\usage{
plotGeneSaturation(object, ...)

\S4method{plotGeneSaturation}{bcbioRNASeq}(object, interestingGroups,
  minCounts = 1L, trendline = FALSE, label = getOption("bcbio.label",
  FALSE), color = getOption("bcbio.discrete.color", NULL),
  title = "gene saturation")
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}. Groups of interest that define the
samples. If left unset, defaults to \code{sampleName}.}

\item{minCounts}{\code{scalar integer}. Minimum number of counts per gene in the
counts matrix.}

\item{trendline}{\code{boolean}. Include a trendline for each group.}

\item{label}{\code{boolean}. Superimpose sample text labels on the plot.}

\item{color}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.
To set the discrete color palette globally, use
\code{options(bcbio.discrete.color = scale_color_viridis_d())}.}

\item{title}{\code{string} or \code{NULL}. Plot title.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
We should observe a linear trend in the number of genes detected with the
number of mapped reads, which indicates that the sample input was not
overloaded.
}
\examples{
plotGeneSaturation(bcb_small, label = FALSE)
plotGeneSaturation(bcb_small, label = TRUE)
}
\seealso{
Other Quality Control Functions: \code{\link{plot5Prime3PrimeBias}},
  \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotDispEsts}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGenesDetected}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}}, \code{\link{plotMeanSD}},
  \code{\link{plotPCACovariates}}, \code{\link{plotPCA}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
\concept{Quality Control Functions}
