% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneSaturation-methods.R
\docType{methods}
\name{plotGeneSaturation}
\alias{plotGeneSaturation}
\alias{plotGeneSaturation,bcbioRNASeq-method}
\title{Plot gene detection saturation}
\usage{
plotGeneSaturation(object, ...)

\S4method{plotGeneSaturation}{bcbioRNASeq}(object,
  interestingGroups = NULL, minCounts = 1L, perMillion = TRUE,
  trendline = FALSE, label = getOption(x = "acid.label", default =
  FALSE), color = getOption(x = "acid.color.discrete", default =
  acidplots::scale_colour_synesthesia_d()), title = "Gene saturation")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
We should observe a linear trend in the number of genes detected with the
number of mapped reads, which indicates that the sample input was not
overloaded.
}
\note{
Updated 2019-08-20.
}
\examples{
data(bcb)
plotGeneSaturation(bcb, label = FALSE)
plotGeneSaturation(bcb, label = TRUE)
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
