% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneSaturation-methods.R
\docType{methods}
\name{plotGeneSaturation}
\alias{plotGeneSaturation}
\alias{plotGeneSaturation,bcbioRNASeq-method}
\title{Plot Gene Detection Saturation}
\usage{
\S4method{plotGeneSaturation}{bcbioRNASeq}(object,
  interestingGroups = NULL, minCounts = 1L, perMillion = TRUE,
  trendline = FALSE, label = getOption("basejump.label", FALSE),
  color = getOption("basejump.color.discrete", NULL),
  title = "gene saturation")
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character} or \code{NULL}. Groups of interest that
define the samples. If left unset, defaults to \code{sampleName}.}

\item{minCounts}{\code{scalar integer}. Minimum number of counts per gene in the
count matrix.}

\item{perMillion}{\code{boolean}. Display as counts per million.}

\item{trendline}{\code{boolean}. Include a trendline for each group.}

\item{label}{\code{boolean}. Superimpose sample text labels on the plot.}

\item{color}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{ggplot2::scale_color_manual()}. To set the discrete color
palette globally, use
\code{options(basejump.discrete.color = ggplot2::scale_color_viridis_d())}.}

\item{title}{\code{string} or \code{NULL}. Plot title.}
}
\value{
\code{ggplot}.
}
\description{
We should observe a linear trend in the number of genes detected with the
number of mapped reads, which indicates that the sample input was not
overloaded.
}
\examples{
data(bcb)
plotGeneSaturation(bcb, label = FALSE)
plotGeneSaturation(bcb, label = TRUE)
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
