% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneSaturation-methods.R
\docType{methods}
\name{plotGeneSaturation}
\alias{plotGeneSaturation}
\alias{plotGeneSaturation,bcbioRNASeq-method}
\title{Plot gene detection saturation}
\usage{
\S4method{plotGeneSaturation}{bcbioRNASeq}(object,
  interestingGroups = NULL, minCounts = 1L, perMillion = TRUE,
  trendline = FALSE, label = getOption(x = "acid.label", default =
  FALSE), color = getOption(x = "acid.color.discrete", default =
  acidplots::scale_colour_synesthesia_d()), title = "Gene saturation")
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{minCounts}{\code{integer(1)}.
Minimum number of counts per gene in the count matrix.}

\item{perMillion}{\code{logical(1)}.
Display as counts per million.}

\item{trendline}{\code{logical(1)}.
Include a trendline for each group.}

\item{label}{\code{logical(1)}.
Superimpose sample text labels on the plot.}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.

To set the discrete color palette globally, use:\preformatted{options(acid.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{title}{\code{character(1)}.
Plot title.}
}
\value{
\code{ggplot}.
}
\description{
We should observe a linear trend in the number of genes detected with the
number of mapped reads, which indicates that the sample input was not
overloaded.
}
\examples{
data(bcb)
plotGeneSaturation(bcb, label = FALSE)
plotGeneSaturation(bcb, label = TRUE)
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
