% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneSaturation-methods.R
\docType{methods}
\name{plotGeneSaturation}
\alias{plotGeneSaturation}
\alias{plotGeneSaturation,bcbioRNASeq-method}
\title{Plot gene detection saturation}
\usage{
\S4method{plotGeneSaturation}{bcbioRNASeq}(object,
  interestingGroups = NULL, minCounts = 1L, perMillion = TRUE,
  trendline = FALSE, label = getOption("acid.label", default = FALSE),
  color = getOption("acid.color.discrete", default = NULL),
  title = "gene saturation")
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{trendline}{\code{logical(1)}.
Include a trendline for each group.}
}
\value{
\code{ggplot}.
}
\description{
We should observe a linear trend in the number of genes detected with the
number of mapped reads, which indicates that the sample input was not
overloaded.
}
\examples{
data(bcb)
plotGeneSaturation(bcb, label = FALSE)
plotGeneSaturation(bcb, label = TRUE)
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
