% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotGeneSaturation-methods.R
\name{plotGeneSaturation}
\alias{plotGeneSaturation}
\alias{plotGeneSaturation,bcbioRNASeq-method}
\title{Plot gene detection saturation}
\usage{
plotGeneSaturation(object, ...)

\S4method{plotGeneSaturation}{bcbioRNASeq}(
  object,
  interestingGroups = NULL,
  minCounts = 1L,
  perMillion = TRUE,
  trendline = FALSE,
  label = getOption(x = "acid.label", default = FALSE),
  color = getOption(x = "acid.color.discrete", default = getOption(x =
    "ggplot2.discrete.colour", default = AcidPlots::scale_color_synesthesia_d())),
  labels = list(title = "Gene saturation", subtitle = NULL, x = "mapped reads", y =
    "gene count")
)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{minCounts}{\code{integer(1)}.
Minimum number of counts per feature (i.e. gene).}

\item{perMillion}{\code{logical(1)}.
Display as counts per million.}

\item{trendline}{\code{logical(1)}.
Include trendline on plot.}

\item{label}{\code{logical(1)}.
Superimpose sample text labels on the plot.}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{ggplot2::scale_color_manual()}.

To set the discrete color palette globally, use:\preformatted{options(acid.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
We should observe a linear trend in the number of genes detected with the
number of mapped reads, which indicates that the sample input was not
overloaded.
}
\note{
Updated 2021-07-21.
}
\examples{
## bcbioRNASeq ====
data(bcb)
plotGeneSaturation(bcb, label = FALSE)
plotGeneSaturation(bcb, label = TRUE)
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
