% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotGenesDetected.R
\docType{methods}
\name{plotGenesDetected}
\alias{plotGenesDetected}
\alias{plotGenesDetected,bcbioRNASeqANY,missing-method}
\alias{plotGenesDetected,data.frame,DESeqDataSet-method}
\alias{plotGenesDetected,data.frame,matrix-method}
\title{Plot Genes Detected}
\usage{
plotGenesDetected(object, counts, ...)

\S4method{plotGenesDetected}{bcbioRNASeqANY,missing}(object, interestingGroups,
  passLimit = 20000, warnLimit = 15000, minCounts = 0, flip = TRUE)

\S4method{plotGenesDetected}{data.frame,DESeqDataSet}(object, counts,
  interestingGroups = "sampleName", passLimit = 20000, warnLimit = 15000,
  minCounts = 0, flip = TRUE)

\S4method{plotGenesDetected}{data.frame,matrix}(object, counts,
  interestingGroups = "sampleName", passLimit = 20000, warnLimit = 15000,
  minCounts = 0, flip = TRUE)
}
\arguments{
\item{object}{Object.}

\item{counts}{Object containing a count matrix.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{interestingGroups}{Category to use to group samples (color and shape).
If unset, this is automatically determined by the metadata set inside the
\link{bcbioRNASeq} object.}

\item{passLimit}{Threshold to plot pass color marker.}

\item{warnLimit}{Threshold to plot warning color marker.}

\item{minCounts}{Numeric value for filtering the counts matrix before
plotting.}

\item{flip}{Flip X and Y axes.}
}
\value{
\link{ggplot}.
}
\description{
Plot Genes Detected
}
\examples{
data(bcb, dds)

# bcbioRNASeq
plotGenesDetected(
    bcb,
    passLimit = NULL,
    warnLimit = NULL)

\dontrun{
plotGenesDetected(bcb, interestingGroups = "group")

# data.frame, DESeqDataSet
plotGenesDetected(
    metrics(bcb),
    counts = dds)

# data.frame, matrix
plotGenesDetected(
    metrics(bcb),
    counts = assay(dds))
}
}
\seealso{
Other Quality Control Plots: \code{\link{plot53Bias}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGeneSaturation}}, \code{\link{plotGene}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
\concept{Quality Control Plots}
