% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotHeatmap.R
\docType{methods}
\name{plotHeatmap}
\alias{plotHeatmap}
\alias{plotHeatmap,bcbioRNASeq-method}
\alias{plotHeatmap,DESeqDataSet-method}
\alias{plotHeatmap,DESeqTransform-method}
\title{Plot Heatmap}
\usage{
\S4method{plotHeatmap}{bcbioRNASeq}(object, normalized = c("rlog", "vst",
  "tmm", "tpm"), samples = NULL, genes = NULL, scale = "row",
  color = viridis, legendColor = viridis, title = NULL, ...)

\S4method{plotHeatmap}{DESeqDataSet}(object, normalized = TRUE,
  samples = NULL, genes = NULL, gene2symbol = NULL,
  annotationCol = NULL, scale = "row", color = viridis,
  legendColor = viridis, title = NULL, ...)

\S4method{plotHeatmap}{DESeqTransform}(object, samples = NULL, genes = NULL,
  gene2symbol = NULL, annotationCol = NULL, scale = "row",
  color = viridis, legendColor = viridis, title = NULL, ...)
}
\arguments{
\item{object}{Object.}

\item{normalized}{Character indicating which normalization method to apply:
\itemize{
\item "\code{tpm}": Transcripts per million (tximport).
\item "\code{tmm}": edgeR trimmed mean of M-values. Calculated on the fly.
\item "\code{rlog}": DESeq2 \strong{log2} regularized log transformation.
\item "\code{vst}": DESeq2 \strong{log2} variance stabilizing transformation.
}}

\item{samples}{Character vector of samples to include.}

\item{genes}{Character vector of genes to include. These must match the
rownames of the object. It is best practice to use the stable gene
identifiers from Ensembl (e.g. "ENSG00000000003") and not the gene symbols.}

\item{scale}{Character indicating if the values should be centered and scaled
in either the row direction or the column direction, or none. Corresponding
values are "row", "column" and "none".}

\item{color}{Colors to use for plot. Defaults to the \code{\link[=viridis]{viridis()}}
palette.}

\item{legendColor}{Colors to use for legend labels. Defaults to the
\code{\link[=viridis]{viridis()}} palette.}

\item{title}{\emph{Optional.} Plot title.}

\item{...}{Passthrough arguments to \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}.}

\item{gene2symbol}{\code{data.frame} containing gene-to-symbol mappings. Must
contain the columns \code{geneID} and \code{geneName}.}

\item{annotationCol}{\emph{Optional.} \code{data.frame} that defines annotation
mappings for the columns.}
}
\value{
Show heatmap and return \code{list}, containing \code{gtable}.
}
\description{
Plot Heatmap
}
\details{
When column clustering is disabled, the columns are sorted by the interesting
groups (\code{interestingGroups}) specified in the \link{bcbioRNASeq} and then the
sample names.
}
\examples{
load(system.file("extdata/bcb_small.rda", package = "bcbioRNASeq"))
load(system.file("extdata/dds_small.rda", package = "bcbioRNASeq"))
load(system.file("extdata/rld_small.rda", package = "bcbioRNASeq"))

# bcbioRNASeq ====
plotHeatmap(bcb_small, genes = head(rownames(bcb_small), 20L))

# Full transcriptome heatmap with default pheatmap colors
plotHeatmap(bcb_small, color = inferno, legendColor = inferno)

# DESeqDataSet ====
plotHeatmap(
    object = dds_small,
    genes = head(rownames(dds_small), 20L),
    gene2symbol = gene2symbol(bcb_small)
)

# DESeqTransform ====
plotHeatmap(
    object = rld_small,
    genes = head(rownames(rld_small), 20L),
    gene2symbol = gene2symbol(bcb_small)
)
}
\seealso{
Other Gene Expression Functions: \code{\link{plotGenderMarkers}},
  \code{\link{plotGene}}
}
\author{
Michael Steinbaugh
}
