% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotHeatmap.R
\docType{methods}
\name{plotHeatmap}
\alias{plotHeatmap}
\alias{plotHeatmap,bcbioRNASeq-method}
\alias{plotHeatmap,DESeqDataSet-method}
\alias{plotHeatmap,DESeqTransform-method}
\title{Plot Heatmap}
\usage{
\S4method{plotHeatmap}{bcbioRNASeq}(object, normalized = "rlog",
  samples = NULL, genes = NULL, scale = "row", color = viridis::viridis,
  legendColor = viridis::viridis, title = NULL, ...)

\S4method{plotHeatmap}{DESeqDataSet}(object, normalized = TRUE,
  samples = NULL, genes = NULL, gene2symbol = NULL,
  annotationCol = NULL, scale = "row", color = viridis::viridis,
  legendColor = viridis::viridis, title = NULL, ...)

\S4method{plotHeatmap}{DESeqTransform}(object, genes = NULL,
  gene2symbol = NULL, annotationCol = NULL, scale = "row",
  color = viridis::viridis(256L), legendColor = viridis::viridis,
  title = NULL, ...)
}
\arguments{
\item{object}{Object.}

\item{normalized}{Select raw counts (\code{FALSE}), DESeq2 normalized counts
(\code{TRUE}), or additional normalization methods:
\itemize{
\item \code{tpm}: Transcripts per million.
\item \code{tmm}: Trimmed mean of M-values (edgeR).
\item \code{rlog}: Regularized log transformation (\code{\link[DESeq2:rlog]{DESeq2::rlog()}}).
\item \code{vst}: Variance stabilizing transformation
(\code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}}).
}}

\item{samples}{\emph{Optional.} Samples (colnames) to plot.}

\item{genes}{\emph{Optional.} Gene identifiers (rownames) to plot. These must be
the stable identifiers (e.g. ENSG00000000003) used on Ensembl and not the
gene symbols.}

\item{scale}{Character indicating if the values should be centered and scaled
in either the row direction or the column direction, or none. Corresponding
values are "row", "column" and "none".}

\item{color}{Desired ggplot color scale. Defaults to
\code{\link[viridis:scale_color_viridis]{viridis::scale_color_viridis()}}. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using
\code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}

\item{legendColor}{Colors to use for legend labels. Defaults to the
\code{\link[viridis:viridis]{viridis::viridis()}}.}

\item{title}{Title.}

\item{...}{Additional arguments.}

\item{gene2symbol}{Apply gene identifier to symbol mappings. A gene2symbol
\link{data.frame} can be passed in, and must contain the columns \code{ensgene} and
\code{symbol}. then the Ensembl gene identifiers will be labeled in place of
gene symbols.}

\item{annotationCol}{\emph{Optional.} \link{data.frame} that defines annotation
mappings for the columns.}
}
\description{
Plot Heatmap
}
\details{
When column clustering is disabled, the columns are sorted by the interesting
groups (\code{interestingGroups}) specified in the \link{bcbioRNASeq} and then the
sample names.
}
\examples{
load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioRNASeq"))

# Use Ensembl identifiers to define genes
ensgene <- rownames(bcb)[1:20]
plotHeatmap(bcb, genes = ensgene)

# Use inferno color palette
plotHeatmap(
    bcb,
    genes = ensgene,
    color = viridis::inferno,
    legendColor = viridis::inferno)

# Transcriptome heatmap with default pheatmap colors
plotHeatmap(
    bcb,
    color = NULL,
    legendColor = NULL)

# DESeqDataSet
dds <- bcbio(bcb, "DESeqDataSet")
plotHeatmap(dds)

# DESeqTransform
rld <- assays(bcb)[["rlog"]]
plotHeatmap(rld)
}
\seealso{
Other Heatmaps: \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotDEGHeatmap}}
}
\author{
Michael Steinbaugh
}
