% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-plotIntronicMappingRate.R
\docType{methods}
\name{plotIntronicMappingRate}
\alias{plotIntronicMappingRate}
\alias{plotIntronicMappingRate}
\alias{plotIntronicMappingRate,bcbioRNASeq-method}
\title{Plot Intronic Mapping Rate}
\usage{
plotIntronicMappingRate(object, ...)

\S4method{plotIntronicMappingRate}{bcbioRNASeq}(object, interestingGroups,
  warnLimit = 20L, fill = scale_fill_viridis(discrete = TRUE),
  flip = TRUE, title = "intronic mapping rate")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{interestingGroups}{Character vector denoting groups of interest that
define the samples. If left unset, defaults to \code{sampleName}.}

\item{warnLimit}{Threshold to plot warning color marker.}

\item{fill}{Desired ggplot fill scale. Must supply discrete values. When set
to \code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.}

\item{flip}{Flip x and y axes.}

\item{title}{Title of plot.}
}
\value{
\code{ggplot}.
}
\description{
Plot Intronic Mapping Rate
}
\examples{
plotIntronicMappingRate(bcb_small)
}
\seealso{
Other Quality Control Functions: \code{\link{metrics}},
  \code{\link{plot5Prime3PrimeBias}},
  \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotDispEsts}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotGenesDetected}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}}, \code{\link{plotMeanSD}},
  \code{\link{plotPCACovariates}}, \code{\link{plotPCA}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
