% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIntronicMappingRate-methods.R
\docType{methods}
\name{plotIntronicMappingRate}
\alias{plotIntronicMappingRate}
\alias{plotIntronicMappingRate,bcbioRNASeq-method}
\title{Plot intronic mapping rate}
\usage{
plotIntronicMappingRate(object, ...)

\S4method{plotIntronicMappingRate}{bcbioRNASeq}(object,
  interestingGroups = NULL, limit = 0.2, fill = getOption(x =
  "acid.fill.discrete", default = acidplots::scale_fill_synesthesia_d()),
  flip = getOption(x = "acid.flip", default = TRUE),
  title = "Intronic mapping rate")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
The majority of reads should map to exons and not introns.
}
\note{
Updated 2019-08-07.
}
\examples{
data(bcb)
plotIntronicMappingRate(bcb)
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
