% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIntronicMappingRate-methods.R
\docType{methods}
\name{plotIntronicMappingRate}
\alias{plotIntronicMappingRate}
\alias{plotIntronicMappingRate,bcbioRNASeq-method}
\title{Plot intronic mapping rate}
\usage{
plotIntronicMappingRate(object, ...)

\S4method{plotIntronicMappingRate}{bcbioRNASeq}(object,
  interestingGroups = NULL, limit = 0.2, fill = getOption(x =
  "acid.fill.discrete", default = acidplots::scale_fill_synesthesia_d()),
  flip = getOption(x = "acid.flip", default = TRUE),
  title = "Intronic mapping rate")
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{limit}{\code{numeric(1)}.
Threshold limit.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.

To set the discrete fill palette globally, use:\preformatted{options(acid.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{flip}{\code{logical(1)}.
Flip x and y axes. Recommended for plots containing many samples.}

\item{title}{\code{character(1)}.
Title.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
The majority of reads should map to exons and not introns.
}
\note{
Updated 2019-09-15.
}
\examples{
data(bcb)
plotIntronicMappingRate(bcb)
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
