% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIntronicMappingRate-methods.R
\docType{methods}
\name{plotIntronicMappingRate}
\alias{plotIntronicMappingRate}
\alias{plotIntronicMappingRate,bcbioRNASeq-method}
\title{Plot Intronic Mapping Rate}
\usage{
\S4method{plotIntronicMappingRate}{bcbioRNASeq}(object,
  interestingGroups = NULL, limit = 0.2,
  fill = getOption("basejump.fill.discrete", NULL),
  flip = getOption("basejump.flip", TRUE),
  title = "intronic mapping rate")
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character} or \code{NULL}. Groups of interest that
define the samples. If left unset, defaults to \code{sampleName}.}

\item{limit}{\code{scalar numeric}. Threshold to denote on the plot, using a
dashed line.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 fill scale.
Must supply discrete values. When set to \code{NULL}, the default ggplot2 color
palette will be used. If manual color definitions are desired, we recommend
using \code{ggplot2::scale_fill_manual()}. To set the discrete fill palette
globally, use \code{options(basejump.fill.discrete = scale_fill_viridis_d())}.}

\item{flip}{\code{boolean}. Flip x and y axes. Recommended for plots containing
many samples.}

\item{title}{\code{string} or \code{NULL}. Plot title.}
}
\value{
\code{ggplot}.
}
\description{
The majority of reads should map to exons and not introns.
}
\examples{
data(bcb)
plotIntronicMappingRate(bcb)
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
