% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotMA.R
\docType{methods}
\name{plotMA}
\alias{plotMA}
\alias{plotMA,DESeqResults-method}
\alias{plotMA,data.frame-method}
\title{Plot Mean Average}
\usage{
\S4method{plotMA}{DESeqResults}(object, labelPoints = NULL,
  labelColumn = "rowname", pointColorScale = c("darkgrey", "red", "green"),
  labelColor = "black")

\S4method{plotMA}{data.frame}(object, alpha = 0.01, labelPoints = NULL,
  labelColumn = "rowname", pointColorScale = c("darkgrey", "red", "green"),
  labelColor = "black", title = TRUE)
}
\arguments{
\item{object}{Object.}

\item{labelPoints}{\emph{Optional}. Label these particular points.}

\item{labelColumn}{Match \code{labelPoints} argument to this column in the
results.}

\item{pointColorScale}{Point color scale. See
\code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}} for more information.}

\item{labelColor}{Text label color.}

\item{alpha}{Alpha level cutoff (Adjusted P value).}

\item{title}{Plot title.}
}
\value{
\link{ggplot}.
}
\description{
Plot Mean Average
}
\examples{
data(bcb, dds, res)
genes <- c("ENSMUSG00000104523", "ENSMUSG00000016918")

# DESeqResults
plotMA(res, labelPoints = genes)

# data.frame
df <- as.data.frame(res)
plotMA(df, labelPoints = genes)
}
\seealso{
Other Differential Expression Plots: \code{\link{plotVolcano}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
