% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMA-methods.R
\docType{methods}
\name{plotMA}
\alias{plotMA}
\alias{plotMA,DESeqAnalysis-method}
\alias{plotMA,DESeqResults-method}
\alias{plotMeanAverage}
\title{MA Plot}
\usage{
\S4method{plotMA}{DESeqAnalysis}(object, results, lfcShrink = TRUE,
  genes = NULL, ntop = 0L, direction = c("both", "up", "down"),
  pointColor = getOption("bcbio.point.color", "gray50"),
  sigPointColor = getOption("bcbio.sig.point.color", c(upregulated =
  "purple", downregulated = "orange")), return = c("ggplot",
  "DataFrame"))

\S4method{plotMA}{DESeqResults}(object, genes = NULL,
  gene2symbol = NULL, ntop = 0L, direction = c("both", "up", "down"),
  pointColor = getOption("bcbio.point.color", "gray50"),
  sigPointColor = getOption("bcbio.sig.point.color", c(upregulated =
  "purple", downregulated = "orange")), return = c("ggplot",
  "DataFrame"))

plotMeanAverage(...)
}
\arguments{
\item{object}{Object.}

\item{results}{\code{scalar}. Position or name of \code{DESeqResults}.}

\item{lfcShrink}{\code{boolean}. Use shrunken log2 fold change (LFC) values.}

\item{genes}{\code{character}. Genes to include. These must match the rownames of
the object. It is best practice to use the stable gene identifiers from
Ensembl (e.g. "ENSG00000000003") and not the gene symbols.}

\item{ntop}{\code{scalar integer}. Number of top genes to label.}

\item{direction}{\code{string}. Plot "\code{both}", "\code{up}", or "\code{down}" directions.}

\item{pointColor}{\code{string}. Default point color for the plot.}

\item{sigPointColor}{\code{character}. Color names for labeling upregulated and
downregulated genes. Also supports a character string for labeling DEGs
with the same color, regardless of direction.}

\item{return}{\code{string}. Object class to return. Uses \code{\link[=match.arg]{match.arg()}} internally
and picks the first item in the vector by default.}

\item{gene2symbol}{\code{gene2symbol}. Gene-to-symbol mappings. Must contain
\code{geneID} and \code{geneName} columns, with rownames defined.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
An MA plot is an application of a Bland–Altman plot for visual representation
of genomic data. The plot visualizes the differences between measurements
taken in two samples, by transforming the data onto M (log ratio) and A
(mean average) scales, then plotting these values.
}
\examples{
g2s <- gene2symbol(bcb_small)
geneIDs <- head(g2s[["geneID"]])
print(geneIDs)
geneNames <- head(g2s[["geneName"]])
print(geneNames)

# DESeqAnalysis ====
# This is the current recommended method.
object <- deseq_small
print(object)
plotMA(object)

# Customize the colors.
plotMA(
    object = object,
    pointColor = "black",
    sigPointColor = "purple"
)
plotMA(
    object = object,
    sigPointColor = c(
        upregulated = "green",
        downregulated = "red"
    )
)

# Directional support (up or down).
plotMA(object, direction = "up", ntop = 5L)
plotMA(object, direction = "down", ntop = 5L)

# Label genes manually.
# Note that either gene IDs or names (symbols) are supported.
plotMA(object = object, genes = geneIDs)
plotMA(object = object, genes = geneNames)
}
\seealso{
\code{\link[DESeq2:plotMA]{DESeq2::plotMA()}}.

Other Differential Expression Functions: \code{\link{alphaSummary}},
  \code{\link{contrastName}}, \code{\link{plotDEGHeatmap}},
  \code{\link{plotDEGPCA}}, \code{\link{plotVolcano}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Differential Expression Functions}
