% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMA-methods.R
\docType{methods}
\name{plotMA}
\alias{plotMA}
\alias{plotMA,DESeqResults-method}
\alias{plotMA,DESeqAnalysis-method}
\alias{plotMeanAverage}
\title{MA-plot: plot differences versus averages for high-throughput data}
\usage{
plotMA(object, ...)

\S4method{plotMA}{DESeqResults}(object, genes = NULL,
  gene2symbol = NULL, ntop = 0L, direction = c("both", "up", "down"),
  pointColor = getOption("basejump.point.color", "gray50"),
  sigPointColor = getOption("basejump.sig.point.color", c(upregulated =
  "purple", downregulated = "orange")), return = c("ggplot",
  "DataFrame"))

\S4method{plotMA}{DESeqAnalysis}(object, results = 1L,
  lfcShrink = TRUE, genes = NULL, ntop = 0L, direction = c("both",
  "up", "down"), pointColor = getOption("basejump.point.color",
  "gray50"), sigPointColor = getOption("basejump.sig.point.color",
  c(upregulated = "purple", downregulated = "orange")),
  return = c("ggplot", "DataFrame"))

plotMeanAverage(...)
}
\arguments{
\item{object}{
    A data object, typically containing count values from an RNA-Seq experiment or microarray intensity values.
  }

\item{...}{
    Additional arguments, for use in specific methods.
  }

\item{genes}{\code{character}. Genes to include. These must match the rownames of
the object. It is best practice to use the stable gene identifiers from
Ensembl (e.g. "ENSG00000000003") and not the gene symbols.}

\item{gene2symbol}{\code{gene2symbol}. Gene-to-symbol mappings. Must contain
\code{geneID} and \code{geneName} columns, with rownames defined.}

\item{ntop}{\code{scalar integer}. Number of top genes to label.}

\item{direction}{\code{string}. Plot "\code{both}", "\code{up}", or "\code{down}" directions.}

\item{pointColor}{\code{string}. Default point color for the plot.}

\item{sigPointColor}{\code{character}. Color names for labeling upregulated and
downregulated genes. Also supports a character string for labeling DEGs
with the same color, regardless of direction.}

\item{return}{\code{string}. Object class to return. Uses \code{\link[=match.arg]{match.arg()}} internally
and picks the first item in the vector by default.}

\item{results}{\code{scalar}. Position or name of \code{DESeqResults}.}

\item{lfcShrink}{\code{boolean}. Use shrunken log2 fold change (LFC) values.}
}
\value{
\code{ggplot}.
}
\description{
A generic function which produces an MA-plot for an object containing microarray, RNA-Seq
  or other data.
}
\details{
An MA plot is an application of a Bland–Altman plot for visual representation
of genomic data. The plot visualizes the differences between measurements
taken in two samples, by transforming the data onto M (log ratio) and A
(mean average) scales, then plotting these values.
}
\examples{
object <- bcbioRNASeq::deseq_small
print(object)

## DESeqAnalysis ====
## This is the current recommended method.

dds <- as(object, "DESeqDataSet")
g2s <- Gene2Symbol(dds)
geneIDs <- head(g2s[["geneID"]])
print(geneIDs)
geneNames <- head(g2s[["geneName"]])
print(geneNames)

plotMA(object)

## Customize the colors.
plotMA(
    object = object,
    pointColor = "black",
    sigPointColor = "purple"
)
plotMA(
    object = object,
    sigPointColor = c(
        upregulated = "green",
        downregulated = "red"
    )
)

## Directional support (up or down).
plotMA(object, direction = "up", ntop = 5L)
plotMA(object, direction = "down", ntop = 5L)

## Label genes manually.
## Note that either gene IDs or names (symbols) are supported.
plotMA(object = object, genes = geneIDs)
plotMA(object = object, genes = geneNames)
}
\seealso{
\code{\link[DESeq2:plotMA]{DESeq2::plotMA()}}.
}
\author{
Michael Steinbaugh, Rory Kirchner
}
