% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotMA.R
\docType{methods}
\name{plotMA}
\alias{plotMA}
\alias{plotMA,DESeqResults-method}
\alias{plotMA,data.frame-method}
\title{Plot Mean Average}
\usage{
\S4method{plotMA}{DESeqResults}(object, genes = NULL, format = "ensgene",
  gene2symbol = NULL, color = "darkgray", sigColor = "red",
  labelColor = "black")

\S4method{plotMA}{data.frame}(object, alpha = 0.01, genes = NULL,
  format = "ensgene", gene2symbol = NULL, color = "darkgray",
  sigColor = "red", labelColor = "black", title = NULL)
}
\arguments{
\item{object}{Object.}

\item{genes}{\emph{Optional}. Genes to label on the plot.}

\item{format}{Gene identifier format (\code{ensgene} or \code{symbol}) used for
matching.}

\item{gene2symbol}{Gene to symbol \link{data.frame} that defines
identifier mappings. Required if \code{genes} is defined.}

\item{color}{Default point color for the plot.}

\item{sigColor}{Color for points corresponding to significant genes that have
passed alpha level cutoffs.}

\item{labelColor}{Text label color.}

\item{alpha}{Alpha level cutoff (Adjusted P value).}

\item{title}{\emph{Optional}. Plot title.}
}
\value{
\link{ggplot}.
}
\description{
Plot Mean Average
}
\examples{
bcb <- examples[["bcb"]]
res <- examples[["res"]]
gene2symbol <- gene2symbol(bcb)

# DESeqResults
plotMA(
    res,
    genes = "ENSMUSG00000016918",
    format = "ensgene",
    gene2symbol = gene2symbol)
plotMA(
    res,
    genes = "Sulf1",
    format = "symbol",
    gene2symbol = gene2symbol)

# data.frame
df <- as.data.frame(res)
plotMA(df, alpha = 0.01)
}
\seealso{
Other Differential Expression Plots: \code{\link{plotVolcano}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
