% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotMA.R
\docType{methods}
\name{plotMA}
\alias{plotMA}
\alias{plotMA,DESeqResults-method}
\title{Plot Mean Average}
\usage{
\S4method{plotMA}{DESeqResults}(object, genes = NULL, gene2symbol = NULL,
  pointColor = "darkgray", sigPointColor = "purple", labelColor = "black",
  title = TRUE)
}
\arguments{
\item{object}{Object.}

\item{genes}{Character vector of genes to include. These must match the
rownames of the object. It is best practice to use the stable gene
identifiers from Ensembl (e.g. "ENSG00000000003") and not the gene symbols.}

\item{gene2symbol}{\code{data.frame} containing gene-to-symbol mappings. Must
contain the columns \code{geneID} and \code{geneName}.}

\item{pointColor}{Default point color for the plot.}

\item{sigPointColor}{Color for points corresponding to significant genes that
have passed alpha level cutoffs.}

\item{labelColor}{Text label color.}

\item{title}{Title of plot.}
}
\value{
\code{ggplot}.
}
\description{
Plot Mean Average
}
\examples{
load(system.file("extdata/bcb_small.rda", package = "bcbioRNASeq"))
load(system.file("extdata/res_small.rda", package = "bcbioRNASeq"))

# DESeqResults ====
plotMA(res_small)
plotMA(
    object = res_small,
    genes = head(rownames(res_small), 4L),
    gene2symbol = gene2symbol(bcb_small)
)
}
\seealso{
Other Differential Expression Functions: \code{\link{alphaSummary}},
  \code{\link{contrastName}}, \code{\link{plotDEGHeatmap}},
  \code{\link{plotDEGPCA}}, \code{\link{plotVolcano}},
  \code{\link{resultsTables}}
}
\author{
Rory Kirchner, Michael Steinbaugh
}
