% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotMappedReads.R
\docType{methods}
\name{plotMappedReads}
\alias{plotMappedReads}
\alias{plotMappedReads,bcbioRNASeq-method}
\alias{plotMappedReads,data.frame-method}
\title{Plot Mapped Reads}
\usage{
plotMappedReads(object, ...)

\S4method{plotMappedReads}{bcbioRNASeq}(object, interestingGroups,
  passLimit = 20, warnLimit = 10, fill = scale_fill_viridis(discrete =
  TRUE), flip = TRUE)

\S4method{plotMappedReads}{data.frame}(object,
  interestingGroups = "sampleName", passLimit = 20, warnLimit = 10,
  fill = scale_fill_viridis(discrete = TRUE), flip = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{interestingGroups}{Category to use to group samples (color and shape).
If unset, this is automatically determined by the metadata set inside the
\link{bcbioRNASeq} object.}

\item{passLimit}{Threshold to plot pass color marker.}

\item{warnLimit}{Threshold to plot warning color marker.}

\item{fill}{Desired ggplot fill scale. Defaults to
\code{\link[viridis:scale_fill_viridis]{viridis::scale_fill_viridis()}}. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.}

\item{flip}{Flip X and Y axes.}
}
\value{
\link{ggplot}.
}
\description{
Plot Mapped Reads
}
\examples{
plotMappedReads(bcb)

\dontrun{
plotMappedReads(
    bcb,
    interestingGroups = "group",
    fill = NULL)
}

# data.frame
\dontrun{
metrics(bcb) \%>\% plotMappedReads()
}
}
\seealso{
Other Quality Control Plots: \code{\link{plot53Bias}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotGenesDetected}}, \code{\link{plotGene}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappingRate}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
\concept{Quality Control Plots}
