% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotMappedReads.R
\docType{methods}
\name{plotMappedReads}
\alias{plotMappedReads}
\alias{plotMappedReads}
\alias{plotMappedReads,bcbioRNADataSet-method}
\alias{plotMappedReads,data.frame-method}
\title{Plot Mapped Reads}
\usage{
plotMappedReads(object, ...)

\S4method{plotMappedReads}{bcbioRNADataSet}(object, ...)

\S4method{plotMappedReads}{data.frame}(object,
  interestingGroup = "sampleName", passLimit = 20L, warnLimit = 10L,
  flip = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{interestingGroup}{\emph{Optional}. Category to use to group samples (color
and shape). If unset, this is automatically determined by the metadata set
inside the \link{bcbioRNADataSet}.}

\item{passLimit}{Threshold to plot pass color marker.}

\item{warnLimit}{Threshold to plot warning color marker.}

\item{flip}{Flip X and Y axes.}
}
\value{
\link{ggplot}.
}
\description{
Plot Mapped Reads
}
\examples{
data(bcb)

# bcbioRNADataSet
plotMappedReads(bcb)

# data.frame
metrics <- metrics(bcb)
plotMappedReads(metrics)
}
\seealso{
Other Quality Control Plots: \code{\link{plot53Bias}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGeneDetectionSaturation}},
  \code{\link{plotGenesDetected}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappingRate}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
