% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMappedReads-methods.R
\docType{methods}
\name{plotMappedReads}
\alias{plotMappedReads}
\alias{plotMappedReads,bcbioRNASeq-method}
\title{Plot mapped reads}
\usage{
plotMappedReads(object, ...)

\S4method{plotMappedReads}{bcbioRNASeq}(object, interestingGroups = NULL,
  limit = 10000000L, perMillion = TRUE, fill = getOption(x =
  "acid.fill.discrete", default = acidplots::scale_fill_synesthesia_d()),
  flip = getOption(x = "acid.flip", default = TRUE),
  title = "Mapped reads")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
The number of mapped reads should correspond to the number of total reads.
}
\note{
Updated 2019-08-07.
}
\examples{
data(bcb)
plotMappedReads(bcb)
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
