% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMappedReads-methods.R
\docType{methods}
\name{plotMappedReads}
\alias{plotMappedReads}
\alias{plotMappedReads,bcbioRNASeq-method}
\title{Plot mapped reads}
\usage{
plotMappedReads(object, ...)

\S4method{plotMappedReads}{bcbioRNASeq}(object, interestingGroups,
  limit = 10000000L, fill = getOption("bcbio.discrete.fill", NULL),
  flip = getOption("bcbio.flip", TRUE), title = "mapped reads")
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}. Groups of interest that define the
samples. If left unset, defaults to \code{sampleName}.}

\item{limit}{\code{scalar numeric}. Threshold to denote on the plot, using a
dashed line.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 fill scale.
Must supply discrete values. When set to \code{NULL}, the default ggplot2 color
palette will be used. If manual color definitions are desired, we recommend
using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.
To set the discrete fill palette globally, use
\code{options(bcbio.discrete.fill = scale_fill_viridis_d())}.}

\item{flip}{\code{boolean}. Flip x and y axes. Recommended for quality control
plots containing many samples.}

\item{title}{\code{string} or \code{NULL}. Plot title.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
The number of mapped reads should correspond to the number of total reads.
}
\examples{
plotMappedReads(bcb_small)
}
\seealso{
\code{\link[=plotMappingRate]{plotMappingRate()}}.

Other Quality Control Functions: \code{\link{plot5Prime3PrimeBias}},
  \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotDispEsts}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotGenesDetected}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappingRate}}, \code{\link{plotMeanSD}},
  \code{\link{plotPCACovariates}}, \code{\link{plotPCA}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
\concept{Quality Control Functions}
