% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMappedReads-methods.R
\docType{methods}
\name{plotMappedReads}
\alias{plotMappedReads}
\alias{plotMappedReads,bcbioRNASeq-method}
\title{Plot mapped reads}
\usage{
\S4method{plotMappedReads}{bcbioRNASeq}(object, interestingGroups = NULL,
  limit = 10000000L, perMillion = TRUE,
  fill = getOption("acid.fill.discrete", default = NULL),
  flip = getOption("acid.flip", default = TRUE),
  title = "mapped reads")
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}
}
\value{
\code{ggplot}.
}
\description{
The number of mapped reads should correspond to the number of total reads.
}
\examples{
data(bcb)
plotMappedReads(bcb)
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
