% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotMappedReads-methods.R
\name{plotMappedReads}
\alias{plotMappedReads}
\alias{plotMappedReads,bcbioRNASeq-method}
\title{Plot mapped reads}
\usage{
plotMappedReads(object, ...)

\S4method{plotMappedReads}{bcbioRNASeq}(
  object,
  interestingGroups = NULL,
  limit = 10000000L,
  perMillion = TRUE,
  labels = list(title = "Mapped reads", subtitle = NULL, sampleAxis = NULL, metricAxis
    = "reads"),
  flip = getOption(x = "acid.flip", default = TRUE)
)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{limit}{\code{numeric(1)}.
Threshold limit.}

\item{perMillion}{\code{logical(1)}.
Display as counts per million.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{flip}{\code{logical(1)}.
Flip x and y axes. Recommended for plots containing many samples.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
The number of mapped reads should correspond to the number of total reads.
}
\note{
Updated 2021-09-10.
}
\examples{
## bcbioRNASeq ====
data(bcb)
plotMappedReads(bcb)
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
