% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMappedReads-methods.R
\docType{methods}
\name{plotMappedReads}
\alias{plotMappedReads}
\alias{plotMappedReads,bcbioRNASeq-method}
\title{Plot mapped reads}
\usage{
plotMappedReads(object, ...)

\S4method{plotMappedReads}{bcbioRNASeq}(object, interestingGroups = NULL,
  limit = 10000000L, perMillion = TRUE, fill = getOption(x =
  "acid.fill.discrete", default = acidplots::scale_fill_synesthesia_d()),
  labels = list(title = "Mapped reads", subtitle = NULL, sampleAxis =
  NULL, metricAxis = "reads"), flip = getOption(x = "acid.flip", default
  = TRUE))
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{limit}{\code{numeric(1)}.
Threshold limit.}

\item{perMillion}{\code{logical(1)}.
Display as counts per million.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.

To set the discrete fill palette globally, use:\preformatted{options(acid.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{flip}{\code{logical(1)}.
Flip x and y axes. Recommended for plots containing many samples.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
The number of mapped reads should correspond to the number of total reads.
}
\note{
Updated 2019-09-16.
}
\examples{
data(bcb)
plotMappedReads(bcb)
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
