% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMappedReads-methods.R
\docType{methods}
\name{plotMappedReads}
\alias{plotMappedReads}
\alias{plotMappedReads,bcbioRNASeq-method}
\title{Plot Mapped Reads}
\usage{
\S4method{plotMappedReads}{bcbioRNASeq}(object, interestingGroups = NULL,
  limit = 10000000L, fill = getOption("basejump.fill.discrete", NULL),
  flip = getOption("basejump.flip", TRUE), title = "mapped reads")
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}. Groups of interest that define the
samples. If left unset, defaults to \code{sampleName}.}

\item{limit}{\code{numeric(1)}. Threshold to denote on the plot, using a dashed
line.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}. Desired ggplot2 fill scale. Must
supply discrete values. When set to \code{NULL}, the default ggplot2 color
palette will be used. If manual color definitions are desired, we recommend
using \code{ggplot2::scale_fill_manual}.

To set the discrete fill palette globally, use:\preformatted{options(basejump.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{flip}{\code{logical(1)}. Flip x and y axes. Recommended for plots containing
many samples.}

\item{title}{\code{character(1)}. Plot title.}
}
\value{
\code{ggplot}.
}
\description{
The number of mapped reads should correspond to the number of total reads.
}
\examples{
data(bcb)
plotMappedReads(bcb)
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
