% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotMappingRate.R
\docType{methods}
\name{plotMappingRate}
\alias{plotMappingRate}
\alias{plotMappingRate}
\alias{plotMappingRate,bcbioRNADataSet-method}
\alias{plotMappingRate,data.frame-method}
\title{Plot Mapping Rate}
\usage{
plotMappingRate(object, ...)

\S4method{plotMappingRate}{bcbioRNADataSet}(object, ...)

\S4method{plotMappingRate}{data.frame}(object,
  interestingGroup = "sampleName", passLimit = 90L, warnLimit = 70L,
  flip = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{interestingGroup}{\emph{Optional}. Category to use to group samples (color
and shape). If unset, this is automatically determined by the metadata set
inside the \link{bcbioRNADataSet}.}

\item{passLimit}{Threshold to plot pass color marker.}

\item{warnLimit}{Threshold to plot warning color marker.}

\item{flip}{Flip X and Y axes.}
}
\value{
\link{ggplot}.
}
\description{
Plot Mapping Rate
}
\examples{
data(bcb)

# bcbioRNADataSet
plotMappingRate(bcb)

# data.frame
metrics(bcb) \%>\% plotMappingRate
}
\seealso{
Other Quality Control Plots: \code{\link{plot53Bias}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGeneDetectionSaturation}},
  \code{\link{plotGenesDetected}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
