% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMappingRate-methods.R
\docType{methods}
\name{plotMappingRate}
\alias{plotMappingRate}
\alias{plotMappingRate,bcbioRNASeq-method}
\title{Plot mapping rate}
\usage{
plotMappingRate(object, ...)

\S4method{plotMappingRate}{bcbioRNASeq}(object, interestingGroups = NULL,
  limit = 0.7, fill = getOption(x = "acid.fill.discrete", default =
  acidplots::scale_fill_synesthesia_d()), flip = getOption(x =
  "acid.flip", default = TRUE), title = "Mapping rate")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
The genomic mapping rate represents the percentage of reads mapping to the
reference genome. Low mapping rates are indicative of sample contamination,
poor sequencing quality or other artifacts.
}
\note{
Updated 2019-08-07.
}
\examples{
data(bcb)
plotMappingRate(bcb)
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
