% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMappingRate-methods.R
\docType{methods}
\name{plotMappingRate}
\alias{plotMappingRate}
\alias{plotMappingRate,bcbioRNASeq-method}
\title{Plot mapping rate}
\usage{
\S4method{plotMappingRate}{bcbioRNASeq}(object, interestingGroups = NULL,
  limit = 0.7, fill = getOption("acid.fill.discrete", default = NULL),
  flip = getOption("acid.flip", default = TRUE),
  title = "Mapping rate")
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}
}
\value{
\code{ggplot}.
}
\description{
The genomic mapping rate represents the percentage of reads mapping to the
reference genome. Low mapping rates are indicative of sample contamination,
poor sequencing quality or other artifacts.
}
\examples{
data(bcb)
plotMappingRate(bcb)
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
