% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMappingRate-methods.R
\docType{methods}
\name{plotMappingRate}
\alias{plotMappingRate}
\alias{plotMappingRate,bcbioRNASeq-method}
\title{Plot mapping rate}
\usage{
\S4method{plotMappingRate}{bcbioRNASeq}(object, interestingGroups = NULL,
  limit = 0.7, fill = getOption("basejump.fill.discrete", default =
  NULL), flip = getOption("basejump.flip", default = TRUE),
  title = "mapping rate")
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{limit}{\code{numeric(1)}.
Threshold to denote on the plot, using a dashed line.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.

To set the discrete fill palette globally, use:\preformatted{options(basejump.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{flip}{\code{logical(1)}.
Flip x and y axes. Recommended for plots containing many samples.}

\item{title}{\code{character(1)}.
Plot title.}
}
\value{
\code{ggplot}.
}
\description{
The genomic mapping rate represents the percentage of reads mapping to the
reference genome. Low mapping rates are indicative of sample contamination,
poor sequencing quality or other artifacts.
}
\examples{
data(bcb)
plotMappingRate(bcb)
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
