% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotMeanSD.R
\docType{methods}
\name{plotMeanSD}
\alias{plotMeanSD}
\alias{plotMeanSD,bcbioRNASeqANY-method}
\alias{plotMeanSD,DESeqDataSet-method}
\title{Plot Row Standard Deviations vs. Row Means}
\usage{
plotMeanSD(object, ...)

\S4method{plotMeanSD}{bcbioRNASeqANY}(object)

\S4method{plotMeanSD}{DESeqDataSet}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
\link{ggplot} grid.
}
\description{
\code{\link[vsn:meanSdPlot]{vsn::meanSdPlot()}} wrapper that plots \code{\link[=log2]{log2()}}, \code{\link[=rlog]{rlog()}}, and
\code{\link[=varianceStabilizingTransformation]{varianceStabilizingTransformation()}} normalized counts.
}
\examples{
data(bcb, dds)

# bcbioRNASeq
plotMeanSD(bcb)

# DESeqDataSet
plotMeanSD(dds)
}
\seealso{
Other Differential Expression Utilities: \code{\link{alphaSummary}},
  \code{\link{plotDispEsts}}
}
\author{
Michael Steinbaugh, Lorena Patano
}
\concept{Differential Expression Utilities}
