% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotMeanSD-methods.R
\docType{methods}
\name{plotMeanSD}
\alias{plotMeanSD}
\alias{plotMeanSD,bcbioRNASeq-method}
\alias{plotMeanSD,DESeqDataSet-method}
\title{Plot Row Standard Deviations vs. Row Means}
\usage{
plotMeanSD(object, ...)

\S4method{plotMeanSD}{bcbioRNASeq}(object,
  legend = getOption("bcbio.legend", FALSE))

\S4method{plotMeanSD}{DESeqDataSet}(object, vst = TRUE, rlog = FALSE,
  legend = getOption("bcbio.legend", FALSE))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{legend}{\code{boolean}. Show plot legend.}

\item{vst}{\code{boolean}. Calculate variance-stabilizing transformation using
\code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}}. Recommended by default
for visualization.}

\item{rlog}{\code{boolean}. Calcualte regularized log transformation using
\code{\link[DESeq2:rlog]{DESeq2::rlog()}}. This calculation is slow for large datasets and now
discouraged by default for visualization.}
}
\value{
\code{ggplot} grid.
}
\description{
\code{\link[vsn:meanSdPlot]{vsn::meanSdPlot()}} wrapper that plots count transformations on a log2 scale.
}
\details{
\itemize{
\item \strong{DESeq2 log2}: log2 normalized counts.
\item \strong{DESeq2 rlog}: regularized log transformation.
\item \strong{DESeq2 vst}: variance stabilizing transformation.
\item \strong{edgeR log2 tmm}: log2 trimmed mean of M-values transformation.
}
}
\examples{
# bcbioRNASeq ====
plotMeanSD(bcb_small)

# DESeqDataSet ====
dds_small <- as(deseq_small, "DESeqDataSet")
plotMeanSD(dds_small)
}
\seealso{
\itemize{
\item \code{\link[vsn:meanSdPlot]{vsn::meanSdPlot()}}.
\item \code{\link[DESeq2:DESeq]{DESeq2::DESeq()}}.
\item \code{\link[DESeq2:rlog]{DESeq2::rlog()}}.
\item \code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}}.
\item \code{\link[=tmm]{tmm()}}.
}

Other Quality Control Functions: \code{\link{plot5Prime3PrimeBias}},
  \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotDispEsts}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}}, \code{\link{plotPCA}},
  \code{\link{plotQC}}, \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh, Lorena Patano
}
\concept{Quality Control Functions}
