% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotMeanSD.R
\docType{methods}
\name{plotMeanSD}
\alias{plotMeanSD}
\alias{plotMeanSD}
\alias{plotMeanSD,bcbioRNASeq-method}
\alias{plotMeanSD,DESeqDataSet-method}
\title{Plot Row Standard Deviations vs. Row Means}
\usage{
plotMeanSD(object, ...)

\S4method{plotMeanSD}{bcbioRNASeq}(object, orientation = "vertical",
  showLegend = FALSE)

\S4method{plotMeanSD}{DESeqDataSet}(object, orientation = "vertical",
  showLegend = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{orientation}{Orientation to use for plot grid, either \code{horizontal} or
\code{vertical}.}

\item{showLegend}{Include the color bar legend. This is typically not that
informative and is disabled by default, to improve the plot appearance.}
}
\value{
\link{ggplot} grid.
}
\description{
\code{\link[vsn:meanSdPlot]{vsn::meanSdPlot()}} wrapper that plots \code{\link[=log2]{log2()}}, \code{\link[=rlog]{rlog()}}, and
\code{\link[=varianceStabilizingTransformation]{varianceStabilizingTransformation()}} normalized counts.
}
\examples{
load(system.file(
    file.path("inst", "extdata", "bcb.rda"),
    package = "bcbioRNASeq"))

# bcbioRNASeq
plotMeanSD(bcb, orientation = "horizontal")
plotMeanSD(bcb, orientation = "vertical")

# DESeqDataSet
dds <- bcbio(bcb, "DESeqDataSet")
plotMeanSD(dds, orientation = "horizontal")
}
\seealso{
Other Differential Expression Utilities: \code{\link{alphaSummary}},
  \code{\link{plotDispEsts}}
}
\author{
Michael Steinbaugh, Lorena Patano
}
