% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotMeanSD-methods.R
\docType{methods}
\name{plotMeanSD}
\alias{plotMeanSD}
\alias{plotMeanSD,bcbioRNASeq-method}
\alias{plotMeanSD,DESeqDataSet-method}
\alias{plotMeanSD,DESeqAnalysis-method}
\title{Plot Row Standard Deviations vs. Row Means}
\usage{
plotMeanSD(object, ...)

\S4method{plotMeanSD}{bcbioRNASeq}(object,
  legend = getOption("basejump.legend", FALSE))

\S4method{plotMeanSD}{DESeqDataSet}(object, vst = TRUE, rlog = FALSE,
  legend = getOption("basejump.legend", FALSE))

\S4method{plotMeanSD}{DESeqAnalysis}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{legend}{\code{boolean}. Show plot legend.}

\item{vst}{\code{boolean}. Calculate variance-stabilizing transformation using
\code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}}. Recommended by default
for visualization.}

\item{rlog}{\code{boolean}. Calcualte regularized log transformation using
\code{\link[DESeq2:rlog]{DESeq2::rlog()}}. This calculation is slow for large datasets and now
discouraged by default for visualization.}
}
\value{
\code{ggplot} grid.
}
\description{
\code{\link[vsn:meanSdPlot]{vsn::meanSdPlot()}} wrapper that plots count transformations on a log2 scale.
}
\details{
\itemize{
\item DESeq2 log2: log2 size factor-adjusted normalized counts.
\item DESeq2 rlog: \strong{r}egularized \strong{log} transformation.
\item DESeq2 VST: \strong{v}ariance-\strong{s}tabilizing \strong{t}ransformation.
\item edgeR log2 TMM: log2 \strong{t}rimmed \strong{m}ean of \strong{M}-values transformation.
}
}
\examples{
data(bcb_small, deseq_small)

## bcbioRNASeq ====
plotMeanSD(bcb_small)

## DESeqAnalysis ====
plotMeanSD(deseq_small)
}
\seealso{
\itemize{
\item \link[vsn:meanSdPlot]{vsn::meanSdPlot}.
\item \link[DESeq2:DESeq]{DESeq2::DESeq}.
\item \link[DESeq2:rlog]{DESeq2::rlog}.
\item \link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation}.
\item \link[edgeR:calcNormFactors]{edgeR::calcNormFactors}.
}
}
\author{
Michael Steinbaugh, Lorena Patano
}
