% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotMeanSd-methods.R
\name{plotMeanSd}
\alias{plotMeanSd}
\alias{plotMeanSd,bcbioRNASeq-method}
\title{Plot row standard deviations vs. row means}
\usage{
plotMeanSd(object, ...)

\S4method{plotMeanSd}{bcbioRNASeq}(
  object,
  fill = ggplot2::scale_fill_gradient(low = AcidPlots::lightPalette[["gray"]], high =
    AcidPlots::lightPalette[["purple"]]),
  lineColor = AcidPlots::lightPalette[["orange"]],
  legend = getOption(x = "acid.legend", default = TRUE)
)
}
\arguments{
\item{object}{Object.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{ggplot2::scale_fill_manual()}.

To set the discrete fill palette globally, use:

\if{html}{\out{<div class="sourceCode">}}\preformatted{options("acid.fill.discrete" = ggplot2::scale_fill_viridis_d())
}\if{html}{\out{</div>}}}

\item{lineColor}{\code{character(1)}.
Line color.}

\item{legend}{\code{logical(1)}.
Include plot legend.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot row standard deviations vs. row means
}
\details{
\code{vsn::meanSdPlot()} wrapper that plots count transformations on a log2 scale.
\itemize{
\item DESeq2 log2: log2 library size factor-adjusted normalized counts.
\item DESeq2 rlog: \strong{r}egularized \strong{log} transformation.
\item DESeq2 VST: \strong{v}ariance-\strong{s}tabilizing \strong{t}ransformation.
\item edgeR log2 TMM: log2 \strong{t}rimmed \strong{m}ean of \strong{M}-values transformation.
}
}
\note{
Requires the vsn package to be installed.

Updated 2022-10-24.
}
\examples{
data(bcb)

## bcbioRNASeq ====
if (requireNamespace("vsn", quietly = TRUE)) {
    plotMeanSd(bcb)
}
}
\seealso{
\itemize{
\item \code{vsn::meanSdPlot()}.
\item \code{DESeq2::DESeq()}.
\item \code{DESeq2::rlog()}.
\item \code{DESeq2::varianceStabilizingTransformation()}.
\item \code{edgeR::calcNormFactors()}.
}
}
\author{
Michael Steinbaugh, Lorena Patano
}
