% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMeanSD-methods.R
\docType{methods}
\name{plotMeanSD}
\alias{plotMeanSD}
\alias{plotMeanSD,bcbioRNASeq-method}
\title{Plot Row Standard Deviations vs. Row Means}
\usage{
\S4method{plotMeanSD}{bcbioRNASeq}(object,
  legend = getOption("basejump.legend", TRUE))
}
\arguments{
\item{object}{Object.}

\item{legend}{\code{boolean}. Show plot legend.}
}
\value{
\code{ggplot}.
}
\description{
Plot Row Standard Deviations vs. Row Means
}
\details{
\code{\link[vsn:meanSdPlot]{vsn::meanSdPlot()}} wrapper that plots count transformations on a log2 scale.
\itemize{
\item DESeq2 log2: log2 size factor-adjusted normalized counts.
\item DESeq2 rlog: \strong{r}egularized \strong{log} transformation.
\item DESeq2 VST: \strong{v}ariance-\strong{s}tabilizing \strong{t}ransformation.
\item edgeR log2 TMM: log2 \strong{t}rimmed \strong{m}ean of \strong{M}-values transformation.
}
}
\examples{
data(bcb)
plotMeanSD(bcb)
}
\seealso{
\itemize{
\item \code{\link[vsn:meanSdPlot]{vsn::meanSdPlot()}}.
\item \code{\link[DESeq2:DESeq]{DESeq2::DESeq()}}.
\item \code{\link[DESeq2:rlog]{DESeq2::rlog()}}.
\item \code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}}.
\item \code{\link[edgeR:calcNormFactors]{edgeR::calcNormFactors()}}.
}
}
\author{
Michael Steinbaugh, Lorena Patano
}
