% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMeanSD-methods.R
\docType{methods}
\name{plotMeanSD}
\alias{plotMeanSD}
\alias{plotMeanSD,bcbioRNASeq-method}
\alias{plotMeanSD,DESeqDataSet-method}
\title{Plot row standard deviations vs. row means}
\usage{
plotMeanSD(object, ...)

\S4method{plotMeanSD}{bcbioRNASeq}(object,
  legend = getOption("bcbio.legend", FALSE))

\S4method{plotMeanSD}{DESeqDataSet}(object,
  legend = getOption("bcbio.legend", FALSE))
}
\arguments{
\item{object}{Object.}

\item{legend}{\code{boolean}. Show plot legend.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot} grid.
}
\description{
Plot row standard deviations vs. row means
}
\details{
\code{\link[vsn:meanSdPlot]{vsn::meanSdPlot()}} wrapper that plots count transformations on a log2 scale.
\itemize{
\item \strong{DESeq2 log2}: log2 normalized counts.
\item \strong{DESeq2 rlog}: regularized log transformation.
\item \strong{DESeq2 vst}: variance stabilizing transformation.
\item \strong{edgeR log2 tmm}: log2 trimmed mean of M-values transformation.
}
}
\examples{
# bcbioRNASeq ====
plotMeanSD(bcb_small)

# DESeqDataSet ====
plotMeanSD(dds_small)
}
\seealso{
\itemize{
\item \code{\link[vsn:meanSdPlot]{vsn::meanSdPlot()}}.
\item \code{\link[DESeq2:DESeq]{DESeq2::DESeq()}}.
\item \code{\link[DESeq2:rlog]{DESeq2::rlog()}}.
\item \code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}}.
\item \code{\link[=tmm]{tmm()}}.
}

Other Quality Control Functions: \code{\link{plot5Prime3PrimeBias}},
  \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotDispEsts}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotGenesDetected}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}},
  \code{\link{plotPCACovariates}}, \code{\link{plotPCA}},
  \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh, Lorena Patano
}
\concept{Quality Control Functions}
