% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMeanSD-methods.R
\name{plotMeanSD}
\alias{plotMeanSD}
\alias{plotMeanSD,bcbioRNASeq-method}
\title{Plot row standard deviations vs. row means}
\usage{
plotMeanSD(object, ...)

\S4method{plotMeanSD}{bcbioRNASeq}(
  object,
  fill = ggplot2::scale_fill_gradient(low = acidplots::lightPalette[["gray"]], high =
    acidplots::lightPalette[["purple"]]),
  lineColor = acidplots::lightPalette[["orange"]],
  legend = getOption(x = "acid.legend", default = TRUE)
)
}
\arguments{
\item{object}{Object.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{ggplot2::scale_fill_manual()}.

To set the discrete fill palette globally, use:\preformatted{options(acid.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{lineColor}{\code{character(1)}.
Line color.}

\item{legend}{\code{logical(1)}.
Include plot legend.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot row standard deviations vs. row means
}
\details{
\code{\link[vsn:meanSdPlot]{vsn::meanSdPlot()}} wrapper that plots count transformations on a log2 scale.
\itemize{
\item DESeq2 log2: log2 library size factor-adjusted normalized counts.
\item DESeq2 rlog: \strong{r}egularized \strong{log} transformation.
\item DESeq2 VST: \strong{v}ariance-\strong{s}tabilizing \strong{t}ransformation.
\item edgeR log2 TMM: log2 \strong{t}rimmed \strong{m}ean of \strong{M}-values transformation.
}
}
\note{
Requires the vsn package to be installed.

Updated 2019-10-30.
}
\examples{
data(bcb)
if (requireNamespace("vsn", quietly = TRUE)) {
    plotMeanSD(bcb)
}
}
\seealso{
\itemize{
\item \code{\link[vsn:meanSdPlot]{vsn::meanSdPlot()}}.
\item \code{\link[DESeq2:DESeq]{DESeq2::DESeq()}}.
\item \code{\link[DESeq2:rlog]{DESeq2::rlog()}}.
\item \code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}}.
\item \code{\link[edgeR:calcNormFactors]{edgeR::calcNormFactors()}}.
}
}
\author{
Michael Steinbaugh, Lorena Patano
}
