% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotPCA.R
\docType{methods}
\name{plotPCA}
\alias{plotPCA}
\alias{plotPCA,bcbioRNASeq-method}
\title{Sample PCA Plot for Transformed Data}
\usage{
\S4method{plotPCA}{bcbioRNASeq}(object, transform = "rlog", interestingGroups,
  genes, censorSamples, color = scale_color_viridis(discrete = TRUE),
  label = FALSE, returnData = FALSE)
}
\arguments{
\item{object}{Object.}

\item{transform}{String specifying \link{rlog} (\strong{recommended}) or \link{vst}
\link{DESeqTransform} slotted inside the \link{bcbioRNASeq} object.}

\item{interestingGroups}{\emph{Optional}. Interesting groups to use for point
appearance. If missing, color defaults to all \code{interestingGroups}
parameters set in the \link{bcbioRNASeq} object.}

\item{genes}{\emph{Optional}. Character vector of gene identifiers to use.}

\item{censorSamples}{\emph{Optional}. Censors to exclude from PCA plot.}

\item{color}{Desired ggplot color scale. Defaults to
\code{\link[viridis:scale_color_viridis]{viridis::scale_color_viridis()}}. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using
\code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}

\item{label}{\emph{Optional}. Superimpose sample text labels on the plot.}

\item{returnData}{Return PCA loadings data instead of plotting.}
}
\value{
\link{ggplot}.
}
\description{
Wrapper for \code{\link[DESeq2:plotPCA]{DESeq2::plotPCA()}} that improves principal component analysis
(PCA) sample coloring and labeling.
}
\examples{
bcb <- examples[["bcb"]]
plotPCA(bcb, label = TRUE)
plotPCA(bcb, label = FALSE)
plotPCA(bcb, interestingGroups = "sampleName", color = NULL)
}
\seealso{
\code{\link[DESeq2:plotPCA]{DESeq2::plotPCA()}}.
}
\author{
Michael Steinbaugh
}
