% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA-methods.R
\docType{methods}
\name{plotPCA}
\alias{plotPCA}
\alias{plotPCA,bcbioRNASeq-method}
\title{PCA-plot: Principal Component Analysis plot}
\usage{
\S4method{plotPCA}{bcbioRNASeq}(object, normalized = c("vst", "rlog",
  "tmm", "rle", "tpm", "fpkm"), interestingGroups = NULL, ntop = 500L,
  color = getOption("basejump.discrete.color", NULL),
  label = getOption("basejump.label", FALSE), title = "PCA",
  subtitle = NULL, return = c("ggplot", "DataFrame"))
}
\arguments{
\item{object}{
    
    A data object, typically containing gene expression information.

  }

\item{normalized}{\code{string}. Which normalization method to apply:
- "\code{tpm}": Transcripts per million (tximport).
- "\code{tmm}": edgeR trimmed mean of M-values. Calculated on the fly.
- "\code{rlog}": DESeq2 \strong{log2} regularized log transformation.
- "\code{vst}": DESeq2 \strong{log2} variance stabilizing transformation.}

\item{interestingGroups}{\code{character}. Groups of interest that define the
samples. If left unset, defaults to \code{sampleName}.}

\item{ntop}{\code{scalar integer} or \code{Inf}. Number of most variable genes to plot.
Use \code{Inf} to include all genes (\emph{not recommended}).}

\item{color}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.
To set the discrete color palette globally, use
\code{options(bcbio.discrete.color = scale_color_viridis_d())}.}

\item{label}{\code{boolean}. Superimpose sample text labels on the plot.}

\item{title}{\code{string} or \code{NULL}. Plot title.}

\item{subtitle}{\code{string} or \code{NULL}. Plot subtitle.}

\item{return}{\code{string}. Object class to return. Uses \code{\link[=match.arg]{match.arg()}} internally
and picks the first item in the vector by default.}
}
\value{
\code{ggplot} or \code{DataFrame}.
}
\description{
A generic function which produces a PCA-plot.
}
\details{
Wrapper for \code{\link[DESeq2:plotPCA]{DESeq2::plotPCA()}} that improves principal component analysis
(PCA) sample coloring and labeling. Always plots using normalized counts.

PCA (Jolliffe, et al., 2002) is a multivariate technique that allows us to
summarize the systematic patterns of variations in the data. PCA takes the
expression levels for genes and transforms it in principal component space,
reducing each sample into one point. Thereby, we can separate samples by
expression variation, and identify potential sample outliers. The PCA plot is
a way to look at how samples are clustering.
}
\examples{
data(bcb_small)
plotPCA(bcb_small, label = FALSE)
plotPCA(bcb_small, label = TRUE)
}
\references{
Jolliffe, et al., 2002.
}
\seealso{
\itemize{
\item \code{\link[DESeq2:plotPCA]{DESeq2::plotPCA()}}.
\item \code{getMethod("plotPCA", "DESeqTransform")}
}
}
\author{
Michael Steinbaugh
}
