% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotPCA.R
\docType{methods}
\name{plotPCA}
\alias{plotPCA}
\alias{plotPCA,bcbioRNASeqANY-method}
\title{Sample PCA Plot for Transformed Data}
\usage{
\S4method{plotPCA}{bcbioRNASeqANY}(object, transform = "rlog", genes = NULL,
  censorSamples = NULL, interestingGroups = NULL, shape = FALSE,
  label = TRUE)
}
\arguments{
\item{object}{Object.}

\item{transform}{String specifying \link{rlog} (\strong{recommended}) or \link{vst}
\link{DESeqTransform} slotted inside the \link{bcbioRNASeq} object.}

\item{genes}{\emph{Optional}. Character vector of gene identifiers to use.}

\item{censorSamples}{\emph{Optional}. Censors to exclude from PCA plot.}

\item{interestingGroups}{\emph{Optional}. Interesting groups to use for point
appearance. If \code{NULL}, color defaults to all \code{interestingGroups}
parameters set in the \link{bcbioRNASeq} object.}

\item{shape}{\emph{Optional}. Make points easier to inspect with differing shapes.}

\item{label}{\emph{Optional}. Superimpose sample text labels on the plot.}
}
\value{
\link{ggplot}.
}
\description{
Wrapper for \code{\link[DESeq2:plotPCA]{DESeq2::plotPCA()}} that improves principal component analysis
(PCA) sample coloring and labeling.
}
\examples{
data(bcb)
plotPCA(bcb)
plotPCA(bcb, label = FALSE)
}
\seealso{
\code{\link[DESeq2:plotPCA]{DESeq2::plotPCA()}}.
}
\author{
Michael Steinbaugh
}
