% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plotPCA.R
\docType{methods}
\name{plotPCA}
\alias{plotPCA}
\alias{plotPCA,bcbioRNASeq-method}
\title{Sample PCA Plot for Transformed Data}
\usage{
\S4method{plotPCA}{bcbioRNASeq}(object, transform = "rlog", interestingGroups,
  genes, censorSamples, label = FALSE, shape = FALSE, returnData = FALSE)
}
\arguments{
\item{object}{Object.}

\item{transform}{String specifying \link{rlog} (\strong{recommended}) or \link{vst}
\link{DESeqTransform} slotted inside the \link{bcbioRNASeq} object.}

\item{interestingGroups}{\emph{Optional}. Interesting groups to use for point
appearance. If missing, color defaults to all \code{interestingGroups}
parameters set in the \link{bcbioRNASeq} object.}

\item{genes}{\emph{Optional}. Character vector of gene identifiers to use.}

\item{censorSamples}{\emph{Optional}. Censors to exclude from PCA plot.}

\item{label}{\emph{Optional}. Superimpose sample text labels on the plot.}

\item{shape}{\emph{Optional}. Make points easier to inspect with differing shapes.
Generally this isn't recommended for PCA and works poorly for more
than 6 discrete factors, as defined by the \code{interestingGroups} argument.
We recommend typically leaving this to \code{FALSE}.}

\item{returnData}{Return PCA loadings data instead of plotting.}
}
\value{
\link{ggplot}.
}
\description{
Wrapper for \code{\link[DESeq2:plotPCA]{DESeq2::plotPCA()}} that improves principal component analysis
(PCA) sample coloring and labeling.
}
\examples{
plotPCA(bcb, label = TRUE)
plotPCA(bcb, label = FALSE)

# Manually set interesting groups
# Note these columns aren't present in our example data
\dontrun{
plotPCA(bcb, interestingGroups = c("genotype", "treatment"))
}
}
\seealso{
\code{\link[DESeq2:plotPCA]{DESeq2::plotPCA()}}.
}
\author{
Michael Steinbaugh
}
