% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA-methods.R
\docType{methods}
\name{plotPCA}
\alias{plotPCA}
\alias{plotPCA,bcbioRNASeq-method}
\title{Plot PCA}
\usage{
\S4method{plotPCA}{bcbioRNASeq}(object, normalized = c("vst", "rlog",
  "tmm", "rle", "tpm", "fpkm"), interestingGroups = NULL, ntop = 500L,
  label = getOption("basejump.label", FALSE),
  pointColor = getOption("basejump.color.discrete", NULL),
  pointSize = getOption("basejump.point.size", 4L), title = "PCA",
  subtitle = NULL, return = c("ggplot", "DataFrame"), ...)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character} or \code{NULL}. Groups of interest that
define the samples. If left unset, defaults to \code{sampleName}.}

\item{ntop}{\code{scalar integer} or \code{Inf}. Number of most variable genes to plot.
Use \code{Inf} to include all genes (\emph{not recommended}).}

\item{label}{\code{boolean}. Superimpose sample text labels on the plot.}

\item{pointColor}{\code{string}. Default point color for the plot.}

\item{pointSize}{\code{scalar numeric}. Cell point size.}

\item{title}{\code{string} or \code{NULL}. Plot title.}

\item{subtitle}{\code{string} or \code{NULL}. Plot subtitle.}

\item{return}{\code{string}. Return type. Uses \code{base::match.arg()}
internally and defaults to the first argument in the \code{character} vector.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot} or \code{DataFrame}.
}
\description{
Principal component analysis (PCA) plot.
}
\section{Principal component analysis}{



PCA (Jolliffe, et al., 2002) is a multivariate technique that allows us to
summarize the systematic patterns of variations in the data. PCA takes the
expression levels for genes and transforms it in principal component space,
reducing each sample into one point. Thereby, we can separate samples by
expression variation, and identify potential sample outliers. The PCA plot is
a way to look at how samples are clustering.

}

\examples{
data(bcb)
plotPCA(bcb, label = FALSE)
plotPCA(bcb, label = TRUE)
}
\references{
Jolliffe, et al., 2002.
}
\seealso{
\code{DESeq2::plotPCA()}.
We're using a modified version of the \code{DESeqTransform} method here.\preformatted{methodFunction(
    f = "plotPCA",
    signature = "DESeqTransform",
    package = "DESeq2"
)
}
}
\author{
Michael Steinbaugh
}
