% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotPCACovariates.R
\docType{methods}
\name{plotPCACovariates}
\alias{plotPCACovariates}
\alias{plotPCACovariates}
\alias{plotPCACovariates,bcbioRNASeq-method}
\title{Find Correlation Between Principal Components (PCs) and Covariates}
\usage{
plotPCACovariates(object, ...)

\S4method{plotPCACovariates}{bcbioRNASeq}(object, metrics = TRUE,
  normalized = "rlog", ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments, passed to \code{\link[DEGreport:degCovariates]{DEGreport::degCovariates()}}.}

\item{metrics}{Include sample summary metrics as covariates. Defaults to
include all metrics columns (\code{TRUE}), but desired columns can be specified
here as a character vector.}

\item{normalized}{Select raw counts (\code{FALSE}), DESeq2 normalized counts
(\code{TRUE}), or additional normalization methods:
\itemize{
\item \code{tpm}: Transcripts per million.
\item \code{tmm}: Trimmed mean of M-values (edgeR).
\item \code{rlog}: Regularized log transformation (\code{\link[DESeq2:rlog]{DESeq2::rlog()}}).
\item \code{vst}: Variance stabilizing transformation
(\code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}}).
}}
}
\value{
\link{ggplot}.
}
\description{
\code{\link[DEGreport:degCovariates]{DEGreport::degCovariates()}} wrapper supporting a \link{bcbioRNASeq} object.
}
\examples{
load(system.file("extdata/bcb.rda", package = "bcbioRNASeq"))

# bcbioRNASeq
plotPCACovariates(bcb, metrics = c("exonicRate", "intronicRate"))
}
\seealso{
\itemize{
\item \code{\link[DEGreport:degCovariates]{DEGreport::degCovariates()}}.
\item \code{\link[DESeq2:rlog]{DESeq2::rlog()}}.
\item \code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}}.
}
}
\author{
Lorena Pantano, Michael Steinbaugh
}
