% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotPCACovariates-methods.R
\docType{methods}
\name{plotPCACovariates}
\alias{plotPCACovariates}
\alias{plotPCACovariates,bcbioRNASeq-method}
\title{Find Correlation Between Principal Components (PCs) and Covariates}
\usage{
plotPCACovariates(object, ...)

\S4method{plotPCACovariates}{bcbioRNASeq}(object, metrics = TRUE,
  normalized = c("vst", "rlog", "tmm", "tpm", "rle"), ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments, passed to \code{\link[DEGreport:degCovariates]{DEGreport::degCovariates()}}.}

\item{metrics}{\code{boolean}. Include sample summary metrics as covariates.
Defaults to include all metrics columns (\code{TRUE}), but desired columns can
be specified here as a character vector.}

\item{normalized}{\code{string}. Which normalization method to apply:
- "\code{tpm}": Transcripts per million (tximport).
- "\code{tmm}": edgeR trimmed mean of M-values. Calculated on the fly.
- "\code{rlog}": DESeq2 \strong{log2} regularized log transformation.
- "\code{vst}": DESeq2 \strong{log2} variance stabilizing transformation.}
}
\value{
\code{ggplot}.
}
\description{
\code{\link[DEGreport:degCovariates]{DEGreport::degCovariates()}} wrapper supporting a \link{bcbioRNASeq} object.
}
\examples{
plotPCACovariates(
    object = bcb_small,
    metrics = c("exonicRate", "intronicRate")
)
}
\seealso{
\itemize{
\item \code{\link[DEGreport:degCovariates]{DEGreport::degCovariates()}}.
\item \code{\link[DESeq2:rlog]{DESeq2::rlog()}}.
\item \code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}}.
}

Other Quality Control Functions: \code{\link{plot5Prime3PrimeBias}},
  \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotDispEsts}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}}, \code{\link{plotMeanSD}},
  \code{\link{plotPCA}}, \code{\link{plotRRNAMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Lorena Pantano, Michael Steinbaugh
}
\concept{Quality Control Functions}
