% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCACovariates-methods.R
\docType{methods}
\name{plotPCACovariates}
\alias{plotPCACovariates}
\alias{plotPCACovariates,bcbioRNASeq-method}
\title{Find correlation between principal components (PCs) and covariates}
\usage{
plotPCACovariates(object, ...)

\S4method{plotPCACovariates}{bcbioRNASeq}(object, metrics = TRUE,
  normalized = c("vst", "rlog", "tmm", "rle", "tpm", "fpkm"), ...)
}
\arguments{
\item{object}{Object.}

\item{metrics}{\code{boolean}. Include sample summary metrics as covariates.
Defaults to include all metrics columns (\code{TRUE}), but desired columns can
be specified here as a character vector.}

\item{normalized}{\code{character(1)}.
Type of normalized counts to visualize:
\itemize{
\item \code{tpm} Transcripts per million.
\item \code{tmm}: edgeR trimmed mean of M-values.
\item \code{rlog}: DESeq2 log2 regularized log transformation.
\item \code{vst}: DESeq2 log2 variance stabilizing transformation.
}}

\item{...}{Additional arguments, passed to \code{\link[DEGreport:degCovariates]{DEGreport::degCovariates()}}.}
}
\value{
\code{ggplot}.
}
\description{
Find correlation between principal components (PCs) and covariates
}
\examples{
data(bcb)
plotPCACovariates(bcb)
}
\seealso{
\itemize{
\item \code{\link[DEGreport:degCovariates]{DEGreport::degCovariates()}}.
\item \code{\link[DESeq2:rlog]{DESeq2::rlog()}}.
\item \code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}}.
}
}
\author{
Lorena Pantano, Michael Steinbaugh, Rory Kirchner
}
