% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCACovariates-methods.R
\docType{methods}
\name{plotPCACovariates}
\alias{plotPCACovariates}
\alias{plotPCACovariates,bcbioRNASeq-method}
\title{Find correlation between principal components (PCs) and covariates}
\usage{
plotPCACovariates(object, ...)

\S4method{plotPCACovariates}{bcbioRNASeq}(object, metrics = TRUE,
  normalized = c("tpm", "sf", "fpkm", "vst", "rlog", "tmm", "rle"), ...)
}
\arguments{
\item{object}{Object.}

\item{metrics}{\code{boolean}. Include sample summary metrics as covariates.
Defaults to include all metrics columns (\code{TRUE}), but desired columns can
be specified here as a character vector.}

\item{normalized}{\code{character(1)} or \code{logical(1)}.
Normalization method to apply:
\itemize{
\item \code{FALSE}: Raw counts.
When using a \href{https://bioconductor.org/packages/tximport/}{tximport}-compatible caller, these are length scaled
by default (see \code{countsFromAbundance} argument).
When using a \href{http://bioinf.wehi.edu.au/featureCounts/}{featureCounts}-compatible caller, these are \code{integer}.
}

\href{https://bioconductor.org/packages/tximport/}{tximport} caller-specific normalizations:
\itemize{
\item \code{"tpm"}: \strong{T}ranscripts \strong{p}er \strong{m}illion.
}

Additional gene-level-specific normalizations:
\itemize{
\item \code{TRUE} / \code{"sf"}: \strong{S}ize \strong{f}actor (i.e. library size) normalized
counts.\cr
See \code{\link[DESeq2:sizeFactors]{DESeq2::sizeFactors()}} for details.
\item \code{"fpkm"}: \strong{F}ragments \strong{p}er \strong{k}ilobase per \strong{m}illion mapped
fragments.\cr
Requires \code{fpkm = TRUE} in \code{\link[=bcbioRNASeq]{bcbioRNASeq()}} call and gene annotations in
\code{\link[SummarizedExperiment:rowRanges]{rowRanges()}} with defined
\code{\link[BiocGenerics:width]{width}}.\cr
See \code{\link[DESeq2:fpkm]{DESeq2::fpkm()}} for details.
\item \code{"vst"}: \strong{V}ariance-\strong{s}tabilizing \strong{t}ransformation (log2).\cr
Requires \code{vst = TRUE} to be set during \code{\link[=bcbioRNASeq]{bcbioRNASeq()}} call.\cr
See \code{DESeq2::varianceStabilizingTransformation} for more information.
\item \code{"rlog"}: \strong{R}egularized \strong{log} transformation (log2).\cr
Requires \code{rlog = TRUE} to be set during \code{\link[=bcbioRNASeq]{bcbioRNASeq()}} call.\cr
See \code{\link[DESeq2:rlog]{DESeq2::rlog()}} for details.
\item \code{"tmm"}: \strong{T}rimmed \strong{m}ean of \strong{M}-values.\cr
Calculated on the fly.\cr
See \code{\link[edgeR:calcNormFactors]{edgeR::calcNormFactors()}} for details.
\item \code{"rle"}: \strong{R}elative \strong{l}og \strong{e}xpression transformation.\cr
Calculated on the fly.\cr
See \code{\link[=relativeLogExpression]{relativeLogExpression()}} for details.

Note that \code{logical(1)} support only applies to \code{counts()}. Other
functions in the package require \code{character(1)} and use
\code{\link[base:match.arg]{match.arg()}} internally.
}}

\item{...}{Additional arguments, passed to \code{\link[DEGreport:degCovariates]{DEGreport::degCovariates()}}.}
}
\value{
\code{ggplot}.
}
\description{
Find correlation between principal components (PCs) and covariates
}
\note{
Requires the DEGreport package to be installed.

Updated 2019-10-30.
}
\examples{
data(bcb)
if (requireNamespace("DEGreport", quietly = TRUE)) {
    plotPCACovariates(bcb)
}
}
\seealso{
\itemize{
\item \code{\link[DEGreport:degCovariates]{DEGreport::degCovariates()}}.
\item \code{\link[DESeq2:rlog]{DESeq2::rlog()}}.
\item \code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}}.
}
}
\author{
Lorena Pantano, Michael Steinbaugh, Rory Kirchner
}
