% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/plotPseudoVsAlignedCounts-methods.R
\name{plotPseudoVsAlignedCounts}
\alias{plotPseudoVsAlignedCounts}
\alias{plotPseudoVsAlignedCounts,bcbioRNASeq-method}
\title{Plot pseudo vs. aligned counts}
\usage{
plotPseudoVsAlignedCounts(object, ...)

\S4method{plotPseudoVsAlignedCounts}{bcbioRNASeq}(
  object,
  genes = NULL,
  title = "Pseudoaligned vs. aligned counts",
  ...
)
}
\arguments{
\item{object}{Object.}

\item{genes}{\code{character}.
Gene identifiers.}

\item{title}{\code{character(1)}.
Title.}

\item{...}{Passthrough to \code{AcidPlots::plotCountsCorrelationHeatmap()} when
\code{genes = NULL} or \code{AcidPlots::plotCountsCorrelation()} when \code{genes} are
defined.}
}
\value{
\code{ggplot}.
}
\description{
Plot pseudo vs. aligned counts
}
\note{
Updated 2022-03-07.

Currently supported for salmon or kallisto. The function will
intentionally error for datasets containing aligned counts in the primary
\code{counts} assay.
}
\examples{
data(bcb)

## bcbioRNASeq ====
## Correlation heatmap.
plotPseudoVsAlignedCounts(bcb)

## Individual genes.
## Checking the most expressed aligned genes here.
aligned <- assay(bcb, i = "aligned")
genes <- names(tail(sort(rowSums(aligned)), n = 2L))
plotPseudoVsAlignedCounts(bcb, genes = genes)
}
\author{
Michael Steinbaugh
}
