% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/plotPseudoVsAlignedCounts-methods.R
\name{plotPseudoVsAlignedCounts}
\alias{plotPseudoVsAlignedCounts}
\alias{plotPseudoVsAlignedCounts,bcbioRNASeq-method}
\title{Compare pseudoaligned counts to aligned counts.}
\usage{
plotPseudoVsAlignedCounts(object, ...)

\S4method{plotPseudoVsAlignedCounts}{bcbioRNASeq}(
  object,
  genes = NULL,
  title = "Pseudoaligned vs. aligned counts",
  ...
)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough to \code{\link[acidplots:plotCountsCorrelationHeatmap]{acidplots::plotCountsCorrelationHeatmap()}} when
\code{genes = NULL} or \code{\link[acidplots:plotCountsCorrelation]{acidplots::plotCountsCorrelation()}} when \code{genes} are
defined.}

\item{genes}{\code{character}.
Gene identifiers.}

\item{title}{\code{character(1)}.
Title.}
}
\description{
Compare pseudoaligned counts to aligned counts.
}
\note{
Currently supported for salmon or kallisto. The function will
intentionally error for datasets containing aligned counts in the primary
\code{counts} assay.

Updated 2019-09-16.
}
\examples{
data(bcb)

## Correlation heatmap.
plotPseudoVsAlignedCounts(bcb)

## Individual genes.
## Checking the most expressed aligned genes here.
aligned <- assay(bcb, i = "aligned")
genes <- rowSums(aligned) \%>\% sort \%>\% tail(n = 2L) \%>\% names()
plotPseudoVsAlignedCounts(bcb, genes = genes)
}
