% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQC-methods.R
\docType{methods}
\name{plotQC}
\alias{plotQC}
\alias{plotQC,bcbioRNASeq-method}
\title{Plot Quality Control}
\usage{
\S4method{plotQC}{bcbioRNASeq}(object)
}
\arguments{
\item{object}{Object.}
}
\description{
Plot Quality Control
}
\examples{
data(bcb_small)
plotQC(bcb_small)
}
\author{
Michael Steinbaugh
}
