% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotQc-methods.R
\name{plotQc}
\alias{plotQc}
\alias{plotQc,bcbioRNASeq-method}
\title{Quality control}
\usage{
plotQc(object, ...)

\S4method{plotQc}{bcbioRNASeq}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Quality control
}
\note{
Updated 2022-03-07.
}
\examples{
data(bcb)

## bcbioRNASeq ====
plotQc(bcb)
}
\author{
Michael Steinbaugh
}
