% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQC-methods.R
\docType{methods}
\name{plotQC}
\alias{plotQC}
\alias{plotQC,bcbioRNASeq-method}
\title{Quality control}
\usage{
plotQC(object, ...)

\S4method{plotQC}{bcbioRNASeq}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Quality control
}
\examples{
data(bcb)
plotQC(bcb)
}
\author{
Michael Steinbaugh
}
