% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotRRNAMappingRate.R
\docType{methods}
\name{plotRRNAMappingRate}
\alias{plotRRNAMappingRate}
\alias{plotRRNAMappingRate}
\alias{plotRRNAMappingRate,bcbioRNASeq-method}
\title{Plot Ribosomal RNA (rRNA) Mapping Rate}
\usage{
plotRRNAMappingRate(object, ...)

\S4method{plotRRNAMappingRate}{bcbioRNASeq}(object, interestingGroups,
  warnLimit = 10L, fill = scale_fill_viridis(discrete = TRUE),
  flip = TRUE, title = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{interestingGroups}{Category to use to group samples. In the plotting
functions, this will define color and shape, where applicable. If unset,
this is automatically determined by the metadata set inside the
\link{bcbioRNASeq} object. When set to \code{NULL}, this will default to
\code{sampleName}.}

\item{warnLimit}{Threshold to plot warning color marker.}

\item{fill}{Desired ggplot fill scale. Defaults to
\code{\link[=scale_fill_viridis]{scale_fill_viridis()}}. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[=scale_fill_manual]{scale_fill_manual()}}.}

\item{flip}{Flip x and y axes.}

\item{title}{Title.}
}
\value{
\code{ggplot}.
}
\description{
Plot Ribosomal RNA (rRNA) Mapping Rate
}
\examples{
load(system.file("extdata/bcb.rda", package = "bcbioRNASeq"))
plotRRNAMappingRate(bcb)
}
\seealso{
Other Quality Control Plots: \code{\link{plot5Prime3PrimeBias}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotGenesDetected}}, \code{\link{plotGene}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}},
  \code{\link{plotTotalReads}}
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
