% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRRNAMappingRate-methods.R
\docType{methods}
\name{plotRRNAMappingRate}
\alias{plotRRNAMappingRate}
\alias{plotRRNAMappingRate,bcbioRNASeq-method}
\title{Plot ribosomal RNA (rRNA) mapping rate}
\usage{
\S4method{plotRRNAMappingRate}{bcbioRNASeq}(object,
  interestingGroups = NULL, limit = 0.1,
  fill = getOption("acid.fill.discrete", default = NULL),
  flip = getOption("acid.flip", default = TRUE),
  title = "rRNA mapping rate")
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}
}
\value{
\code{ggplot}.
}
\description{
Clean, high-quality samples should have an rRNA mapping rate below 10%.
Higher rates are likely indicative of the polyA enrichment or ribo depletion
protocol not having removed all ribosomal RNA (rRNA) transcripts. This will
reduce the number of biologically meaningful reads in the experiment and is
best avoided.
}
\examples{
data(bcb)
plotRRNAMappingRate(bcb)
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
