% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRRNAMappingRate-methods.R
\docType{methods}
\name{plotRRNAMappingRate}
\alias{plotRRNAMappingRate}
\alias{plotRRNAMappingRate,bcbioRNASeq-method}
\title{Plot ribosomal RNA (rRNA) mapping rate}
\usage{
\S4method{plotRRNAMappingRate}{bcbioRNASeq}(object,
  interestingGroups = NULL, limit = 0.1,
  fill = getOption("basejump.fill.discrete", NULL),
  flip = getOption("basejump.flip", TRUE), title = "rRNA mapping rate")
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{limit}{\code{numeric(1)}.
Threshold to denote on the plot, using a dashed line.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.

To set the discrete fill palette globally, use:\preformatted{options(basejump.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{flip}{\code{logical(1)}.
Flip x and y axes. Recommended for plots containing many samples.}

\item{title}{\code{character(1)}.
Plot title.}
}
\value{
\code{ggplot}.
}
\description{
Clean, high-quality samples should have an rRNA mapping rate below 10%.
Higher rates are likely indicative of the polyA enrichment or ribo depletion
protocol not having removed all ribosomal RNA (rRNA) transcripts. This will
reduce the number of biologically meaningful reads in the experiment and is
best avoided.
}
\examples{
data(bcb)
plotRRNAMappingRate(bcb)
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
