% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRRNAMappingRate-methods.R
\name{plotRRNAMappingRate}
\alias{plotRRNAMappingRate}
\alias{plotRRNAMappingRate,bcbioRNASeq-method}
\title{Plot ribosomal RNA (rRNA) mapping rate}
\usage{
plotRRNAMappingRate(object, ...)

\S4method{plotRRNAMappingRate}{bcbioRNASeq}(
  object,
  interestingGroups = NULL,
  limit = 0.1,
  fill = getOption(x = "acid.fill.discrete", default = getOption(x =
    "ggplot2.discrete.fill", default = acidplots::scale_fill_synesthesia_d())),
  labels = list(title = "rRNA mapping rate", subtitle = NULL, sampleAxis = NULL,
    metricAxis = "rRNA mapping rate (\%)"),
  flip = getOption(x = "acid.flip", default = TRUE)
)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{limit}{\code{numeric(1)}.
Threshold limit.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{ggplot2::scale_fill_manual()}.

To set the discrete fill palette globally, use:\preformatted{options(acid.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{flip}{\code{logical(1)}.
Flip x and y axes. Recommended for plots containing many samples.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Clean, high-quality samples should have an rRNA mapping rate below 10\%.
Higher rates are likely indicative of the polyA enrichment or ribo depletion
protocol not having removed all ribosomal RNA (rRNA) transcripts. This will
reduce the number of biologically meaningful reads in the experiment and is
best avoided.
}
\note{
Updated 2019-09-16.
}
\examples{
data(bcb)
plotRRNAMappingRate(bcb)
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
