% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-metrics.R
\name{plot_total_reads}
\alias{plot_total_reads}
\alias{plot_mapped_reads}
\alias{plot_mapping_rate}
\alias{plot_genes_detected}
\alias{plot_gene_detection_saturation}
\alias{plot_exonic_mapping_rate}
\alias{plot_intronic_mapping_rate}
\alias{plot_rrna_mapping_rate}
\alias{plot_counts_per_gene}
\alias{plot_count_density}
\title{Quality control metrics plots}
\usage{
plot_total_reads(bcb, pass_limit = 20L, warn_limit = 10L,
  interesting_groups = NULL)

plot_mapped_reads(bcb, pass_limit = 20L, warn_limit = 10L,
  interesting_groups = NULL)

plot_mapping_rate(bcb, pass_limit = 90L, warn_limit = 70L,
  interesting_groups = NULL)

plot_genes_detected(bcb, pass_limit = 20000L, interesting_groups = NULL,
  filter_value = 0L)

plot_gene_detection_saturation(bcb, interesting_groups = NULL)

plot_exonic_mapping_rate(bcb, pass_limit = 60L, interesting_groups = NULL)

plot_intronic_mapping_rate(bcb, warn_limit = 20L, interesting_groups = NULL)

plot_rrna_mapping_rate(bcb, warn_limit = 10L, interesting_groups = NULL)

plot_counts_per_gene(bcb, normalized = "tmm", interesting_groups = NULL)

plot_count_density(bcb, normalized = "tmm")
}
\arguments{
\item{bcb}{\link{bcbioRnaDataSet}.}

\item{pass_limit}{Threshold to plot pass color marker.}

\item{warn_limit}{Threshold to plot warning color marker.}

\item{interesting_groups}{(\emph{Optional}). Category to use to group samples
(color and shape). If unset, this is automatically determined by
the metadata set inside the \link{bcbioRnaDataSet}.}

\item{filter_value}{Numeric value for filtering the counts matrix before plotting. Default 0.}

\item{normalized}{Count normalization method. See \code{\link[=counts]{counts()}} documentation
for more information.}
}
\value{
\link{ggplot}.
}
\description{
These functions are a collection of gene-level quality control plots.
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
