% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotTotalReads.R
\docType{methods}
\name{plotTotalReads}
\alias{plotTotalReads}
\alias{plotTotalReads}
\alias{plotTotalReads,bcbioRNASeq-method}
\title{Plot Total Reads}
\usage{
plotTotalReads(object, ...)

\S4method{plotTotalReads}{bcbioRNASeq}(object, interestingGroups,
  passLimit = 20L, warnLimit = 10L, fill = scale_fill_viridis(discrete =
  TRUE), flip = TRUE, title = "total reads")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{interestingGroups}{Character vector denoting groups of interest that
define the samples. If left unset, defaults to \code{sampleName}.}

\item{passLimit}{Threshold to plot pass color marker.}

\item{warnLimit}{Threshold to plot warning color marker.}

\item{fill}{Desired ggplot fill scale. Must supply discrete values. When set
to \code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.}

\item{flip}{Flip x and y axes.}

\item{title}{Title of plot.}
}
\value{
\code{ggplot}.
}
\description{
Plot Total Reads
}
\examples{
plotTotalReads(bcb_small)
}
\seealso{
Other Quality Control Functions: \code{\link{plot5Prime3PrimeBias}},
  \code{\link{plotCorrelationHeatmap}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotDispEsts}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotGenesDetected}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}}, \code{\link{plotMeanSD}},
  \code{\link{plotPCACovariates}}, \code{\link{plotPCA}},
  \code{\link{plotRRNAMappingRate}}
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
