% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotTotalReads.R
\docType{methods}
\name{plotTotalReads}
\alias{plotTotalReads}
\alias{plotTotalReads,bcbioRNASeqANY-method}
\alias{plotTotalReads,data.frame-method}
\title{Plot Total Reads}
\usage{
plotTotalReads(object, ...)

\S4method{plotTotalReads}{bcbioRNASeqANY}(object, interestingGroup,
  passLimit = 20L, warnLimit = 10L, flip = TRUE)

\S4method{plotTotalReads}{data.frame}(object, interestingGroup = "sampleName",
  passLimit = 20L, warnLimit = 10L, flip = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{interestingGroup}{Category to use to group samples (color and shape).
If unset, this is automatically determined by the metadata set inside the
\link{bcbioRNADataSet}.}

\item{passLimit}{Threshold to plot pass color marker.}

\item{warnLimit}{Threshold to plot warning color marker.}

\item{flip}{Flip X and Y axes.}
}
\value{
\link{ggplot}.
}
\description{
Plot Total Reads
}
\examples{
data(bcb)

# bcbioRNASeq
plotTotalReads(bcb)

# data.frame
metrics(bcb) \%>\% plotTotalReads
}
\seealso{
Other Quality Control Plots: \code{\link{plot53Bias}},
  \code{\link{plotCountDensity}},
  \code{\link{plotCountsPerGene}},
  \code{\link{plotExonicMappingRate}},
  \code{\link{plotGenderMarkers}},
  \code{\link{plotGeneSaturation}},
  \code{\link{plotGenesDetected}}, \code{\link{plotGene}},
  \code{\link{plotIntronicMappingRate}},
  \code{\link{plotMappedReads}},
  \code{\link{plotMappingRate}},
  \code{\link{plotRRNAMappingRate}}
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
\concept{Quality Control Plots}
