% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTotalReads-methods.R
\docType{methods}
\name{plotTotalReads}
\alias{plotTotalReads}
\alias{plotTotalReads,bcbioRNASeq-method}
\title{Plot total reads}
\usage{
\S4method{plotTotalReads}{bcbioRNASeq}(object, interestingGroups = NULL,
  limit = 20000000L, perMillion = TRUE,
  fill = getOption("acid.fill.discrete", default = NULL),
  flip = getOption("acid.flip", default = TRUE), title = "total reads")
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}
}
\value{
\code{ggplot}.
}
\description{
High quality RNA-seq samples ideally should have at least 10 million reads
per sample.
}
\examples{
data(bcb)
plotTotalReads(bcb)
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
