% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotTotalReads-methods.R
\name{plotTotalReads}
\alias{plotTotalReads}
\alias{plotTotalReads,bcbioRNASeq-method}
\title{Plot total reads}
\usage{
plotTotalReads(object, ...)

\S4method{plotTotalReads}{bcbioRNASeq}(
  object,
  interestingGroups = NULL,
  limit = 20000000L,
  perMillion = TRUE,
  labels = list(title = "Total reads", subtitle = NULL, sampleAxis = NULL, metricAxis =
    "reads"),
  flip = getOption(x = "acid.flip", default = TRUE)
)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{limit}{\code{numeric(1)}.
Threshold limit.}

\item{perMillion}{\code{logical(1)}.
Display as counts per million.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{ggplot2::labs()} for details.}

\item{flip}{\code{logical(1)}.
Flip x and y axes. Recommended for plots containing many samples.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
High quality RNA-seq samples ideally should have at least 10 million reads
per sample.
}
\note{
Updated 2022-05-09.
}
\examples{
data(bcb)

## bcbioRNASeq ====
plotTotalReads(bcb)
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
