% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTotalReads-methods.R
\docType{methods}
\name{plotTotalReads}
\alias{plotTotalReads}
\alias{plotTotalReads,bcbioRNASeq-method}
\title{Plot total reads}
\usage{
plotTotalReads(object, ...)

\S4method{plotTotalReads}{bcbioRNASeq}(object, interestingGroups = NULL,
  limit = 20000000L, perMillion = TRUE, fill = getOption(x =
  "acid.fill.discrete", default = acidplots::scale_fill_synesthesia_d()),
  flip = getOption(x = "acid.flip", default = TRUE),
  title = "Total reads")
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{limit}{\code{numeric(1)}.
Threshold limit.}

\item{perMillion}{\code{logical(1)}.
Display as counts per million.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.

To set the discrete fill palette globally, use:\preformatted{options(acid.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{flip}{\code{logical(1)}.
Flip x and y axes. Recommended for plots containing many samples.}

\item{title}{\code{character(1)}.
Title.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
High quality RNA-seq samples ideally should have at least 10 million reads
per sample.
}
\note{
Updated 2019-09-15.
}
\examples{
data(bcb)
plotTotalReads(bcb)
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
