% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotTotalReads-methods.R
\name{plotTotalReads}
\alias{plotTotalReads}
\alias{plotTotalReads,bcbioRNASeq-method}
\title{Plot total reads}
\usage{
plotTotalReads(object, ...)

\S4method{plotTotalReads}{bcbioRNASeq}(
  object,
  interestingGroups = NULL,
  limit = 20000000L,
  perMillion = TRUE,
  fill = getOption(x = "acid.fill.discrete", default = getOption(x =
    "ggplot2.discrete.fill", default = AcidPlots::scale_fill_synesthesia_d())),
  labels = list(title = "Total reads", subtitle = NULL, sampleAxis = NULL, metricAxis =
    "reads"),
  flip = getOption(x = "acid.flip", default = TRUE)
)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{limit}{\code{numeric(1)}.
Threshold limit.}

\item{perMillion}{\code{logical(1)}.
Display as counts per million.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete}.
Desired ggplot2 fill scale. Must supply discrete values. When set to
\code{NULL}, the default ggplot2 color palette will be used. If manual color
definitions are desired, we recommend using \code{ggplot2::scale_fill_manual()}.

To set the discrete fill palette globally, use:\preformatted{options(acid.fill.discrete = ggplot2::scale_fill_viridis_d())
}}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{flip}{\code{logical(1)}.
Flip x and y axes. Recommended for plots containing many samples.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
High quality RNA-seq samples ideally should have at least 10 million reads
per sample.
}
\note{
Updated 2021-07-21.
}
\examples{
## bcbioRNASeq ====
data(bcb)
plotTotalReads(bcb)
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
