% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTotalReads-methods.R
\docType{methods}
\name{plotTotalReads}
\alias{plotTotalReads}
\alias{plotTotalReads,bcbioRNASeq-method}
\title{Plot total reads}
\usage{
plotTotalReads(object, ...)

\S4method{plotTotalReads}{bcbioRNASeq}(object, interestingGroups = NULL,
  limit = 20000000L, perMillion = TRUE, fill = getOption(x =
  "acid.fill.discrete", default = acidplots::scale_fill_synesthesia_d()),
  flip = getOption(x = "acid.flip", default = TRUE),
  title = "Total reads")
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
High quality RNA-seq samples ideally should have at least 10 million reads
per sample.
}
\note{
Updated 2019-08-07.
}
\examples{
data(bcb)
plotTotalReads(bcb)
}
\author{
Michael Steinbaugh, Rory Kirchner, Victor Barrera
}
