% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotVolcano.R
\docType{methods}
\name{plotVolcano}
\alias{plotVolcano}
\alias{plotVolcano}
\alias{plotVolcano,DESeqResults-method}
\title{Plot Volcano}
\usage{
plotVolcano(object, ...)

\S4method{plotVolcano}{DESeqResults}(object, alpha, lfc = 0L, genes = NULL,
  gene2symbol = NULL, ntop = 0L, direction = c("both", "up", "down"),
  pointColor = "gray50", sigPointColor = c(upregulated = "purple",
  downregulated = "orange"), histograms = FALSE, return = c("ggplot",
  "data.frame"))
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}}. The names of the
arguments must be formatted in camel case, not snake case.}

\item{alpha}{Adjusted P value ("alpha") cutoff.}

\item{lfc}{Log fold change ratio (base 2) cutoff.}

\item{genes}{Character vector of genes to include. These must match the
rownames of the object. It is best practice to use the stable gene
identifiers from Ensembl (e.g. "ENSG00000000003") and not the gene symbols.}

\item{gene2symbol}{\code{data.frame} containing gene-to-symbol mappings. Must
contain the columns \code{geneID} and \code{geneName}.}

\item{ntop}{Number of top genes to label.}

\item{direction}{Plot "\code{both}", "\code{up}", or "\code{down}" directions.}

\item{pointColor}{Point color.}

\item{histograms}{Show LFC and P value histograms.}

\item{return}{Return \code{ggplot} or \code{data.frame}.}

\item{padj}{Use P values adjusted for multiple comparisions.}

\item{pointAlpha}{Point transparency alpha.}

\item{pointOutlineColor}{Point outline color.}

\item{shadeColor}{Shading color for bounding box.}

\item{shadeAlpha}{Shading transparency alpha.}

\item{labelColor}{Gene label color.}
}
\value{
\code{ggplot}.
}
\description{
Plot Volcano
}
\examples{
# DESeqResults ====
gene2symbol <- gene2symbol(bcb_small)
plotVolcano(
    object = res_small,
    ntop = 5L,
    gene2symbol = gene2symbol,
    histograms = TRUE
)

# Return coordinates as a data.frame
x <- plotVolcano(res_small, return = "data.frame")
glimpse(x)
}
\seealso{
This function is an updated variant of
\code{CHBUtils::volcano_density_plot()}.

Other Differential Expression Functions: \code{\link{alphaSummary}},
  \code{\link{contrastName}}, \code{\link{plotDEGHeatmap}},
  \code{\link{plotDEGPCA}}, \code{\link{plotMA}},
  \code{\link{resultsTables}}
}
\author{
John Hutchinson, Michael Steinbaugh, Lorena Pantano
}
