% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotVolcano.R
\docType{methods}
\name{plotVolcano}
\alias{plotVolcano}
\alias{plotVolcano}
\alias{plotVolcano,DESeqResults-method}
\alias{plotVolcano,data.frame-method}
\title{Plot Volcano}
\usage{
plotVolcano(object, ...)

\S4method{plotVolcano}{DESeqResults}(object, alpha, padj = TRUE, lfc = 0L,
  genes = NULL, gene2symbol = NULL, ntop = 0L, direction = "both",
  shadeColor = "green", shadeAlpha = 0.25, pointColor = "gray",
  pointAlpha = 0.75, pointOutlineColor = "darkgray", labelColor = "black",
  histograms = TRUE)

\S4method{plotVolcano}{data.frame}(object, alpha = 0.01, padj = TRUE,
  lfc = 0L, genes = NULL, gene2symbol = NULL, ntop = 0L,
  direction = "both", pointColor = "gray", pointAlpha = 0.75,
  pointOutlineColor = "darkgray", shadeColor = "green", shadeAlpha = 0.25,
  labelColor = "black", histograms = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{alpha}{Alpha level cutoff used for coloring.}

\item{padj}{Use P values adjusted for multiple comparisions.}

\item{lfc}{Log fold change ratio (base 2) cutoff for coloring.}

\item{genes}{Gene identifiers (rownames) to plot. These must be the stable
identifiers (e.g. ENSG00000000003) used on Ensembl and not the gene
symbols.}

\item{gene2symbol}{Apply gene identifier to symbol mappings. A gene2symbol
\link{data.frame} can be passed in, and must contain the columns \code{ensgene} and
\code{symbol}. then the Ensembl gene identifiers will be labeled in place of
gene symbols.}

\item{ntop}{Number of top genes to label.}

\item{direction}{Plot \code{up}, \code{down}, or \code{both} (\strong{default}) directions.}

\item{shadeColor}{Shading color for bounding box.}

\item{shadeAlpha}{Shading transparency alpha.}

\item{pointColor}{Point color.}

\item{pointAlpha}{Point transparency alpha.}

\item{pointOutlineColor}{Point outline color.}

\item{labelColor}{Gene label color.}

\item{histograms}{Show LFC and P value histograms.}
}
\value{
Volcano plot arranged as grid (\code{grid = TRUE}), or \code{\link[=show]{show()}}
individual \link{ggplot} (\code{grid = FALSE}).
}
\description{
Plot Volcano
}
\examples{
load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioRNASeq"))
load(system.file(
    file.path("extdata", "res.rda"),
    package = "bcbioRNASeq"))

# Use a defined gene2symbol data.frame for better speed
gene2symbol <- gene2symbol(bcb)

# Label the top genes
plotVolcano(res, ntop = 5, gene2symbol = gene2symbol)

# Label specific genes
genes <- rownames(res) \%>\% head()
plotVolcano(res, genes = genes, gene2symbol = gene2symbol)

# Label with Ensembl gene identifiers
plotVolcano(res, ntop = 1, gene2symbol = FALSE)

# data.frame
df <- as.data.frame(res)
plotVolcano(df)
}
\seealso{
This function is an updated variant of
\code{CHBUtils::volcano_density_plot()}.

Other Differential Expression Plots: \code{\link{plotMA}}
}
\author{
John Hutchinson, Michael Steinbaugh, Lorena Pantano
}
