% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotVolcano.R
\docType{methods}
\name{plotVolcano}
\alias{plotVolcano}
\alias{plotVolcano}
\alias{plotVolcano,DESeqResults-method}
\title{Plot Volcano}
\usage{
plotVolcano(object, ...)

\S4method{plotVolcano}{DESeqResults}(object, alpha, padj = TRUE, lfc = 0L,
  genes = NULL, gene2symbol = NULL, ntop = 0L, direction = c("both",
  "up", "down"), pointColor = "gray", pointAlpha = 0.75,
  pointOutlineColor = "darkgray", shadeColor = "green", shadeAlpha = 0.25,
  labelColor = "black", histograms = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{alpha}{Alpha level cutoff used for coloring.}

\item{padj}{Use P values adjusted for multiple comparisions.}

\item{lfc}{Log fold change ratio (base 2) cutoff for coloring.}

\item{genes}{Gene identifiers (rownames) to plot. These must be the stable
identifiers (e.g. "ENSG00000000003") used on Ensembl and not the gene
symbols.}

\item{gene2symbol}{Apply gene identifier to symbol mappings. A gene2symbol
\code{data.frame} can be passed in, and must contain the columns \code{geneID} and
\code{geneName}.}

\item{ntop}{Number of top genes to label.}

\item{direction}{Plot \code{up}, \code{down}, or \code{both} (\strong{default}) directions.}

\item{pointColor}{Point color.}

\item{pointAlpha}{Point transparency alpha.}

\item{pointOutlineColor}{Point outline color.}

\item{shadeColor}{Shading color for bounding box.}

\item{shadeAlpha}{Shading transparency alpha.}

\item{labelColor}{Gene label color.}

\item{histograms}{Show LFC and P value histograms.}
}
\value{
Volcano plot arranged as grid (\code{grid = TRUE}), or \code{\link[=show]{show()}}
individual \code{ggplot} (\code{grid = FALSE}).
}
\description{
Plot Volcano
}
\examples{
# DESeqResults ====
plotVolcano(
    object = res_small,
    ntop = 5L,
    gene2symbol = gene2symbol(bcb_small)
)
}
\seealso{
This function is an updated variant of
\code{CHBUtils::volcano_density_plot()}.

Other Differential Expression Functions: \code{\link{alphaSummary}},
  \code{\link{contrastName}}, \code{\link{plotDEGHeatmap}},
  \code{\link{plotDEGPCA}}, \code{\link{plotMA}},
  \code{\link{resultsTables}}
}
\author{
John Hutchinson, Michael Steinbaugh, Lorena Pantano
}
