% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotVolcano.R
\docType{methods}
\name{plotVolcano}
\alias{plotVolcano}
\alias{plotVolcano}
\alias{plotVolcano,DESeqResults-method}
\alias{plotVolcano,data.frame-method}
\title{Plot Volcano}
\usage{
plotVolcano(object, ...)

\S4method{plotVolcano}{DESeqResults}(object, alpha, padj = TRUE, lfc = 1,
  genes = NULL, gene2symbol = TRUE, ntop = 0, direction = "both",
  shadeColor = "green", shadeAlpha = 0.25, pointColor = "gray",
  pointAlpha = 0.75, pointOutlineColor = "darkgray", labelColor = "black",
  histograms = TRUE)

\S4method{plotVolcano}{data.frame}(object, alpha = 0.01, padj = TRUE,
  lfc = 1, genes = NULL, gene2symbol = TRUE, ntop = 0,
  direction = "both", pointColor = "gray", pointAlpha = 0.75,
  pointOutlineColor = "darkgray", shadeColor = "green", shadeAlpha = 0.25,
  labelColor = "black", histograms = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{alpha}{Alpha level cutoff used for coloring.}

\item{padj}{Use P values adjusted for multiple comparisions.}

\item{lfc}{Log fold change ratio (base 2) cutoff for coloring.}

\item{genes}{\emph{Optional}. Character vector of gene identifiers to plot. These
must be the stable identifiers (e.g. ENSG00000000003) used on Ensembl and
not the gene symbols.}

\item{gene2symbol}{Apply gene identifier to symbol mappings. If set \code{TRUE},
the function will attempt to automatically map gene identifiers to symbols
from Ensembl using \code{\link[basejump:annotable]{basejump::annotable()}}. If set \code{FALSE}/\code{NULL}, then
gene2symbol mapping will be disabled. This is useful when working with a
poorly annotated genome. Alternatively, a gene2symbol \link{data.frame} can be
passed in, and must contain the columns \code{ensgene} and \code{symbol}. then the
Ensembl gene identifiers will be labeled in place of gene symbols.}

\item{ntop}{Number of top genes to label.}

\item{direction}{Plot \code{up}, \code{down}, or \code{both} (\strong{default}) directions.}

\item{shadeColor}{Shading color for bounding box.}

\item{shadeAlpha}{Shading transparency alpha.}

\item{pointColor}{Point color.}

\item{pointAlpha}{Point transparency alpha.}

\item{pointOutlineColor}{Point outline color.}

\item{labelColor}{Gene label color.}

\item{histograms}{Show LFC and P value histograms.}
}
\value{
Volcano plot arranged as grid (\code{grid = TRUE}), or \code{\link[=show]{show()}}
individual \link{ggplot} (\code{grid = FALSE}).
}
\description{
Plot Volcano
}
\examples{
load(system.file(
    file.path("inst", "extdata", "bcb.rda"),
    package = "bcbioRNASeq"))
load(system.file(
    file.path("inst", "extdata", "res.rda"),
    package = "bcbioRNASeq"))

# Use a defined gene2symbol data.frame for better speed
gene2symbol <- gene2symbol(bcb)

# Label the top genes
plotVolcano(res, ntop = 5, gene2symbol = gene2symbol)

# Label specific genes
ensgene <- rownames(res) \%>\% head()
print(ensgene)
plotVolcano(res, genes = ensgene, gene2symbol = gene2symbol)

# Label with Ensembl gene identifiers
plotVolcano(res, ntop = 1, gene2symbol = FALSE)

# data.frame
df <- as.data.frame(res)
plotVolcano(df)
}
\seealso{
This function is an updated variant of
\code{CHBUtils::volcano_density_plot()}.

Other Differential Expression Plots: \code{\link{plotMA}}
}
\author{
John Hutchinson, Michael Steinbaugh, Lorena Pantano
}
