% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-plotVolcano.R
\docType{methods}
\name{plotVolcano}
\alias{plotVolcano}
\alias{plotVolcano}
\alias{plotVolcano,DESeqResults-method}
\alias{plotVolcano,data.frame-method}
\title{Plot Volcano}
\usage{
plotVolcano(object, ...)

\S4method{plotVolcano}{DESeqResults}(object, alpha = NULL, ...)

\S4method{plotVolcano}{data.frame}(object, alpha = 0.1, padj = TRUE,
  lfc = 1L, genes = NULL, ntop = 0L, direction = "both",
  shadeColor = "green", shadeAlpha = 0.25, pointColor = "gray",
  pointAlpha = 0.75, pointOutlineColor = "darkgray", histograms = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{alpha}{Alpha level cutoff used for coloring.}

\item{padj}{Use P values adjusted for multiple comparisions.}

\item{lfc}{Log fold change ratio (base 2) cutoff for coloring.}

\item{genes}{Character vector of gene symbols to label.}

\item{ntop}{Number of top genes to label.}

\item{direction}{Plot \code{up}, \code{down}, or \code{both} (\strong{default}) directions.}

\item{shadeColor}{Shading color for bounding box.}

\item{shadeAlpha}{Shading transparency alpha.}

\item{pointColor}{Point color.}

\item{pointAlpha}{Point transparency alpha.}

\item{pointOutlineColor}{Point outline color.}

\item{histograms}{Show LFC and P value histograms.}
}
\value{
Volcano plot arranged as grid (\code{grid = TRUE}), or \code{\link[=show]{show()}}
individual \link{ggplot} (\code{grid = FALSE}).
}
\description{
Plot Volcano
}
\examples{
data(res)

# DESeqResults
plotVolcano(res, genes = "Sulf1")

# data.frame
as.data.frame(res) \%>\% plotVolcano
}
\seealso{
This function is an updated variant of
\code{CHBUtils::volcano_density_plot()}.

Other Differential Expression Plots: \code{\link{plotMA}}
}
\author{
John Hutchinson, Michael Steinbaugh, Lorena Pantano
}
