% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVolcano-methods.R
\docType{methods}
\name{plotVolcano}
\alias{plotVolcano}
\alias{plotVolcano,DESeqResults-method}
\alias{plotVolcano,DESeqAnalysis-method}
\title{Plot Volcano}
\usage{
\S4method{plotVolcano}{DESeqResults}(object, ylim = 1e-10,
  genes = NULL, gene2symbol = NULL, ntop = 0L,
  direction = c("both", "up", "down"),
  pointColor = getOption("basejump.color", "gray50"),
  sigPointColor = getOption("basejump.point.color", c(upregulated =
  "purple", downregulated = "orange")), histograms = FALSE,
  return = c("ggplot", "DataFrame"))

\S4method{plotVolcano}{DESeqAnalysis}(object, results = 1L,
  lfcShrink = TRUE, ylim = 1e-10, genes = NULL, ntop = 0L,
  direction = c("both", "up", "down"),
  pointColor = getOption("basejump.color", "gray50"),
  sigPointColor = getOption("basejump.point.color", c(upregulated =
  "purple", downregulated = "orange")), histograms = FALSE,
  return = c("ggplot", "DataFrame"))
}
\arguments{
\item{object}{Object.}

\item{ylim}{\code{scalar numeric}. Upper boundary limit for y-axis. Helps preserve
dynamic range for gene sets containing highly significant P values (e.g.
\code{1e-100}).}

\item{genes}{\code{character}. Gene identifiers. It is considered better practice
to input the stable gene identifiers from Ensembl (e.g. "ENSG00000000003")
and not the (HGNC) gene symbols (e.g. "TSPN6"), if possible.}

\item{gene2symbol}{\code{Gene2Symbol}. Gene-to-symbol mappings. Must contain
\code{geneID} and \code{geneName} columns. See \code{Gene2Symbol()} for more information.}

\item{ntop}{\code{scalar integer}. Number of top genes to label.}

\item{direction}{\code{string}. Plot "\code{both}", "\code{up}", or "\code{down}" directions.}

\item{pointColor}{\code{string}. Default point color for the plot.}

\item{sigPointColor}{\code{character}. Color names for labeling upregulated and
downregulated genes. Also supports a character string for labeling DEGs
with the same color, regardless of direction.}

\item{histograms}{\code{boolean}. Show LFC and P value histograms.}

\item{return}{\code{string}. Return type. Uses \code{base::match.arg()}
internally and defaults to the first argument in the \code{character} vector.}

\item{results}{\code{scalar}. Position or name of \code{DESeqResults}.}

\item{lfcShrink}{\code{boolean}. Use shrunken log2 fold change (LFC) values.}
}
\value{
\code{ggplot}.
}
\description{
Plot Volcano
}
\examples{
data(deseq)

object <- deseq
print(object)

## Get genes from DESeqDataSet.
dds <- as(object, "DESeqDataSet")
g2s <- Gene2Symbol(dds)
geneIDs <- head(g2s[["geneID"]])
print(geneIDs)
geneNames <- head(g2s[["geneName"]])
print(geneNames)

plotVolcano(object)

## Customize the colors.
plotVolcano(
    object = object,
    pointColor = "black",
    sigPointColor = "purple"
)
plotVolcano(
    object = object,
    sigPointColor = c(
        upregulated = "green",
        downregulated = "red"
    )
)

## Directional support (up or down).
plotVolcano(
    object = object,
    direction = "up",
    ntop = 5L
)
plotVolcano(
    object = object,
    direction = "down",
    ntop = 5L
)

## Label genes manually.
## Note that either gene IDs or names (symbols) are supported.
plotVolcano(object, genes = geneIDs)
plotVolcano(object, genes = geneNames)
}
\seealso{
This method is an updated variant of
\code{CHBUtils::volcano_density_plot()}.
}
\author{
Michael Steinbaugh, John Hutchinson, Lorena Pantano
}
