% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVolcano-methods.R
\docType{methods}
\name{plotVolcano}
\alias{plotVolcano}
\alias{plotVolcano,DESeqResults-method}
\title{Volcano plot}
\usage{
plotVolcano(object, ...)

\S4method{plotVolcano}{DESeqResults}(object, alpha, lfcThreshold = 0L,
  ylim = 1e-10, genes = NULL, gene2symbol = NULL, ntop = 0L,
  direction = c("both", "up", "down"), pointColor = "gray50",
  sigPointColor = c(upregulated = "purple", downregulated = "orange"),
  histograms = FALSE, return = c("ggplot", "data.frame"))
}
\arguments{
\item{object}{Object.}

\item{alpha}{\code{scalar numeric}. Adjusted P value ("alpha") cutoff.}

\item{lfcThreshold}{\code{scalar numeric}. Log fold change ratio (base 2) cutoff
threshold.}

\item{ylim}{\code{scalar numeric}. Upper boundary limit for y-axis. Helps preserve
dynamic range for gene sets containing highly significant P values (e.g.
\code{1e-100}).}

\item{genes}{\code{character}. Genes to include. These must match the rownames of
the object. It is best practice to use the stable gene identifiers from
Ensembl (e.g. "ENSG00000000003") and not the gene symbols.}

\item{gene2symbol}{\code{data.frame}. Gene-to-symbol mappings. Must contain the
columns \code{geneID} and \code{geneName}.}

\item{ntop}{\code{scalar integer}. Number of top genes to label.}

\item{direction}{\code{string}. Plot "\code{both}", "\code{up}", or "\code{down}" directions.}

\item{pointColor}{\code{string}. Default point color for the plot.}

\item{sigPointColor}{\code{character}. Color names for labeling upregulated and
downregulated genes. Also supports a character string for labeling DEGs
with the same color, regardless of direction.}

\item{histograms}{\code{boolean}. Show LFC and P value histograms.}

\item{return}{\code{string}. Object class to return. Uses \code{\link[=match.arg]{match.arg()}} internally
and picks the first item in the vector by default.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Volcano plot
}
\examples{
gene2symbol <- gene2symbol(bcb_small)

# DESeqResults ====
# Color DEGs in each direction separately
plotVolcano(
    object = res_small,
    sigPointColor = c(
        upregulated = "purple",
        downregulated = "orange"
    )
)

# Label DEGs with a single color
plotVolcano(res_small, sigPointColor = "purple")

# Directional support
plotVolcano(
    object = res_small,
    direction = "up",
    ntop = 5L,
    gene2symbol = gene2symbol,
    histograms = TRUE
)
plotVolcano(
    object = res_small,
    direction = "down",
    ntop = 5L,
    gene2symbol = gene2symbol,
    histograms = TRUE
)

# Return coordinates as a data.frame
x <- plotVolcano(res_small, return = "data.frame")
glimpse(x)
}
\seealso{
This function is an updated variant of
\code{CHBUtils::volcano_density_plot()}.

Other Differential Expression Functions: \code{\link{alphaSummary}},
  \code{\link{contrastName}}, \code{\link{plotDEGHeatmap}},
  \code{\link{plotDEGPCA}}, \code{\link{plotMA}},
  \code{\link{resultsTables}}
}
\author{
John Hutchinson, Michael Steinbaugh, Lorena Pantano
}
\concept{Differential Expression Functions}
