% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{pool_counts}
\alias{pool_counts}
\alias{pool_dds}
\title{Lane-split FASTQ aggregation functions}
\usage{
pool_counts(counts, pattern = "_L\\\\d+")

pool_dds(dds, pattern = "_L\\\\d+", save = FALSE)
}
\arguments{
\item{counts}{Raw counts matrix.}

\item{pattern}{Grep pattern to match lane identifiers in file name.}

\item{dds}{\link{DESeqDataSet}.}

\item{save}{Whether to assign into environment and write CSV files.}
}
\value{
Object of same class, with pooled technical replicates.
}
\description{
Pool lane-split counts matrix or \link{DESeqDataSet} using \code{\link[=rowSums]{rowSums()}}.
}
\seealso{
A new \link{DESeqDataSet} is returned from \code{\link[=DESeqDataSetFromMatrix]{DESeqDataSetFromMatrix()}}.
}
\author{
Michael Steinbaugh
}
