% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{pool_lane_split_counts}
\alias{pool_lane_split_counts}
\alias{pool_lane_split_dds}
\title{Lane-split FASTQ aggregation functions}
\usage{
pool_lane_split_counts(raw_counts, lane_grep = "_L\\\\d+$")

pool_lane_split_dds(run, dds, save = TRUE)
}
\arguments{
\item{raw_counts}{Raw counts matrix.}

\item{lane_grep}{Grep string to match lane identifiers in file name.}

\item{run}{\link{bcbioRnaDataSet}.}

\item{dds}{\link{DESeqDataSet}.}

\item{save}{Whether to save and export counts.}
}
\value{
Pooled raw counts matrix.

\link{DESeqDataSet} using pooled technical replicates.
}
\description{
Pool lane split counts.

Pool \link{DESeqDataSet} counts for technical replicates. Extract
lane split technical replicate counts from a \link{DESeqDataSet}, perform
\code{\link[=rowSums]{rowSums()}}, and create a new \link{DESeqDataSet} using
\code{\link[=DESeqDataSetFromMatrix]{DESeqDataSetFromMatrix()}}.
}
\author{
Michael Steinbaugh
}
