% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{pool_counts}
\alias{pool_counts}
\alias{pool_dds}
\title{Lane-split FASTQ aggregation functions}
\usage{
pool_counts(counts, pattern = "_L\\\\d+")

pool_dds(dds, pattern = "_L\\\\d+", save = FALSE)
}
\arguments{
\item{counts}{Raw counts matrix.}

\item{pattern}{Grep pattern to match lane identifiers in file name.}

\item{dds}{\link{DESeqDataSet}.}

\item{save}{Whether to save and export counts.}
}
\value{
Pooled counts matrix.

\link{DESeqDataSet} using pooled technical replicates.
}
\description{
Pool lane split counts.

Pool \link{DESeqDataSet} counts for technical replicates. Extract
lane split technical replicate counts from a \link{DESeqDataSet}, perform
\code{\link[=rowSums]{rowSums()}}, and create a new \link{DESeqDataSet} using
\code{\link[=DESeqDataSetFromMatrix]{DESeqDataSetFromMatrix()}}.
}
\author{
Michael Steinbaugh
}
