% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareRNASeqTemplate-methods.R
\name{prepareRNASeqTemplate}
\alias{prepareRNASeqTemplate}
\title{Prepare RNA-Seq R Markdown Template}
\usage{
prepareRNASeqTemplate(overwrite = FALSE)
}
\arguments{
\item{overwrite}{\code{logical}. Should existing destination files be overwritten?}
}
\value{
Invisible \code{logical} vector indicating which files were copied.
}
\description{
If the required template dependency files aren't present, download latest
versions from the package website. Existing files are never overwritten.
}
\details{
By default, this function will create local copies of these files:
\itemize{
\item \code{_footer.Rmd}
\item \code{_header.Rmd}
\item \code{_output.yaml}
\item \code{_setup.R}
\item \code{bibliography.bib}
}

This code is used internally by:
\itemize{
\item \code{bcbioRNASeq::prepareRNASeqTemplate()}.
\item \code{bcbioSingleCell::prepareSingleCellTemplate()}.
}
}
\examples{
x <- prepareRNASeqTemplate()
x

# Clean up
unlink(c(
    "_footer.Rmd",
    "_header.Rmd",
    "_output.yaml",
    "_setup.R",
    "bibliography.bib"
))
}
\seealso{
Other R Markdown Functions: \code{\link{topTables}}
}
\author{
Michael Steinbaugh
}
\concept{R Markdown Functions}
