% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/methods-prepareRNASeqTemplate.R
\docType{methods}
\name{prepareRNASeqTemplate}
\alias{prepareRNASeqTemplate}
\alias{prepareRNASeqTemplate}
\alias{prepareRNASeqTemplate,missing-method}
\title{Prepare RNA-Seq R Markdown Template}
\usage{
prepareRNASeqTemplate(object, ...)

\S4method{prepareRNASeqTemplate}{missing}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
No value.
}
\description{
Prepare RNA-Seq R Markdown Template
}
\examples{
prepareRNASeqTemplate()

# Clean up
unlink(c(
    "_footer.Rmd",
    "_header.Rmd",
    "_output.yaml",
    "_setup.R",
    "bibliography.bib",
))
}
\author{
Michael Steinbaugh
}
