% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareRNASeqTemplate.R
\name{prepareRNASeqTemplate}
\alias{prepareRNASeqTemplate}
\title{Prepare RNA-Seq R Markdown Template}
\usage{
prepareRNASeqTemplate(overwrite = FALSE)
}
\arguments{
\item{overwrite}{\code{logical(1)}. Should existing destination files be overwritten?}
}
\value{
Invisible \code{logical}. Was the file copied?.
}
\description{
If the required template dependency files aren't present, copy them from the
requested package. Existing files are not overwritten by default. This
function will copy dependency files from a requested package inside the
\code{rmarkdown/shared/} directory. If a package doesn't contain this
subdirectory, the function will return an error.
}
\details{
This code is used internally by:
\itemize{
\item \code{bcbioRNASeq::prepareRNASeqTemplate}.
\item \code{bcbioSingleCell::prepareSingleCellTemplate}.
}
}
\examples{
x <- prepareRNASeqTemplate()
x

## Clean up.
unlink(c(
    "_footer.Rmd",
    "_header.Rmd",
    "_output.yaml",
    "_setup.R",
    "bibliography.bib"
))
}
\author{
Michael Steinbaugh
}
