% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-qc.R, R/plot_mean_sd.R
\name{plot_total_reads}
\alias{plot_total_reads}
\alias{plot_mapped_reads}
\alias{plot_mapping_rate}
\alias{plot_genes_detected}
\alias{plot_gene_detection_saturation}
\alias{plot_exonic_mapping_rate}
\alias{plot_intronic_mapping_rate}
\alias{plot_rrna_mapping_rate}
\alias{plot_counts_per_gene}
\alias{plot_count_density}
\alias{plot_gender_markers}
\alias{plot_mean_sd}
\title{Gene-level quality control plots}
\usage{
plot_total_reads(bcb, pass_limit = 20L, warn_limit = 10L,
  interesting_groups = NULL)

plot_mapped_reads(bcb, pass_limit = 20L, warn_limit = 10L,
  interesting_groups = NULL)

plot_mapping_rate(bcb, pass_limit = 90, warn_limit = 70,
  interesting_groups = NULL)

plot_genes_detected(bcb, pass_limit = 20000, interesting_groups = NULL)

plot_gene_detection_saturation(bcb, interesting_groups = NULL)

plot_exonic_mapping_rate(bcb, pass_limit = 60L, interesting_groups = NULL)

plot_intronic_mapping_rate(bcb, warn_limit = 20L, interesting_groups = NULL)

plot_rrna_mapping_rate(bcb, warn_limit = 10L, interesting_groups = NULL)

plot_counts_per_gene(bcb, counts = "tmm", interesting_groups = NULL)

plot_count_density(bcb, counts = "tmm")

plot_gender_markers(bcb, normalized_counts = NULL)

plot_mean_sd(dds, rld = NULL, vsd = NULL)
}
\arguments{
\item{bcb}{\link{bcbioRnaDataSet}.}

\item{pass_limit}{Threshold to plot pass color marker.}

\item{warn_limit}{Threshold to plot warning color marker.}

\item{interesting_groups}{(\emph{Optional}). Category to use to group samples
(color and shape). If unset, this is automatically determined by
the metadata set inside the \link{bcbioRnaDataSet}.}

\item{dds}{\link{DESeqDataSet}.}

\item{rld}{\link{DESeqTransform} resulting from \code{\link[=rlog]{rlog()}} applied to
\link{DESeqDataSet}.}

\item{vsd}{\link{DESeqTransform} resulting from
\code{\link[=varianceStabilizingTransformation]{varianceStabilizingTransformation()}} applied to \link{DESeqDataSet}.}

\item{normalized}{Apply normalization to counts.}
}
\value{
\link{ggplot}.
}
\description{
Gene-level quality control plots

Modified wrapper function of \code{\link[vsn:meanSdPlot]{vsn::meanSdPlot()}} that plots \code{\link[=log2]{log2()}},
\code{\link[=rlog]{rlog()}}, and \code{\link[=varianceStabilizingTransformation]{varianceStabilizingTransformation()}} count data.
}
\author{
Michael Steinbaugh

Rory Kirchner

Victor Barrera
}
