% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{qc_plots}
\alias{qc_plots}
\title{Quality Control Plots}
\arguments{
\item{counts}{Counts matrix.}

\item{flip}{Flip X and Y axes.}

\item{interesting_group}{\emph{Optional}. Category to use to group samples (color
and shape). If unset, this is automatically determined by the metadata set
inside the \link{bcbioRNADataSet}.}

\item{melted}{Melted counts generated by \code{\link[=melt_log10]{melt_log10()}}.}

\item{metrics}{Metrics \link{data.frame} generated by \code{\link[=metrics]{metrics()}}.}

\item{min_counts}{Numeric value for filtering the counts matrix before
plotting.}

\item{normalized}{Count normalization method. See \code{\link[=counts]{counts()}} documentation
for more information.}

\item{pass_limit}{Threshold to plot pass color marker.}

\item{warn_limit}{Threshold to plot warning color marker.}
}
\value{
\link{ggplot}.
}
\description{
Quality Control Plots
}
\keyword{internal}
