% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{plot_total_reads}
\alias{plot_total_reads}
\alias{plot_mapped_reads}
\alias{plot_mapping_rate}
\alias{plot_genes_detected}
\alias{plot_gene_detection_saturation}
\alias{plot_exonic_mapping_rate}
\alias{plot_intronic_mapping_rate}
\alias{plot_rrna_mapping_rate}
\alias{plot_counts_per_gene}
\alias{plot_count_density}
\alias{plot_gender_markers}
\title{Standard RNA-seq quality control plots}
\usage{
plot_total_reads(run)

plot_mapped_reads(run)

plot_mapping_rate(run)

plot_genes_detected(run, raw_counts)

plot_gene_detection_saturation(run, raw_counts)

plot_exonic_mapping_rate(run)

plot_intronic_mapping_rate(run)

plot_rrna_mapping_rate(run)

plot_counts_per_gene(run, normalized_counts)

plot_count_density(run, normalized_counts)

plot_gender_markers(run, normalized_counts)
}
\arguments{
\item{run}{bcbio-nextgen run}

\item{raw_counts}{Raw counts matrix. Can be obtained from \code{DESeqDataSet}
by running \code{counts(normalized = FALSE)}.}

\item{normalized_counts}{Normalized counts matrix. Can be obtained from
\code{DESeqDataSet} by running \code{counts(normalized = TRUE)}.
Transcripts per million (TPM) are also acceptable.}
}
\value{
Bar plot

Bar plot

Bar plot

Bar plot

Smooth plot

Bar plot

Bar plot

Bar plot

Boxplot

Density plot

Scatter plot
}
\description{
Plot total reads

Mapped reads plot

Mapping rate plot

Genes detected plot

Genes detection saturation plot

Exonic mapping rate plot

Intronic mapping rate plot

rRNA contamination mapping rate

Counts per gene plot

Plot count density

Plot sexually dimorphic gender marker genes
}
\author{
Michael Steinbaugh

Rory Kirchner
}
