% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{qc_plots}
\alias{qc_plots}
\title{Quality Control Plots}
\arguments{
\item{pass_limit}{Threshold to plot pass color marker.}

\item{warn_limit}{Threshold to plot warning color marker.}

\item{interesting_group}{\emph{Optional}. Category to use to group samples (color
and shape). If unset, this is automatically determined by the metadata set
inside the \link{bcbioRNADataSet}.}

\item{min_counts}{Numeric value for filtering the counts matrix before
plotting.}

\item{normalized}{Count normalization method. See \code{\link[=counts]{counts()}} documentation
for more information.}
}
\value{
\link{ggplot}.
}
\description{
Quality Control Plots
}
