% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-qc.R
\name{plot_total_reads}
\alias{plot_total_reads}
\alias{plot_mapped_reads}
\alias{plot_mapping_rate}
\alias{plot_genes_detected}
\alias{plot_gene_detection_saturation}
\alias{plot_exonic_mapping_rate}
\alias{plot_intronic_mapping_rate}
\alias{plot_rrna_mapping_rate}
\alias{plot_counts_per_gene}
\alias{plot_count_density}
\alias{plot_gender_markers}
\title{Gene-level quality control plots}
\usage{
plot_total_reads(run, pass_limit = 20, warn_limit = 10)

plot_mapped_reads(run, pass_limit = 20, warn_limit = 10)

plot_mapping_rate(run, pass_limit = 90, warn_limit = 70)

plot_genes_detected(run, raw_counts, pass_limit = 20000)

plot_gene_detection_saturation(run, raw_counts)

plot_exonic_mapping_rate(run, pass_limit = 60)

plot_intronic_mapping_rate(run, warn_limit = 20)

plot_rrna_mapping_rate(run, warn_limit = 10)

plot_counts_per_gene(run, normalized_counts)

plot_count_density(run, normalized_counts)

plot_gender_markers(run, normalized_counts)
}
\arguments{
\item{run}{bcbio-nextgen run.}

\item{pass_limit}{Threshold to plot pass color marker.}

\item{warn_limit}{Threshold to plot warning color marker.}

\item{raw_counts}{Raw counts matrix. Can be obtained from
\link{DESeqDataSet} by running \code{counts(normalized = FALSE)}.}

\item{normalized_counts}{Normalized counts matrix. Can be obtained from
\link{DESeqDataSet} by running \code{counts(normalized = TRUE)}.
Transcripts per million (TPM) are also acceptable.}
}
\value{
ggplot object.
}
\description{
Gene-level quality control plots
}
\author{
Michael Steinbaugh

Rory Kirchner

Victor Barrera
}
