% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-qc.R, R/plot_mean_sd.R
\name{plot_total_reads}
\alias{plot_total_reads}
\alias{plot_mapped_reads}
\alias{plot_mapping_rate}
\alias{plot_genes_detected}
\alias{plot_gene_detection_saturation}
\alias{plot_exonic_mapping_rate}
\alias{plot_intronic_mapping_rate}
\alias{plot_rrna_mapping_rate}
\alias{plot_counts_per_gene}
\alias{plot_count_density}
\alias{plot_gender_markers}
\alias{plot_mean_sd}
\title{Gene-level quality control plots}
\usage{
plot_total_reads(bcb, pass_limit = 20, warn_limit = 10, intgroup = NULL)

plot_mapped_reads(bcb, pass_limit = 20, warn_limit = 10, intgroup = NULL)

plot_mapping_rate(bcb, pass_limit = 90, warn_limit = 70, intgroup = NULL)

plot_genes_detected(bcb, raw_counts, pass_limit = 20000, intgroup = NULL)

plot_gene_detection_saturation(bcb, raw_counts, intgroup = NULL)

plot_exonic_mapping_rate(bcb, pass_limit = 60, intgroup = NULL)

plot_intronic_mapping_rate(bcb, warn_limit = 20, intgroup = NULL)

plot_rrna_mapping_rate(bcb, warn_limit = 10, intgroup = NULL)

plot_counts_per_gene(bcb, normalized_counts, intgroup = NULL)

plot_count_density(bcb, normalized_counts)

plot_gender_markers(bcb, normalized_counts)

plot_mean_sd(dds, rld = NULL, vsd = NULL)
}
\arguments{
\item{bcb}{\link{bcbioRnaDataSet}.}

\item{pass_limit}{Threshold to plot pass color marker.}

\item{warn_limit}{Threshold to plot warning color marker.}

\item{intgroup}{Category to use to group samples (color and shape). Default: \code{NULL}}

\item{raw_counts}{Raw counts matrix. Can be obtained from
\link{DESeqDataSet} using \code{\link[=counts]{counts()}} with \code{normalized = FALSE}.}

\item{normalized_counts}{Normalized counts matrix. Can be obtained from
\link{DESeqDataSet} using \code{\link[=counts]{counts()}} with \code{normalized = TRUE}.
Transcripts per million (TPM) are also acceptable.}

\item{dds}{\link{DESeqDataSet}.}

\item{rld}{\link{DESeqTransform} resulting from \code{\link[=rlog]{rlog()}} applied to
\link{DESeqDataSet}.}

\item{vsd}{\link{DESeqTransform} resulting from
\code{\link[=varianceStabilizingTransformation]{varianceStabilizingTransformation()}} applied to \link{DESeqDataSet}.}
}
\value{
\link{ggplot}.
}
\description{
Gene-level quality control plots

Modified wrapper function of \code{\link[vsn:meanSdPlot]{vsn::meanSdPlot()}} that plots \code{\link[=log2]{log2()}},
\code{\link[=rlog]{rlog()}}, and \code{\link[=varianceStabilizingTransformation]{varianceStabilizingTransformation()}} count data.
}
\author{
Michael Steinbaugh

Rory Kirchner

Victor Barrera
}
