% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_bcbio_counts}
\alias{read_bcbio_counts}
\alias{read_bcbio_file}
\alias{read_bcbio_metadata}
\alias{read_bcbio_metrics}
\alias{read_bcbio_samples_yaml}
\title{Read data from bcbio-nextgen run}
\usage{
read_bcbio_counts(run, grep = NULL, samples = NULL)

read_bcbio_file(run, file, row_names = NULL, ...)

read_bcbio_metadata(run)

read_bcbio_metrics(run)

read_bcbio_samples_yaml(run, ...)
}
\arguments{
\item{run}{bcbio-nextgen run.}

\item{grep}{Use grep pattern to match samples. This will override the
\code{samples} parameter if set.}

\item{samples}{Character vector of sample names in bcbio final directory to
include. If \code{NULL} (default), all samples will be imported.}

\item{file}{File name.}

\item{row_names}{Column identifier to use for row names.}

\item{...}{Passthrough parameters for \link{readr} importers (\code{\link[readr:read_csv]{readr::read_csv()}},
\code{\link[readr:read_tsv]{readr::read_tsv()}}).}

\item{...}{Nested operator keys supplied as dot objects.}
}
\value{
Counts saved in txi list object.

Miscellaneous data frame.

Metadata data frame.

Summary statistics data frame.

Sample information data frame.
}
\description{
Read RNA-seq counts using \code{\link[tximport:tximport]{tximport::tximport()}}. Currently supports
\href{https://combine-lab.github.io/salmon/}{salmon} (preferred) and
\href{http://www.cs.cmu.edu/~ckingsf/software/sailfish/}{sailfish}.

Read a project data file.

Read sample metadata from YAML.

Read summary metrics from YAML. These statistics are only
generated for a standard RNA-seq run with aligned counts. Fast RNA-seq mode
with lightweight counts (pseudocounts) doesn't output the same metrics into
the YAML.

Read bcbio sample information from YAML.
}
\seealso{
\itemize{
\item \code{\link[tximport:tximport]{tximport::tximport()}}.
}

\itemize{
\item \code{\link[readr:read_csv]{readr::read_csv()}}.
\item \code{\link[readr:read_tsv]{readr::read_tsv()}}.
}
}
\author{
Michael Steinbaugh

Rory Kirchner
}
\keyword{internal}
