% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bcbio.R
\name{read_bcbio_counts}
\alias{read_bcbio_counts}
\alias{read_bcbio_file}
\alias{read_bcbio_metadata}
\alias{read_bcbio_metrics}
\title{Read data from bcbio-nextgen run}
\usage{
read_bcbio_counts(run, type = "salmon", grep = NULL, samples = NULL)

read_bcbio_file(run, file, row_names = NULL, ...)

read_bcbio_metadata(run)

read_bcbio_metrics(run)
}
\arguments{
\item{run}{\code{bcbio-nextgen} run}

\item{type}{The type of software used to generate the abundances. Follows the
conventions of \code{tximport}.}

\item{grep}{Apply grep pattern matching to samples}

\item{samples}{Specify the names of samples in bcbio final directory to
input. If \code{NULL} (default), all samples will be imported.}

\item{file}{File name}

\item{row_names}{Column identifier to use for row names}

\item{...}{Optional parameters for \code{readr}}
}
\value{
txi \code{tximport} list

bcbio run data

Metadata data frame

Summary statistics data frame
}
\description{
Read RNA-Seq counts using \code{tximport}

Read a project data file

Read sample metadata from YAML

Read summary metrics from YAML
}
\author{
Michael Steinbaugh

Rory Kirchner
}
