% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-files.R
\name{.read_file}
\alias{.read_file}
\title{Read file}
\usage{
.read_file(file, column_to_rownames = NULL, ...)
}
\arguments{
\item{file}{File path.}

\item{column_to_rownames}{Column identifier to use for row names.}

\item{...}{Additional parameters.}
}
\value{
\link{DataFrame}.
}
\description{
Supports automatic loading of \code{.csv}, \code{.tsv}, \code{.xlsx}, and \code{.counts} files.
}
\seealso{
\itemize{
\item \href{http://readr.tidyverse.org}{readr}
\item \href{http://readxl.tidyverse.org}{readxl}
}
}
\author{
Michael Steinbaugh
}
\keyword{internal}
