% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-read_file.R
\name{.read_file}
\alias{.read_file}
\title{Read file}
\usage{
.read_file(file, ...)
}
\arguments{
\item{file}{File path.}

\item{...}{Additional parameters.}
}
\value{
\link{tibble}.
}
\description{
Supports automatic loading of standard \code{.csv}, \code{.tsv}, and \code{.xlsx}
files. Also supports bcbio-nextgen pipeline-specific \code{.counts}, \code{.colnames},
and \code{.rownames} files.
}
\seealso{
\href{http://readr.tidyverse.org}{readr},
\href{http://readxl.tidyverse.org}{readxl}.
}
\author{
Michael Steinbaugh
}
\keyword{internal}
