% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Read metadata}
\usage{
read_metadata(csv, grep = NULL, pool = FALSE, save = FALSE)
}
\arguments{
\item{csv}{Comma separated values file}

\item{grep}{Apply grep pattern matching to samples}

\item{pool}{Pool lane split samples}

\item{save}{Save csv and rda files}
}
\value{
Metadata data frame
}
\description{
Read metadata
}
