% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{read_metadata}
\alias{read_metadata}
\title{Read metadata}
\usage{
read_metadata(file, pattern = NULL, pattern_col = "description",
  lanes = NULL)
}
\arguments{
\item{file}{Metadata file. CSV and XLSX formats are supported.}

\item{pattern}{Apply grep pattern matching to samples}

\item{pattern_col}{Column in data frame used for pattern subsetting}

\item{lanes}{Number of lanes used to split the samples into technical
replicates (\code{_LXXX}) suffix.}
}
\value{
Metadata data frame.
}
\description{
Read metadata
}
\author{
Michael Steinbaugh
}
\keyword{internal}
