% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bcbio.R
\name{read_smallrna_counts}
\alias{read_smallrna_counts}
\title{Create isomiRs object from bcbio output}
\usage{
read_smallrna_counts(bcbiods)
}
\arguments{
\item{bcbiods}{bcbioRnaDataSet object}
}
\description{
Read bcbio sample information from YAML to get
isomiR object
}
\details{
Read bcbio output to create isomiRs object
}
