% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{reexports-S4}
\alias{reexports-S4}
\alias{plotFeaturesDetected}
\alias{plotHeatmap}
\alias{plotQuantileHeatmap}
\title{S4 generics re-exported from other packages}
\usage{
plotFeaturesDetected(object, ...)

plotHeatmap(object, ...)

plotQuantileHeatmap(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\description{
S4 generics re-exported from other packages
}
