% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateReplicates-methods.R,
%   R/counts-methods.R, R/metrics-methods.R, R/plotCorrelationHeatmap-methods.R,
%   R/sampleData-methods.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{aggregateReplicates}
\alias{reexports}
\alias{counts}
\alias{reexports}
\alias{metrics}
\alias{reexports}
\alias{plotCorrelationHeatmap}
\alias{reexports}
\alias{sampleData}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump}{\code{\link[basejump]{aggregateReplicates}}, \code{\link[basejump]{sampleData}}}

  \item{bcbioBase}{\code{\link[bcbioBase]{metrics}}, \code{\link[bcbioBase]{plotCorrelationHeatmap}}}

  \item{BiocGenerics}{\code{\link[BiocGenerics]{counts}}}
}}

