% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaSummary-methods.R,
%   R/contrastName-methods.R, R/counts-methods.R, R/export-methods.R,
%   R/markdown-methods.R, R/package.R, R/plot5Prime3PrimeBias-methods.R,
%   R/plotCorrelationHeatmap-methods.R, R/plotCountsPerBiotype-methods.R,
%   R/plotCountsPerGene-methods.R, R/plotDEGHeatmap-methods.R,
%   R/plotPCA-methods.R, R/plotDEGPCA-methods.R, R/plotDispEsts-methods.R,
%   R/plotExonicMappingRate-methods.R, R/plotGenderMarkers-methods.R,
%   R/plotGene-methods.R, R/plotGeneSaturation-methods.R,
%   R/plotIntronicMappingRate-methods.R, R/plotMA-methods.R,
%   R/plotMappedReads-methods.R, R/plotMappingRate-methods.R,
%   R/plotMeanSD-methods.R, R/plotQC-methods.R, R/plotRRNAMappingRate-methods.R,
%   R/plotTotalReads-methods.R, R/plotVolcano-methods.R,
%   R/relativeLogExpression-methods.R, R/sampleData-methods.R, R/show-methods.R,
%   R/tmm-methods.R, R/topTables-methods.R, R/updateObject-methods.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump}{\code{\link[basejump]{alphaSummary}}, \code{\link[basejump]{contrastName}}, \code{\link[basejump]{export}}, \code{\link[basejump]{markdown}}, \code{\link[basejump]{plot5Prime3PrimeBias}}, \code{\link[basejump]{plotCorrelationHeatmap}}, \code{\link[basejump]{plotCountsPerBiotype}}, \code{\link[basejump]{plotCountsPerBroadClass}}, \code{\link[basejump]{plotCountsPerGene}}, \code{\link[basejump]{plotDEGHeatmap}}, \code{\link[basejump]{plotDEGPCA}}, \code{\link[basejump]{plotExonicMappingRate}}, \code{\link[basejump]{plotGenderMarkers}}, \code{\link[basejump]{plotGene}}, \code{\link[basejump]{plotGeneSaturation}}, \code{\link[basejump]{plotIntronicMappingRate}}, \code{\link[basejump]{plotMappedReads}}, \code{\link[basejump]{plotMappingRate}}, \code{\link[basejump]{plotMeanSD}}, \code{\link[basejump]{plotQC}}, \code{\link[basejump]{plotRRNAMappingRate}}, \code{\link[basejump]{plotTotalReads}}, \code{\link[basejump]{plotVolcano}}, \code{\link[basejump]{relativeLogExpression}}, \code{\link[basejump]{sampleData}}, \code{\link[basejump]{tmm}}, \code{\link[basejump]{topTables}}}

  \item{BiocGenerics}{\code{\link[BiocGenerics]{counts}}, \code{\link[BiocGenerics]{counts<-}}, \code{\link[BiocGenerics]{plotPCA}}, \code{\link[BiocGenerics]{plotDispEsts}}, \code{\link[BiocGenerics]{plotMA}}, \code{\link[BiocGenerics]{updateObject}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{methods}{\code{\link[methods]{show}}}
}}

