% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{counts}
\alias{reexports}
\alias{design}
\alias{reexports}
\alias{meanSdPlot}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{DESeq2}{\code{\link[DESeq2]{counts}}, \code{\link[DESeq2]{design}}}

  \item{vsn}{\code{\link[vsn]{meanSdPlot}}}
}}

