% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{ggplot}
\alias{reexports}
\alias{scale_color_hue}
\alias{reexports}
\alias{scale_color_manual}
\alias{reexports}
\alias{scale_colour_hue}
\alias{reexports}
\alias{scale_colour_manual}
\alias{reexports}
\alias{scale_fill_hue}
\alias{reexports}
\alias{scale_fill_manual}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{glimpse}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{scale_color_hue}}, \code{\link[ggplot2]{scale_color_manual}}, \code{\link[ggplot2]{scale_colour_hue}}, \code{\link[ggplot2]{scale_colour_manual}}, \code{\link[ggplot2]{scale_fill_hue}}, \code{\link[ggplot2]{scale_fill_manual}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{tibble}{\code{\link[tibble]{glimpse}}}
}}

