% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaSummary-methods.R,
%   R/contrastName-methods.R, R/counts-methods.R, R/export-methods.R,
%   R/markdown-methods.R, R/plot5Prime3PrimeBias-methods.R,
%   R/plotCorrelationHeatmap-methods.R, R/plotCountsPerGene-methods.R,
%   R/plotDEGHeatmap-methods.R, R/plotPCA-methods.R, R/plotDEGPCA-methods.R,
%   R/plotDispEsts-methods.R, R/plotExonicMappingRate-methods.R,
%   R/plotGenderMarkers-methods.R, R/plotGene-methods.R,
%   R/plotGeneSaturation-methods.R, R/plotIntronicMappingRate-methods.R,
%   R/plotMA-methods.R, R/plotMappedReads-methods.R,
%   R/plotMappingRate-methods.R, R/plotMeanSD-methods.R, R/plotQC-methods.R,
%   R/plotRRNAMappingRate-methods.R, R/plotTotalReads-methods.R,
%   R/plotVolcano-methods.R, R/relativeLogExpression-methods.R,
%   R/sampleData-methods.R, R/show-methods.R, R/tmm-methods.R,
%   R/topTables-methods.R, R/updateObject-methods.R
\docType{import}
\name{reexports}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump.generics}{\code{\link[basejump.generics]{alphaSummary}}, \code{\link[basejump.generics]{contrastName}}, \code{\link[basejump.generics]{markdown}}, \code{\link[basejump.generics]{plot5Prime3PrimeBias}}, \code{\link[basejump.generics]{plotCorrelationHeatmap}}, \code{\link[basejump.generics]{plotCountsPerGene}}, \code{\link[basejump.generics]{plotDEGHeatmap}}, \code{\link[basejump.generics]{plotDEGPCA}}, \code{\link[basejump.generics]{plotExonicMappingRate}}, \code{\link[basejump.generics]{plotGenderMarkers}}, \code{\link[basejump.generics]{plotGene}}, \code{\link[basejump.generics]{plotGeneSaturation}}, \code{\link[basejump.generics]{plotIntronicMappingRate}}, \code{\link[basejump.generics]{plotMappedReads}}, \code{\link[basejump.generics]{plotMappingRate}}, \code{\link[basejump.generics]{plotMeanSD}}, \code{\link[basejump.generics]{plotQC}}, \code{\link[basejump.generics]{plotRRNAMappingRate}}, \code{\link[basejump.generics]{plotTotalReads}}, \code{\link[basejump.generics]{plotVolcano}}, \code{\link[basejump.generics]{relativeLogExpression}}, \code{\link[basejump.generics]{sampleData}}, \code{\link[basejump.generics]{tmm}}, \code{\link[basejump.generics]{topTables}}}

  \item{basejump.io}{\code{\link[basejump.io]{export}}}

  \item{BiocGenerics}{\code{\link[BiocGenerics]{counts}}, \code{\link[BiocGenerics]{counts<-}}, \code{\link[BiocGenerics]{plotPCA}}, \code{\link[BiocGenerics]{plotDispEsts}}, \code{\link[BiocGenerics]{plotMA}}, \code{\link[BiocGenerics]{updateObject}}}

  \item{methods}{\code{\link[methods]{show}}}
}}

