% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts-methods.R, R/export-methods.R,
%   R/markdown-methods.R, R/plotCorrelationHeatmap-methods.R,
%   R/plotCountsPerGene-methods.R, R/plotPCA-methods.R,
%   R/plotDispEsts-methods.R, R/plotGenderMarkers-methods.R,
%   R/plotGene-methods.R, R/plotMA-methods.R, R/plotQC-methods.R,
%   R/sampleData-methods.R, R/show-methods.R, R/updateObject-methods.R
\docType{import}
\name{reexports}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump}{\code{\link[basejump]{export}}}

  \item{basejump.generics}{\code{\link[basejump.generics]{markdown}}, \code{\link[basejump.generics]{plotCorrelationHeatmap}}, \code{\link[basejump.generics]{plotCountsPerGene}}, \code{\link[basejump.generics]{plotGenderMarkers}}, \code{\link[basejump.generics]{plotGene}}, \code{\link[basejump.generics]{plotQC}}, \code{\link[basejump.generics]{sampleData}}}

  \item{BiocGenerics}{\code{\link[BiocGenerics]{counts}}, \code{\link[BiocGenerics]{counts<-}}, \code{\link[BiocGenerics]{plotPCA}}, \code{\link[BiocGenerics]{plotDispEsts}}, \code{\link[BiocGenerics]{plotMA}}, \code{\link[BiocGenerics]{updateObject}}}

  \item{methods}{\code{\link[methods]{show}}}
}}

