% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbioRnaseq.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{counts}
\alias{reexports}
\alias{design}
\alias{reexports}
\alias{meanSdPlot}
\alias{reexports}
\alias{!!!}
\alias{reexports}
\alias{!!}
\alias{reexports}
\alias{sym}
\alias{reexports}
\alias{syms}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{glimpse}
\alias{reexports}
\alias{kable}
\alias{reexports}
\alias{opts_chunk}
\alias{reexports}
\alias{set_rownames}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{DESeq2}{\code{\link[DESeq2]{counts}}, \code{\link[DESeq2]{design}}}

  \item{knitr}{\code{\link[knitr]{kable}}, \code{\link[knitr]{opts_chunk}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}, \code{\link[magrittr]{set_rownames}}}

  \item{rlang}{\code{\link[rlang]{!!!}}, \code{\link[rlang]{!!}}, \code{\link[rlang]{sym}}, \code{\link[rlang]{syms}}}

  \item{tibble}{\code{\link[tibble]{glimpse}}}

  \item{vsn}{\code{\link[vsn]{meanSdPlot}}}
}}

