% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateReplicates-methods.R,
%   R/counts-methods.R, R/metrics-methods.R, R/plotCorrelationHeatmap-methods.R,
%   R/plotGene-methods.R, R/plotMA-methods.R, R/plotPCA-methods.R,
%   R/sampleData-methods.R, R/updateObject-methods.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{aggregateReplicates}
\alias{reexports}
\alias{counts}
\alias{reexports}
\alias{metrics}
\alias{reexports}
\alias{plotCorrelationHeatmap}
\alias{reexports}
\alias{plotGene}
\alias{reexports}
\alias{plotMA}
\alias{reexports}
\alias{plotPCA}
\alias{reexports}
\alias{sampleData}
\alias{reexports}
\alias{updateObject}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump}{\code{\link[basejump]{aggregateReplicates}}, \code{\link[basejump]{sampleData}}}

  \item{bcbioBase}{\code{\link[bcbioBase]{metrics}}, \code{\link[bcbioBase]{plotCorrelationHeatmap}}, \code{\link[bcbioBase]{plotGene}}}

  \item{BiocGenerics}{\code{\link[BiocGenerics]{counts}}, \code{\link[BiocGenerics]{plotMA}}, \code{\link[BiocGenerics]{plotPCA}}, \code{\link[BiocGenerics]{updateObject}}}
}}

