% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbioRnaseq.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{!!!}
\alias{reexports}
\alias{!!}
\alias{reexports}
\alias{sym}
\alias{reexports}
\alias{syms}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{kable}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{knitr}{\code{\link[knitr]{kable}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{rlang}{\code{\link[rlang]{!!!}}, \code{\link[rlang]{!!}}, \code{\link[rlang]{sym}}, \code{\link[rlang]{syms}}}
}}

