% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basejump.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{assay}
\alias{reexports}
\alias{assays}
\alias{reexports}
\alias{colData}
\alias{reexports}
\alias{design}
\alias{reexports}
\alias{metadata}
\alias{reexports}
\alias{rowData}
\alias{reexports}
\alias{save_data}
\alias{reexports}
\alias{write_counts}
\alias{reexports}
\alias{degCovariates}
\alias{reexports}
\alias{DESeq}
\alias{reexports}
\alias{DESeqDataSet}
\alias{reexports}
\alias{DESeqDataSetFromMatrix}
\alias{reexports}
\alias{DESeqResults}
\alias{reexports}
\alias{DESeqTransform}
\alias{reexports}
\alias{results}
\alias{reexports}
\alias{rlog}
\alias{reexports}
\alias{vst}
\alias{reexports}
\alias{tximport}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{DEGreport}{\code{\link[DEGreport]{degCovariates}}}

  \item{DESeq2}{\code{\link[DESeq2]{DESeq}}, \code{\link[DESeq2]{DESeqDataSet}}, \code{\link[DESeq2]{DESeqDataSetFromMatrix}}, \code{\link[DESeq2]{DESeqResults}}, \code{\link[DESeq2]{DESeqTransform}}, \code{\link[DESeq2]{results}}, \code{\link[DESeq2]{rlog}}, \code{\link[DESeq2]{vst}}}

  \item{basejump}{\code{\link[basejump]{\%>\%}}, \code{\link[basejump]{assay}}, \code{\link[basejump]{assays}}, \code{\link[basejump]{colData}}, \code{\link[basejump]{design}}, \code{\link[basejump]{metadata}}, \code{\link[basejump]{rowData}}, \code{\link[basejump]{save_data}}, \code{\link[basejump]{write_counts}}}

  \item{tximport}{\code{\link[tximport]{tximport}}}
}}

