% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{aggregateReplicates}
\alias{contrastName}
\alias{plotMA}
\alias{plotMeanAverage}
\alias{plotVolcano}
\alias{topTables}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump}{\code{\link[basejump]{aggregateReplicates}}}

  \item{DESeqAnalysis}{\code{\link[DESeqAnalysis]{contrastName}}, \code{\link[DESeqAnalysis]{plotMA}}, \code{\link[DESeqAnalysis]{plotMeanAverage}}, \code{\link[DESeqAnalysis]{plotVolcano}}, \code{\link[DESeqAnalysis]{topTables}}}
}}

