% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{sampleData}
\alias{plotCountsPerGene}
\alias{plotGenesDetected}
\alias{plotCountsPerFeature}
\alias{plotFeaturesDetected}
\alias{plotHeatmap}
\alias{plotQuantileHeatmap}
\alias{basenameSansExt}
\alias{export}
\alias{import}
\alias{interestingGroups}
\alias{interestingGroups<-}
\alias{loadData}
\alias{loadRemoteData}
\alias{pasteURL}
\alias{sampleData<-}
\alias{saveData}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{acidplots}{\code{\link[acidplots]{plotCountsPerGene}}, \code{\link[acidplots]{plotGenesDetected}}, \code{\link[acidplots]{plotCountsPerFeature}}, \code{\link[acidplots]{plotFeaturesDetected}}, \code{\link[acidplots]{plotHeatmap}}, \code{\link[acidplots]{plotQuantileHeatmap}}}

  \item{basejump}{\code{\link[basejump]{sampleData}}, \code{\link[basejump]{basenameSansExt}}, \code{\link[basejump]{export}}, \code{\link[basejump]{import}}, \code{\link[basejump]{interestingGroups}}, \code{\link[basejump]{loadData}}, \code{\link[basejump]{loadRemoteData}}, \code{\link[basejump]{pasteURL}}, \code{\link[basejump]{sampleData}}, \code{\link[basejump]{saveData}}}
}}

