% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{sampleData}
\alias{basenameSansExt}
\alias{export}
\alias{import}
\alias{interestingGroups}
\alias{interestingGroups<-}
\alias{loadData}
\alias{loadRemoteData}
\alias{pasteURL}
\alias{sampleData<-}
\alias{saveData}
\alias{plotHeatmap}
\alias{plotQuantileHeatmap}
\alias{plotGenesDetected}
\alias{theme_paperwhite}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump}{\code{\link[basejump]{sampleData}}, \code{\link[basejump]{basenameSansExt}}, \code{\link[basejump]{export}}, \code{\link[basejump]{import}}, \code{\link[basejump]{interestingGroups}}, \code{\link[basejump]{loadData}}, \code{\link[basejump]{loadRemoteData}}, \code{\link[basejump]{pasteURL}}, \code{\link[basejump]{sampleData}}, \code{\link[basejump]{saveData}}}

  \item{firestarter}{\code{\link[firestarter]{plotHeatmap}}, \code{\link[firestarter]{plotQuantileHeatmap}}}

  \item{minimalism}{\code{\link[minimalism]{plotGenesDetected}}, \code{\link[minimalism]{theme_paperwhite}}}
}}

