% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{plotFeaturesDetected}
\alias{plotHeatmap}
\alias{plotQuantileHeatmap}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{AcidPlots}{\code{\link[AcidPlots]{plotFeaturesDetected}}, \code{\link[AcidPlots]{plotHeatmap}}, \code{\link[AcidPlots]{plotQuantileHeatmap}}}
}}

