% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{loadRemoteData}
\alias{plotFeaturesDetected}
\alias{plotHeatmap}
\alias{plotQuantileHeatmap}
\alias{saveData}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{AcidGenerics}{\code{\link[AcidGenerics]{plotFeaturesDetected}}, \code{\link[AcidGenerics]{plotHeatmap}}, \code{\link[AcidGenerics]{plotQuantileHeatmap}}}

  \item{pipette}{\code{\link[pipette]{loadRemoteData}}, \code{\link[pipette]{saveData}}}
}}

