% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts-methods.R,
%   R/plot5Prime3PrimeBias-methods.R, R/plotCorrelationHeatmap-methods.R,
%   R/plotCountsPerBiotype-methods.R, R/plotCountsPerGene-methods.R,
%   R/plotDispEsts-methods.R, R/plotExonicMappingRate-methods.R,
%   R/plotGenderMarkers-methods.R, R/plotGene-methods.R,
%   R/plotGeneSaturation-methods.R, R/plotIntronicMappingRate-methods.R,
%   R/plotMappedReads-methods.R, R/plotMappingRate-methods.R,
%   R/plotMeanSD-methods.R, R/plotPCA-methods.R,
%   R/plotPCACovariates-methods.R, R/plotQC-methods.R,
%   R/plotRRNAMappingRate-methods.R, R/plotTotalReads-methods.R, R/reexports.R,
%   R/relativeLogExpression-methods.R, R/sampleData-methods.R, R/show-methods.R,
%   R/tmm-methods.R, R/updateObject-methods.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{Gene2Symbol}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump}{\code{\link[basejump]{Gene2Symbol}}, \code{\link[basejump]{sampleData}}}

  \item{BiocGenerics}{\code{\link[BiocGenerics]{counts}}, \code{\link[BiocGenerics]{counts<-}}, \code{\link[BiocGenerics]{plotDispEsts}}, \code{\link[BiocGenerics]{plotPCA}}, \code{\link[BiocGenerics]{updateObject}}}

  \item{bioverbs}{\code{\link[bioverbs]{plot5Prime3PrimeBias}}, \code{\link[bioverbs]{plotCorrelationHeatmap}}, \code{\link[bioverbs]{plotCountsPerBiotype}}, \code{\link[bioverbs]{plotCountsPerBroadClass}}, \code{\link[bioverbs]{plotCountsPerGene}}, \code{\link[bioverbs]{plotExonicMappingRate}}, \code{\link[bioverbs]{plotGenderMarkers}}, \code{\link[bioverbs]{plotGene}}, \code{\link[bioverbs]{plotGeneSaturation}}, \code{\link[bioverbs]{plotIntronicMappingRate}}, \code{\link[bioverbs]{plotMappedReads}}, \code{\link[bioverbs]{plotMappingRate}}, \code{\link[bioverbs]{plotMeanSD}}, \code{\link[bioverbs]{plotPCACovariates}}, \code{\link[bioverbs]{plotQC}}, \code{\link[bioverbs]{plotRRNAMappingRate}}, \code{\link[bioverbs]{plotTotalReads}}, \code{\link[bioverbs]{relativeLogExpression}}, \code{\link[bioverbs]{tmm}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{methods}{\code{\link[methods]{show}}}
}}

