% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relativeLogExpression-methods.R
\docType{methods}
\name{relativeLogExpression}
\alias{relativeLogExpression}
\alias{relativeLogExpression,matrix-method}
\alias{relativeLogExpression,SummarizedExperiment-method}
\title{Relative log expression}
\usage{
\S4method{relativeLogExpression}{matrix}(object)

\S4method{relativeLogExpression}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{matrix}.
}
\description{
Relative log expression
}
\examples{
data(bcb)
relativeLogExpression(bcb)
}
\references{
Anders and Huber (2010).
}
\seealso{
\code{edgeR::calcNormFactors()}.
}
\author{
Lorena Pantano, Michael Steinbaugh
}
