% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{res_contrast_name}
\alias{res_contrast_name}
\title{Get contrast name from \link{DESeqResults}}
\usage{
res_contrast_name(res)
}
\description{
Get contrast name from \link{DESeqResults}
}
\author{
Michael Steinbaugh
}
\keyword{internal}
