% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{res_small}
\alias{res_small}
\title{\code{DESeqResults} Example}
\format{An object of class \code{DESeqResults} with 500 rows and 5 columns.}
\usage{
res_small
}
\description{
\code{DESeqResults} Example
}
\seealso{
Other GSE65267: \code{\link{bcb_small}},
  \code{\link{dds_small}}, \code{\link{rld_small}}
}
\author{
Lorena Pantano, Michael Steinbaugh
}
\keyword{datasets}
