% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{res_small}
\alias{res_small}
\title{\code{DESeqResults} example}
\format{An object of class \code{DESeqResults} with 500 rows and 6 columns.}
\usage{
res_small
}
\description{
\code{DESeqResults} example
}
\seealso{
Other GSE65267: \code{\link{bcb_small}},
  \code{\link{dds_small}}
}
\author{
Lorena Pantano, Michael Steinbaugh
}
\concept{GSE65267}
\keyword{datasets}
