% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_tables.R
\name{res_tables}
\alias{res_tables}
\title{Differential expression results tables}
\usage{
res_tables(run, res, lfc = 0)
}
\arguments{
\item{run}{\code{bcbio-nextgen} run object}

\item{res}{\code{DESeqResults}}

\item{lfc}{Log fold change ratio (base 2) cutoff. Does not apply to
statistical hypothesis testing, only gene filtering in the results tables.
See \code{?DESeq2::results} for additional information about using
\code{lfcThreshold} and \code{altHypothesis} to set an alternative
hypothesis based on expected fold changes.}
}
\value{
Results list
}
\description{
Differential expression results tables
}
\examples{
\dontrun{
res_tables(run, res, lfc = 0.25)
}
}
\author{
Michael Steinbaugh
}
