% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{res_tables}
\alias{res_tables}
\title{Differential expression results tables}
\usage{
res_tables(bcb, res, lfc = 0, write = TRUE, dir = "results/de",
  print = TRUE)
}
\arguments{
\item{bcb}{\link{bcbioRnaDataSet}.}

\item{res}{\link{DESeqResults}.}

\item{lfc}{Log fold change ratio (base 2) cutoff. Does not apply to
statistical hypothesis testing, only gene filtering in the results tables.
See \code{\link[=results]{results()}} for additional information about using \code{lfcThreshold} and
\code{altHypothesis} to set an alternative hypothesis based on expected fold
changes.}

\item{write}{Write CSV files to disk.}

\item{dir}{Directory path where to write files.}

\item{print}{Print summary statistics and links to files.}
}
\value{
Results list.
}
\description{
Differential expression results tables
}
\examples{
\dontrun{
res_tables(run, res, lfc = 0.25)
}
}
\author{
Michael Steinbaugh
}
