% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_tables.R
\name{results_tables}
\alias{results_tables}
\title{Differential gene expression results tables}
\usage{
results_tables(results, alpha = 0.05, lfc = 0, organism, print = TRUE)
}
\arguments{
\item{results}{\code{DESeqResults} or \code{DESeqDataSet} object}

\item{alpha}{Alpha cutoff level}

\item{lfc}{Log fold change ratio (base 2) cutoff level}

\item{organism}{Organism}

\item{print}{Print summary}
}
\value{
List of DGE results tables
}
\description{
Differential gene expression results tables
}
\examples{
\dontrun{
results_tables(res, alpha = 0.05, lfc = 1, organism = "hsapiens")
}
}
\author{
Michael Steinbaugh
}
