% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_tables.R
\name{results_tables}
\alias{results_tables}
\title{Differential gene expression results tables}
\usage{
results_tables(bcbio, results, lfc = 0)
}
\arguments{
\item{bcbio}{bcbio list object}

\item{results}{\code{DESeqResults} object}

\item{lfc}{Log fold change ratio (base 2) cutoff level}
}
\value{
List of DGE results tables
}
\description{
Differential gene expression results tables
}
\examples{
\dontrun{
results_tables(bcbio, res, lfc = 1)
}
}
\author{
Michael Steinbaugh
}
