% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-resultsTables.R
\docType{methods}
\name{resultsTables}
\alias{resultsTables}
\alias{resultsTables}
\alias{resultsTables,DESeqResults-method}
\title{Differential Expression Results Tables}
\usage{
resultsTables(object, ...)

\S4method{resultsTables}{DESeqResults}(object, lfcThreshold = 0L,
  rowData = NULL, summary = TRUE, headerLevel = 2L, write = FALSE,
  dir = ".", dropboxDir = NULL, rdsToken = NULL, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{lfcThreshold}{Log fold change ratio (base 2) cutoff threshold.}

\item{rowData}{Join Ensembl gene annotations to the results. Apply gene
identifier to symbol mappings. A previously saved \code{data.frame} is
recommended. Alternatively if set \code{NULL}, then gene annotations will not be
added to the results.}

\item{summary}{Show summary statistics.}

\item{headerLevel}{R Markdown header level.}

\item{write}{Write CSV files to disk.}

\item{dir}{Local directory path.}

\item{dropboxDir}{Dropbox directory path where to archive the results tables
for permanent storage (e.g. Stem Cell Commons). When this option is
enabled, unique links per file are generated internally with the rdrop2
package.}

\item{rdsToken}{RDS file token to use for Dropbox authentication.}
}
\value{
Results \code{list}.
}
\description{
Differential Expression Results Tables
}
\note{
Log fold change cutoff threshold ("\code{lfcThreshold}") does not apply to
statistical hypothesis testing, only gene filtering in the results tables.
See \code{\link[DESeq2:results]{DESeq2::results()}} for additional information about using
\code{lfcThreshold} and \code{altHypothesis} to set an alternative hypothesis based
on expected fold changes.
}
\examples{
# DESeqResults ====
x <- resultsTables(
    object = res_small,
    lfcThreshold = 0.25,
    rowData = rowData(bcb_small),
    summary = TRUE,
    headerLevel = 2L,
    write = FALSE
)
names(x)
}
\seealso{
Other Differential Expression Functions: \code{\link{alphaSummary}},
  \code{\link{contrastName}}, \code{\link{plotDEGHeatmap}},
  \code{\link{plotDEGPCA}}, \code{\link{plotMeanAverage}},
  \code{\link{plotVolcano}}
}
\author{
Michael Steinbaugh
}
