% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/resultsTables-methods.R
\docType{methods}
\name{resultsTables}
\alias{resultsTables}
\alias{resultsTables,DESeqResults-method}
\alias{resultsTables,DESeqAnalysis-method}
\title{Differential Expression Results Tables}
\usage{
resultsTables(object, ...)

\S4method{resultsTables}{DESeqResults}(object)

\S4method{resultsTables}{DESeqAnalysis}(object, summary = TRUE,
  write = FALSE, headerLevel = 2L, dir = ".", dropboxDir = NULL,
  rdsToken = NULL)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{summary}{\code{boolean}. Show summary statistics.}

\item{write}{\code{boolean}. Write CSV files to disk.}

\item{headerLevel}{\code{scalar integer} (\code{1}-\code{7}). R Markdown header level.}

\item{dir}{\code{string}. Local directory path.}

\item{dropboxDir}{\code{string} or \code{NULL}. Dropbox directory path where to archive
the results tables for permanent storage (e.g. Stem Cell Commons). When
this option is enabled, unique links per file are generated internally with
the rdrop2 package. Note that local files are written to \code{\link[base:tempdir]{base::tempdir()}}
and the \code{dir} argument is ignored, if this is enabled.}

\item{rdsToken}{\code{string} or \code{NULL}. RDS file token to use for Dropbox
authentication. If set \code{NULL} and \code{dropboxDir} is defined, then an
interactive prompt will appear requesting authorization.}
}
\value{
\code{DESeqResultsTables}.
}
\description{
Differential Expression Results Tables
}
\note{
Log fold change cutoff threshold ("\code{lfcThreshold}") does not apply to
statistical hypothesis testing, only gene filtering in the results tables.
See \code{\link[DESeq2:results]{DESeq2::results()}} for additional information about using
\code{lfcThreshold} and \code{altHypothesis} to set an alternative hypothesis based
on expected fold changes.
}
\examples{
# DESeqResults ====
x <- resultsTables(res_small)
class(x)
slotNames(x)
}
\seealso{
Other Differential Expression Functions: \code{\link{alphaSummary}},
  \code{\link{contrastName}}, \code{\link{plotDEGHeatmap}},
  \code{\link{plotDEGPCA}}, \code{\link{plotMeanAverage}},
  \code{\link{plotVolcano}}
}
\author{
Michael Steinbaugh
}
\concept{Differential Expression Functions}
