% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/resultsTables-methods.R
\docType{methods}
\name{resultsTables}
\alias{resultsTables}
\alias{resultsTables,DESeqResults-method}
\alias{resultsTables,DESeqAnalysis-method}
\title{Differential Expression Results Tables}
\usage{
resultsTables(object, ...)

\S4method{resultsTables}{DESeqResults}(object)

\S4method{resultsTables}{DESeqAnalysis}(object, results,
  lfcShrink = TRUE, rowData = TRUE, counts = TRUE, write = FALSE,
  dir = ".", dropboxDir = NULL, rdsToken = NULL, markdown = FALSE,
  headerLevel = 2L)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{results}{\code{scalar}. Position or name of \code{DESeqResults}.}

\item{lfcShrink}{\code{boolean}. Use shrunken log2 fold change (LFC) values.}

\item{rowData}{\code{boolean}. Include gene annotations.}

\item{counts}{\code{boolean}. Include DESeq2 normalized counts.}

\item{write}{\code{boolean}. Write CSV files to disk.}

\item{dir}{\code{string}. Local directory path.}

\item{dropboxDir}{\code{string} or \code{NULL}. Dropbox directory path where to archive
the results tables for permanent storage (e.g. Stem Cell Commons). When
this option is enabled, unique links per file are generated internally with
the rdrop2 package. Note that local files are written to \code{\link[base:tempdir]{base::tempdir()}}
and the \code{dir} argument is ignored, if this is enabled.}

\item{rdsToken}{\code{string} or \code{NULL}. RDS file token to use for Dropbox
authentication. If set \code{NULL} and \code{dropboxDir} is defined, then an
interactive prompt will appear requesting authorization.}

\item{markdown}{\code{boolean}. Include Markdown links to file paths. Only applies
when \code{write = TRUE}.}

\item{headerLevel}{\code{scalar integer} (\code{1}-\code{7}). R Markdown header level.}
}
\value{
\code{DESeqResultsTables}.
}
\description{
Differential Expression Results Tables
}
\note{
Log fold change cutoff threshold ("\code{lfcThreshold}") does not apply to
statistical hypothesis testing, only gene filtering in the results tables.
See \code{\link[DESeq2:results]{DESeq2::results()}} for additional information about using
\code{lfcThreshold} and \code{altHypothesis} to set an alternative hypothesis based
on expected fold changes.
}
\examples{
# DESeqAnalysis ====
# This is the recommended default method.
x <- resultsTables(res)

# DESeqResults ====
res <- deseq_small@lfcShrink[[1L]]
x <- resultsTables(res)
class(x)
slotNames(x)
}
\seealso{
Other Differential Expression Functions: \code{\link{alphaSummary}},
  \code{\link{contrastName}}, \code{\link{plotDEGHeatmap}},
  \code{\link{plotDEGPCA}}, \code{\link{plotMeanAverage}},
  \code{\link{plotVolcano}}
}
\author{
Michael Steinbaugh
}
\concept{Differential Expression Functions}
