% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsTables.R
\name{resultsTables}
\alias{resultsTables}
\title{Differential Expression Results Tables}
\usage{
resultsTables(bcb, res, lfc = 0L, write = TRUE, dir = file.path("results",
  "de"))
}
\arguments{
\item{bcb}{\link{bcbioRNADataSet}.}

\item{res}{\link{DESeqResults}.}

\item{lfc}{Log fold change ratio (base 2) cutoff. Does not apply to
statistical hypothesis testing, only gene filtering in the results tables.
See \code{\link[=results]{results()}} for additional information about using \code{lfcThreshold} and
\code{altHypothesis} to set an alternative hypothesis based on expected fold
changes.}

\item{write}{Write CSV files to disk.}

\item{dir}{Directory path where to write files.}
}
\value{
Results list.
}
\description{
Differential Expression Results Tables
}
\examples{
\dontrun{
resultsTables(bcb, res, lfc = 0.25)
}
}
\author{
Michael Steinbaugh
}
