% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/resultsTables-methods.R
\docType{methods}
\name{resultsTables}
\alias{resultsTables}
\alias{resultsTables,DESeqResults-method}
\alias{resultsTables,DESeqAnalysis-method}
\title{Differential Expression Results Tables}
\usage{
resultsTables(object, ...)

\S4method{resultsTables}{DESeqResults}(object)

\S4method{resultsTables}{DESeqAnalysis}(object, results,
  lfcShrink = TRUE, rowData = TRUE, counts = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{results}{\code{scalar}. Position or name of \code{DESeqResults}.}

\item{lfcShrink}{\code{boolean}. Use shrunken log2 fold change (LFC) values.}

\item{rowData}{\code{boolean}. Include gene annotations.}

\item{counts}{\code{boolean}. Include DESeq2 normalized counts.}
}
\value{
\code{DESeqResultsTables}.
}
\description{
Differential Expression Results Tables
}
\note{
Log fold change cutoff threshold ("\code{lfcThreshold}") does not apply to
statistical hypothesis testing, only gene filtering in the results tables.
See \code{\link[DESeq2:results]{DESeq2::results()}} for additional information about using
\code{lfcThreshold} and \code{altHypothesis} to set an alternative hypothesis based
on expected fold changes.
}
\examples{
# DESeqAnalysis ====
# This is the recommended default method.
x <- resultsTables(deseq_small)
print(x)

# DESeqResults ====
res <- as(deseq_small, "DESeqResults")
x <- resultsTables(res)
print(x)
}
\seealso{
\itemize{
\item \code{\link[=markdown]{markdown()}}.
\item \code{\link[=write]{write()}}.
}

Other Differential Expression Functions: \code{\link{alphaSummary}},
  \code{\link{contrastName}}, \code{\link{plotDEGHeatmap}},
  \code{\link{plotDEGPCA}}, \code{\link{plotMeanAverage}},
  \code{\link{plotVolcano}}
}
\author{
Michael Steinbaugh
}
\concept{Differential Expression Functions}
