% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-resultsTables.R
\docType{methods}
\name{resultsTables}
\alias{resultsTables}
\alias{resultsTables}
\alias{resultsTables,DESeqResults-method}
\title{Differential Expression Results Tables}
\usage{
resultsTables(object, ...)

\S4method{resultsTables}{DESeqResults}(object, lfc = 0L, write = TRUE,
  headerLevel = 3L, dir = file.path("results", "differential_expression"),
  genomeBuild = NULL)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{lfc}{Log fold change ratio (base 2) cutoff. Does not apply to
statistical hypothesis testing, only gene filtering in the results tables.
See \code{\link[=results]{results()}} for additional information about using \code{lfcThreshold} and
\code{altHypothesis} to set an alternative hypothesis based on expected fold
changes.}

\item{write}{Write CSV files to disk.}

\item{headerLevel}{Markdown header level.}

\item{dir}{Directory path where to write files.}

\item{genomeBuild}{\emph{Optional}. Override automatic genome detection. By
default the function matches the genome annotations based on the first
Ensembl gene identifier row in the object. If a custom FASTA spike-in
is provided, then this may need to be manually set.}

\item{res}{\link{DESeqResults}.}
}
\value{
Results list.
}
\description{
Differential Expression Results Tables
}
\examples{
data(res)
resTbl <- resultsTables(res, lfc = 0.25, write = FALSE)
class(resTbl)
names(resTbl)
}
\author{
Michael Steinbaugh
}
