% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-resultsTables.R
\docType{methods}
\name{resultsTables}
\alias{resultsTables}
\alias{resultsTables}
\alias{resultsTables,DESeqResults-method}
\title{Differential Expression Results Tables}
\usage{
resultsTables(object, ...)

\S4method{resultsTables}{DESeqResults}(object, lfc = 0L, annotable = TRUE,
  summary = TRUE, headerLevel = 2L, write = FALSE, dir = getwd(),
  quiet = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{lfc}{Log fold change ratio (base 2) cutoff. Does not apply to
statistical hypothesis testing, only gene filtering in the results tables.
See \code{\link[=results]{results()}} for additional information about using \code{lfcThreshold} and
\code{altHypothesis} to set an alternative hypothesis based on expected fold
changes.}

\item{annotable}{Join Ensembl gene annotations to the results. Apply gene
identifier to symbol mappings. If \code{TRUE}, the function will attempt to
detect the organism from the gene identifiers (rownames; "ensgene" column)
and automatically obtain the latest annotations from Ensembl using
\code{\link[=annotable]{annotable()}}. If set \code{FALSE}/\code{NULL}, then gene annotations will
not be added to the results. This is useful when working with a poorly
annotated genome. Alternatively, a previously saved annotable \link{data.frame}
can be passed in.}

\item{summary}{Show summary statistics.}

\item{headerLevel}{R Markdown header level. Applies only when
\code{summary = TRUE}.}

\item{write}{Write CSV files to disk.}

\item{dir}{Directory path where to write files.}

\item{quiet}{If \code{TRUE}, suppress any status messages and/or progress bars.}
}
\value{
Results \link{list}.
}
\description{
Differential Expression Results Tables
}
\examples{
load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioRNASeq"))
load(system.file(
    file.path("extdata", "res.rda"),
    package = "bcbioRNASeq"))

annotable <- annotable(bcb)

resTbl <- resultsTables(
    res,
    lfc = 0.25,
    annotable = annotable,
    summary = TRUE,
    headerLevel = 2L,
    write = FALSE)
names(resTbl)
}
\author{
Michael Steinbaugh
}
