% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-resultsTables.R
\docType{methods}
\name{resultsTables}
\alias{resultsTables}
\alias{resultsTables}
\alias{resultsTables,DESeqResults,DESeqDataSet-method}
\title{Differential Expression Results Tables}
\usage{
resultsTables(results, counts, ...)

\S4method{resultsTables}{DESeqResults,DESeqDataSet}(results, counts,
  lfcThreshold = 0L, summary = TRUE, write = FALSE, headerLevel = 2L,
  dir = ".", dropboxDir = NULL, rdsToken = NULL)
}
\arguments{
\item{results}{\code{DESeqResults} object.}

\item{counts}{\code{DESeqDataSet} containing the counts that were used to generate
the \code{DESeqResults}.}

\item{...}{Additional arguments.}

\item{lfcThreshold}{Log fold change ratio (base 2) cutoff threshold.}

\item{summary}{Show summary statistics.}

\item{write}{Write CSV files to disk.}

\item{headerLevel}{R Markdown header level.}

\item{dir}{Local directory path.}

\item{dropboxDir}{Dropbox directory path where to archive the results tables
for permanent storage (e.g. Stem Cell Commons). When this option is
enabled, unique links per file are generated internally with the rdrop2
package.}

\item{rdsToken}{RDS file token to use for Dropbox authentication.}
}
\value{
\code{list} containing modified \code{DESeqResults} return, including
additional gene-level metadata and normalized counts.
}
\description{
Differential Expression Results Tables
}
\note{
Log fold change cutoff threshold ("\code{lfcThreshold}") does not apply to
statistical hypothesis testing, only gene filtering in the results tables.
See \code{\link[DESeq2:results]{DESeq2::results()}} for additional information about using
\code{lfcThreshold} and \code{altHypothesis} to set an alternative hypothesis based
on expected fold changes.
}
\examples{
# DESeqResults, DESeqDataSet ====
x <- resultsTables(
    results = res_small,
    counts = dds_small,
    lfcThreshold = 0.25,
    summary = TRUE,
    write = FALSE
)
names(x)
}
\seealso{
Other Differential Expression Functions: \code{\link{alphaSummary}},
  \code{\link{contrastName}}, \code{\link{plotDEGHeatmap}},
  \code{\link{plotDEGPCA}}, \code{\link{plotMeanAverage}},
  \code{\link{plotVolcano}}
}
\author{
Michael Steinbaugh
}
