% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-resultsTables.R
\docType{methods}
\name{resultsTables}
\alias{resultsTables}
\alias{resultsTables}
\alias{resultsTables,DESeqResults-method}
\title{Differential Expression Results Tables}
\usage{
resultsTables(object, ...)

\S4method{resultsTables}{DESeqResults}(object, lfc = 0L, annotable = TRUE,
  summary = TRUE, headerLevel = 2L, write = FALSE, dir = getwd(),
  dropboxDir = NULL, rdsToken = NULL, quiet = FALSE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}

\item{lfc}{Log fold change ratio (base 2) cutoff. Does not apply to
statistical hypothesis testing, only gene filtering in the results tables.
See \code{\link[=results]{results()}} for additional information about using \code{lfcThreshold} and
\code{altHypothesis} to set an alternative hypothesis based on expected fold
changes.}

\item{annotable}{Join Ensembl gene annotations to the results. Apply gene
identifier to symbol mappings. If \code{TRUE}, the function will attempt to
detect the organism from the gene identifiers (rownames; "ensgene" column)
and automatically obtain the latest annotations from Ensembl using
\code{\link[=annotable]{annotable()}}. If set \code{FALSE}/\code{NULL}, then gene annotations will
not be added to the results. This is useful when working with a poorly
annotated genome. Alternatively, a previously saved annotable \link{data.frame}
can be passed in.}

\item{summary}{Show summary statistics.}

\item{headerLevel}{R Markdown header level. Applies only when
\code{summary = TRUE}.}

\item{write}{Write CSV files to disk.}

\item{dir}{Local directory path where to write the results tables.}

\item{dropboxDir}{Dropbox directory path where to archive the results tables
for permanent storage (e.g. Stem Cell Commons). When this option is
enabled, unique links per file are generated internally with the rdrop2
package.}

\item{rdsToken}{RDS file token to use for Dropbox authentication.}

\item{quiet}{If \code{TRUE}, suppress any status messages and/or progress bars.}
}
\value{
Results \link{list}.
}
\description{
Differential Expression Results Tables
}
\examples{
load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioRNASeq"))
load(system.file(
    file.path("extdata", "res.rda"),
    package = "bcbioRNASeq"))

# Can use a stashed annotable for improved speed
annotable <- annotable(bcb)

resTbl <- resultsTables(
    res,
    lfc = 0.25,
    annotable = annotable,
    summary = TRUE,
    headerLevel = 2L,
    write = TRUE,
    dir = "resultsTables",
    dropboxDir = file.path("bcbioRNASeq_examples", "resultsTables"),
    rdsToken = system.file("token.rds", package = "bcbioRNASeq")
)
names(resTbl)
dir("resultsTables")
unlink("resultsTables", recursive = TRUE)
}
\author{
Michael Steinbaugh
}
