% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rld_small}
\alias{rld_small}
\title{Minimal \code{DESeqTransform} Example}
\format{An object of class \code{DESeqTransform} with 500 rows and 6 columns.}
\usage{
rld_small
}
\description{
Minimal \code{DESeqTransform} Example
}
\seealso{
Other GSE65267: \code{\link{bcb_small}},
  \code{\link{dds_small}}, \code{\link{res_small}}
}
\author{
Michael Steinbaugh, Lorena Pantano
}
\keyword{internal}
