% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-rowData.R
\name{rowData}
\alias{rowData}
\title{Row Data}
\value{
\code{DataFrame} or \code{data.frame}.
}
\description{
Row Data
}
\examples{
load(system.file("extdata/bcb.rda", package = "bcbioRNASeq"))

# Return as data.frame
rowData(bcb, return = "data.frame")
}
\seealso{
\code{help("rowData", "SummarizedExperiment")}
}
