% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-rowData.R
\docType{methods}
\name{rowData}
\alias{rowData}
\alias{rowData,bcbioRNASeq-method}
\title{Row Data}
\usage{
\S4method{rowData}{bcbioRNASeq}(x, return = c("data.frame", "DataFrame"))
}
\arguments{
\item{x}{Object.}

\item{return}{Object class to return. Uses \code{\link[=match.arg]{match.arg()}} internally and picks
the first item in the vector by default.}
}
\value{
Data describing the rows of the object.
}
\description{
Row Data
}
\examples{
rowData(bcb_small) \%>\% glimpse()
}
\seealso{
Other Data Functions: \code{\link{colData}},
  \code{\link{counts}}, \code{\link{selectSamples}},
  \code{\link{subset}}, \code{\link{tmm}},
  \code{\link{tpm}}
}
\author{
Michael Steinbaugh
}
