% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-rowData.R
\docType{methods}
\name{rowData}
\alias{rowData}
\alias{rowData,bcbioRNASeq-method}
\title{Row Data}
\usage{
\S4method{rowData}{bcbioRNASeq}(x, return = c("data.frame", "DataFrame"))
}
\arguments{
\item{return}{Return as "\code{data.frame}" or "\code{DataFrame}".}
}
\value{
Data describing the rows of the object.
}
\description{
Row Data
}
\examples{
rowData(bcb_small) \%>\% glimpse()
}
\seealso{
Other Data Functions: \code{\link{colData}},
  \code{\link{counts}}, \code{\link{selectSamples}},
  \code{\link{tmm}}, \code{\link{tpm}}
}
\author{
Michael Steinbaugh
}
