% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_david.R
\name{run_david}
\alias{run_david}
\title{Run DAVID analysis}
\usage{
run_david(res_tables, direction = "both", count = 3, save = TRUE)
}
\arguments{
\item{res_tables}{\code{res_tables()} list object}

\item{direction}{Gene direction: up, down, or both}

\item{count}{Minimum hit count}

\item{save}{Save files to disk (\code{TRUE/FALSE})}
}
\value{
List of \code{RDAVIDWebService} report objects
}
\description{
Wrapper function that performs gene set enrichment analysis (GSEA) with the
\code{RDAVIDWebService} package, using simplified input options.
}
\author{
Michael Steinbaugh
}
