% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-sample_dirs.R
\docType{methods}
\name{sample_dirs}
\alias{sample_dirs}
\alias{sample_dirs,bcbioRNADataSet-method}
\title{Sample directories of \link{bcbioRNADataSet} object.}
\usage{
\S4method{sample_dirs}{bcbioRNADataSet}(object)
}
\arguments{
\item{object}{\link{bcbioRNADataSet} object.}
}
\value{
Folders where samples are kept.
}
\description{
This method will be used to access folders where sample information is kept.
}
