% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/all_methods.R
\docType{methods}
\name{bcbcols}
\alias{bcbcols}
\alias{bcbcols,bcbioRnaDataSet}
\alias{sample_dirs,bcbioRnaDataSet-method}
\title{Sample directories of \link{bcbioRnaDataSet} object.}
\usage{
sample_dirs(object)

\S4method{sample_dirs}{bcbioRnaDataSet}(object)
}
\arguments{
\item{object}{\link{bcbioRnaDataSet} object.}
}
\value{
Folders where samples are kept.
}
\description{
This method will be used to access folders where sample information is kept.
}
