% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/internal-sample_dirs.R,
%   R/methods-sample_dirs.R
\docType{methods}
\name{sample_dirs}
\alias{sample_dirs}
\alias{.sample_dirs}
\alias{sample_dirs,bcbioRNADataSet-method}
\title{Detect Sample Directories}
\usage{
sample_dirs(object)

.sample_dirs(upload_dir)

\S4method{sample_dirs}{bcbioRNADataSet}(object)
}
\arguments{
\item{object}{Object.}

\item{upload_dir}{Upload directory.}
}
\value{
Named character vector containing sample directory paths. Function
will \code{\link[=stop]{stop()}} if no complete sample directories match.

Named character vector containing sample directory paths.
}
\description{
Detect Sample Directories

This method will be used to access folders where sample information is kept.
}
\examples{
data(bcb)
sample_dirs(bcb)
}
\author{
Michael Steinbaugh

Michael Steinbaugh
}
\keyword{internal}
