% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleData-methods.R
\docType{methods}
\name{sampleData}
\alias{sampleData}
\alias{sampleData,bcbioRNASeq-method}
\title{Sample Data}
\usage{
\S4method{sampleData}{bcbioRNASeq}(object, clean = FALSE)
}
\arguments{
\item{object}{Object.}

\item{clean}{\code{logical}. Only return \code{factor} columns not defined in
\link[bcbioBase:metadataBlacklist]{bcbioBase::metadataBlacklist}.}
}
\value{
\code{DataFrame}.
}
\description{
Return the sample metadata. Columns are always sanitized to factor.
}
\note{
This is a complement to the standard \code{\link[=colData]{colData()}} function, but improves
support for accessing sample metadata for datasets where multiple items in
the columns map to a single sample (e.g. cells for a single-cell RNA-seq
experiment).
}
\examples{
# SummarizedExperiment ====
sampleData(bcb_small, clean = TRUE) \%>\% glimpse()
sampleData(bcb_small, clean = FALSE) \%>\% glimpse()

# Assignment support
x <- bcb_small
sampleData(x)[["test"]] <- seq_len(ncol(x))
# `test` column should be now defined
"test" \%in\% colnames(sampleData(x))
}
\seealso{
\link[bcbioBase:metadataBlacklist]{bcbioBase::metadataBlacklist}.

Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{counts}}, \code{\link{metrics}},
  \code{\link{tmm}}, \code{\link{tpm}}
}
\author{
Michael Steinbaugh
}
