% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleData-methods.R
\docType{methods}
\name{sampleData}
\alias{sampleData}
\alias{sampleData,bcbioRNASeq-method}
\title{Sample data}
\usage{
\S4method{sampleData}{bcbioRNASeq}(object, clean = FALSE)
}
\arguments{
\item{object}{Object.}

\item{clean}{\code{logical(1)}. Only return \code{factor} columns not defined in
\link[bcbioBase:metadataBlacklist]{bcbioBase::metadataBlacklist}. This removes metrics columns used for
quality control analysis, which are often not informative as sample
metadata.}
}
\value{
\code{DataFrame}.
}
\description{
Metadata that describes the samples.
}
\details{
This is a complement to the standard \code{colData()} function, but improves
support for accessing sample metadata for datasets where multiple items in
the columns map to a single sample (e.g. cells for a single-cell RNA-seq
experiment).
}
\examples{
data(bcb)
sampleData(bcb, clean = TRUE) \%>\% lapply(head)

## Assignment support.
x <- bcb
sampleData(x)[["test"]] <- seq_len(ncol(x))
## `test` column should be now defined.
"test" \%in\% colnames(sampleData(x))
}
\author{
Michael Steinbaugh
}
