% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/internal-sampleDirs.R,
%   R/methods-sampleDirs.R
\docType{methods}
\name{sampleDirs}
\alias{sampleDirs}
\alias{.sampleDirs}
\alias{sampleDirs}
\alias{sampleDirs,bcbioRNADataSet-method}
\title{Detect Sample Directories}
\usage{
sampleDirs(object)

.sampleDirs(uploadDir)

\S4method{sampleDirs}{bcbioRNADataSet}(object)
}
\arguments{
\item{object}{Object.}

\item{uploadDir}{Upload directory.}
}
\value{
Named character vector containing sample directory paths. Function
will \code{\link[=stop]{stop()}} if no complete sample directories match.

Named character vector containing sample directory paths.
}
\description{
Detect Sample Directories

This method will be used to access folders where sample information is kept.
}
\examples{
data(bcb)
sampleDirs(bcb)
}
\author{
Michael Steinbaugh

Michael Steinbaugh
}
\keyword{internal}
