% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-sampleDirs.R
\docType{methods}
\name{sampleDirs}
\alias{sampleDirs}
\alias{sampleDirs,bcbioRNADataSet-method}
\title{Sample Directories}
\usage{
\S4method{sampleDirs}{bcbioRNADataSet}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Named character vector containing sample directory paths.
}
\description{
This method will be used to access folders where sample information is kept.
}
\examples{
data(bcb)
sampleDirs(bcb) \%>\% basename
}
