% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-sampleMetadata.R
\docType{methods}
\name{sampleMetadata}
\alias{sampleMetadata}
\alias{sampleMetadata,bcbioRNASeq-method}
\alias{sampleMetadata,DESeqDataSet-method}
\alias{sampleMetadata,DESeqTransform-method}
\title{Sample Metadata}
\usage{
\S4method{sampleMetadata}{bcbioRNASeq}(object, ...)

\S4method{sampleMetadata}{DESeqDataSet}(object, ...)

\S4method{sampleMetadata}{DESeqTransform}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\link{data.frame}.
}
\description{
Returns the sample metadata.
}
\examples{
load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioRNASeq"))

# bcbioRNASeq
sampleMetadata(bcb) \%>\% glimpse()

# DESeqDataSet
dds <- bcbio(bcb, "DESeqDataSet")
sampleMetadata(dds) \%>\% glimpse()

# DESeqTransform
rld <- assays(bcb)[["rlog"]]
sampleMetadata(rld) \%>\% glimpse()
}
