% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-sampleMetadata.R
\docType{methods}
\name{sampleMetadata}
\alias{sampleMetadata}
\alias{sampleMetadata,bcbioRNASeq-method}
\alias{sampleMetadata,DESeqDataSet-method}
\alias{sampleMetadata,DESeqTransform-method}
\title{Sample Metadata}
\usage{
\S4method{sampleMetadata}{bcbioRNASeq}(object, ...)

\S4method{sampleMetadata}{DESeqDataSet}(object, ...)

\S4method{sampleMetadata}{DESeqTransform}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
\link{data.frame}.
}
\description{
Returns the sample metadata.
}
\examples{
# bcbioRNASeq
bcb <- examples[["bcb"]]
sampleMetadata(bcb)

# DESeqDataSet
dds <- examples[["dds"]]
sampleMetadata(dds)

# DESeqTransform
rld <- examples[["rld"]]
sampleMetadata(rld)
}
