% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-selectSamples.R
\docType{methods}
\name{selectSamples}
\alias{selectSamples}
\alias{selectSamples,bcbioRNASeq-method}
\title{Select Samples}
\usage{
\S4method{selectSamples}{bcbioRNASeq}(object, ..., transform = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{\emph{Additional arguments (for the S4 generic definition).}}
}
\value{
\link{bcbioRNASeq}.
}
\description{
Select Samples
}
\examples{
load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioRNASeq"))
}
\author{
Michael Steinbaugh
}
