% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-selectSamples.R
\docType{methods}
\name{selectSamples}
\alias{selectSamples}
\alias{selectSamples,bcbioRNASeq-method}
\alias{selectSamples,DESeqDataSet-method}
\title{Select Samples}
\usage{
\S4method{selectSamples}{bcbioRNASeq}(object, ..., transform = TRUE)

\S4method{selectSamples}{DESeqDataSet}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{transform}{Apply CPU-intensive DESeq2 transformations. This can
take a long time for large datasets.}
}
\value{
\link{bcbioRNASeq}.
}
\description{
Select Samples
}
\examples{
load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioRNASeq"))
load(system.file(
    file.path("extdata", "dds.rda"),
    package = "bcbioRNASeq"))

# bcbioRNASeq
selectSamples(bcb, group = "ko")

# DESeqDataSet
selectSamples(dds, group = c("ctrl", "ko"))
}
\author{
Michael Steinbaugh
}
