% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-selectSamples.R
\docType{methods}
\name{selectSamples}
\alias{selectSamples}
\alias{selectSamples,bcbioRNASeq-method}
\alias{selectSamples,DESeqDataSet-method}
\title{Select Samples}
\usage{
\S4method{selectSamples}{bcbioRNASeq}(object, ..., transform = TRUE)

\S4method{selectSamples}{DESeqDataSet}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{transform}{Apply CPU-intensive DESeq2 transformations. This can
take a long time for large datasets.}
}
\value{
\code{bcbioRNASeq}.
}
\description{
Select Samples
}
\examples{
load(system.file("extdata/bcb.rda", package = "bcbioRNASeq"))
load(system.file("extdata/dds.rda", package = "bcbioRNASeq"))

# bcbioRNASeq ====
subset <- selectSamples(bcb, group = "ko")
print(subset)
metrics(subset) \%>\% glimpse()

# DESeqDataSet ====
subset <- selectSamples(dds, group = "ko")
print(subset)
}
\author{
Michael Steinbaugh
}
