% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\name{show}
\alias{show}
\alias{show,bcbioRNASeq-method}
\title{Show an object}
\usage{
\S4method{show}{bcbioRNASeq}(object)
}
\arguments{
\item{object}{Any R object}
}
\value{
\code{show} returns an invisible \code{NULL}.
}
\description{
Show an object
}
\note{
Updated 2020-09-15.
}
\examples{
data(bcb)
show(bcb)
}
\author{
Michael Steinbuagh
}
