% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/show-methods.R
\name{show}
\alias{show}
\alias{show,bcbioRNASeq-method}
\title{Show an object}
\usage{
show(object, ...)

\S4method{show}{bcbioRNASeq}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Console output.
}
\description{
Show an object
}
\note{
Updated 2020-12-22.
}
\examples{
data(bcb)
show(bcb)
}
\author{
Michael Steinbaugh
}
