% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/slotAlignedCounts-methods.R
\docType{methods}
\name{slotAlignedCounts}
\alias{slotAlignedCounts}
\alias{slotAlignedCounts,bcbioRNASeq-method}
\title{Slot aligned counts into object}
\usage{
slotAlignedCounts(object, ...)

\S4method{slotAlignedCounts}{bcbioRNASeq}(object)
}
\description{
This function loads aligned counts (e.g. STAR, HISAT2) from the bcbio final
output directory into the bcbioRNASeq object, so we can visually inspect
correlations with the primary pseudoaligned counts.
}
\examples{
uploadDir <- system.file("extdata/bcbio", package = "bcbioRNASeq")
## Fast mode skips import of aligned counts.
bcb <- bcbioRNASeq(uploadDir, fast = TRUE)
"aligned" \%in\% assayNames(bcb)
bcb <- slotAlignedCounts(bcb)
"aligned" \%in\% assayNames(bcb)
}
