% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-smallrna.R
\name{plot_size_distribution}
\alias{plot_size_distribution}
\alias{plot_mirna_counts}
\alias{plot_srna_clusters}
\title{Small RNA-seq quality control plots}
\usage{
plot_size_distribution(run)

plot_mirna_counts(run)

plot_srna_clusters(run)
}
\arguments{
\item{run}{bcbio-nextgen run.}
}
\value{
ggplot object.
}
\description{
Plot size distribution of small RNA-seq data

Plot of total miRNA counts

Clustering small RNA samples
}
\author{
Lorena Pantano

Michael Steinbaugh
}
