% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallrna_qc_plots.R
\name{plot_size_distribution}
\alias{plot_size_distribution}
\alias{plot_mirna_counts}
\title{Small RNA-seq quality control plots.}
\usage{
plot_size_distribution(run)

plot_mirna_counts(run)
}
\arguments{
\item{run}{bcbio-nextgen run}
}
\value{
ggplot figure.
}
\description{
Plot size distribution of small RNA-seq data

Plot of total miRNA counts
}
\author{
Lorena Pantano
}
