% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-summarized_experiment.R
\name{.summarized_experiment}
\alias{.summarized_experiment}
\title{Package RNA-Seq counts into \link{SummarizedExperiment}}
\usage{
.summarized_experiment(txi, col_data, row_data, metadata = NULL)
}
\arguments{
\item{txi}{\link{tximport} list.}

\item{col_data}{Sample metadata.}

\item{row_data}{\href{http://www.ensembl.org/}{Ensembl} gene annotations.}

\item{metadata}{Custom metadata.}
}
\value{
\link{SummarizedExperiment}.
}
\description{
Contains counts generated by \code{\link[=tximport]{tximport()}}, along with
\href{https://bcbio-nextgen.readthedocs.io/}{bcbio-nextgen} run metadata.
}
\author{
Michael Steinbaugh
}
\keyword{internal}
