% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-SummarizedExperiment.R
\name{.SummarizedExperiment}
\alias{.SummarizedExperiment}
\title{Package RNA-Seq counts into \link{SummarizedExperiment}}
\usage{
.SummarizedExperiment(tximport, colData, rowData, metadata = NULL)
}
\arguments{
\item{tximport}{\link{tximport} list.}

\item{colData}{Sample metadata.}

\item{rowData}{\href{http://www.ensembl.org/}{Ensembl} gene annotations.}

\item{metadata}{Custom metadata.}
}
\value{
\link{SummarizedExperiment}.
}
\description{
Contains counts generated by \code{\link[=tximport]{tximport()}}, along with
\href{https://bcbio-nextgen.readthedocs.io/}{bcbio-nextgen} run metadata.
}
\author{
Michael Steinbaugh
}
\keyword{internal}
