% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-SummarizedExperiment.R
\docType{methods}
\name{SummarizedExperiment}
\alias{SummarizedExperiment}
\alias{assays<-,bcbioRNASeq,list-method}
\alias{assays<-,bcbioRNASeq,SimpleList-method}
\alias{metadata<-,bcbioRNASeq-method}
\title{Extend SummarizedExperiment Assignment Methods}
\usage{
\S4method{assays}{bcbioRNASeq,list}(x, ..., withDimnames = TRUE) <- value

\S4method{assays}{bcbioRNASeq,SimpleList}(x, ...,
  withDimnames = TRUE) <- value

\S4method{metadata}{bcbioRNASeq}(x) <- value
}
\description{
Extend SummarizedExperiment Assignment Methods
}
\examples{
# bcbioRNASeq ====
metadata(bcb_small)[["stash"]] <- "test"
metadata(bcb_small)[["stash"]]
}
\author{
Michael Steinbaugh
}
\keyword{internal}
