% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tmm-methods.R
\docType{methods}
\name{tmm}
\alias{tmm}
\alias{tmm,matrix-method}
\alias{tmm,SummarizedExperiment-method}
\alias{tmm,bcbioRNASeq-method}
\title{\strong{T}rimmed \strong{M}ean of \strong{M}-Values (TMM) Normalization}
\usage{
tmm(object, ...)

\S4method{tmm}{matrix}(object)

\S4method{tmm}{SummarizedExperiment}(object, assay = 1L)

\S4method{tmm}{bcbioRNASeq}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{assay}{\code{scalar}. Name or index of counts matrix slotted in \code{\link[=assays]{assays()}}.
When passing in a \code{string}, the name must be defined in \code{\link[=assayNames]{assayNames()}}.}
}
\value{
\code{matrix}.
}
\description{
TMM normalization is recommended for RNA-seq data generally when the majority
of genes are not differentially expressed. We use this as a quality control
tool when plotting counts per gene.
}
\note{
Only gene-level counts are supported.
}
\examples{
x <- tmm(bcb_small)
summary(x)
}
\seealso{
Other Data Functions: \code{\link{counts}},
  \code{\link{sampleData}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
