% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tmm-methods.R
\docType{methods}
\name{tmm}
\alias{tmm}
\alias{tmm,matrix-method}
\alias{tmm,SummarizedExperiment-method}
\title{\strong{T}rimmed \strong{M}ean of \strong{M}-Values (TMM) Normalization}
\usage{
tmm(object, ...)

\S4method{tmm}{matrix}(object)

\S4method{tmm}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{matrix}.
}
\description{
TMM normalization is recommended for RNA-seq data generally when the majority
of genes are not differentially expressed.
}
\note{
Only recommended for gene-level counts.
}
\examples{
data(bcb_small)
x <- tmm(bcb_small)
summary(x)
}
\references{
Robinson and Oshlack (2010).
}
\seealso{
\itemize{
\item \link[edgeR:calcNormFactors]{edgeR::calcNormFactors}.
\item \link[edgeR:cpm]{edgeR::cpm}.
}
}
\author{
Michael Steinbaugh
}
