% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tmm-methods.R
\docType{methods}
\name{tmm}
\alias{tmm}
\alias{tmm,bcbioRNASeq-method}
\alias{tmm,DESeqDataSet-method}
\alias{tmm,matrix-method}
\title{\strong{T}rimmed \strong{M}ean of \strong{M}-Values (TMM) Normalization}
\usage{
tmm(object)

\S4method{tmm}{bcbioRNASeq}(object)

\S4method{tmm}{DESeqDataSet}(object)

\S4method{tmm}{matrix}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{matrix}.
}
\description{
TMM normalization is recommended for RNA-seq data generally when the majority
of genes are not differentially expressed. We use this as a quality control
tool when plotting counts per gene.
}
\examples{
# bcbioRNASeq ====
tmm(bcb_small) \%>\% summary()

# DESeqDataSet ====
tmm(dds_small) \%>\% summary()

# matrix ====
counts(bcb_small) \%>\% tmm() \%>\% summary()
}
\seealso{
Other Data Functions: \code{\link{aggregateReplicates}},
  \code{\link{counts}}, \code{\link{metrics}},
  \code{\link{sampleData}}, \code{\link{tpm}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
