% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-tmm.R
\docType{methods}
\name{tmm}
\alias{tmm}
\alias{tmm,bcbioRNADataSet-method}
\alias{tmm,matrix-method}
\alias{tmm,DESeqDataSet-method}
\title{Trimmed mean of M-values (TMM) normalization}
\usage{
\S4method{tmm}{bcbioRNADataSet}(object)

\S4method{tmm}{matrix}(object)

\S4method{tmm}{DESeqDataSet}(object)
}
\arguments{
\item{object}{Object containing counts.}
}
\description{
TMM normalization is recommended for RNA-seq data generally when the majority
of genes are not differentially expressed. We use this as a quality control
tool when plotting counts per gene.
}
\author{
Michael Steinbaugh
}
