% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmm-methods.R
\docType{methods}
\name{tmm}
\alias{tmm}
\alias{tmm,bcbioRNASeq-method}
\alias{tmm,DESeqDataSet-method}
\alias{tmm,matrix-method}
\title{Trimmed mean of M-values}
\usage{
tmm(object, ...)

\S4method{tmm}{bcbioRNASeq}(object)

\S4method{tmm}{DESeqDataSet}(object)

\S4method{tmm}{matrix}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{matrix}.
}
\description{
TMM normalization is recommended for RNA-seq data generally when the majority
of genes are not differentially expressed.
}
\examples{
# bcbioRNASeq ====
tmm(bcb_small) \%>\% summary()

# DESeqDataSet ====
tmm(dds_small) \%>\% summary()

# matrix ====
counts(bcb_small) \%>\% tmm() \%>\% summary()
}
\references{
Robinson and Oshlack (2010).
}
\seealso{
Other Data Functions: \code{\link{counts}},
  \code{\link{metrics}}
}
\author{
Michael Steinbaugh
}
\concept{Data Functions}
