% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmm-methods.R
\docType{methods}
\name{tmm}
\alias{tmm}
\alias{tmm,matrix-method}
\alias{tmm,SummarizedExperiment-method}
\title{Trimmed mean of M-values}
\usage{
tmm(object, ...)

\S4method{tmm}{matrix}(object)

\S4method{tmm}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{matrix}.
}
\description{
TMM normalization is recommended for RNA-seq data generally when the majority
of genes are not differentially expressed.
}
\note{
Updated 2019-08-07.
}
\examples{
data(bcb)
x <- tmm(bcb)
summary(x)
}
\references{
Robinson and Oshlack (2010).
}
\seealso{
\itemize{
\item \code{edgeR::calcNormFactors()}.
\item \code{edgeR::cpm()}.
}
}
\author{
Michael Steinbaugh
}
