% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_table.R
\name{top_table}
\alias{top_table}
\title{Print top table}
\usage{
top_table(df, caption = NULL)
}
\arguments{
\item{df}{Data frame}

\item{caption}{Caption}
}
\value{
Data frame or kable, depending on the call
}
\description{
Print top table
}
\author{
Michael Steinbaugh
}
