% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_tables.R
\name{top_tables}
\alias{top_tables}
\title{Top tables of differential expression results}
\usage{
top_tables(res_tables)
}
\arguments{
\item{res_tables}{Results tables generated by \code{res_tables()}}
}
\value{
Data frame or kable, depending on the call
}
\description{
Top tables of differential expression results
}
\author{
Michael Steinbaugh
}
