% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_tables.R
\name{top_tables}
\alias{top_tables}
\title{Top tables of differential expression results}
\usage{
top_tables(res_tbl)
}
\arguments{
\item{res_tbl}{Results tables generated by \code{\link{res_tables}}.}
}
\value{
Top tables list.
}
\description{
Top tables of differential expression results
}
\author{
Michael Steinbaugh
}
