% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{top_tables}
\alias{top_tables}
\title{Top tables of differential expression results}
\usage{
top_tables(res_tbl, n = 50L, coding = FALSE)
}
\arguments{
\item{res_tbl}{DESeq2 results tables list generated by \code{\link[=results_tables]{results_tables()}}.}

\item{n}{Number genes to report.}

\item{coding}{Whether to only return coding genes.}
}
\value{
Top table kables, for knit report.
}
\description{
Top tables of differential expression results
}
\author{
Michael Steinbaugh
}
