% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-topTables.R
\docType{methods}
\name{topTables}
\alias{topTables}
\alias{topTables}
\alias{topTables,list-method}
\title{Top Tables of Differential Expression Results}
\usage{
topTables(object, ...)

\S4method{topTables}{list}(object, n = 50L, coding = FALSE)
}
\arguments{
\item{object}{\code{\link[=resultsTables]{resultsTables()}} return \code{list}.}

\item{...}{Additional arguments.}

\item{n}{Number genes to report.}

\item{coding}{Whether to only return coding genes.}
}
\value{
\code{kable}.
}
\description{
Top Tables of Differential Expression Results
}
\examples{
resTbl <- resultsTables(
    object = res_small,
    rowData = rowData(bcb_small),
    summary = FALSE,
    write = FALSE
)
topTables(resTbl)
}
\seealso{
Other R Markdown Functions: \code{\link{prepareRNASeqTemplate}}
}
\author{
Michael Steinbaugh
}
