% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-topTables.R
\docType{methods}
\name{topTables}
\alias{topTables}
\alias{topTables}
\alias{topTables,list-method}
\title{Top Tables of Differential Expression Results}
\usage{
topTables(object, ...)

\S4method{topTables}{list}(object, n = 50L, coding = FALSE)
}
\arguments{
\item{object}{DESeq2 results tables list generated by \code{\link[=resultsTables]{resultsTables()}}.}

\item{...}{Additional arguments.}

\item{n}{Number genes to report.}

\item{coding}{Whether to only return coding genes.}
}
\value{
Top table kables, for knit report.
}
\description{
Top Tables of Differential Expression Results
}
\examples{
data(res)
resTbl <- resultsTables(res, write = FALSE)
topTables(resTbl)
}
\author{
Michael Steinbaugh
}
