% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topTables-methods.R
\docType{methods}
\name{topTables}
\alias{topTables}
\alias{topTables,DESeqResults-method}
\alias{topTables,list-method}
\title{Top tables}
\usage{
topTables(object, ...)

\S4method{topTables}{DESeqResults}(object, n = 50L, coding = FALSE)

\S4method{topTables}{list}(object, n = 50L, coding = FALSE)
}
\arguments{
\item{object}{\code{DESeqResults} or \code{list}. For \code{list} method, must use
\code{\link[=resultsTables]{resultsTables()}} return.}

\item{n}{\code{scalar integer}. Number genes to report.}

\item{coding}{\code{boolean}. Whether to only return coding genes.}

\item{...}{Additional arguments.}
}
\value{
\code{kable}.
}
\description{
Top tables of differentially expressed genes.
}
\examples{
# DESeqResults ====
# Minimal return
topTables(res_small, n = 5L)

# `resultsTables()` return list ====
# Return with gene annotations and DESeq2 normalized counts
res_tbl <- resultsTables(
    results = res_small,
    counts = dds_small
)
topTables(res_tbl, n = 5L)
}
\author{
Michael Steinbaugh
}
\concept{R Markdown Functions}
