% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-tpm.R
\docType{methods}
\name{tpm}
\alias{tpm}
\alias{tpm,bcbioRNADataSet-method}
\title{Transcripts per million (TPM)}
\usage{
tpm(object)

\S4method{tpm}{bcbioRNADataSet}(object)
}
\arguments{
\item{object}{\link{bcbioRNADataSet} object.}
}
\description{
Transcripts per million (TPM)
}
