% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpm.R
\name{tpm}
\alias{tpm}
\title{Transcripts per million.}
\usage{
tpm(txi)
}
\arguments{
\item{txi}{\code{\link[tximport]{tximport}} list, containing counts in the
\code{abundance} slot.}
}
\value{
TPM (transcripts per million) matrix.
}
\description{
Save TPM values from a tximport counts object.
}
\author{
Michael Steinbaugh
}
