% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-tpm.R
\docType{methods}
\name{tpm}
\alias{tpm}
\alias{tpm,bcbioRNASeq-method}
\title{Transcripts Per Million (TPM)}
\usage{
tpm(object)

\S4method{tpm}{bcbioRNASeq}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\link{matrix}.
}
\description{
Transcripts Per Million (TPM)
}
\examples{
bcb <- examples[["bcb"]]
tpm(bcb) \%>\% head()
}
\author{
Michael Steinbaugh
}
