% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-tpm.R
\docType{methods}
\name{tpm,bcbioRNADataSet-method}
\alias{tpm,bcbioRNADataSet-method}
\title{Transcripts per million (TPM)}
\usage{
\S4method{tpm}{bcbioRNADataSet}(object)
}
\arguments{
\item{object}{Object.}
}
\description{
Transcripts per million (TPM)
}
\author{
Michael Steinbaugh
}
