% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-tpm.R
\docType{methods}
\name{tpm}
\alias{tpm}
\alias{tpm,bcbioRNASeq-method}
\title{Transcripts Per Million (TPM)}
\usage{
\S4method{tpm}{bcbioRNASeq}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{matrix}.
}
\description{
Transcripts Per Million (TPM)
}
\examples{
load(system.file("extdata/bcb.rda", package = "bcbioRNASeq"))

# bcbioRNASeq
tpm(bcb) \%>\% glimpse()
}
\author{
Michael Steinbaugh
}
