% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpm.R
\name{tpm}
\alias{tpm}
\title{Transcripts per million}
\usage{
tpm(txi)
}
\arguments{
\item{txi}{\code{tximport} object}
}
\value{
TPM (transcripts per million) matrix
}
\description{
Save TPM values from a tximport counts object
}
\author{
Michael Steinbaugh
}
