% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-tpm.R
\docType{methods}
\name{tpm}
\alias{tpm}
\alias{tpm}
\alias{tpm,bcbioRNASeq-method}
\title{Transcripts Per Million (TPM)}
\usage{
tpm(object)

\S4method{tpm}{bcbioRNASeq}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{matrix}.
}
\description{
Transcripts Per Million (TPM)
}
\examples{
tpm(bcb_small) \%>\% glimpse()
}
\seealso{
Other Data Functions: \code{\link{counts}},
  \code{\link{metrics}}, \code{\link{selectSamples}},
  \code{\link{tmm}}
}
\author{
Michael Steinbaugh
}
