% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-tximport.R
\name{.tx2gene}
\alias{.tx2gene}
\title{Read transcript to gene (tx2gene) annotation file}
\usage{
.tx2gene(project_dir, genome_build = NULL)
}
\arguments{
\item{project_dir}{Project directory.}
}
\description{
Read transcript to gene (tx2gene) annotation file
}
\author{
Michael Steinbaugh
}
\keyword{internal}
