% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-tximport.R
\docType{methods}
\name{txi,bcbioRNADataSet-method}
\alias{txi,bcbioRNADataSet-method}
\title{tximport list accessor}
\usage{
\S4method{txi}{bcbioRNADataSet}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{\link[tximport:tximport]{tximport::tximport()}} list.
}
\description{
tximport list accessor
}
\author{
Michael Steinbaugh
}
