% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods-tximport.R
\docType{methods}
\name{txi}
\alias{txi}
\alias{txi,bcbioRNADataSet-method}
\title{Access tximport data in bcbioRNADataSet}
\usage{
\S4method{txi}{bcbioRNADataSet}(object)
}
\value{
\code{\link[tximport:tximport]{tximport::tximport()}} list.
}
\description{
Access tximport data in bcbioRNADataSet
}
\author{
Michael Steinbaugh
}
\keyword{Object.}
\keyword{object}
