% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_methods.R
\docType{methods}
\name{txi}
\alias{txi}
\alias{txi,bcbioRnaDataSet-method}
\alias{txi<-,bcbioRnaDataSet,matrix-method}
\alias{txi<-}
\alias{txi.bcbioRnaDataSet}
\alias{txi,bcbioRnaDataSet-method}
\alias{"<-txi"}
\title{Accessors for the count matrix of a \link{bcbioRnaDataSet} object}
\usage{
txi(object, ...)

txi(object, ...) <- value

txi.bcbioRnaDataSet(object)

\S4method{txi}{bcbioRnaDataSet}(object)

\S4method{txi}{bcbioRnaDataSet,matrix}(object) <- value
}
\arguments{
\item{object}{\link{bcbioRnaDataSet} object.}

\item{...}{Matrix count data.}

\item{value}{An integer matrix.}
}
\value{
\link[base:matrix]{base::matrix} with raw count data.
}
\description{
Accessors for the count matrix of a \link{bcbioRnaDataSet} object
}
\keyword{internal}
