% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-txi.R
\docType{methods}
\name{txi}
\alias{txi}
\alias{txi}
\alias{txi,bcbioRNASeq-method}
\title{\code{tximport} Accessor Shortcut}
\usage{
txi(object)

\S4method{txi}{bcbioRNASeq}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{\link[tximport:tximport]{tximport::tximport()}} list.
}
\description{
\code{tximport} Accessor Shortcut
}
\examples{
bcb <- examples[["bcb"]]
txi(bcb) \%>\% names()
}
\author{
Michael Steinbaugh
}
\keyword{internal}
