% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-updateObject.R
\docType{methods}
\name{updateObject}
\alias{updateObject}
\alias{updateObject,bcbioRNASeq-method}
\title{Update Object}
\usage{
\S4method{updateObject}{bcbioRNASeq}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\link{bcbioRNASeq}.
}
\description{
Update old objects created by the bcbioRNASeq package. The session
information metadata is preserved from the time when the bcbio data was
originally loaded into R.
}
\note{
The previous bcbioRnaseq package (note lowercase "c") must be
reinstalled to load objects from versions <= 0.0.24.
}
\section{Upgrade bcbioRNADataSet to bcbioRNASeq object}{

This method adds support for upgrading \link{bcbioRNADataSet} objects to the
latest \link{bcbioRNASeq} class version. This should be backwards compatible to
\link{bcbioRNASeq} version 0.0.26.
}

\examples{
load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioRNASeq"))

metadata(bcb)[["version"]]

updated <- updateObject(bcb)
metadata(updated)[["version"]]
metadata(updated)[["previousVersion"]]
}
\author{
Michael Steinbaugh
}
