% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateObject.R
\docType{methods}
\name{updateObject}
\alias{updateObject}
\alias{updateObject,bcbioRNADataSet-method}
\title{Update Object}
\usage{
\S4method{updateObject}{bcbioRNADataSet}(object)
}
\value{
\link{bcbioRNASeq}.
}
\description{
Compatible with old versions created by bcbioRNASeq package.
}
\note{
The previous bcbioRnaseq package (note lowercase "c") must be
reinstalled to load objects from versions <= 0.0.24.
}
\section{Upgrade bcbioRNADataSet to bcbioRNASeq object}{

This method adds support for upgrading \link{bcbioRNADataSet} objects to the
latest \link{bcbioRNASeq} class version. This should be backwards compatible to
\link{bcbioRNASeq} version 0.0.26. Previous objects saved using \code{bcbioRnaseq}
(note case) will likely fail to load with newer versions of the package.
}

\examples{
# FIXME Needs working example
}
\author{
Michael Steinbaugh
}
