% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/updateObject-methods.R
\name{updateObject}
\alias{updateObject}
\alias{updateObject,bcbioRNASeq-method}
\title{Update object}
\usage{
updateObject(object, ..., verbose = FALSE)

\S4method{updateObject}{bcbioRNASeq}(object, rowRanges = NULL, ..., verbose = FALSE)
}
\arguments{
\item{object}{Object.}

\item{rowRanges}{\code{GRanges} or \code{NULL}.
Row annotations. Since we converted to \code{RangedSummarizedExperiment} in
v0.2.0, this option had to be added to enable updating of newly required
\code{rowRanges} slot. Objects that are >= v0.2 don't require this argument and
it can be left \code{NULL}.}

\item{...}{Additional arguments.}

\item{verbose}{\code{logical(1)}.
Run the function with verbose output.}
}
\value{
Modified object.
}
\description{
Update object
}
\details{
Update old objects created by the bcbioRNASeq package. The session
information metadata is preserved from the time when the bcbio data was
originally loaded into R.
}
\note{
Updated 2022-05-07.
}
\section{Legacy \code{bcbioRNADataSet} class}{


Support for \code{bcbioRNADataSet} objects was dropped in v0.2.0 of the package.
If you need to load one of these objects, please install an older release.
}

\section{Legacy \code{bcbioRnaseq} package}{


The previous \code{bcbioRnaseq} package (note case) must be reinstalled to load
objects from versions <= 0.0.20. We changed the name of the package to
\code{bcbioRNASeq} starting in v0.0.21.
}

\examples{
data(bcb)

## bcbioRNASeq ====
object <- bcb
object <- updateObject(object)
validObject(object)

## Example that depends on remote file.
object <- import(file = file.path(
    bcbioRNASeqTestsURL,
    "bcbioRNASeq_0.1.4.rds"
))
object <- updateObject(object, verbose = TRUE)
validObject(object)
}
\author{
Michael Steinbaugh
}
