% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/updateObject-methods.R
\name{updateObject}
\alias{updateObject}
\alias{updateObject,bcbioRNASeq-method}
\title{Update an object to its current class definition}
\usage{
updateObject(object, ..., verbose = TRUE)

\S4method{updateObject}{bcbioRNASeq}(object, rowRanges = NULL, ..., verbose = FALSE)
}
\arguments{
\item{object}{
    Object to be updated for \code{updateObject} and
    \code{updateObjectFromSlots}.

    Object for slot information to be extracted from for
    \code{getObjectSlots}.
  }

\item{rowRanges}{\code{GRanges} or \code{NULL}.
Row annotations. Since we converted to \code{RangedSummarizedExperiment} in
v0.2.0, this option had to be added to enable updating of newly required
\code{rowRanges} slot. Objects that are >= v0.2 don't require this argument and
it can be left \code{NULL}.}

\item{...}{
    Additional arguments, for use in specific \code{updateObject}
    methods.
  }

\item{verbose}{
    \code{TRUE} or \code{FALSE}, indicating whether information about
    the update should be reported. Use \code{\link[base]{message}} to
    report this information.
  }
}
\value{
\code{bcbioRNASeq}.
}
\description{
\code{updateObject} is a generic function that returns an instance
  of \code{object} updated to its current class definition.
}
\details{
Update old objects created by the bcbioRNASeq package. The session
information metadata is preserved from the time when the bcbio data was
originally loaded into R.
}
\note{
Updated 2020-05-11.
}
\section{Legacy \code{bcbioRNADataSet} class}{


Support for \code{bcbioRNADataSet} objects was dropped in v0.2.0 of the package.
If you need to load one of these objects, please install an older release.
}

\section{Legacy \code{bcbioRnaseq} package}{


The previous \code{bcbioRnaseq} package (note case) must be reinstalled to load
objects from versions <= 0.0.20. We changed the name of the package to
\code{bcbioRNASeq} starting in v0.0.21.
}

\examples{
data(bcb)
updateObject(bcb)

## Example that depends on remote file.
## > x <- import(file.path(bcbioRNASeqTestsURL, "bcbioRNASeq_0.1.4.rds"))
## > x <- updateObject(x)
## > x
}
\seealso{
\itemize{
    \item \code{\link[Biobase]{updateObjectTo}} in the \pkg{Biobase} package
          for updating an object to the class definition of a template (might
          be useful for updating a virtual superclass).

    \item \code{\link[methods]{validObject}} for testing the validity of an
          object.

    \item \code{\link[methods]{showMethods}} for displaying a summary of the
          methods defined for a given generic function.

    \item \code{\link[methods]{selectMethod}} for getting the definition of
          a specific method.

    \item \link[BiocGenerics]{BiocGenerics} for a summary of all the generics defined
          in the \pkg{BiocGenerics} package.
  }
}
\author{
Michael Steinbaugh
}
