% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_counts.R
\name{write_counts}
\alias{write_counts}
\title{Write gzipped counts.}
\usage{
write_counts(..., dir = file.path("results", "counts"))
}
\arguments{
\item{...}{Count matrices, passed in as dots.}

\item{dir}{Output directory.}
}
\description{
Write gzipped counts.
}
\author{
Michael Steinbaugh
}
