% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-yaml.R
\name{.sample_yaml}
\alias{.sample_yaml}
\alias{.sample_yaml_metadata}
\alias{.sample_yaml_metrics}
\title{Sample YAML metadata utilities}
\usage{
.sample_yaml_metrics(yaml)
}
\arguments{
\item{yaml}{Project summary YAML.}

\item{...}{Nested operator keys supplied as dot objects.}
}
\value{
\link{tibble}.
}
\description{
Sample YAML metadata utilities
}
\note{
Metrics are only generated for a standard RNA-seq run with aligned
counts. Fast RNA-seq mode with lightweight counts (pseudocounts) doesn't
output the same metrics into the YAML.
}
\author{
Michael Steinbaugh
}
\keyword{internal}
